% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ActiveLearningConfig}
\alias{GoogleCloudAiplatformV1ActiveLearningConfig}
\title{GoogleCloudAiplatformV1ActiveLearningConfig Object}
\usage{
GoogleCloudAiplatformV1ActiveLearningConfig(
  maxDataItemPercentage = NULL,
  maxDataItemCount = NULL,
  trainingConfig = NULL,
  sampleConfig = NULL
)
}
\arguments{
\item{maxDataItemPercentage}{Max percent of total DataItems for human labeling}

\item{maxDataItemCount}{Max number of human labeled DataItems}

\item{trainingConfig}{CMLE training config}

\item{sampleConfig}{Active learning data sampling config}
}
\value{
GoogleCloudAiplatformV1ActiveLearningConfig object
}
\description{
GoogleCloudAiplatformV1ActiveLearningConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Parameters that configure the active learning pipeline. Active learning will label the data incrementally by several iterations. For every iteration, it will select a batch of data based on the sampling strategy.
}
\concept{GoogleCloudAiplatformV1ActiveLearningConfig functions}
