% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1AddContextArtifactsAndExecutionsRequest}
\alias{GoogleCloudAiplatformV1AddContextArtifactsAndExecutionsRequest}
\title{GoogleCloudAiplatformV1AddContextArtifactsAndExecutionsRequest Object}
\usage{
GoogleCloudAiplatformV1AddContextArtifactsAndExecutionsRequest(
  artifacts = NULL,
  executions = NULL
)
}
\arguments{
\item{artifacts}{The resource names of the Artifacts to attribute to the Context}

\item{executions}{The resource names of the Executions to associate with the Context}
}
\value{
GoogleCloudAiplatformV1AddContextArtifactsAndExecutionsRequest object
}
\description{
GoogleCloudAiplatformV1AddContextArtifactsAndExecutionsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for MetadataService.AddContextArtifactsAndExecutions.
}
\seealso{
Other GoogleCloudAiplatformV1AddContextArtifactsAndExecutionsRequest functions: 
\code{\link{projects.locations.metadataStores.contexts.addContextArtifactsAndExecutions}()}
}
\concept{GoogleCloudAiplatformV1AddContextArtifactsAndExecutionsRequest functions}
