% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1AddContextChildrenRequest}
\alias{GoogleCloudAiplatformV1AddContextChildrenRequest}
\title{GoogleCloudAiplatformV1AddContextChildrenRequest Object}
\usage{
GoogleCloudAiplatformV1AddContextChildrenRequest(childContexts = NULL)
}
\arguments{
\item{childContexts}{The resource names of the child Contexts}
}
\value{
GoogleCloudAiplatformV1AddContextChildrenRequest object
}
\description{
GoogleCloudAiplatformV1AddContextChildrenRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for MetadataService.AddContextChildren.
}
\seealso{
Other GoogleCloudAiplatformV1AddContextChildrenRequest functions: 
\code{\link{projects.locations.metadataStores.contexts.addContextChildren}()}
}
\concept{GoogleCloudAiplatformV1AddContextChildrenRequest functions}
