% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1AddExecutionEventsRequest}
\alias{GoogleCloudAiplatformV1AddExecutionEventsRequest}
\title{GoogleCloudAiplatformV1AddExecutionEventsRequest Object}
\usage{
GoogleCloudAiplatformV1AddExecutionEventsRequest(events = NULL)
}
\arguments{
\item{events}{The Events to create and add}
}
\value{
GoogleCloudAiplatformV1AddExecutionEventsRequest object
}
\description{
GoogleCloudAiplatformV1AddExecutionEventsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for MetadataService.AddExecutionEvents.
}
\seealso{
Other GoogleCloudAiplatformV1AddExecutionEventsRequest functions: 
\code{\link{projects.locations.metadataStores.executions.addExecutionEvents}()}
}
\concept{GoogleCloudAiplatformV1AddExecutionEventsRequest functions}
