% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1AddTrialMeasurementRequest}
\alias{GoogleCloudAiplatformV1AddTrialMeasurementRequest}
\title{GoogleCloudAiplatformV1AddTrialMeasurementRequest Object}
\usage{
GoogleCloudAiplatformV1AddTrialMeasurementRequest(measurement = NULL)
}
\arguments{
\item{measurement}{Required}
}
\value{
GoogleCloudAiplatformV1AddTrialMeasurementRequest object
}
\description{
GoogleCloudAiplatformV1AddTrialMeasurementRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for VizierService.AddTrialMeasurement.
}
\seealso{
Other GoogleCloudAiplatformV1AddTrialMeasurementRequest functions: 
\code{\link{projects.locations.studies.trials.addTrialMeasurement}()}
}
\concept{GoogleCloudAiplatformV1AddTrialMeasurementRequest functions}
