% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1Annotation.labels}
\alias{GoogleCloudAiplatformV1Annotation.labels}
\title{GoogleCloudAiplatformV1Annotation.labels Object}
\usage{
GoogleCloudAiplatformV1Annotation.labels()
}
\value{
GoogleCloudAiplatformV1Annotation.labels object
}
\description{
GoogleCloudAiplatformV1Annotation.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Optional. The labels with user-defined metadata to organize your Annotations. Label keys and values can be no longer than 64 characters (Unicode codepoints), can only contain lowercase letters, numeric characters, underscores and dashes. International characters are allowed. No more than 64 user labels can be associated with one Annotation(System labels are excluded). See https://goo.gl/xmQnxf for more information and examples of labels. System reserved label keys are prefixed with 'aiplatform.googleapis.com/' and are immutable. Following system labels exist for each Annotation: * 'aiplatform.googleapis.com/annotation_set_name': optional, name of the UI's annotation set this Annotation belongs to. If not set, the Annotation is not visible in the UI. * 'aiplatform.googleapis.com/payload_schema': output only, its value is the payload_schema's title.
}
\seealso{
Other GoogleCloudAiplatformV1Annotation functions: 
\code{\link{GoogleCloudAiplatformV1Annotation}()}
}
\concept{GoogleCloudAiplatformV1Annotation functions}
