% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1Annotation}
\alias{GoogleCloudAiplatformV1Annotation}
\title{GoogleCloudAiplatformV1Annotation Object}
\usage{
GoogleCloudAiplatformV1Annotation(
  GoogleCloudAiplatformV1Annotation.labels = NULL,
  payload = NULL,
  labels = NULL,
  payloadSchemaUri = NULL,
  etag = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1Annotation.labels}{The \link{GoogleCloudAiplatformV1Annotation.labels} object or list of objects}

\item{payload}{Required}

\item{labels}{Optional}

\item{payloadSchemaUri}{Required}

\item{etag}{Optional}
}
\value{
GoogleCloudAiplatformV1Annotation object
}
\description{
GoogleCloudAiplatformV1Annotation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Used to assign specific AnnotationSpec to a particular area of a DataItem or the whole part of the DataItem.
}
\seealso{
Other GoogleCloudAiplatformV1Annotation functions: 
\code{\link{GoogleCloudAiplatformV1Annotation.labels}()}
}
\concept{GoogleCloudAiplatformV1Annotation functions}
