% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1AnnotationSpec}
\alias{GoogleCloudAiplatformV1AnnotationSpec}
\title{GoogleCloudAiplatformV1AnnotationSpec Object}
\usage{
GoogleCloudAiplatformV1AnnotationSpec(displayName = NULL, etag = NULL)
}
\arguments{
\item{displayName}{Required}

\item{etag}{Optional}
}
\value{
GoogleCloudAiplatformV1AnnotationSpec object
}
\description{
GoogleCloudAiplatformV1AnnotationSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Identifies a concept with which DataItems may be annotated with.
}
\concept{GoogleCloudAiplatformV1AnnotationSpec functions}
