% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1Artifact.labels}
\alias{GoogleCloudAiplatformV1Artifact.labels}
\title{GoogleCloudAiplatformV1Artifact.labels Object}
\usage{
GoogleCloudAiplatformV1Artifact.labels()
}
\value{
GoogleCloudAiplatformV1Artifact.labels object
}
\description{
GoogleCloudAiplatformV1Artifact.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The labels with user-defined metadata to organize your Artifacts. Label keys and values can be no longer than 64 characters (Unicode codepoints), can only contain lowercase letters, numeric characters, underscores and dashes. International characters are allowed. No more than 64 user labels can be associated with one Artifact (System labels are excluded).
}
\seealso{
Other GoogleCloudAiplatformV1Artifact functions: 
\code{\link{GoogleCloudAiplatformV1Artifact.metadata}()},
\code{\link{GoogleCloudAiplatformV1Artifact}()},
\code{\link{projects.locations.metadataStores.artifacts.create}()},
\code{\link{projects.locations.metadataStores.artifacts.patch}()}
}
\concept{GoogleCloudAiplatformV1Artifact functions}
