% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1Artifact.metadata}
\alias{GoogleCloudAiplatformV1Artifact.metadata}
\title{GoogleCloudAiplatformV1Artifact.metadata Object}
\usage{
GoogleCloudAiplatformV1Artifact.metadata()
}
\value{
GoogleCloudAiplatformV1Artifact.metadata object
}
\description{
GoogleCloudAiplatformV1Artifact.metadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Properties of the Artifact. Top level metadata keys' heading and trailing spaces will be trimmed. The size of this field should not exceed 200KB.
}
\seealso{
Other GoogleCloudAiplatformV1Artifact functions: 
\code{\link{GoogleCloudAiplatformV1Artifact.labels}()},
\code{\link{GoogleCloudAiplatformV1Artifact}()},
\code{\link{projects.locations.metadataStores.artifacts.create}()},
\code{\link{projects.locations.metadataStores.artifacts.patch}()}
}
\concept{GoogleCloudAiplatformV1Artifact functions}
