% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1Artifact}
\alias{GoogleCloudAiplatformV1Artifact}
\title{GoogleCloudAiplatformV1Artifact Object}
\usage{
GoogleCloudAiplatformV1Artifact(
  GoogleCloudAiplatformV1Artifact.labels = NULL,
  GoogleCloudAiplatformV1Artifact.metadata = NULL,
  schemaVersion = NULL,
  displayName = NULL,
  schemaTitle = NULL,
  state = NULL,
  uri = NULL,
  labels = NULL,
  etag = NULL,
  description = NULL,
  metadata = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1Artifact.labels}{The \link{GoogleCloudAiplatformV1Artifact.labels} object or list of objects}

\item{GoogleCloudAiplatformV1Artifact.metadata}{The \link{GoogleCloudAiplatformV1Artifact.metadata} object or list of objects}

\item{schemaVersion}{The version of the schema in schema_name to use}

\item{displayName}{User provided display name of the Artifact}

\item{schemaTitle}{The title of the schema describing the metadata}

\item{state}{The state of this Artifact}

\item{uri}{The uniform resource identifier of the artifact file}

\item{labels}{The labels with user-defined metadata to organize your Artifacts}

\item{etag}{An eTag used to perform consistent read-modify-write updates}

\item{description}{Description of the Artifact}

\item{metadata}{Properties of the Artifact}
}
\value{
GoogleCloudAiplatformV1Artifact object
}
\description{
GoogleCloudAiplatformV1Artifact Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Instance of a general artifact.
}
\seealso{
Other GoogleCloudAiplatformV1Artifact functions: 
\code{\link{GoogleCloudAiplatformV1Artifact.labels}()},
\code{\link{GoogleCloudAiplatformV1Artifact.metadata}()},
\code{\link{projects.locations.metadataStores.artifacts.create}()},
\code{\link{projects.locations.metadataStores.artifacts.patch}()}
}
\concept{GoogleCloudAiplatformV1Artifact functions}
