% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1AutomaticResources}
\alias{GoogleCloudAiplatformV1AutomaticResources}
\title{GoogleCloudAiplatformV1AutomaticResources Object}
\usage{
GoogleCloudAiplatformV1AutomaticResources(
  minReplicaCount = NULL,
  maxReplicaCount = NULL
)
}
\arguments{
\item{minReplicaCount}{Immutable}

\item{maxReplicaCount}{Immutable}
}
\value{
GoogleCloudAiplatformV1AutomaticResources object
}
\description{
GoogleCloudAiplatformV1AutomaticResources Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A description of resources that to large degree are decided by Vertex AI, and require only a modest additional configuration. Each Model supporting these resources documents its specific guidelines.
}
\concept{GoogleCloudAiplatformV1AutomaticResources functions}
