% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1AutoscalingMetricSpec}
\alias{GoogleCloudAiplatformV1AutoscalingMetricSpec}
\title{GoogleCloudAiplatformV1AutoscalingMetricSpec Object}
\usage{
GoogleCloudAiplatformV1AutoscalingMetricSpec(target = NULL, metricName = NULL)
}
\arguments{
\item{target}{The target resource utilization in percentage (1\% - 100\%) for the given metric; once the real usage deviates from the target by a certain percentage, the machine replicas change}

\item{metricName}{Required}
}
\value{
GoogleCloudAiplatformV1AutoscalingMetricSpec object
}
\description{
GoogleCloudAiplatformV1AutoscalingMetricSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The metric specification that defines the target resource utilization (CPU utilization, accelerator's duty cycle, and so on) for calculating the desired replica count.
}
\concept{GoogleCloudAiplatformV1AutoscalingMetricSpec functions}
