% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1AvroSource}
\alias{GoogleCloudAiplatformV1AvroSource}
\title{GoogleCloudAiplatformV1AvroSource Object}
\usage{
GoogleCloudAiplatformV1AvroSource(gcsSource = NULL)
}
\arguments{
\item{gcsSource}{Required}
}
\value{
GoogleCloudAiplatformV1AvroSource object
}
\description{
GoogleCloudAiplatformV1AvroSource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The storage details for Avro input content.
}
\concept{GoogleCloudAiplatformV1AvroSource functions}
