% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1BatchCreateFeaturesOperationMetadata}
\alias{GoogleCloudAiplatformV1BatchCreateFeaturesOperationMetadata}
\title{GoogleCloudAiplatformV1BatchCreateFeaturesOperationMetadata Object}
\usage{
GoogleCloudAiplatformV1BatchCreateFeaturesOperationMetadata(
  genericMetadata = NULL
)
}
\arguments{
\item{genericMetadata}{Operation metadata for Feature}
}
\value{
GoogleCloudAiplatformV1BatchCreateFeaturesOperationMetadata object
}
\description{
GoogleCloudAiplatformV1BatchCreateFeaturesOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details of operations that perform batch create Features.
}
\concept{GoogleCloudAiplatformV1BatchCreateFeaturesOperationMetadata functions}
