% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1BatchCreateFeaturesRequest}
\alias{GoogleCloudAiplatformV1BatchCreateFeaturesRequest}
\title{GoogleCloudAiplatformV1BatchCreateFeaturesRequest Object}
\usage{
GoogleCloudAiplatformV1BatchCreateFeaturesRequest(requests = NULL)
}
\arguments{
\item{requests}{Required}
}
\value{
GoogleCloudAiplatformV1BatchCreateFeaturesRequest object
}
\description{
GoogleCloudAiplatformV1BatchCreateFeaturesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for FeaturestoreService.BatchCreateFeatures.
}
\seealso{
Other GoogleCloudAiplatformV1BatchCreateFeaturesRequest functions: 
\code{\link{projects.locations.featurestores.entityTypes.features.batchCreate}()}
}
\concept{GoogleCloudAiplatformV1BatchCreateFeaturesRequest functions}
