% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1BatchCreateTensorboardRunsRequest}
\alias{GoogleCloudAiplatformV1BatchCreateTensorboardRunsRequest}
\title{GoogleCloudAiplatformV1BatchCreateTensorboardRunsRequest Object}
\usage{
GoogleCloudAiplatformV1BatchCreateTensorboardRunsRequest(requests = NULL)
}
\arguments{
\item{requests}{Required}
}
\value{
GoogleCloudAiplatformV1BatchCreateTensorboardRunsRequest object
}
\description{
GoogleCloudAiplatformV1BatchCreateTensorboardRunsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for TensorboardService.BatchCreateTensorboardRuns.
}
\seealso{
Other GoogleCloudAiplatformV1BatchCreateTensorboardRunsRequest functions: 
\code{\link{projects.locations.tensorboards.experiments.runs.batchCreate}()}
}
\concept{GoogleCloudAiplatformV1BatchCreateTensorboardRunsRequest functions}
