% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1BatchCreateTensorboardRunsResponse}
\alias{GoogleCloudAiplatformV1BatchCreateTensorboardRunsResponse}
\title{GoogleCloudAiplatformV1BatchCreateTensorboardRunsResponse Object}
\usage{
GoogleCloudAiplatformV1BatchCreateTensorboardRunsResponse(
  tensorboardRuns = NULL
)
}
\arguments{
\item{tensorboardRuns}{The created TensorboardRuns}
}
\value{
GoogleCloudAiplatformV1BatchCreateTensorboardRunsResponse object
}
\description{
GoogleCloudAiplatformV1BatchCreateTensorboardRunsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for TensorboardService.BatchCreateTensorboardRuns.
}
\concept{GoogleCloudAiplatformV1BatchCreateTensorboardRunsResponse functions}
