% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1BatchCreateTensorboardTimeSeriesRequest}
\alias{GoogleCloudAiplatformV1BatchCreateTensorboardTimeSeriesRequest}
\title{GoogleCloudAiplatformV1BatchCreateTensorboardTimeSeriesRequest Object}
\usage{
GoogleCloudAiplatformV1BatchCreateTensorboardTimeSeriesRequest(requests = NULL)
}
\arguments{
\item{requests}{Required}
}
\value{
GoogleCloudAiplatformV1BatchCreateTensorboardTimeSeriesRequest object
}
\description{
GoogleCloudAiplatformV1BatchCreateTensorboardTimeSeriesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for TensorboardService.BatchCreateTensorboardTimeSeries.
}
\seealso{
Other GoogleCloudAiplatformV1BatchCreateTensorboardTimeSeriesRequest functions: 
\code{\link{projects.locations.tensorboards.experiments.runs.timeSeries.batchCreate}()}
}
\concept{GoogleCloudAiplatformV1BatchCreateTensorboardTimeSeriesRequest functions}
