% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1BatchCreateTensorboardTimeSeriesResponse}
\alias{GoogleCloudAiplatformV1BatchCreateTensorboardTimeSeriesResponse}
\title{GoogleCloudAiplatformV1BatchCreateTensorboardTimeSeriesResponse Object}
\usage{
GoogleCloudAiplatformV1BatchCreateTensorboardTimeSeriesResponse(
  tensorboardTimeSeries = NULL
)
}
\arguments{
\item{tensorboardTimeSeries}{The created TensorboardTimeSeries}
}
\value{
GoogleCloudAiplatformV1BatchCreateTensorboardTimeSeriesResponse object
}
\description{
GoogleCloudAiplatformV1BatchCreateTensorboardTimeSeriesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for TensorboardService.BatchCreateTensorboardTimeSeries.
}
\concept{GoogleCloudAiplatformV1BatchCreateTensorboardTimeSeriesResponse functions}
