% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1BatchDedicatedResources}
\alias{GoogleCloudAiplatformV1BatchDedicatedResources}
\title{GoogleCloudAiplatformV1BatchDedicatedResources Object}
\usage{
GoogleCloudAiplatformV1BatchDedicatedResources(
  machineSpec = NULL,
  startingReplicaCount = NULL,
  maxReplicaCount = NULL
)
}
\arguments{
\item{machineSpec}{Required}

\item{startingReplicaCount}{Immutable}

\item{maxReplicaCount}{Immutable}
}
\value{
GoogleCloudAiplatformV1BatchDedicatedResources object
}
\description{
GoogleCloudAiplatformV1BatchDedicatedResources Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A description of resources that are used for performing batch operations, are dedicated to a Model, and need manual configuration.
}
\concept{GoogleCloudAiplatformV1BatchDedicatedResources functions}
