% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1BatchImportEvaluatedAnnotationsRequest}
\alias{GoogleCloudAiplatformV1BatchImportEvaluatedAnnotationsRequest}
\title{GoogleCloudAiplatformV1BatchImportEvaluatedAnnotationsRequest Object}
\usage{
GoogleCloudAiplatformV1BatchImportEvaluatedAnnotationsRequest(
  evaluatedAnnotations = NULL
)
}
\arguments{
\item{evaluatedAnnotations}{Required}
}
\value{
GoogleCloudAiplatformV1BatchImportEvaluatedAnnotationsRequest object
}
\description{
GoogleCloudAiplatformV1BatchImportEvaluatedAnnotationsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for ModelService.BatchImportEvaluatedAnnotations
}
\seealso{
Other GoogleCloudAiplatformV1BatchImportEvaluatedAnnotationsRequest functions: 
\code{\link{projects.locations.models.evaluations.slices.batchImport}()}
}
\concept{GoogleCloudAiplatformV1BatchImportEvaluatedAnnotationsRequest functions}
