% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1BatchImportEvaluatedAnnotationsResponse}
\alias{GoogleCloudAiplatformV1BatchImportEvaluatedAnnotationsResponse}
\title{GoogleCloudAiplatformV1BatchImportEvaluatedAnnotationsResponse Object}
\usage{
GoogleCloudAiplatformV1BatchImportEvaluatedAnnotationsResponse()
}
\value{
GoogleCloudAiplatformV1BatchImportEvaluatedAnnotationsResponse object
}
\description{
GoogleCloudAiplatformV1BatchImportEvaluatedAnnotationsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for ModelService.BatchImportEvaluatedAnnotations
}
\concept{GoogleCloudAiplatformV1BatchImportEvaluatedAnnotationsResponse functions}
