% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1BatchImportModelEvaluationSlicesRequest}
\alias{GoogleCloudAiplatformV1BatchImportModelEvaluationSlicesRequest}
\title{GoogleCloudAiplatformV1BatchImportModelEvaluationSlicesRequest Object}
\usage{
GoogleCloudAiplatformV1BatchImportModelEvaluationSlicesRequest(
  modelEvaluationSlices = NULL
)
}
\arguments{
\item{modelEvaluationSlices}{Required}
}
\value{
GoogleCloudAiplatformV1BatchImportModelEvaluationSlicesRequest object
}
\description{
GoogleCloudAiplatformV1BatchImportModelEvaluationSlicesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for ModelService.BatchImportModelEvaluationSlices
}
\concept{GoogleCloudAiplatformV1BatchImportModelEvaluationSlicesRequest functions}
