% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1BatchImportModelEvaluationSlicesResponse}
\alias{GoogleCloudAiplatformV1BatchImportModelEvaluationSlicesResponse}
\title{GoogleCloudAiplatformV1BatchImportModelEvaluationSlicesResponse Object}
\usage{
GoogleCloudAiplatformV1BatchImportModelEvaluationSlicesResponse()
}
\value{
GoogleCloudAiplatformV1BatchImportModelEvaluationSlicesResponse object
}
\description{
GoogleCloudAiplatformV1BatchImportModelEvaluationSlicesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for ModelService.BatchImportModelEvaluationSlices
}
\concept{GoogleCloudAiplatformV1BatchImportModelEvaluationSlicesResponse functions}
