% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1BatchMigrateResourcesOperationMetadata}
\alias{GoogleCloudAiplatformV1BatchMigrateResourcesOperationMetadata}
\title{GoogleCloudAiplatformV1BatchMigrateResourcesOperationMetadata Object}
\usage{
GoogleCloudAiplatformV1BatchMigrateResourcesOperationMetadata(
  partialResults = NULL,
  genericMetadata = NULL
)
}
\arguments{
\item{partialResults}{Partial results that reflect the latest migration operation progress}

\item{genericMetadata}{The common part of the operation metadata}
}
\value{
GoogleCloudAiplatformV1BatchMigrateResourcesOperationMetadata object
}
\description{
GoogleCloudAiplatformV1BatchMigrateResourcesOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Runtime operation information for MigrationService.BatchMigrateResources.
}
\concept{GoogleCloudAiplatformV1BatchMigrateResourcesOperationMetadata functions}
