% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1BatchMigrateResourcesOperationMetadataPartialResult}
\alias{GoogleCloudAiplatformV1BatchMigrateResourcesOperationMetadataPartialResult}
\title{GoogleCloudAiplatformV1BatchMigrateResourcesOperationMetadataPartialResult Object}
\usage{
GoogleCloudAiplatformV1BatchMigrateResourcesOperationMetadataPartialResult(
  model = NULL,
  dataset = NULL,
  error = NULL,
  request = NULL
)
}
\arguments{
\item{model}{Migrated model resource name}

\item{dataset}{Migrated dataset resource name}

\item{error}{The error result of the migration request in case of failure}

\item{request}{It's the same as the value in MigrateResourceRequest}
}
\value{
GoogleCloudAiplatformV1BatchMigrateResourcesOperationMetadataPartialResult object
}
\description{
GoogleCloudAiplatformV1BatchMigrateResourcesOperationMetadataPartialResult Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a partial result in batch migration operation for one MigrateResourceRequest.
}
\concept{GoogleCloudAiplatformV1BatchMigrateResourcesOperationMetadataPartialResult functions}
