% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1BatchMigrateResourcesRequest}
\alias{GoogleCloudAiplatformV1BatchMigrateResourcesRequest}
\title{GoogleCloudAiplatformV1BatchMigrateResourcesRequest Object}
\usage{
GoogleCloudAiplatformV1BatchMigrateResourcesRequest(
  migrateResourceRequests = NULL
)
}
\arguments{
\item{migrateResourceRequests}{Required}
}
\value{
GoogleCloudAiplatformV1BatchMigrateResourcesRequest object
}
\description{
GoogleCloudAiplatformV1BatchMigrateResourcesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for MigrationService.BatchMigrateResources.
}
\seealso{
Other GoogleCloudAiplatformV1BatchMigrateResourcesRequest functions: 
\code{\link{projects.locations.migratableResources.batchMigrate}()}
}
\concept{GoogleCloudAiplatformV1BatchMigrateResourcesRequest functions}
