% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1BatchMigrateResourcesResponse}
\alias{GoogleCloudAiplatformV1BatchMigrateResourcesResponse}
\title{GoogleCloudAiplatformV1BatchMigrateResourcesResponse Object}
\usage{
GoogleCloudAiplatformV1BatchMigrateResourcesResponse(
  migrateResourceResponses = NULL
)
}
\arguments{
\item{migrateResourceResponses}{Successfully migrated resources}
}
\value{
GoogleCloudAiplatformV1BatchMigrateResourcesResponse object
}
\description{
GoogleCloudAiplatformV1BatchMigrateResourcesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for MigrationService.BatchMigrateResources.
}
\concept{GoogleCloudAiplatformV1BatchMigrateResourcesResponse functions}
