% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1BatchPredictionJob}
\alias{GoogleCloudAiplatformV1BatchPredictionJob}
\title{GoogleCloudAiplatformV1BatchPredictionJob Object}
\usage{
GoogleCloudAiplatformV1BatchPredictionJob(
  GoogleCloudAiplatformV1BatchPredictionJob.labels = NULL,
  displayName = NULL,
  encryptionSpec = NULL,
  modelParameters = NULL,
  unmanagedContainerModel = NULL,
  explanationSpec = NULL,
  serviceAccount = NULL,
  instanceConfig = NULL,
  model = NULL,
  labels = NULL,
  dedicatedResources = NULL,
  inputConfig = NULL,
  generateExplanation = NULL,
  disableContainerLogging = NULL,
  manualBatchTuningParameters = NULL,
  outputConfig = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1BatchPredictionJob.labels}{The \link{GoogleCloudAiplatformV1BatchPredictionJob.labels} object or list of objects}

\item{displayName}{Required}

\item{encryptionSpec}{Customer-managed encryption key options for a BatchPredictionJob}

\item{modelParameters}{The parameters that govern the predictions}

\item{unmanagedContainerModel}{Contains model information necessary to perform batch prediction without requiring uploading to model registry}

\item{explanationSpec}{Explanation configuration for this BatchPredictionJob}

\item{serviceAccount}{The service account that the DeployedModel's container runs as}

\item{instanceConfig}{Configuration for how to convert batch prediction input instances to the prediction instances that are sent to the Model}

\item{model}{The name of the Model resource that produces the predictions via this job, must share the same ancestor Location}

\item{labels}{The labels with user-defined metadata to organize BatchPredictionJobs}

\item{dedicatedResources}{The config of resources used by the Model during the batch prediction}

\item{inputConfig}{Required}

\item{generateExplanation}{Generate explanation with the batch prediction results}

\item{disableContainerLogging}{For custom-trained Models and AutoML Tabular Models, the container of the DeployedModel instances will send \code{stderr} and \code{stdout} streams to Cloud Logging by default}

\item{manualBatchTuningParameters}{Immutable}

\item{outputConfig}{Required}
}
\value{
GoogleCloudAiplatformV1BatchPredictionJob object
}
\description{
GoogleCloudAiplatformV1BatchPredictionJob Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A job that uses a Model to produce predictions on multiple input instances. If predictions for significant portion of the instances fail, the job may finish without attempting predictions for all remaining instances.
}
\seealso{
Other GoogleCloudAiplatformV1BatchPredictionJob functions: 
\code{\link{GoogleCloudAiplatformV1BatchPredictionJob.labels}()},
\code{\link{projects.locations.batchPredictionJobs.create}()}
}
\concept{GoogleCloudAiplatformV1BatchPredictionJob functions}
