% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1BatchPredictionJobInputConfig}
\alias{GoogleCloudAiplatformV1BatchPredictionJobInputConfig}
\title{GoogleCloudAiplatformV1BatchPredictionJobInputConfig Object}
\usage{
GoogleCloudAiplatformV1BatchPredictionJobInputConfig(
  bigquerySource = NULL,
  instancesFormat = NULL,
  gcsSource = NULL
)
}
\arguments{
\item{bigquerySource}{The BigQuery location of the input table}

\item{instancesFormat}{Required}

\item{gcsSource}{The Cloud Storage location for the input instances}
}
\value{
GoogleCloudAiplatformV1BatchPredictionJobInputConfig object
}
\description{
GoogleCloudAiplatformV1BatchPredictionJobInputConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configures the input to BatchPredictionJob. See Model.supported_input_storage_formats for Model's supported input formats, and how instances should be expressed via any of them.
}
\concept{GoogleCloudAiplatformV1BatchPredictionJobInputConfig functions}
