% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1BatchPredictionJobOutputConfig}
\alias{GoogleCloudAiplatformV1BatchPredictionJobOutputConfig}
\title{GoogleCloudAiplatformV1BatchPredictionJobOutputConfig Object}
\usage{
GoogleCloudAiplatformV1BatchPredictionJobOutputConfig(
  bigqueryDestination = NULL,
  gcsDestination = NULL,
  predictionsFormat = NULL
)
}
\arguments{
\item{bigqueryDestination}{The BigQuery project or dataset location where the output is to be written to}

\item{gcsDestination}{The Cloud Storage location of the directory where the output is to be written to}

\item{predictionsFormat}{Required}
}
\value{
GoogleCloudAiplatformV1BatchPredictionJobOutputConfig object
}
\description{
GoogleCloudAiplatformV1BatchPredictionJobOutputConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configures the output of BatchPredictionJob. See Model.supported_output_storage_formats for supported output formats, and how predictions are expressed via any of them.
}
\concept{GoogleCloudAiplatformV1BatchPredictionJobOutputConfig functions}
