% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1BatchPredictionJobOutputInfo}
\alias{GoogleCloudAiplatformV1BatchPredictionJobOutputInfo}
\title{GoogleCloudAiplatformV1BatchPredictionJobOutputInfo Object}
\usage{
GoogleCloudAiplatformV1BatchPredictionJobOutputInfo()
}
\value{
GoogleCloudAiplatformV1BatchPredictionJobOutputInfo object
}
\description{
GoogleCloudAiplatformV1BatchPredictionJobOutputInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Further describes this job's output. Supplements output_config.
}
\concept{GoogleCloudAiplatformV1BatchPredictionJobOutputInfo functions}
