% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1BatchReadFeatureValuesOperationMetadata}
\alias{GoogleCloudAiplatformV1BatchReadFeatureValuesOperationMetadata}
\title{GoogleCloudAiplatformV1BatchReadFeatureValuesOperationMetadata Object}
\usage{
GoogleCloudAiplatformV1BatchReadFeatureValuesOperationMetadata(
  genericMetadata = NULL
)
}
\arguments{
\item{genericMetadata}{Operation metadata for Featurestore batch read Features values}
}
\value{
GoogleCloudAiplatformV1BatchReadFeatureValuesOperationMetadata object
}
\description{
GoogleCloudAiplatformV1BatchReadFeatureValuesOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details of operations that batch reads Feature values.
}
\concept{GoogleCloudAiplatformV1BatchReadFeatureValuesOperationMetadata functions}
