% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1BatchReadFeatureValuesRequestEntityTypeSpec}
\alias{GoogleCloudAiplatformV1BatchReadFeatureValuesRequestEntityTypeSpec}
\title{GoogleCloudAiplatformV1BatchReadFeatureValuesRequestEntityTypeSpec Object}
\usage{
GoogleCloudAiplatformV1BatchReadFeatureValuesRequestEntityTypeSpec(
  settings = NULL,
  entityTypeId = NULL,
  featureSelector = NULL
)
}
\arguments{
\item{settings}{Per-Feature settings for the batch read}

\item{entityTypeId}{Required}

\item{featureSelector}{Required}
}
\value{
GoogleCloudAiplatformV1BatchReadFeatureValuesRequestEntityTypeSpec object
}
\description{
GoogleCloudAiplatformV1BatchReadFeatureValuesRequestEntityTypeSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Selects Features of an EntityType to read values of and specifies read settings.
}
\concept{GoogleCloudAiplatformV1BatchReadFeatureValuesRequestEntityTypeSpec functions}
