% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1BatchReadTensorboardTimeSeriesDataResponse}
\alias{GoogleCloudAiplatformV1BatchReadTensorboardTimeSeriesDataResponse}
\title{GoogleCloudAiplatformV1BatchReadTensorboardTimeSeriesDataResponse Object}
\usage{
GoogleCloudAiplatformV1BatchReadTensorboardTimeSeriesDataResponse(
  timeSeriesData = NULL
)
}
\arguments{
\item{timeSeriesData}{The returned time series data}
}
\value{
GoogleCloudAiplatformV1BatchReadTensorboardTimeSeriesDataResponse object
}
\description{
GoogleCloudAiplatformV1BatchReadTensorboardTimeSeriesDataResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for TensorboardService.BatchReadTensorboardTimeSeriesData.
}
\concept{GoogleCloudAiplatformV1BatchReadTensorboardTimeSeriesDataResponse functions}
