% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1BigQueryDestination}
\alias{GoogleCloudAiplatformV1BigQueryDestination}
\title{GoogleCloudAiplatformV1BigQueryDestination Object}
\usage{
GoogleCloudAiplatformV1BigQueryDestination(outputUri = NULL)
}
\arguments{
\item{outputUri}{Required}
}
\value{
GoogleCloudAiplatformV1BigQueryDestination object
}
\description{
GoogleCloudAiplatformV1BigQueryDestination Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The BigQuery location for the output content.
}
\concept{GoogleCloudAiplatformV1BigQueryDestination functions}
