% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1BigQuerySource}
\alias{GoogleCloudAiplatformV1BigQuerySource}
\title{GoogleCloudAiplatformV1BigQuerySource Object}
\usage{
GoogleCloudAiplatformV1BigQuerySource(inputUri = NULL)
}
\arguments{
\item{inputUri}{Required}
}
\value{
GoogleCloudAiplatformV1BigQuerySource object
}
\description{
GoogleCloudAiplatformV1BigQuerySource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The BigQuery location for the input content.
}
\concept{GoogleCloudAiplatformV1BigQuerySource functions}
