% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1BlurBaselineConfig}
\alias{GoogleCloudAiplatformV1BlurBaselineConfig}
\title{GoogleCloudAiplatformV1BlurBaselineConfig Object}
\usage{
GoogleCloudAiplatformV1BlurBaselineConfig(maxBlurSigma = NULL)
}
\arguments{
\item{maxBlurSigma}{The standard deviation of the blur kernel for the blurred baseline}
}
\value{
GoogleCloudAiplatformV1BlurBaselineConfig object
}
\description{
GoogleCloudAiplatformV1BlurBaselineConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Config for blur baseline. When enabled, a linear path from the maximally blurred image to the input image is created. Using a blurred baseline instead of zero (black image) is motivated by the BlurIG approach explained here: https://arxiv.org/abs/2004.03383
}
\concept{GoogleCloudAiplatformV1BlurBaselineConfig functions}
