% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1BoolArray}
\alias{GoogleCloudAiplatformV1BoolArray}
\title{GoogleCloudAiplatformV1BoolArray Object}
\usage{
GoogleCloudAiplatformV1BoolArray(values = NULL)
}
\arguments{
\item{values}{A list of bool values}
}
\value{
GoogleCloudAiplatformV1BoolArray object
}
\description{
GoogleCloudAiplatformV1BoolArray Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of boolean values.
}
\concept{GoogleCloudAiplatformV1BoolArray functions}
