% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1CancelBatchPredictionJobRequest}
\alias{GoogleCloudAiplatformV1CancelBatchPredictionJobRequest}
\title{GoogleCloudAiplatformV1CancelBatchPredictionJobRequest Object}
\usage{
GoogleCloudAiplatformV1CancelBatchPredictionJobRequest()
}
\value{
GoogleCloudAiplatformV1CancelBatchPredictionJobRequest object
}
\description{
GoogleCloudAiplatformV1CancelBatchPredictionJobRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for JobService.CancelBatchPredictionJob.
}
\seealso{
Other GoogleCloudAiplatformV1CancelBatchPredictionJobRequest functions: 
\code{\link{projects.locations.batchPredictionJobs.cancel}()}
}
\concept{GoogleCloudAiplatformV1CancelBatchPredictionJobRequest functions}
