% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1CancelCustomJobRequest}
\alias{GoogleCloudAiplatformV1CancelCustomJobRequest}
\title{GoogleCloudAiplatformV1CancelCustomJobRequest Object}
\usage{
GoogleCloudAiplatformV1CancelCustomJobRequest()
}
\value{
GoogleCloudAiplatformV1CancelCustomJobRequest object
}
\description{
GoogleCloudAiplatformV1CancelCustomJobRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for JobService.CancelCustomJob.
}
\seealso{
Other GoogleCloudAiplatformV1CancelCustomJobRequest functions: 
\code{\link{projects.locations.customJobs.cancel}()}
}
\concept{GoogleCloudAiplatformV1CancelCustomJobRequest functions}
