% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1CancelDataLabelingJobRequest}
\alias{GoogleCloudAiplatformV1CancelDataLabelingJobRequest}
\title{Vertex AI API Objects
Train high-quality custom machine learning models with minimal machine learning expertise and effort.}
\usage{
GoogleCloudAiplatformV1CancelDataLabelingJobRequest()
}
\value{
GoogleCloudAiplatformV1CancelDataLabelingJobRequest object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
at 2023-08-09 17:20:44
filename: /Users/justin/Dropbox/projects/github.com/justinjm/autoGoogleAPI/googleaiplatformv1.auto/R/aiplatform_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
GoogleCloudAiplatformV1CancelDataLabelingJobRequest Object

Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for JobService.CancelDataLabelingJob.
}
\seealso{
Other GoogleCloudAiplatformV1CancelDataLabelingJobRequest functions: 
\code{\link{projects.locations.dataLabelingJobs.cancel}()}
}
\concept{GoogleCloudAiplatformV1CancelDataLabelingJobRequest functions}
