% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1CancelTrainingPipelineRequest}
\alias{GoogleCloudAiplatformV1CancelTrainingPipelineRequest}
\title{GoogleCloudAiplatformV1CancelTrainingPipelineRequest Object}
\usage{
GoogleCloudAiplatformV1CancelTrainingPipelineRequest()
}
\value{
GoogleCloudAiplatformV1CancelTrainingPipelineRequest object
}
\description{
GoogleCloudAiplatformV1CancelTrainingPipelineRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for PipelineService.CancelTrainingPipeline.
}
\seealso{
Other GoogleCloudAiplatformV1CancelTrainingPipelineRequest functions: 
\code{\link{projects.locations.trainingPipelines.cancel}()}
}
\concept{GoogleCloudAiplatformV1CancelTrainingPipelineRequest functions}
