% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1CheckTrialEarlyStoppingStateMetatdata}
\alias{GoogleCloudAiplatformV1CheckTrialEarlyStoppingStateMetatdata}
\title{GoogleCloudAiplatformV1CheckTrialEarlyStoppingStateMetatdata Object}
\usage{
GoogleCloudAiplatformV1CheckTrialEarlyStoppingStateMetatdata(
  genericMetadata = NULL,
  study = NULL,
  trial = NULL
)
}
\arguments{
\item{genericMetadata}{Operation metadata for suggesting Trials}

\item{study}{The name of the Study that the Trial belongs to}

\item{trial}{The Trial name}
}
\value{
GoogleCloudAiplatformV1CheckTrialEarlyStoppingStateMetatdata object
}
\description{
GoogleCloudAiplatformV1CheckTrialEarlyStoppingStateMetatdata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This message will be placed in the metadata field of a google.longrunning.Operation associated with a CheckTrialEarlyStoppingState request.
}
\concept{GoogleCloudAiplatformV1CheckTrialEarlyStoppingStateMetatdata functions}
