% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1CompleteTrialRequest}
\alias{GoogleCloudAiplatformV1CompleteTrialRequest}
\title{GoogleCloudAiplatformV1CompleteTrialRequest Object}
\usage{
GoogleCloudAiplatformV1CompleteTrialRequest(
  finalMeasurement = NULL,
  infeasibleReason = NULL,
  trialInfeasible = NULL
)
}
\arguments{
\item{finalMeasurement}{Optional}

\item{infeasibleReason}{Optional}

\item{trialInfeasible}{Optional}
}
\value{
GoogleCloudAiplatformV1CompleteTrialRequest object
}
\description{
GoogleCloudAiplatformV1CompleteTrialRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for VizierService.CompleteTrial.
}
\seealso{
Other GoogleCloudAiplatformV1CompleteTrialRequest functions: 
\code{\link{projects.locations.studies.trials.complete}()}
}
\concept{GoogleCloudAiplatformV1CompleteTrialRequest functions}
