% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1CompletionStats}
\alias{GoogleCloudAiplatformV1CompletionStats}
\title{GoogleCloudAiplatformV1CompletionStats Object}
\usage{
GoogleCloudAiplatformV1CompletionStats()
}
\value{
GoogleCloudAiplatformV1CompletionStats object
}
\description{
GoogleCloudAiplatformV1CompletionStats Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Success and error statistics of processing multiple entities (for example, DataItems or structured data rows) in batch.
}
\concept{GoogleCloudAiplatformV1CompletionStats functions}
