% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ContainerRegistryDestination}
\alias{GoogleCloudAiplatformV1ContainerRegistryDestination}
\title{GoogleCloudAiplatformV1ContainerRegistryDestination Object}
\usage{
GoogleCloudAiplatformV1ContainerRegistryDestination(outputUri = NULL)
}
\arguments{
\item{outputUri}{Required}
}
\value{
GoogleCloudAiplatformV1ContainerRegistryDestination object
}
\description{
GoogleCloudAiplatformV1ContainerRegistryDestination Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The Container Registry location for the container image.
}
\concept{GoogleCloudAiplatformV1ContainerRegistryDestination functions}
