% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ContainerSpec}
\alias{GoogleCloudAiplatformV1ContainerSpec}
\title{GoogleCloudAiplatformV1ContainerSpec Object}
\usage{
GoogleCloudAiplatformV1ContainerSpec(
  command = NULL,
  imageUri = NULL,
  env = NULL,
  args = NULL
)
}
\arguments{
\item{command}{The command to be invoked when the container is started}

\item{imageUri}{Required}

\item{env}{Environment variables to be passed to the container}

\item{args}{The arguments to be passed when starting the container}
}
\value{
GoogleCloudAiplatformV1ContainerSpec object
}
\description{
GoogleCloudAiplatformV1ContainerSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The spec of a Container.
}
\concept{GoogleCloudAiplatformV1ContainerSpec functions}
