% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1Context.labels}
\alias{GoogleCloudAiplatformV1Context.labels}
\title{GoogleCloudAiplatformV1Context.labels Object}
\usage{
GoogleCloudAiplatformV1Context.labels()
}
\value{
GoogleCloudAiplatformV1Context.labels object
}
\description{
GoogleCloudAiplatformV1Context.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The labels with user-defined metadata to organize your Contexts. Label keys and values can be no longer than 64 characters (Unicode codepoints), can only contain lowercase letters, numeric characters, underscores and dashes. International characters are allowed. No more than 64 user labels can be associated with one Context (System labels are excluded).
}
\seealso{
Other GoogleCloudAiplatformV1Context functions: 
\code{\link{GoogleCloudAiplatformV1Context.metadata}()},
\code{\link{GoogleCloudAiplatformV1Context}()},
\code{\link{projects.locations.metadataStores.contexts.create}()},
\code{\link{projects.locations.metadataStores.contexts.patch}()}
}
\concept{GoogleCloudAiplatformV1Context functions}
