% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1Context.metadata}
\alias{GoogleCloudAiplatformV1Context.metadata}
\title{GoogleCloudAiplatformV1Context.metadata Object}
\usage{
GoogleCloudAiplatformV1Context.metadata()
}
\value{
GoogleCloudAiplatformV1Context.metadata object
}
\description{
GoogleCloudAiplatformV1Context.metadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Properties of the Context. Top level metadata keys' heading and trailing spaces will be trimmed. The size of this field should not exceed 200KB.
}
\seealso{
Other GoogleCloudAiplatformV1Context functions: 
\code{\link{GoogleCloudAiplatformV1Context.labels}()},
\code{\link{GoogleCloudAiplatformV1Context}()},
\code{\link{projects.locations.metadataStores.contexts.create}()},
\code{\link{projects.locations.metadataStores.contexts.patch}()}
}
\concept{GoogleCloudAiplatformV1Context functions}
