% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1Context}
\alias{GoogleCloudAiplatformV1Context}
\title{GoogleCloudAiplatformV1Context Object}
\usage{
GoogleCloudAiplatformV1Context(
  GoogleCloudAiplatformV1Context.labels = NULL,
  GoogleCloudAiplatformV1Context.metadata = NULL,
  schemaTitle = NULL,
  displayName = NULL,
  etag = NULL,
  labels = NULL,
  schemaVersion = NULL,
  metadata = NULL,
  name = NULL,
  description = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1Context.labels}{The \link{GoogleCloudAiplatformV1Context.labels} object or list of objects}

\item{GoogleCloudAiplatformV1Context.metadata}{The \link{GoogleCloudAiplatformV1Context.metadata} object or list of objects}

\item{schemaTitle}{The title of the schema describing the metadata}

\item{displayName}{User provided display name of the Context}

\item{etag}{An eTag used to perform consistent read-modify-write updates}

\item{labels}{The labels with user-defined metadata to organize your Contexts}

\item{schemaVersion}{The version of the schema in schema_name to use}

\item{metadata}{Properties of the Context}

\item{name}{Immutable}

\item{description}{Description of the Context}
}
\value{
GoogleCloudAiplatformV1Context object
}
\description{
GoogleCloudAiplatformV1Context Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Instance of a general context.
}
\seealso{
Other GoogleCloudAiplatformV1Context functions: 
\code{\link{GoogleCloudAiplatformV1Context.labels}()},
\code{\link{GoogleCloudAiplatformV1Context.metadata}()},
\code{\link{projects.locations.metadataStores.contexts.create}()},
\code{\link{projects.locations.metadataStores.contexts.patch}()}
}
\concept{GoogleCloudAiplatformV1Context functions}
