% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1CopyModelOperationMetadata}
\alias{GoogleCloudAiplatformV1CopyModelOperationMetadata}
\title{GoogleCloudAiplatformV1CopyModelOperationMetadata Object}
\usage{
GoogleCloudAiplatformV1CopyModelOperationMetadata(genericMetadata = NULL)
}
\arguments{
\item{genericMetadata}{The common part of the operation metadata}
}
\value{
GoogleCloudAiplatformV1CopyModelOperationMetadata object
}
\description{
GoogleCloudAiplatformV1CopyModelOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details of ModelService.CopyModel operation.
}
\concept{GoogleCloudAiplatformV1CopyModelOperationMetadata functions}
