% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1CopyModelRequest}
\alias{GoogleCloudAiplatformV1CopyModelRequest}
\title{GoogleCloudAiplatformV1CopyModelRequest Object}
\usage{
GoogleCloudAiplatformV1CopyModelRequest(
  sourceModel = NULL,
  encryptionSpec = NULL,
  parentModel = NULL,
  modelId = NULL
)
}
\arguments{
\item{sourceModel}{Required}

\item{encryptionSpec}{Customer-managed encryption key options}

\item{parentModel}{Optional}

\item{modelId}{Optional}
}
\value{
GoogleCloudAiplatformV1CopyModelRequest object
}
\description{
GoogleCloudAiplatformV1CopyModelRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for ModelService.CopyModel.
}
\seealso{
Other GoogleCloudAiplatformV1CopyModelRequest functions: 
\code{\link{projects.locations.models.copy}()}
}
\concept{GoogleCloudAiplatformV1CopyModelRequest functions}
