% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1CopyModelResponse}
\alias{GoogleCloudAiplatformV1CopyModelResponse}
\title{GoogleCloudAiplatformV1CopyModelResponse Object}
\usage{
GoogleCloudAiplatformV1CopyModelResponse(model = NULL)
}
\arguments{
\item{model}{The name of the copied Model resource}
}
\value{
GoogleCloudAiplatformV1CopyModelResponse object
}
\description{
GoogleCloudAiplatformV1CopyModelResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message of ModelService.CopyModel operation.
}
\concept{GoogleCloudAiplatformV1CopyModelResponse functions}
