% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1CreateDatasetOperationMetadata}
\alias{GoogleCloudAiplatformV1CreateDatasetOperationMetadata}
\title{GoogleCloudAiplatformV1CreateDatasetOperationMetadata Object}
\usage{
GoogleCloudAiplatformV1CreateDatasetOperationMetadata(genericMetadata = NULL)
}
\arguments{
\item{genericMetadata}{The operation generic information}
}
\value{
GoogleCloudAiplatformV1CreateDatasetOperationMetadata object
}
\description{
GoogleCloudAiplatformV1CreateDatasetOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Runtime operation information for DatasetService.CreateDataset.
}
\concept{GoogleCloudAiplatformV1CreateDatasetOperationMetadata functions}
