% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1CreateDatasetVersionOperationMetadata}
\alias{GoogleCloudAiplatformV1CreateDatasetVersionOperationMetadata}
\title{GoogleCloudAiplatformV1CreateDatasetVersionOperationMetadata Object}
\usage{
GoogleCloudAiplatformV1CreateDatasetVersionOperationMetadata(
  genericMetadata = NULL
)
}
\arguments{
\item{genericMetadata}{The common part of the operation metadata}
}
\value{
GoogleCloudAiplatformV1CreateDatasetVersionOperationMetadata object
}
\description{
GoogleCloudAiplatformV1CreateDatasetVersionOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Runtime operation information for DatasetService.CreateDatasetVersion.
}
\concept{GoogleCloudAiplatformV1CreateDatasetVersionOperationMetadata functions}
