% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1CreateEndpointOperationMetadata}
\alias{GoogleCloudAiplatformV1CreateEndpointOperationMetadata}
\title{GoogleCloudAiplatformV1CreateEndpointOperationMetadata Object}
\usage{
GoogleCloudAiplatformV1CreateEndpointOperationMetadata(genericMetadata = NULL)
}
\arguments{
\item{genericMetadata}{The operation generic information}
}
\value{
GoogleCloudAiplatformV1CreateEndpointOperationMetadata object
}
\description{
GoogleCloudAiplatformV1CreateEndpointOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Runtime operation information for EndpointService.CreateEndpoint.
}
\concept{GoogleCloudAiplatformV1CreateEndpointOperationMetadata functions}
