% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1CreateEntityTypeOperationMetadata}
\alias{GoogleCloudAiplatformV1CreateEntityTypeOperationMetadata}
\title{GoogleCloudAiplatformV1CreateEntityTypeOperationMetadata Object}
\usage{
GoogleCloudAiplatformV1CreateEntityTypeOperationMetadata(
  genericMetadata = NULL
)
}
\arguments{
\item{genericMetadata}{Operation metadata for EntityType}
}
\value{
GoogleCloudAiplatformV1CreateEntityTypeOperationMetadata object
}
\description{
GoogleCloudAiplatformV1CreateEntityTypeOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details of operations that perform create EntityType.
}
\concept{GoogleCloudAiplatformV1CreateEntityTypeOperationMetadata functions}
