% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1CreateFeatureRequest}
\alias{GoogleCloudAiplatformV1CreateFeatureRequest}
\title{GoogleCloudAiplatformV1CreateFeatureRequest Object}
\usage{
GoogleCloudAiplatformV1CreateFeatureRequest(
  feature = NULL,
  parent = NULL,
  featureId = NULL
)
}
\arguments{
\item{feature}{Required}

\item{parent}{Required}

\item{featureId}{Required}
}
\value{
GoogleCloudAiplatformV1CreateFeatureRequest object
}
\description{
GoogleCloudAiplatformV1CreateFeatureRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for FeaturestoreService.CreateFeature and FeatureRegistryService.CreateFeature.
}
\concept{GoogleCloudAiplatformV1CreateFeatureRequest functions}
