% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1CreateFeaturestoreOperationMetadata}
\alias{GoogleCloudAiplatformV1CreateFeaturestoreOperationMetadata}
\title{GoogleCloudAiplatformV1CreateFeaturestoreOperationMetadata Object}
\usage{
GoogleCloudAiplatformV1CreateFeaturestoreOperationMetadata(
  genericMetadata = NULL
)
}
\arguments{
\item{genericMetadata}{Operation metadata for Featurestore}
}
\value{
GoogleCloudAiplatformV1CreateFeaturestoreOperationMetadata object
}
\description{
GoogleCloudAiplatformV1CreateFeaturestoreOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details of operations that perform create Featurestore.
}
\concept{GoogleCloudAiplatformV1CreateFeaturestoreOperationMetadata functions}
