% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1CreateIndexEndpointOperationMetadata}
\alias{GoogleCloudAiplatformV1CreateIndexEndpointOperationMetadata}
\title{GoogleCloudAiplatformV1CreateIndexEndpointOperationMetadata Object}
\usage{
GoogleCloudAiplatformV1CreateIndexEndpointOperationMetadata(
  genericMetadata = NULL
)
}
\arguments{
\item{genericMetadata}{The operation generic information}
}
\value{
GoogleCloudAiplatformV1CreateIndexEndpointOperationMetadata object
}
\description{
GoogleCloudAiplatformV1CreateIndexEndpointOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Runtime operation information for IndexEndpointService.CreateIndexEndpoint.
}
\concept{GoogleCloudAiplatformV1CreateIndexEndpointOperationMetadata functions}
