% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1CreateIndexOperationMetadata}
\alias{GoogleCloudAiplatformV1CreateIndexOperationMetadata}
\title{GoogleCloudAiplatformV1CreateIndexOperationMetadata Object}
\usage{
GoogleCloudAiplatformV1CreateIndexOperationMetadata(
  genericMetadata = NULL,
  nearestNeighborSearchOperationMetadata = NULL
)
}
\arguments{
\item{genericMetadata}{The operation generic information}

\item{nearestNeighborSearchOperationMetadata}{The operation metadata with regard to Matching Engine Index operation}
}
\value{
GoogleCloudAiplatformV1CreateIndexOperationMetadata object
}
\description{
GoogleCloudAiplatformV1CreateIndexOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Runtime operation information for IndexService.CreateIndex.
}
\concept{GoogleCloudAiplatformV1CreateIndexOperationMetadata functions}
