% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1CreateMetadataStoreOperationMetadata}
\alias{GoogleCloudAiplatformV1CreateMetadataStoreOperationMetadata}
\title{GoogleCloudAiplatformV1CreateMetadataStoreOperationMetadata Object}
\usage{
GoogleCloudAiplatformV1CreateMetadataStoreOperationMetadata(
  genericMetadata = NULL
)
}
\arguments{
\item{genericMetadata}{Operation metadata for creating a MetadataStore}
}
\value{
GoogleCloudAiplatformV1CreateMetadataStoreOperationMetadata object
}
\description{
GoogleCloudAiplatformV1CreateMetadataStoreOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details of operations that perform MetadataService.CreateMetadataStore.
}
\concept{GoogleCloudAiplatformV1CreateMetadataStoreOperationMetadata functions}
