% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1CreatePipelineJobRequest}
\alias{GoogleCloudAiplatformV1CreatePipelineJobRequest}
\title{GoogleCloudAiplatformV1CreatePipelineJobRequest Object}
\usage{
GoogleCloudAiplatformV1CreatePipelineJobRequest(
  pipelineJobId = NULL,
  parent = NULL,
  pipelineJob = NULL
)
}
\arguments{
\item{pipelineJobId}{The ID to use for the PipelineJob, which will become the final component of the PipelineJob name}

\item{parent}{Required}

\item{pipelineJob}{Required}
}
\value{
GoogleCloudAiplatformV1CreatePipelineJobRequest object
}
\description{
GoogleCloudAiplatformV1CreatePipelineJobRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for PipelineService.CreatePipelineJob.
}
\concept{GoogleCloudAiplatformV1CreatePipelineJobRequest functions}
