% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1CreateSpecialistPoolOperationMetadata}
\alias{GoogleCloudAiplatformV1CreateSpecialistPoolOperationMetadata}
\title{GoogleCloudAiplatformV1CreateSpecialistPoolOperationMetadata Object}
\usage{
GoogleCloudAiplatformV1CreateSpecialistPoolOperationMetadata(
  genericMetadata = NULL
)
}
\arguments{
\item{genericMetadata}{The operation generic information}
}
\value{
GoogleCloudAiplatformV1CreateSpecialistPoolOperationMetadata object
}
\description{
GoogleCloudAiplatformV1CreateSpecialistPoolOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Runtime operation information for SpecialistPoolService.CreateSpecialistPool.
}
\concept{GoogleCloudAiplatformV1CreateSpecialistPoolOperationMetadata functions}
