% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1CreateTensorboardOperationMetadata}
\alias{GoogleCloudAiplatformV1CreateTensorboardOperationMetadata}
\title{GoogleCloudAiplatformV1CreateTensorboardOperationMetadata Object}
\usage{
GoogleCloudAiplatformV1CreateTensorboardOperationMetadata(
  genericMetadata = NULL
)
}
\arguments{
\item{genericMetadata}{Operation metadata for Tensorboard}
}
\value{
GoogleCloudAiplatformV1CreateTensorboardOperationMetadata object
}
\description{
GoogleCloudAiplatformV1CreateTensorboardOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details of operations that perform create Tensorboard.
}
\concept{GoogleCloudAiplatformV1CreateTensorboardOperationMetadata functions}
