% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1CreateTensorboardRunRequest}
\alias{GoogleCloudAiplatformV1CreateTensorboardRunRequest}
\title{GoogleCloudAiplatformV1CreateTensorboardRunRequest Object}
\usage{
GoogleCloudAiplatformV1CreateTensorboardRunRequest(
  tensorboardRun = NULL,
  parent = NULL,
  tensorboardRunId = NULL
)
}
\arguments{
\item{tensorboardRun}{Required}

\item{parent}{Required}

\item{tensorboardRunId}{Required}
}
\value{
GoogleCloudAiplatformV1CreateTensorboardRunRequest object
}
\description{
GoogleCloudAiplatformV1CreateTensorboardRunRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for TensorboardService.CreateTensorboardRun.
}
\concept{GoogleCloudAiplatformV1CreateTensorboardRunRequest functions}
