% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1CreateTensorboardTimeSeriesRequest}
\alias{GoogleCloudAiplatformV1CreateTensorboardTimeSeriesRequest}
\title{GoogleCloudAiplatformV1CreateTensorboardTimeSeriesRequest Object}
\usage{
GoogleCloudAiplatformV1CreateTensorboardTimeSeriesRequest(
  tensorboardTimeSeriesId = NULL,
  tensorboardTimeSeries = NULL,
  parent = NULL
)
}
\arguments{
\item{tensorboardTimeSeriesId}{Optional}

\item{tensorboardTimeSeries}{Required}

\item{parent}{Required}
}
\value{
GoogleCloudAiplatformV1CreateTensorboardTimeSeriesRequest object
}
\description{
GoogleCloudAiplatformV1CreateTensorboardTimeSeriesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for TensorboardService.CreateTensorboardTimeSeries.
}
\concept{GoogleCloudAiplatformV1CreateTensorboardTimeSeriesRequest functions}
