% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1CsvDestination}
\alias{GoogleCloudAiplatformV1CsvDestination}
\title{GoogleCloudAiplatformV1CsvDestination Object}
\usage{
GoogleCloudAiplatformV1CsvDestination(gcsDestination = NULL)
}
\arguments{
\item{gcsDestination}{Required}
}
\value{
GoogleCloudAiplatformV1CsvDestination object
}
\description{
GoogleCloudAiplatformV1CsvDestination Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The storage details for CSV output content.
}
\concept{GoogleCloudAiplatformV1CsvDestination functions}
