% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1CsvSource}
\alias{GoogleCloudAiplatformV1CsvSource}
\title{GoogleCloudAiplatformV1CsvSource Object}
\usage{
GoogleCloudAiplatformV1CsvSource(gcsSource = NULL)
}
\arguments{
\item{gcsSource}{Required}
}
\value{
GoogleCloudAiplatformV1CsvSource object
}
\description{
GoogleCloudAiplatformV1CsvSource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The storage details for CSV input content.
}
\concept{GoogleCloudAiplatformV1CsvSource functions}
