% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1CustomJob}
\alias{GoogleCloudAiplatformV1CustomJob}
\title{GoogleCloudAiplatformV1CustomJob Object}
\usage{
GoogleCloudAiplatformV1CustomJob(
  GoogleCloudAiplatformV1CustomJob.webAccessUris = NULL,
  GoogleCloudAiplatformV1CustomJob.labels = NULL,
  displayName = NULL,
  jobSpec = NULL,
  encryptionSpec = NULL,
  labels = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1CustomJob.webAccessUris}{The \link{GoogleCloudAiplatformV1CustomJob.webAccessUris} object or list of objects}

\item{GoogleCloudAiplatformV1CustomJob.labels}{The \link{GoogleCloudAiplatformV1CustomJob.labels} object or list of objects}

\item{displayName}{Required}

\item{jobSpec}{Required}

\item{encryptionSpec}{Customer-managed encryption key options for a CustomJob}

\item{labels}{The labels with user-defined metadata to organize CustomJobs}
}
\value{
GoogleCloudAiplatformV1CustomJob object
}
\description{
GoogleCloudAiplatformV1CustomJob Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a job that runs custom workloads such as a Docker container or a Python package. A CustomJob can have multiple worker pools and each worker pool can have its own machine and input spec. A CustomJob will be cleaned up once the job enters terminal state (failed or succeeded).
}
\seealso{
Other GoogleCloudAiplatformV1CustomJob functions: 
\code{\link{GoogleCloudAiplatformV1CustomJob.labels}()},
\code{\link{GoogleCloudAiplatformV1CustomJob.webAccessUris}()},
\code{\link{projects.locations.customJobs.create}()}
}
\concept{GoogleCloudAiplatformV1CustomJob functions}
