% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1CustomJob.webAccessUris}
\alias{GoogleCloudAiplatformV1CustomJob.webAccessUris}
\title{GoogleCloudAiplatformV1CustomJob.webAccessUris Object}
\usage{
GoogleCloudAiplatformV1CustomJob.webAccessUris()
}
\value{
GoogleCloudAiplatformV1CustomJob.webAccessUris object
}
\description{
GoogleCloudAiplatformV1CustomJob.webAccessUris Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Output only. URIs for accessing \href{https://cloud.google.com/vertex-ai/docs/training/monitor-debug-interactive-shell}{interactive shells} (one URI for each training node). Only available if job_spec.enable_web_access is \code{true}. The keys are names of each node in the training job; for example, \code{workerpool0-0} for the primary node, \code{workerpool1-0} for the first node in the second worker pool, and \code{workerpool1-1} for the second node in the second worker pool. The values are the URIs for each node's interactive shell.
}
\seealso{
Other GoogleCloudAiplatformV1CustomJob functions: 
\code{\link{GoogleCloudAiplatformV1CustomJob.labels}()},
\code{\link{GoogleCloudAiplatformV1CustomJob}()},
\code{\link{projects.locations.customJobs.create}()}
}
\concept{GoogleCloudAiplatformV1CustomJob functions}
