% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1CustomJobSpec}
\alias{GoogleCloudAiplatformV1CustomJobSpec}
\title{GoogleCloudAiplatformV1CustomJobSpec Object}
\usage{
GoogleCloudAiplatformV1CustomJobSpec(
  enableDashboardAccess = NULL,
  reservedIpRanges = NULL,
  serviceAccount = NULL,
  enableWebAccess = NULL,
  tensorboard = NULL,
  scheduling = NULL,
  workerPoolSpecs = NULL,
  network = NULL,
  experimentRun = NULL,
  baseOutputDirectory = NULL,
  experiment = NULL
)
}
\arguments{
\item{enableDashboardAccess}{Optional}

\item{reservedIpRanges}{Optional}

\item{serviceAccount}{Specifies the service account for workload run-as account}

\item{enableWebAccess}{Optional}

\item{tensorboard}{Optional}

\item{scheduling}{Scheduling options for a CustomJob}

\item{workerPoolSpecs}{Required}

\item{network}{Optional}

\item{experimentRun}{Optional}

\item{baseOutputDirectory}{The Cloud Storage location to store the output of this CustomJob or HyperparameterTuningJob}

\item{experiment}{Optional}
}
\value{
GoogleCloudAiplatformV1CustomJobSpec object
}
\description{
GoogleCloudAiplatformV1CustomJobSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents the spec of a CustomJob.
}
\concept{GoogleCloudAiplatformV1CustomJobSpec functions}
