% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1DataItem}
\alias{GoogleCloudAiplatformV1DataItem}
\title{GoogleCloudAiplatformV1DataItem Object}
\usage{
GoogleCloudAiplatformV1DataItem(
  GoogleCloudAiplatformV1DataItem.labels = NULL,
  etag = NULL,
  labels = NULL,
  payload = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1DataItem.labels}{The \link{GoogleCloudAiplatformV1DataItem.labels} object or list of objects}

\item{etag}{Optional}

\item{labels}{Optional}

\item{payload}{Required}
}
\value{
GoogleCloudAiplatformV1DataItem object
}
\description{
GoogleCloudAiplatformV1DataItem Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A piece of data in a Dataset. Could be an image, a video, a document or plain text.
}
\seealso{
Other GoogleCloudAiplatformV1DataItem functions: 
\code{\link{GoogleCloudAiplatformV1DataItem.labels}()}
}
\concept{GoogleCloudAiplatformV1DataItem functions}
