% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1DataItemView}
\alias{GoogleCloudAiplatformV1DataItemView}
\title{GoogleCloudAiplatformV1DataItemView Object}
\usage{
GoogleCloudAiplatformV1DataItemView(
  annotations = NULL,
  hasTruncatedAnnotations = NULL,
  dataItem = NULL
)
}
\arguments{
\item{annotations}{The Annotations on the DataItem}

\item{hasTruncatedAnnotations}{True if and only if the Annotations field has been truncated}

\item{dataItem}{The DataItem}
}
\value{
GoogleCloudAiplatformV1DataItemView object
}
\description{
GoogleCloudAiplatformV1DataItemView Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A container for a single DataItem and Annotations on it.
}
\concept{GoogleCloudAiplatformV1DataItemView functions}
