% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1DataLabelingJob.annotationLabels}
\alias{GoogleCloudAiplatformV1DataLabelingJob.annotationLabels}
\title{GoogleCloudAiplatformV1DataLabelingJob.annotationLabels Object}
\usage{
GoogleCloudAiplatformV1DataLabelingJob.annotationLabels()
}
\value{
GoogleCloudAiplatformV1DataLabelingJob.annotationLabels object
}
\description{
GoogleCloudAiplatformV1DataLabelingJob.annotationLabels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Labels to assign to annotations generated by this DataLabelingJob. Label keys and values can be no longer than 64 characters (Unicode codepoints), can only contain lowercase letters, numeric characters, underscores and dashes. International characters are allowed. See https://goo.gl/xmQnxf for more information and examples of labels. System reserved label keys are prefixed with 'aiplatform.googleapis.com/' and are immutable.
}
\seealso{
Other GoogleCloudAiplatformV1DataLabelingJob functions: 
\code{\link{GoogleCloudAiplatformV1DataLabelingJob.labels}()},
\code{\link{GoogleCloudAiplatformV1DataLabelingJob}()},
\code{\link{projects.locations.dataLabelingJobs.create}()}
}
\concept{GoogleCloudAiplatformV1DataLabelingJob functions}
