% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1DataLabelingJob.labels}
\alias{GoogleCloudAiplatformV1DataLabelingJob.labels}
\title{GoogleCloudAiplatformV1DataLabelingJob.labels Object}
\usage{
GoogleCloudAiplatformV1DataLabelingJob.labels()
}
\value{
GoogleCloudAiplatformV1DataLabelingJob.labels object
}
\description{
GoogleCloudAiplatformV1DataLabelingJob.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The labels with user-defined metadata to organize your DataLabelingJobs. Label keys and values can be no longer than 64 characters (Unicode codepoints), can only contain lowercase letters, numeric characters, underscores and dashes. International characters are allowed. See https://goo.gl/xmQnxf for more information and examples of labels. System reserved label keys are prefixed with 'aiplatform.googleapis.com/' and are immutable. Following system labels exist for each DataLabelingJob: * 'aiplatform.googleapis.com/schema': output only, its value is the inputs_schema's title.
}
\seealso{
Other GoogleCloudAiplatformV1DataLabelingJob functions: 
\code{\link{GoogleCloudAiplatformV1DataLabelingJob.annotationLabels}()},
\code{\link{GoogleCloudAiplatformV1DataLabelingJob}()},
\code{\link{projects.locations.dataLabelingJobs.create}()}
}
\concept{GoogleCloudAiplatformV1DataLabelingJob functions}
