% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1DataLabelingJob}
\alias{GoogleCloudAiplatformV1DataLabelingJob}
\title{GoogleCloudAiplatformV1DataLabelingJob Object}
\usage{
GoogleCloudAiplatformV1DataLabelingJob(
  GoogleCloudAiplatformV1DataLabelingJob.labels = NULL,
  GoogleCloudAiplatformV1DataLabelingJob.annotationLabels = NULL,
  datasets = NULL,
  displayName = NULL,
  activeLearningConfig = NULL,
  encryptionSpec = NULL,
  specialistPools = NULL,
  labels = NULL,
  inputsSchemaUri = NULL,
  instructionUri = NULL,
  inputs = NULL,
  annotationLabels = NULL,
  labelerCount = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1DataLabelingJob.labels}{The \link{GoogleCloudAiplatformV1DataLabelingJob.labels} object or list of objects}

\item{GoogleCloudAiplatformV1DataLabelingJob.annotationLabels}{The \link{GoogleCloudAiplatformV1DataLabelingJob.annotationLabels} object or list of objects}

\item{datasets}{Required}

\item{displayName}{Required}

\item{activeLearningConfig}{Parameters that configure the active learning pipeline}

\item{encryptionSpec}{Customer-managed encryption key spec for a DataLabelingJob}

\item{specialistPools}{The SpecialistPools' resource names associated with this job}

\item{labels}{The labels with user-defined metadata to organize your DataLabelingJobs}

\item{inputsSchemaUri}{Required}

\item{instructionUri}{Required}

\item{inputs}{Required}

\item{annotationLabels}{Labels to assign to annotations generated by this DataLabelingJob}

\item{labelerCount}{Required}
}
\value{
GoogleCloudAiplatformV1DataLabelingJob object
}
\description{
GoogleCloudAiplatformV1DataLabelingJob Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
DataLabelingJob is used to trigger a human labeling job on unlabeled data from the following Dataset:
}
\seealso{
Other GoogleCloudAiplatformV1DataLabelingJob functions: 
\code{\link{GoogleCloudAiplatformV1DataLabelingJob.annotationLabels}()},
\code{\link{GoogleCloudAiplatformV1DataLabelingJob.labels}()},
\code{\link{projects.locations.dataLabelingJobs.create}()}
}
\concept{GoogleCloudAiplatformV1DataLabelingJob functions}
