% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1Dataset.labels}
\alias{GoogleCloudAiplatformV1Dataset.labels}
\title{GoogleCloudAiplatformV1Dataset.labels Object}
\usage{
GoogleCloudAiplatformV1Dataset.labels()
}
\value{
GoogleCloudAiplatformV1Dataset.labels object
}
\description{
GoogleCloudAiplatformV1Dataset.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The labels with user-defined metadata to organize your Datasets. Label keys and values can be no longer than 64 characters (Unicode codepoints), can only contain lowercase letters, numeric characters, underscores and dashes. International characters are allowed. No more than 64 user labels can be associated with one Dataset (System labels are excluded). See https://goo.gl/xmQnxf for more information and examples of labels. System reserved label keys are prefixed with 'aiplatform.googleapis.com/' and are immutable. Following system labels exist for each Dataset: * 'aiplatform.googleapis.com/dataset_metadata_schema': output only, its value is the metadata_schema's title.
}
\seealso{
Other GoogleCloudAiplatformV1Dataset functions: 
\code{\link{GoogleCloudAiplatformV1Dataset}()},
\code{\link{projects.locations.datasets.create}()},
\code{\link{projects.locations.datasets.patch}()}
}
\concept{GoogleCloudAiplatformV1Dataset functions}
