% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1Dataset}
\alias{GoogleCloudAiplatformV1Dataset}
\title{GoogleCloudAiplatformV1Dataset Object}
\usage{
GoogleCloudAiplatformV1Dataset(
  GoogleCloudAiplatformV1Dataset.labels = NULL,
  encryptionSpec = NULL,
  etag = NULL,
  labels = NULL,
  displayName = NULL,
  metadata = NULL,
  metadataSchemaUri = NULL,
  savedQueries = NULL,
  description = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1Dataset.labels}{The \link{GoogleCloudAiplatformV1Dataset.labels} object or list of objects}

\item{encryptionSpec}{Customer-managed encryption key spec for a Dataset}

\item{etag}{Used to perform consistent read-modify-write updates}

\item{labels}{The labels with user-defined metadata to organize your Datasets}

\item{displayName}{Required}

\item{metadata}{Required}

\item{metadataSchemaUri}{Required}

\item{savedQueries}{All SavedQueries belong to the Dataset will be returned in List/Get Dataset response}

\item{description}{The description of the Dataset}
}
\value{
GoogleCloudAiplatformV1Dataset object
}
\description{
GoogleCloudAiplatformV1Dataset Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A collection of DataItems and Annotations on them.
}
\seealso{
Other GoogleCloudAiplatformV1Dataset functions: 
\code{\link{GoogleCloudAiplatformV1Dataset.labels}()},
\code{\link{projects.locations.datasets.create}()},
\code{\link{projects.locations.datasets.patch}()}
}
\concept{GoogleCloudAiplatformV1Dataset functions}
