% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1DedicatedResources}
\alias{GoogleCloudAiplatformV1DedicatedResources}
\title{GoogleCloudAiplatformV1DedicatedResources Object}
\usage{
GoogleCloudAiplatformV1DedicatedResources(
  autoscalingMetricSpecs = NULL,
  maxReplicaCount = NULL,
  minReplicaCount = NULL,
  machineSpec = NULL
)
}
\arguments{
\item{autoscalingMetricSpecs}{Immutable}

\item{maxReplicaCount}{Immutable}

\item{minReplicaCount}{Required}

\item{machineSpec}{Required}
}
\value{
GoogleCloudAiplatformV1DedicatedResources object
}
\description{
GoogleCloudAiplatformV1DedicatedResources Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A description of resources that are dedicated to a DeployedModel, and that need a higher degree of manual configuration.
}
\concept{GoogleCloudAiplatformV1DedicatedResources functions}
