% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1DeleteFeatureValuesRequest}
\alias{GoogleCloudAiplatformV1DeleteFeatureValuesRequest}
\title{GoogleCloudAiplatformV1DeleteFeatureValuesRequest Object}
\usage{
GoogleCloudAiplatformV1DeleteFeatureValuesRequest(
  selectTimeRangeAndFeature = NULL,
  selectEntity = NULL
)
}
\arguments{
\item{selectTimeRangeAndFeature}{Select feature values to be deleted by specifying time range and features}

\item{selectEntity}{Select feature values to be deleted by specifying entities}
}
\value{
GoogleCloudAiplatformV1DeleteFeatureValuesRequest object
}
\description{
GoogleCloudAiplatformV1DeleteFeatureValuesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for FeaturestoreService.DeleteFeatureValues.
}
\seealso{
Other GoogleCloudAiplatformV1DeleteFeatureValuesRequest functions: 
\code{\link{projects.locations.featurestores.entityTypes.deleteFeatureValues}()}
}
\concept{GoogleCloudAiplatformV1DeleteFeatureValuesRequest functions}
