% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1DeleteFeatureValuesRequestSelectEntity}
\alias{GoogleCloudAiplatformV1DeleteFeatureValuesRequestSelectEntity}
\title{GoogleCloudAiplatformV1DeleteFeatureValuesRequestSelectEntity Object}
\usage{
GoogleCloudAiplatformV1DeleteFeatureValuesRequestSelectEntity(
  entityIdSelector = NULL
)
}
\arguments{
\item{entityIdSelector}{Required}
}
\value{
GoogleCloudAiplatformV1DeleteFeatureValuesRequestSelectEntity object
}
\description{
GoogleCloudAiplatformV1DeleteFeatureValuesRequestSelectEntity Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Message to select entity. If an entity id is selected, all the feature values corresponding to the entity id will be deleted, including the entityId.
}
\concept{GoogleCloudAiplatformV1DeleteFeatureValuesRequestSelectEntity functions}
