% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1DeleteFeatureValuesRequestSelectTimeRangeAndFeature}
\alias{GoogleCloudAiplatformV1DeleteFeatureValuesRequestSelectTimeRangeAndFeature}
\title{GoogleCloudAiplatformV1DeleteFeatureValuesRequestSelectTimeRangeAndFeature Object}
\usage{
GoogleCloudAiplatformV1DeleteFeatureValuesRequestSelectTimeRangeAndFeature(
  skipOnlineStorageDelete = NULL,
  timeRange = NULL,
  featureSelector = NULL
)
}
\arguments{
\item{skipOnlineStorageDelete}{If set, data will not be deleted from online storage}

\item{timeRange}{Required}

\item{featureSelector}{Required}
}
\value{
GoogleCloudAiplatformV1DeleteFeatureValuesRequestSelectTimeRangeAndFeature object
}
\description{
GoogleCloudAiplatformV1DeleteFeatureValuesRequestSelectTimeRangeAndFeature Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Message to select time range and feature. Values of the selected feature generated within an inclusive time range will be deleted. Using this option permanently deletes the feature values from the specified feature IDs within the specified time range. This might include data from the online storage. If you want to retain any deleted historical data in the online storage, you must re-ingest it.
}
\concept{GoogleCloudAiplatformV1DeleteFeatureValuesRequestSelectTimeRangeAndFeature functions}
