% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1DeleteFeatureValuesResponseSelectEntity}
\alias{GoogleCloudAiplatformV1DeleteFeatureValuesResponseSelectEntity}
\title{GoogleCloudAiplatformV1DeleteFeatureValuesResponseSelectEntity Object}
\usage{
GoogleCloudAiplatformV1DeleteFeatureValuesResponseSelectEntity(
  offlineStorageDeletedEntityRowCount = NULL,
  onlineStorageDeletedEntityCount = NULL
)
}
\arguments{
\item{offlineStorageDeletedEntityRowCount}{The count of deleted entity rows in the offline storage}

\item{onlineStorageDeletedEntityCount}{The count of deleted entities in the online storage}
}
\value{
GoogleCloudAiplatformV1DeleteFeatureValuesResponseSelectEntity object
}
\description{
GoogleCloudAiplatformV1DeleteFeatureValuesResponseSelectEntity Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message if the request uses the SelectEntity option.
}
\concept{GoogleCloudAiplatformV1DeleteFeatureValuesResponseSelectEntity functions}
