% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1DeleteFeatureValuesResponseSelectTimeRangeAndFeature}
\alias{GoogleCloudAiplatformV1DeleteFeatureValuesResponseSelectTimeRangeAndFeature}
\title{GoogleCloudAiplatformV1DeleteFeatureValuesResponseSelectTimeRangeAndFeature Object}
\usage{
GoogleCloudAiplatformV1DeleteFeatureValuesResponseSelectTimeRangeAndFeature(
  offlineStorageModifiedEntityRowCount = NULL,
  impactedFeatureCount = NULL,
  onlineStorageModifiedEntityCount = NULL
)
}
\arguments{
\item{offlineStorageModifiedEntityRowCount}{The count of modified entity rows in the offline storage}

\item{impactedFeatureCount}{The count of the features or columns impacted}

\item{onlineStorageModifiedEntityCount}{The count of modified entities in the online storage}
}
\value{
GoogleCloudAiplatformV1DeleteFeatureValuesResponseSelectTimeRangeAndFeature object
}
\description{
GoogleCloudAiplatformV1DeleteFeatureValuesResponseSelectTimeRangeAndFeature Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message if the request uses the SelectTimeRangeAndFeature option.
}
\concept{GoogleCloudAiplatformV1DeleteFeatureValuesResponseSelectTimeRangeAndFeature functions}
