% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1DeleteMetadataStoreOperationMetadata}
\alias{GoogleCloudAiplatformV1DeleteMetadataStoreOperationMetadata}
\title{GoogleCloudAiplatformV1DeleteMetadataStoreOperationMetadata Object}
\usage{
GoogleCloudAiplatformV1DeleteMetadataStoreOperationMetadata(
  genericMetadata = NULL
)
}
\arguments{
\item{genericMetadata}{Operation metadata for deleting a MetadataStore}
}
\value{
GoogleCloudAiplatformV1DeleteMetadataStoreOperationMetadata object
}
\description{
GoogleCloudAiplatformV1DeleteMetadataStoreOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details of operations that perform MetadataService.DeleteMetadataStore.
}
\concept{GoogleCloudAiplatformV1DeleteMetadataStoreOperationMetadata functions}
