% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1DeleteOperationMetadata}
\alias{GoogleCloudAiplatformV1DeleteOperationMetadata}
\title{GoogleCloudAiplatformV1DeleteOperationMetadata Object}
\usage{
GoogleCloudAiplatformV1DeleteOperationMetadata(genericMetadata = NULL)
}
\arguments{
\item{genericMetadata}{The common part of the operation metadata}
}
\value{
GoogleCloudAiplatformV1DeleteOperationMetadata object
}
\description{
GoogleCloudAiplatformV1DeleteOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details of operations that perform deletes of any entities.
}
\concept{GoogleCloudAiplatformV1DeleteOperationMetadata functions}
