% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1DeployedIndex}
\alias{GoogleCloudAiplatformV1DeployedIndex}
\title{GoogleCloudAiplatformV1DeployedIndex Object}
\usage{
GoogleCloudAiplatformV1DeployedIndex(
  reservedIpRanges = NULL,
  automaticResources = NULL,
  deploymentGroup = NULL,
  displayName = NULL,
  index = NULL,
  dedicatedResources = NULL,
  deployedIndexAuthConfig = NULL,
  id = NULL,
  enableAccessLogging = NULL
)
}
\arguments{
\item{reservedIpRanges}{Optional}

\item{automaticResources}{Optional}

\item{deploymentGroup}{Optional}

\item{displayName}{The display name of the DeployedIndex}

\item{index}{Required}

\item{dedicatedResources}{Optional}

\item{deployedIndexAuthConfig}{Optional}

\item{id}{Required}

\item{enableAccessLogging}{Optional}
}
\value{
GoogleCloudAiplatformV1DeployedIndex object
}
\description{
GoogleCloudAiplatformV1DeployedIndex Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A deployment of an Index. IndexEndpoints contain one or more DeployedIndexes.
}
\seealso{
Other GoogleCloudAiplatformV1DeployedIndex functions: 
\code{\link{projects.locations.indexEndpoints.mutateDeployedIndex}()}
}
\concept{GoogleCloudAiplatformV1DeployedIndex functions}
