% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1DeployedIndexAuthConfig}
\alias{GoogleCloudAiplatformV1DeployedIndexAuthConfig}
\title{GoogleCloudAiplatformV1DeployedIndexAuthConfig Object}
\usage{
GoogleCloudAiplatformV1DeployedIndexAuthConfig(authProvider = NULL)
}
\arguments{
\item{authProvider}{Defines the authentication provider that the DeployedIndex uses}
}
\value{
GoogleCloudAiplatformV1DeployedIndexAuthConfig object
}
\description{
GoogleCloudAiplatformV1DeployedIndexAuthConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Used to set up the auth on the DeployedIndex's private endpoint.
}
\concept{GoogleCloudAiplatformV1DeployedIndexAuthConfig functions}
