% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1DeployedIndexAuthConfigAuthProvider}
\alias{GoogleCloudAiplatformV1DeployedIndexAuthConfigAuthProvider}
\title{GoogleCloudAiplatformV1DeployedIndexAuthConfigAuthProvider Object}
\usage{
GoogleCloudAiplatformV1DeployedIndexAuthConfigAuthProvider(
  audiences = NULL,
  allowedIssuers = NULL
)
}
\arguments{
\item{audiences}{The list of JWT \link{audiences}(https://tools}

\item{allowedIssuers}{A list of allowed JWT issuers}
}
\value{
GoogleCloudAiplatformV1DeployedIndexAuthConfigAuthProvider object
}
\description{
GoogleCloudAiplatformV1DeployedIndexAuthConfigAuthProvider Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration for an authentication provider, including support for \href{https://tools.ietf.org/html/draft-ietf-oauth-json-web-token-32}{JSON Web Token (JWT)}.
}
\concept{GoogleCloudAiplatformV1DeployedIndexAuthConfigAuthProvider functions}
