% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1DeployedIndexRef}
\alias{GoogleCloudAiplatformV1DeployedIndexRef}
\title{GoogleCloudAiplatformV1DeployedIndexRef Object}
\usage{
GoogleCloudAiplatformV1DeployedIndexRef(
  indexEndpoint = NULL,
  deployedIndexId = NULL
)
}
\arguments{
\item{indexEndpoint}{Immutable}

\item{deployedIndexId}{Immutable}
}
\value{
GoogleCloudAiplatformV1DeployedIndexRef object
}
\description{
GoogleCloudAiplatformV1DeployedIndexRef Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Points to a DeployedIndex.
}
\concept{GoogleCloudAiplatformV1DeployedIndexRef functions}
