% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1DeployedModelRef}
\alias{GoogleCloudAiplatformV1DeployedModelRef}
\title{GoogleCloudAiplatformV1DeployedModelRef Object}
\usage{
GoogleCloudAiplatformV1DeployedModelRef(
  deployedModelId = NULL,
  endpoint = NULL
)
}
\arguments{
\item{deployedModelId}{Immutable}

\item{endpoint}{Immutable}
}
\value{
GoogleCloudAiplatformV1DeployedModelRef object
}
\description{
GoogleCloudAiplatformV1DeployedModelRef Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Points to a DeployedModel.
}
\concept{GoogleCloudAiplatformV1DeployedModelRef functions}
