% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1DeployIndexOperationMetadata}
\alias{GoogleCloudAiplatformV1DeployIndexOperationMetadata}
\title{GoogleCloudAiplatformV1DeployIndexOperationMetadata Object}
\usage{
GoogleCloudAiplatformV1DeployIndexOperationMetadata(
  genericMetadata = NULL,
  deployedIndexId = NULL
)
}
\arguments{
\item{genericMetadata}{The operation generic information}

\item{deployedIndexId}{The unique index id specified by user}
}
\value{
GoogleCloudAiplatformV1DeployIndexOperationMetadata object
}
\description{
GoogleCloudAiplatformV1DeployIndexOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Runtime operation information for IndexEndpointService.DeployIndex.
}
\concept{GoogleCloudAiplatformV1DeployIndexOperationMetadata functions}
