% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1DeployIndexRequest}
\alias{GoogleCloudAiplatformV1DeployIndexRequest}
\title{GoogleCloudAiplatformV1DeployIndexRequest Object}
\usage{
GoogleCloudAiplatformV1DeployIndexRequest(deployedIndex = NULL)
}
\arguments{
\item{deployedIndex}{Required}
}
\value{
GoogleCloudAiplatformV1DeployIndexRequest object
}
\description{
GoogleCloudAiplatformV1DeployIndexRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for IndexEndpointService.DeployIndex.
}
\seealso{
Other GoogleCloudAiplatformV1DeployIndexRequest functions: 
\code{\link{projects.locations.indexEndpoints.deployIndex}()}
}
\concept{GoogleCloudAiplatformV1DeployIndexRequest functions}
