% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1DeployIndexResponse}
\alias{GoogleCloudAiplatformV1DeployIndexResponse}
\title{GoogleCloudAiplatformV1DeployIndexResponse Object}
\usage{
GoogleCloudAiplatformV1DeployIndexResponse(deployedIndex = NULL)
}
\arguments{
\item{deployedIndex}{The DeployedIndex that had been deployed in the IndexEndpoint}
}
\value{
GoogleCloudAiplatformV1DeployIndexResponse object
}
\description{
GoogleCloudAiplatformV1DeployIndexResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for IndexEndpointService.DeployIndex.
}
\concept{GoogleCloudAiplatformV1DeployIndexResponse functions}
