% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1DeployModelOperationMetadata}
\alias{GoogleCloudAiplatformV1DeployModelOperationMetadata}
\title{GoogleCloudAiplatformV1DeployModelOperationMetadata Object}
\usage{
GoogleCloudAiplatformV1DeployModelOperationMetadata(genericMetadata = NULL)
}
\arguments{
\item{genericMetadata}{The operation generic information}
}
\value{
GoogleCloudAiplatformV1DeployModelOperationMetadata object
}
\description{
GoogleCloudAiplatformV1DeployModelOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Runtime operation information for EndpointService.DeployModel.
}
\concept{GoogleCloudAiplatformV1DeployModelOperationMetadata functions}
