% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1DeployModelRequest}
\alias{GoogleCloudAiplatformV1DeployModelRequest}
\title{GoogleCloudAiplatformV1DeployModelRequest Object}
\usage{
GoogleCloudAiplatformV1DeployModelRequest(
  GoogleCloudAiplatformV1DeployModelRequest.trafficSplit = NULL,
  deployedModel = NULL,
  trafficSplit = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1DeployModelRequest.trafficSplit}{The \link{GoogleCloudAiplatformV1DeployModelRequest.trafficSplit} object or list of objects}

\item{deployedModel}{Required}

\item{trafficSplit}{A map from a DeployedModel's ID to the percentage of this Endpoint's traffic that should be forwarded to that DeployedModel}
}
\value{
GoogleCloudAiplatformV1DeployModelRequest object
}
\description{
GoogleCloudAiplatformV1DeployModelRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for EndpointService.DeployModel.
}
\seealso{
Other GoogleCloudAiplatformV1DeployModelRequest functions: 
\code{\link{GoogleCloudAiplatformV1DeployModelRequest.trafficSplit}()},
\code{\link{projects.locations.endpoints.deployModel}()}
}
\concept{GoogleCloudAiplatformV1DeployModelRequest functions}
