% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1DeployModelRequest.trafficSplit}
\alias{GoogleCloudAiplatformV1DeployModelRequest.trafficSplit}
\title{GoogleCloudAiplatformV1DeployModelRequest.trafficSplit Object}
\usage{
GoogleCloudAiplatformV1DeployModelRequest.trafficSplit()
}
\value{
GoogleCloudAiplatformV1DeployModelRequest.trafficSplit object
}
\description{
GoogleCloudAiplatformV1DeployModelRequest.trafficSplit Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A map from a DeployedModel's ID to the percentage of this Endpoint's traffic that should be forwarded to that DeployedModel. If this field is non-empty, then the Endpoint's traffic_split will be overwritten with it. To refer to the ID of the just being deployed Model, a '0' should be used, and the actual ID of the new DeployedModel will be filled in its place by this method. The traffic percentage values must add up to 100. If this field is empty, then the Endpoint's traffic_split is not updated.
}
\seealso{
Other GoogleCloudAiplatformV1DeployModelRequest functions: 
\code{\link{GoogleCloudAiplatformV1DeployModelRequest}()},
\code{\link{projects.locations.endpoints.deployModel}()}
}
\concept{GoogleCloudAiplatformV1DeployModelRequest functions}
