% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1DeployModelResponse}
\alias{GoogleCloudAiplatformV1DeployModelResponse}
\title{GoogleCloudAiplatformV1DeployModelResponse Object}
\usage{
GoogleCloudAiplatformV1DeployModelResponse(deployedModel = NULL)
}
\arguments{
\item{deployedModel}{The DeployedModel that had been deployed in the Endpoint}
}
\value{
GoogleCloudAiplatformV1DeployModelResponse object
}
\description{
GoogleCloudAiplatformV1DeployModelResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for EndpointService.DeployModel.
}
\concept{GoogleCloudAiplatformV1DeployModelResponse functions}
