% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1DestinationFeatureSetting}
\alias{GoogleCloudAiplatformV1DestinationFeatureSetting}
\title{GoogleCloudAiplatformV1DestinationFeatureSetting Object}
\usage{
GoogleCloudAiplatformV1DestinationFeatureSetting(
  destinationField = NULL,
  featureId = NULL
)
}
\arguments{
\item{destinationField}{Specify the field name in the export destination}

\item{featureId}{Required}
}
\value{
GoogleCloudAiplatformV1DestinationFeatureSetting object
}
\description{
GoogleCloudAiplatformV1DestinationFeatureSetting Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\concept{GoogleCloudAiplatformV1DestinationFeatureSetting functions}
