% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1DiskSpec}
\alias{GoogleCloudAiplatformV1DiskSpec}
\title{GoogleCloudAiplatformV1DiskSpec Object}
\usage{
GoogleCloudAiplatformV1DiskSpec(bootDiskSizeGb = NULL, bootDiskType = NULL)
}
\arguments{
\item{bootDiskSizeGb}{Size in GB of the boot disk (default is 100GB)}

\item{bootDiskType}{Type of the boot disk (default is 'pd-ssd')}
}
\value{
GoogleCloudAiplatformV1DiskSpec object
}
\description{
GoogleCloudAiplatformV1DiskSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents the spec of disk options.
}
\concept{GoogleCloudAiplatformV1DiskSpec functions}
