% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1DoubleArray}
\alias{GoogleCloudAiplatformV1DoubleArray}
\title{GoogleCloudAiplatformV1DoubleArray Object}
\usage{
GoogleCloudAiplatformV1DoubleArray(values = NULL)
}
\arguments{
\item{values}{A list of double values}
}
\value{
GoogleCloudAiplatformV1DoubleArray object
}
\description{
GoogleCloudAiplatformV1DoubleArray Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of double values.
}
\concept{GoogleCloudAiplatformV1DoubleArray functions}
