% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1EncryptionSpec}
\alias{GoogleCloudAiplatformV1EncryptionSpec}
\title{GoogleCloudAiplatformV1EncryptionSpec Object}
\usage{
GoogleCloudAiplatformV1EncryptionSpec(kmsKeyName = NULL)
}
\arguments{
\item{kmsKeyName}{Required}
}
\value{
GoogleCloudAiplatformV1EncryptionSpec object
}
\description{
GoogleCloudAiplatformV1EncryptionSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a customer-managed encryption key spec that can be applied to a top-level resource.
}
\concept{GoogleCloudAiplatformV1EncryptionSpec functions}
