% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1Endpoint.labels}
\alias{GoogleCloudAiplatformV1Endpoint.labels}
\title{GoogleCloudAiplatformV1Endpoint.labels Object}
\usage{
GoogleCloudAiplatformV1Endpoint.labels()
}
\value{
GoogleCloudAiplatformV1Endpoint.labels object
}
\description{
GoogleCloudAiplatformV1Endpoint.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The labels with user-defined metadata to organize your Endpoints. Label keys and values can be no longer than 64 characters (Unicode codepoints), can only contain lowercase letters, numeric characters, underscores and dashes. International characters are allowed. See https://goo.gl/xmQnxf for more information and examples of labels.
}
\seealso{
Other GoogleCloudAiplatformV1Endpoint functions: 
\code{\link{GoogleCloudAiplatformV1Endpoint.trafficSplit}()},
\code{\link{GoogleCloudAiplatformV1Endpoint}()},
\code{\link{projects.locations.endpoints.create}()},
\code{\link{projects.locations.endpoints.patch}()}
}
\concept{GoogleCloudAiplatformV1Endpoint functions}
