% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1Endpoint}
\alias{GoogleCloudAiplatformV1Endpoint}
\title{GoogleCloudAiplatformV1Endpoint Object}
\usage{
GoogleCloudAiplatformV1Endpoint(
  GoogleCloudAiplatformV1Endpoint.labels = NULL,
  GoogleCloudAiplatformV1Endpoint.trafficSplit = NULL,
  encryptionSpec = NULL,
  description = NULL,
  predictRequestResponseLoggingConfig = NULL,
  labels = NULL,
  etag = NULL,
  trafficSplit = NULL,
  enablePrivateServiceConnect = NULL,
  network = NULL,
  displayName = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1Endpoint.labels}{The \link{GoogleCloudAiplatformV1Endpoint.labels} object or list of objects}

\item{GoogleCloudAiplatformV1Endpoint.trafficSplit}{The \link{GoogleCloudAiplatformV1Endpoint.trafficSplit} object or list of objects}

\item{encryptionSpec}{Customer-managed encryption key spec for an Endpoint}

\item{description}{The description of the Endpoint}

\item{predictRequestResponseLoggingConfig}{Configures the request-response logging for online prediction}

\item{labels}{The labels with user-defined metadata to organize your Endpoints}

\item{etag}{Used to perform consistent read-modify-write updates}

\item{trafficSplit}{A map from a DeployedModel's ID to the percentage of this Endpoint's traffic that should be forwarded to that DeployedModel}

\item{enablePrivateServiceConnect}{Deprecated: If true, expose the Endpoint via private service connect}

\item{network}{Optional}

\item{displayName}{Required}
}
\value{
GoogleCloudAiplatformV1Endpoint object
}
\description{
GoogleCloudAiplatformV1Endpoint Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Models are deployed into it, and afterwards Endpoint is called to obtain predictions and explanations.
}
\seealso{
Other GoogleCloudAiplatformV1Endpoint functions: 
\code{\link{GoogleCloudAiplatformV1Endpoint.labels}()},
\code{\link{GoogleCloudAiplatformV1Endpoint.trafficSplit}()},
\code{\link{projects.locations.endpoints.create}()},
\code{\link{projects.locations.endpoints.patch}()}
}
\concept{GoogleCloudAiplatformV1Endpoint functions}
