% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1Endpoint.trafficSplit}
\alias{GoogleCloudAiplatformV1Endpoint.trafficSplit}
\title{GoogleCloudAiplatformV1Endpoint.trafficSplit Object}
\usage{
GoogleCloudAiplatformV1Endpoint.trafficSplit()
}
\value{
GoogleCloudAiplatformV1Endpoint.trafficSplit object
}
\description{
GoogleCloudAiplatformV1Endpoint.trafficSplit Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A map from a DeployedModel's ID to the percentage of this Endpoint's traffic that should be forwarded to that DeployedModel. If a DeployedModel's ID is not listed in this map, then it receives no traffic. The traffic percentage values must add up to 100, or map must be empty if the Endpoint is to not accept any traffic at a moment.
}
\seealso{
Other GoogleCloudAiplatformV1Endpoint functions: 
\code{\link{GoogleCloudAiplatformV1Endpoint.labels}()},
\code{\link{GoogleCloudAiplatformV1Endpoint}()},
\code{\link{projects.locations.endpoints.create}()},
\code{\link{projects.locations.endpoints.patch}()}
}
\concept{GoogleCloudAiplatformV1Endpoint functions}
