% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1EntityIdSelector}
\alias{GoogleCloudAiplatformV1EntityIdSelector}
\title{GoogleCloudAiplatformV1EntityIdSelector Object}
\usage{
GoogleCloudAiplatformV1EntityIdSelector(csvSource = NULL, entityIdField = NULL)
}
\arguments{
\item{csvSource}{Source of Csv}

\item{entityIdField}{Source column that holds entity IDs}
}
\value{
GoogleCloudAiplatformV1EntityIdSelector object
}
\description{
GoogleCloudAiplatformV1EntityIdSelector Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Selector for entityId. Getting ids from the given source.
}
\concept{GoogleCloudAiplatformV1EntityIdSelector functions}
