% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1EnvVar}
\alias{GoogleCloudAiplatformV1EnvVar}
\title{GoogleCloudAiplatformV1EnvVar Object}
\usage{
GoogleCloudAiplatformV1EnvVar(value = NULL, name = NULL)
}
\arguments{
\item{value}{Required}

\item{name}{Required}
}
\value{
GoogleCloudAiplatformV1EnvVar object
}
\description{
GoogleCloudAiplatformV1EnvVar Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents an environment variable present in a Container or Python Module.
}
\concept{GoogleCloudAiplatformV1EnvVar functions}
