% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ErrorAnalysisAnnotation}
\alias{GoogleCloudAiplatformV1ErrorAnalysisAnnotation}
\title{GoogleCloudAiplatformV1ErrorAnalysisAnnotation Object}
\usage{
GoogleCloudAiplatformV1ErrorAnalysisAnnotation(
  queryType = NULL,
  outlierScore = NULL,
  attributedItems = NULL,
  outlierThreshold = NULL
)
}
\arguments{
\item{queryType}{The query type used for finding the attributed items}

\item{outlierScore}{The outlier score of this annotated item}

\item{attributedItems}{Attributed items for a given annotation, typically representing neighbors from the training sets constrained by the query type}

\item{outlierThreshold}{The threshold used to determine if this annotation is an outlier or not}
}
\value{
GoogleCloudAiplatformV1ErrorAnalysisAnnotation object
}
\description{
GoogleCloudAiplatformV1ErrorAnalysisAnnotation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Model error analysis for each annotation.
}
\concept{GoogleCloudAiplatformV1ErrorAnalysisAnnotation functions}
