% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ErrorAnalysisAnnotationAttributedItem}
\alias{GoogleCloudAiplatformV1ErrorAnalysisAnnotationAttributedItem}
\title{GoogleCloudAiplatformV1ErrorAnalysisAnnotationAttributedItem Object}
\usage{
GoogleCloudAiplatformV1ErrorAnalysisAnnotationAttributedItem(
  annotationResourceName = NULL,
  distance = NULL
)
}
\arguments{
\item{annotationResourceName}{The unique ID for each annotation}

\item{distance}{The distance of this item to the annotation}
}
\value{
GoogleCloudAiplatformV1ErrorAnalysisAnnotationAttributedItem object
}
\description{
GoogleCloudAiplatformV1ErrorAnalysisAnnotationAttributedItem Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Attributed items for a given annotation, typically representing neighbors from the training sets constrained by the query type.
}
\concept{GoogleCloudAiplatformV1ErrorAnalysisAnnotationAttributedItem functions}
