% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1EvaluatedAnnotation}
\alias{GoogleCloudAiplatformV1EvaluatedAnnotation}
\title{GoogleCloudAiplatformV1EvaluatedAnnotation Object}
\usage{
GoogleCloudAiplatformV1EvaluatedAnnotation(
  errorAnalysisAnnotations = NULL,
  explanations = NULL
)
}
\arguments{
\item{errorAnalysisAnnotations}{Annotations of model error analysis results}

\item{explanations}{Explanations of predictions}
}
\value{
GoogleCloudAiplatformV1EvaluatedAnnotation object
}
\description{
GoogleCloudAiplatformV1EvaluatedAnnotation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
True positive, false positive, or false negative. EvaluatedAnnotation is only available under ModelEvaluationSlice with slice of \code{annotationSpec} dimension.
}
\concept{GoogleCloudAiplatformV1EvaluatedAnnotation functions}
