% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1EvaluatedAnnotationExplanation}
\alias{GoogleCloudAiplatformV1EvaluatedAnnotationExplanation}
\title{GoogleCloudAiplatformV1EvaluatedAnnotationExplanation Object}
\usage{
GoogleCloudAiplatformV1EvaluatedAnnotationExplanation(
  explanationType = NULL,
  explanation = NULL
)
}
\arguments{
\item{explanationType}{Explanation type}

\item{explanation}{Explanation attribution response details}
}
\value{
GoogleCloudAiplatformV1EvaluatedAnnotationExplanation object
}
\description{
GoogleCloudAiplatformV1EvaluatedAnnotationExplanation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Explanation result of the prediction produced by the Model.
}
\concept{GoogleCloudAiplatformV1EvaluatedAnnotationExplanation functions}
