% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1Event.labels}
\alias{GoogleCloudAiplatformV1Event.labels}
\title{GoogleCloudAiplatformV1Event.labels Object}
\usage{
GoogleCloudAiplatformV1Event.labels()
}
\value{
GoogleCloudAiplatformV1Event.labels object
}
\description{
GoogleCloudAiplatformV1Event.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The labels with user-defined metadata to annotate Events. Label keys and values can be no longer than 64 characters (Unicode codepoints), can only contain lowercase letters, numeric characters, underscores and dashes. International characters are allowed. No more than 64 user labels can be associated with one Event (System labels are excluded). See https://goo.gl/xmQnxf for more information and examples of labels. System reserved label keys are prefixed with 'aiplatform.googleapis.com/' and are immutable.
}
\seealso{
Other GoogleCloudAiplatformV1Event functions: 
\code{\link{GoogleCloudAiplatformV1Event}()}
}
\concept{GoogleCloudAiplatformV1Event functions}
