% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1Event}
\alias{GoogleCloudAiplatformV1Event}
\title{GoogleCloudAiplatformV1Event Object}
\usage{
GoogleCloudAiplatformV1Event(
  GoogleCloudAiplatformV1Event.labels = NULL,
  labels = NULL,
  artifact = NULL,
  type = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1Event.labels}{The \link{GoogleCloudAiplatformV1Event.labels} object or list of objects}

\item{labels}{The labels with user-defined metadata to annotate Events}

\item{artifact}{Required}

\item{type}{Required}
}
\value{
GoogleCloudAiplatformV1Event object
}
\description{
GoogleCloudAiplatformV1Event Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An edge describing the relationship between an Artifact and an Execution in a lineage graph.
}
\seealso{
Other GoogleCloudAiplatformV1Event functions: 
\code{\link{GoogleCloudAiplatformV1Event.labels}()}
}
\concept{GoogleCloudAiplatformV1Event functions}
