% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1Examples}
\alias{GoogleCloudAiplatformV1Examples}
\title{GoogleCloudAiplatformV1Examples Object}
\usage{
GoogleCloudAiplatformV1Examples(
  nearestNeighborSearchConfig = NULL,
  exampleGcsSource = NULL,
  presets = NULL,
  neighborCount = NULL
)
}
\arguments{
\item{nearestNeighborSearchConfig}{The full configuration for the generated index, the semantics are the same as metadata and should match \link{NearestNeighborSearchConfig}(https://cloud}

\item{exampleGcsSource}{The Cloud Storage input instances}

\item{presets}{Simplified preset configuration, which automatically sets configuration values based on the desired query speed-precision trade-off and modality}

\item{neighborCount}{The number of neighbors to return when querying for examples}
}
\value{
GoogleCloudAiplatformV1Examples object
}
\description{
GoogleCloudAiplatformV1Examples Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Example-based explainability that returns the nearest neighbors from the provided dataset.
}
\concept{GoogleCloudAiplatformV1Examples functions}
