% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ExamplesExampleGcsSource}
\alias{GoogleCloudAiplatformV1ExamplesExampleGcsSource}
\title{GoogleCloudAiplatformV1ExamplesExampleGcsSource Object}
\usage{
GoogleCloudAiplatformV1ExamplesExampleGcsSource(
  gcsSource = NULL,
  dataFormat = NULL
)
}
\arguments{
\item{gcsSource}{The Cloud Storage location for the input instances}

\item{dataFormat}{The format in which instances are given, if not specified, assume it's JSONL format}
}
\value{
GoogleCloudAiplatformV1ExamplesExampleGcsSource object
}
\description{
GoogleCloudAiplatformV1ExamplesExampleGcsSource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The Cloud Storage input instances.
}
\concept{GoogleCloudAiplatformV1ExamplesExampleGcsSource functions}
