% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ExamplesOverride}
\alias{GoogleCloudAiplatformV1ExamplesOverride}
\title{GoogleCloudAiplatformV1ExamplesOverride Object}
\usage{
GoogleCloudAiplatformV1ExamplesOverride(
  crowdingCount = NULL,
  restrictions = NULL,
  neighborCount = NULL,
  dataFormat = NULL,
  returnEmbeddings = NULL
)
}
\arguments{
\item{crowdingCount}{The number of neighbors to return that have the same crowding tag}

\item{restrictions}{Restrict the resulting nearest neighbors to respect these constraints}

\item{neighborCount}{The number of neighbors to return}

\item{dataFormat}{The format of the data being provided with each call}

\item{returnEmbeddings}{If true, return the embeddings instead of neighbors}
}
\value{
GoogleCloudAiplatformV1ExamplesOverride object
}
\description{
GoogleCloudAiplatformV1ExamplesOverride Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Overrides for example-based explanations.
}
\concept{GoogleCloudAiplatformV1ExamplesOverride functions}
