% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ExamplesRestrictionsNamespace}
\alias{GoogleCloudAiplatformV1ExamplesRestrictionsNamespace}
\title{GoogleCloudAiplatformV1ExamplesRestrictionsNamespace Object}
\usage{
GoogleCloudAiplatformV1ExamplesRestrictionsNamespace(
  deny = NULL,
  allow = NULL,
  namespaceName = NULL
)
}
\arguments{
\item{deny}{The list of deny tags}

\item{allow}{The list of allowed tags}

\item{namespaceName}{The namespace name}
}
\value{
GoogleCloudAiplatformV1ExamplesRestrictionsNamespace object
}
\description{
GoogleCloudAiplatformV1ExamplesRestrictionsNamespace Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Restrictions namespace for example-based explanations overrides.
}
\concept{GoogleCloudAiplatformV1ExamplesRestrictionsNamespace functions}
