% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1Execution.labels}
\alias{GoogleCloudAiplatformV1Execution.labels}
\title{GoogleCloudAiplatformV1Execution.labels Object}
\usage{
GoogleCloudAiplatformV1Execution.labels()
}
\value{
GoogleCloudAiplatformV1Execution.labels object
}
\description{
GoogleCloudAiplatformV1Execution.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The labels with user-defined metadata to organize your Executions. Label keys and values can be no longer than 64 characters (Unicode codepoints), can only contain lowercase letters, numeric characters, underscores and dashes. International characters are allowed. No more than 64 user labels can be associated with one Execution (System labels are excluded).
}
\seealso{
Other GoogleCloudAiplatformV1Execution functions: 
\code{\link{GoogleCloudAiplatformV1Execution.metadata}()},
\code{\link{GoogleCloudAiplatformV1Execution}()},
\code{\link{projects.locations.metadataStores.executions.create}()},
\code{\link{projects.locations.metadataStores.executions.patch}()}
}
\concept{GoogleCloudAiplatformV1Execution functions}
