% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1Execution.metadata}
\alias{GoogleCloudAiplatformV1Execution.metadata}
\title{GoogleCloudAiplatformV1Execution.metadata Object}
\usage{
GoogleCloudAiplatformV1Execution.metadata()
}
\value{
GoogleCloudAiplatformV1Execution.metadata object
}
\description{
GoogleCloudAiplatformV1Execution.metadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Properties of the Execution. Top level metadata keys' heading and trailing spaces will be trimmed. The size of this field should not exceed 200KB.
}
\seealso{
Other GoogleCloudAiplatformV1Execution functions: 
\code{\link{GoogleCloudAiplatformV1Execution.labels}()},
\code{\link{GoogleCloudAiplatformV1Execution}()},
\code{\link{projects.locations.metadataStores.executions.create}()},
\code{\link{projects.locations.metadataStores.executions.patch}()}
}
\concept{GoogleCloudAiplatformV1Execution functions}
