% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1Execution}
\alias{GoogleCloudAiplatformV1Execution}
\title{GoogleCloudAiplatformV1Execution Object}
\usage{
GoogleCloudAiplatformV1Execution(
  GoogleCloudAiplatformV1Execution.metadata = NULL,
  GoogleCloudAiplatformV1Execution.labels = NULL,
  displayName = NULL,
  description = NULL,
  etag = NULL,
  schemaTitle = NULL,
  schemaVersion = NULL,
  metadata = NULL,
  state = NULL,
  labels = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1Execution.metadata}{The \link{GoogleCloudAiplatformV1Execution.metadata} object or list of objects}

\item{GoogleCloudAiplatformV1Execution.labels}{The \link{GoogleCloudAiplatformV1Execution.labels} object or list of objects}

\item{displayName}{User provided display name of the Execution}

\item{description}{Description of the Execution}

\item{etag}{An eTag used to perform consistent read-modify-write updates}

\item{schemaTitle}{The title of the schema describing the metadata}

\item{schemaVersion}{The version of the schema in \code{schema_title} to use}

\item{metadata}{Properties of the Execution}

\item{state}{The state of this Execution}

\item{labels}{The labels with user-defined metadata to organize your Executions}
}
\value{
GoogleCloudAiplatformV1Execution object
}
\description{
GoogleCloudAiplatformV1Execution Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Instance of a general execution.
}
\seealso{
Other GoogleCloudAiplatformV1Execution functions: 
\code{\link{GoogleCloudAiplatformV1Execution.labels}()},
\code{\link{GoogleCloudAiplatformV1Execution.metadata}()},
\code{\link{projects.locations.metadataStores.executions.create}()},
\code{\link{projects.locations.metadataStores.executions.patch}()}
}
\concept{GoogleCloudAiplatformV1Execution functions}
