% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ExplainRequest}
\alias{GoogleCloudAiplatformV1ExplainRequest}
\title{GoogleCloudAiplatformV1ExplainRequest Object}
\usage{
GoogleCloudAiplatformV1ExplainRequest(
  explanationSpecOverride = NULL,
  parameters = NULL,
  instances = NULL,
  deployedModelId = NULL
)
}
\arguments{
\item{explanationSpecOverride}{If specified, overrides the explanation_spec of the DeployedModel}

\item{parameters}{The parameters that govern the prediction}

\item{instances}{Required}

\item{deployedModelId}{If specified, this ExplainRequest will be served by the chosen DeployedModel, overriding Endpoint}
}
\value{
GoogleCloudAiplatformV1ExplainRequest object
}
\description{
GoogleCloudAiplatformV1ExplainRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for PredictionService.Explain.
}
\seealso{
Other GoogleCloudAiplatformV1ExplainRequest functions: 
\code{\link{projects.locations.endpoints.explain}()}
}
\concept{GoogleCloudAiplatformV1ExplainRequest functions}
