% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ExplainResponse}
\alias{GoogleCloudAiplatformV1ExplainResponse}
\title{GoogleCloudAiplatformV1ExplainResponse Object}
\usage{
GoogleCloudAiplatformV1ExplainResponse(
  explanations = NULL,
  predictions = NULL,
  deployedModelId = NULL
)
}
\arguments{
\item{explanations}{The explanations of the Model's PredictResponse}

\item{predictions}{The predictions that are the output of the predictions call}

\item{deployedModelId}{ID of the Endpoint's DeployedModel that served this explanation}
}
\value{
GoogleCloudAiplatformV1ExplainResponse object
}
\description{
GoogleCloudAiplatformV1ExplainResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for PredictionService.Explain.
}
\concept{GoogleCloudAiplatformV1ExplainResponse functions}
