% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ExplanationMetadata.inputs}
\alias{GoogleCloudAiplatformV1ExplanationMetadata.inputs}
\title{GoogleCloudAiplatformV1ExplanationMetadata.inputs Object}
\usage{
GoogleCloudAiplatformV1ExplanationMetadata.inputs()
}
\value{
GoogleCloudAiplatformV1ExplanationMetadata.inputs object
}
\description{
GoogleCloudAiplatformV1ExplanationMetadata.inputs Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Required. Map from feature names to feature input metadata. Keys are the name of the features. Values are the specification of the feature. An empty InputMetadata is valid. It describes a text feature which has the name specified as the key in ExplanationMetadata.inputs. The baseline of the empty feature is chosen by Vertex AI. For Vertex AI-provided Tensorflow images, the key can be any friendly name of the feature. Once specified, featureAttributions are keyed by this key (if not grouped with another feature). For custom images, the key must match with the key in instance.
}
\seealso{
Other GoogleCloudAiplatformV1ExplanationMetadata functions: 
\code{\link{GoogleCloudAiplatformV1ExplanationMetadata.outputs}()},
\code{\link{GoogleCloudAiplatformV1ExplanationMetadata}()}
}
\concept{GoogleCloudAiplatformV1ExplanationMetadata functions}
