% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ExplanationMetadata.outputs}
\alias{GoogleCloudAiplatformV1ExplanationMetadata.outputs}
\title{GoogleCloudAiplatformV1ExplanationMetadata.outputs Object}
\usage{
GoogleCloudAiplatformV1ExplanationMetadata.outputs()
}
\value{
GoogleCloudAiplatformV1ExplanationMetadata.outputs object
}
\description{
GoogleCloudAiplatformV1ExplanationMetadata.outputs Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Required. Map from output names to output metadata. For Vertex AI-provided Tensorflow images, keys can be any user defined string that consists of any UTF-8 characters. For custom images, keys are the name of the output field in the prediction to be explained. Currently only one key is allowed.
}
\seealso{
Other GoogleCloudAiplatformV1ExplanationMetadata functions: 
\code{\link{GoogleCloudAiplatformV1ExplanationMetadata.inputs}()},
\code{\link{GoogleCloudAiplatformV1ExplanationMetadata}()}
}
\concept{GoogleCloudAiplatformV1ExplanationMetadata functions}
