% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ExplanationMetadata}
\alias{GoogleCloudAiplatformV1ExplanationMetadata}
\title{GoogleCloudAiplatformV1ExplanationMetadata Object}
\usage{
GoogleCloudAiplatformV1ExplanationMetadata(
  GoogleCloudAiplatformV1ExplanationMetadata.outputs = NULL,
  GoogleCloudAiplatformV1ExplanationMetadata.inputs = NULL,
  outputs = NULL,
  inputs = NULL,
  featureAttributionsSchemaUri = NULL,
  latentSpaceSource = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1ExplanationMetadata.outputs}{The \link{GoogleCloudAiplatformV1ExplanationMetadata.outputs} object or list of objects}

\item{GoogleCloudAiplatformV1ExplanationMetadata.inputs}{The \link{GoogleCloudAiplatformV1ExplanationMetadata.inputs} object or list of objects}

\item{outputs}{Required}

\item{inputs}{Required}

\item{featureAttributionsSchemaUri}{Points to a YAML file stored on Google Cloud Storage describing the format of the feature attributions}

\item{latentSpaceSource}{Name of the source to generate embeddings for example based explanations}
}
\value{
GoogleCloudAiplatformV1ExplanationMetadata object
}
\description{
GoogleCloudAiplatformV1ExplanationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata describing the Model's input and output for explanation.
}
\seealso{
Other GoogleCloudAiplatformV1ExplanationMetadata functions: 
\code{\link{GoogleCloudAiplatformV1ExplanationMetadata.inputs}()},
\code{\link{GoogleCloudAiplatformV1ExplanationMetadata.outputs}()}
}
\concept{GoogleCloudAiplatformV1ExplanationMetadata functions}
