% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ExplanationMetadataInputMetadataFeatureValueDomain}
\alias{GoogleCloudAiplatformV1ExplanationMetadataInputMetadataFeatureValueDomain}
\title{GoogleCloudAiplatformV1ExplanationMetadataInputMetadataFeatureValueDomain Object}
\usage{
GoogleCloudAiplatformV1ExplanationMetadataInputMetadataFeatureValueDomain(
  originalMean = NULL,
  originalStddev = NULL,
  maxValue = NULL,
  minValue = NULL
)
}
\arguments{
\item{originalMean}{If this input feature has been normalized to a mean value of 0, the original_mean specifies the mean value of the domain prior to normalization}

\item{originalStddev}{If this input feature has been normalized to a standard deviation of 1}

\item{maxValue}{The maximum permissible value for this feature}

\item{minValue}{The minimum permissible value for this feature}
}
\value{
GoogleCloudAiplatformV1ExplanationMetadataInputMetadataFeatureValueDomain object
}
\description{
GoogleCloudAiplatformV1ExplanationMetadataInputMetadataFeatureValueDomain Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Domain details of the input feature value. Provides numeric information about the feature, such as its range (min, max). If the feature has been pre-processed, for example with z-scoring, then it provides information about how to recover the original feature. For example, if the input feature is an image and it has been pre-processed to obtain 0-mean and stddev = 1 values, then original_mean, and original_stddev refer to the mean and stddev of the original feature (e.g. image tensor) from which input feature (with mean = 0 and stddev = 1) was obtained.
}
\concept{GoogleCloudAiplatformV1ExplanationMetadataInputMetadataFeatureValueDomain functions}
