% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ExplanationMetadataInputMetadataVisualization}
\alias{GoogleCloudAiplatformV1ExplanationMetadataInputMetadataVisualization}
\title{GoogleCloudAiplatformV1ExplanationMetadataInputMetadataVisualization Object}
\usage{
GoogleCloudAiplatformV1ExplanationMetadataInputMetadataVisualization(
  type = NULL,
  polarity = NULL,
  clipPercentUpperbound = NULL,
  clipPercentLowerbound = NULL,
  overlayType = NULL,
  colorMap = NULL
)
}
\arguments{
\item{type}{Type of the image visualization}

\item{polarity}{Whether to only highlight pixels with positive contributions, negative or both}

\item{clipPercentUpperbound}{Excludes attributions above the specified percentile from the highlighted areas}

\item{clipPercentLowerbound}{Excludes attributions below the specified percentile, from the highlighted areas}

\item{overlayType}{How the original image is displayed in the visualization}

\item{colorMap}{The color scheme used for the highlighted areas}
}
\value{
GoogleCloudAiplatformV1ExplanationMetadataInputMetadataVisualization object
}
\description{
GoogleCloudAiplatformV1ExplanationMetadataInputMetadataVisualization Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Visualization configurations for image explanation.
}
\concept{GoogleCloudAiplatformV1ExplanationMetadataInputMetadataVisualization functions}
