% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ExplanationMetadataOutputMetadata}
\alias{GoogleCloudAiplatformV1ExplanationMetadataOutputMetadata}
\title{GoogleCloudAiplatformV1ExplanationMetadataOutputMetadata Object}
\usage{
GoogleCloudAiplatformV1ExplanationMetadataOutputMetadata(
  outputTensorName = NULL,
  displayNameMappingKey = NULL,
  indexDisplayNameMapping = NULL
)
}
\arguments{
\item{outputTensorName}{Name of the output tensor}

\item{displayNameMappingKey}{Specify a field name in the prediction to look for the display name}

\item{indexDisplayNameMapping}{Static mapping between the index and display name}
}
\value{
GoogleCloudAiplatformV1ExplanationMetadataOutputMetadata object
}
\description{
GoogleCloudAiplatformV1ExplanationMetadataOutputMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata of the prediction output to be explained.
}
\concept{GoogleCloudAiplatformV1ExplanationMetadataOutputMetadata functions}
