% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ExplanationMetadataOverride.inputs}
\alias{GoogleCloudAiplatformV1ExplanationMetadataOverride.inputs}
\title{GoogleCloudAiplatformV1ExplanationMetadataOverride.inputs Object}
\usage{
GoogleCloudAiplatformV1ExplanationMetadataOverride.inputs()
}
\value{
GoogleCloudAiplatformV1ExplanationMetadataOverride.inputs object
}
\description{
GoogleCloudAiplatformV1ExplanationMetadataOverride.inputs Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Required. Overrides the input metadata of the features. The key is the name of the feature to be overridden. The keys specified here must exist in the input metadata to be overridden. If a feature is not specified here, the corresponding feature's input metadata is not overridden.
}
\seealso{
Other GoogleCloudAiplatformV1ExplanationMetadataOverride functions: 
\code{\link{GoogleCloudAiplatformV1ExplanationMetadataOverride}()}
}
\concept{GoogleCloudAiplatformV1ExplanationMetadataOverride functions}
