% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ExplanationMetadataOverride}
\alias{GoogleCloudAiplatformV1ExplanationMetadataOverride}
\title{GoogleCloudAiplatformV1ExplanationMetadataOverride Object}
\usage{
GoogleCloudAiplatformV1ExplanationMetadataOverride(
  GoogleCloudAiplatformV1ExplanationMetadataOverride.inputs = NULL,
  inputs = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1ExplanationMetadataOverride.inputs}{The \link{GoogleCloudAiplatformV1ExplanationMetadataOverride.inputs} object or list of objects}

\item{inputs}{Required}
}
\value{
GoogleCloudAiplatformV1ExplanationMetadataOverride object
}
\description{
GoogleCloudAiplatformV1ExplanationMetadataOverride Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The ExplanationMetadata entries that can be overridden at online explanation time.
}
\seealso{
Other GoogleCloudAiplatformV1ExplanationMetadataOverride functions: 
\code{\link{GoogleCloudAiplatformV1ExplanationMetadataOverride.inputs}()}
}
\concept{GoogleCloudAiplatformV1ExplanationMetadataOverride functions}
