% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ExplanationParameters}
\alias{GoogleCloudAiplatformV1ExplanationParameters}
\title{GoogleCloudAiplatformV1ExplanationParameters Object}
\usage{
GoogleCloudAiplatformV1ExplanationParameters(
  examples = NULL,
  sampledShapleyAttribution = NULL,
  outputIndices = NULL,
  integratedGradientsAttribution = NULL,
  xraiAttribution = NULL,
  topK = NULL
)
}
\arguments{
\item{examples}{Example-based explanations that returns the nearest neighbors from the provided dataset}

\item{sampledShapleyAttribution}{An attribution method that approximates Shapley values for features that contribute to the label being predicted}

\item{outputIndices}{If populated, only returns attributions that have output_index contained in output_indices}

\item{integratedGradientsAttribution}{An attribution method that computes Aumann-Shapley values taking advantage of the model's fully differentiable structure}

\item{xraiAttribution}{An attribution method that redistributes Integrated Gradients attribution to segmented regions, taking advantage of the model's fully differentiable structure}

\item{topK}{If populated, returns attributions for top K indices of outputs (defaults to 1)}
}
\value{
GoogleCloudAiplatformV1ExplanationParameters object
}
\description{
GoogleCloudAiplatformV1ExplanationParameters Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Parameters to configure explaining for Model's predictions.
}
\concept{GoogleCloudAiplatformV1ExplanationParameters functions}
