% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ExplanationSpec}
\alias{GoogleCloudAiplatformV1ExplanationSpec}
\title{GoogleCloudAiplatformV1ExplanationSpec Object}
\usage{
GoogleCloudAiplatformV1ExplanationSpec(metadata = NULL, parameters = NULL)
}
\arguments{
\item{metadata}{Optional}

\item{parameters}{Required}
}
\value{
GoogleCloudAiplatformV1ExplanationSpec object
}
\description{
GoogleCloudAiplatformV1ExplanationSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Specification of Model explanation.
}
\concept{GoogleCloudAiplatformV1ExplanationSpec functions}
