% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ExportDataConfig}
\alias{GoogleCloudAiplatformV1ExportDataConfig}
\title{GoogleCloudAiplatformV1ExportDataConfig Object}
\usage{
GoogleCloudAiplatformV1ExportDataConfig(
  gcsDestination = NULL,
  annotationsFilter = NULL,
  fractionSplit = NULL
)
}
\arguments{
\item{gcsDestination}{The Google Cloud Storage location where the output is to be written to}

\item{annotationsFilter}{An expression for filtering what part of the Dataset is to be exported}

\item{fractionSplit}{Split based on fractions defining the size of each set}
}
\value{
GoogleCloudAiplatformV1ExportDataConfig object
}
\description{
GoogleCloudAiplatformV1ExportDataConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describes what part of the Dataset is to be exported, the destination of the export and how to export.
}
\concept{GoogleCloudAiplatformV1ExportDataConfig functions}
