% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ExportDataOperationMetadata}
\alias{GoogleCloudAiplatformV1ExportDataOperationMetadata}
\title{GoogleCloudAiplatformV1ExportDataOperationMetadata Object}
\usage{
GoogleCloudAiplatformV1ExportDataOperationMetadata(
  gcsOutputDirectory = NULL,
  genericMetadata = NULL
)
}
\arguments{
\item{gcsOutputDirectory}{A Google Cloud Storage directory which path ends with '/'}

\item{genericMetadata}{The common part of the operation metadata}
}
\value{
GoogleCloudAiplatformV1ExportDataOperationMetadata object
}
\description{
GoogleCloudAiplatformV1ExportDataOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Runtime operation information for DatasetService.ExportData.
}
\concept{GoogleCloudAiplatformV1ExportDataOperationMetadata functions}
