% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ExportDataRequest}
\alias{GoogleCloudAiplatformV1ExportDataRequest}
\title{GoogleCloudAiplatformV1ExportDataRequest Object}
\usage{
GoogleCloudAiplatformV1ExportDataRequest(exportConfig = NULL)
}
\arguments{
\item{exportConfig}{Required}
}
\value{
GoogleCloudAiplatformV1ExportDataRequest object
}
\description{
GoogleCloudAiplatformV1ExportDataRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for DatasetService.ExportData.
}
\seealso{
Other GoogleCloudAiplatformV1ExportDataRequest functions: 
\code{\link{projects.locations.datasets.export}()}
}
\concept{GoogleCloudAiplatformV1ExportDataRequest functions}
