% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ExportDataResponse}
\alias{GoogleCloudAiplatformV1ExportDataResponse}
\title{GoogleCloudAiplatformV1ExportDataResponse Object}
\usage{
GoogleCloudAiplatformV1ExportDataResponse(exportedFiles = NULL)
}
\arguments{
\item{exportedFiles}{All of the files that are exported in this export operation}
}
\value{
GoogleCloudAiplatformV1ExportDataResponse object
}
\description{
GoogleCloudAiplatformV1ExportDataResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for DatasetService.ExportData.
}
\concept{GoogleCloudAiplatformV1ExportDataResponse functions}
