% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ExportFeatureValuesRequest}
\alias{GoogleCloudAiplatformV1ExportFeatureValuesRequest}
\title{GoogleCloudAiplatformV1ExportFeatureValuesRequest Object}
\usage{
GoogleCloudAiplatformV1ExportFeatureValuesRequest(
  featureSelector = NULL,
  destination = NULL,
  snapshotExport = NULL,
  fullExport = NULL,
  settings = NULL
)
}
\arguments{
\item{featureSelector}{Required}

\item{destination}{Required}

\item{snapshotExport}{Exports the latest Feature values of all entities of the EntityType within a time range}

\item{fullExport}{Exports all historical values of all entities of the EntityType within a time range}

\item{settings}{Per-Feature export settings}
}
\value{
GoogleCloudAiplatformV1ExportFeatureValuesRequest object
}
\description{
GoogleCloudAiplatformV1ExportFeatureValuesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for FeaturestoreService.ExportFeatureValues.
}
\seealso{
Other GoogleCloudAiplatformV1ExportFeatureValuesRequest functions: 
\code{\link{projects.locations.featurestores.entityTypes.exportFeatureValues}()}
}
\concept{GoogleCloudAiplatformV1ExportFeatureValuesRequest functions}
