% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ExportFeatureValuesRequestSnapshotExport}
\alias{GoogleCloudAiplatformV1ExportFeatureValuesRequestSnapshotExport}
\title{GoogleCloudAiplatformV1ExportFeatureValuesRequestSnapshotExport Object}
\usage{
GoogleCloudAiplatformV1ExportFeatureValuesRequestSnapshotExport(
  startTime = NULL,
  snapshotTime = NULL
)
}
\arguments{
\item{startTime}{Excludes Feature values with feature generation timestamp before this timestamp}

\item{snapshotTime}{Exports Feature values as of this timestamp}
}
\value{
GoogleCloudAiplatformV1ExportFeatureValuesRequestSnapshotExport object
}
\description{
GoogleCloudAiplatformV1ExportFeatureValuesRequestSnapshotExport Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describes exporting the latest Feature values of all entities of the EntityType between \link{start_time, snapshot_time}.
}
\concept{GoogleCloudAiplatformV1ExportFeatureValuesRequestSnapshotExport functions}
