% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ExportModelOperationMetadata}
\alias{GoogleCloudAiplatformV1ExportModelOperationMetadata}
\title{GoogleCloudAiplatformV1ExportModelOperationMetadata Object}
\usage{
GoogleCloudAiplatformV1ExportModelOperationMetadata(genericMetadata = NULL)
}
\arguments{
\item{genericMetadata}{The common part of the operation metadata}
}
\value{
GoogleCloudAiplatformV1ExportModelOperationMetadata object
}
\description{
GoogleCloudAiplatformV1ExportModelOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details of ModelService.ExportModel operation.
}
\concept{GoogleCloudAiplatformV1ExportModelOperationMetadata functions}
