% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ExportModelOperationMetadataOutputInfo}
\alias{GoogleCloudAiplatformV1ExportModelOperationMetadataOutputInfo}
\title{GoogleCloudAiplatformV1ExportModelOperationMetadataOutputInfo Object}
\usage{
GoogleCloudAiplatformV1ExportModelOperationMetadataOutputInfo()
}
\value{
GoogleCloudAiplatformV1ExportModelOperationMetadataOutputInfo object
}
\description{
GoogleCloudAiplatformV1ExportModelOperationMetadataOutputInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Further describes the output of the ExportModel. Supplements ExportModelRequest.OutputConfig.
}
\concept{GoogleCloudAiplatformV1ExportModelOperationMetadataOutputInfo functions}
