% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ExportModelRequest}
\alias{GoogleCloudAiplatformV1ExportModelRequest}
\title{GoogleCloudAiplatformV1ExportModelRequest Object}
\usage{
GoogleCloudAiplatformV1ExportModelRequest(outputConfig = NULL)
}
\arguments{
\item{outputConfig}{Required}
}
\value{
GoogleCloudAiplatformV1ExportModelRequest object
}
\description{
GoogleCloudAiplatformV1ExportModelRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for ModelService.ExportModel.
}
\seealso{
Other GoogleCloudAiplatformV1ExportModelRequest functions: 
\code{\link{projects.locations.models.export}()}
}
\concept{GoogleCloudAiplatformV1ExportModelRequest functions}
