% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ExportModelRequestOutputConfig}
\alias{GoogleCloudAiplatformV1ExportModelRequestOutputConfig}
\title{GoogleCloudAiplatformV1ExportModelRequestOutputConfig Object}
\usage{
GoogleCloudAiplatformV1ExportModelRequestOutputConfig(
  artifactDestination = NULL,
  imageDestination = NULL,
  exportFormatId = NULL
)
}
\arguments{
\item{artifactDestination}{The Cloud Storage location where the Model artifact is to be written to}

\item{imageDestination}{The Google Container Registry or Artifact Registry uri where the Model container image will be copied to}

\item{exportFormatId}{The ID of the format in which the Model must be exported}
}
\value{
GoogleCloudAiplatformV1ExportModelRequestOutputConfig object
}
\description{
GoogleCloudAiplatformV1ExportModelRequestOutputConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Output configuration for the Model export.
}
\concept{GoogleCloudAiplatformV1ExportModelRequestOutputConfig functions}
