% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ExportTensorboardTimeSeriesDataRequest}
\alias{GoogleCloudAiplatformV1ExportTensorboardTimeSeriesDataRequest}
\title{GoogleCloudAiplatformV1ExportTensorboardTimeSeriesDataRequest Object}
\usage{
GoogleCloudAiplatformV1ExportTensorboardTimeSeriesDataRequest(
  pageSize = NULL,
  filter = NULL,
  pageToken = NULL,
  orderBy = NULL
)
}
\arguments{
\item{pageSize}{The maximum number of data points to return per page}

\item{filter}{Exports the TensorboardTimeSeries' data that match the filter expression}

\item{pageToken}{A page token, received from a previous ExportTensorboardTimeSeriesData call}

\item{orderBy}{Field to use to sort the TensorboardTimeSeries' data}
}
\value{
GoogleCloudAiplatformV1ExportTensorboardTimeSeriesDataRequest object
}
\description{
GoogleCloudAiplatformV1ExportTensorboardTimeSeriesDataRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for TensorboardService.ExportTensorboardTimeSeriesData.
}
\seealso{
Other GoogleCloudAiplatformV1ExportTensorboardTimeSeriesDataRequest functions: 
\code{\link{projects.locations.tensorboards.experiments.runs.timeSeries.exportTensorboardTimeSeries}()}
}
\concept{GoogleCloudAiplatformV1ExportTensorboardTimeSeriesDataRequest functions}
