% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ExportTensorboardTimeSeriesDataResponse}
\alias{GoogleCloudAiplatformV1ExportTensorboardTimeSeriesDataResponse}
\title{GoogleCloudAiplatformV1ExportTensorboardTimeSeriesDataResponse Object}
\usage{
GoogleCloudAiplatformV1ExportTensorboardTimeSeriesDataResponse(
  nextPageToken = NULL,
  timeSeriesDataPoints = NULL
)
}
\arguments{
\item{nextPageToken}{A token, which can be sent as page_token to retrieve the next page}

\item{timeSeriesDataPoints}{The returned time series data points}
}
\value{
GoogleCloudAiplatformV1ExportTensorboardTimeSeriesDataResponse object
}
\description{
GoogleCloudAiplatformV1ExportTensorboardTimeSeriesDataResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for TensorboardService.ExportTensorboardTimeSeriesData.
}
\concept{GoogleCloudAiplatformV1ExportTensorboardTimeSeriesDataResponse functions}
