% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1Feature}
\alias{GoogleCloudAiplatformV1Feature}
\title{GoogleCloudAiplatformV1Feature Object}
\usage{
GoogleCloudAiplatformV1Feature(
  GoogleCloudAiplatformV1Feature.versionLabels = NULL,
  GoogleCloudAiplatformV1Feature.labels = NULL,
  valueType = NULL,
  name = NULL,
  disableMonitoring = NULL,
  versionLabels = NULL,
  description = NULL,
  labels = NULL,
  etag = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1Feature.versionLabels}{The \link{GoogleCloudAiplatformV1Feature.versionLabels} object or list of objects}

\item{GoogleCloudAiplatformV1Feature.labels}{The \link{GoogleCloudAiplatformV1Feature.labels} object or list of objects}

\item{valueType}{Required}

\item{name}{Immutable}

\item{disableMonitoring}{Optional}

\item{versionLabels}{The labels with user-defined metadata to organize your versions}

\item{description}{Description of the Feature}

\item{labels}{Optional}

\item{etag}{Used to perform a consistent read-modify-write updates}
}
\value{
GoogleCloudAiplatformV1Feature object
}
\description{
GoogleCloudAiplatformV1Feature Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Feature Metadata information that describes an attribute of an entity type. For example, apple is an entity type, and color is a feature that describes apple.
}
\seealso{
Other GoogleCloudAiplatformV1Feature functions: 
\code{\link{GoogleCloudAiplatformV1Feature.labels}()},
\code{\link{GoogleCloudAiplatformV1Feature.versionLabels}()},
\code{\link{projects.locations.featurestores.entityTypes.features.create}()},
\code{\link{projects.locations.featurestores.entityTypes.features.patch}()}
}
\concept{GoogleCloudAiplatformV1Feature functions}
