% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1Feature.versionLabels}
\alias{GoogleCloudAiplatformV1Feature.versionLabels}
\title{GoogleCloudAiplatformV1Feature.versionLabels Object}
\usage{
GoogleCloudAiplatformV1Feature.versionLabels()
}
\value{
GoogleCloudAiplatformV1Feature.versionLabels object
}
\description{
GoogleCloudAiplatformV1Feature.versionLabels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The labels with user-defined metadata to organize your versions. Label keys and values can be no longer than 64 characters (Unicode codepoints), can only contain lowercase letters, numeric characters, underscores and dashes. International characters are allowed. See https://goo.gl/xmQnxf for more information and examples of labels.
}
\seealso{
Other GoogleCloudAiplatformV1Feature functions: 
\code{\link{GoogleCloudAiplatformV1Feature.labels}()},
\code{\link{GoogleCloudAiplatformV1Feature}()},
\code{\link{projects.locations.featurestores.entityTypes.features.create}()},
\code{\link{projects.locations.featurestores.entityTypes.features.patch}()}
}
\concept{GoogleCloudAiplatformV1Feature functions}
