% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1FeatureMonitoringStatsAnomaly}
\alias{GoogleCloudAiplatformV1FeatureMonitoringStatsAnomaly}
\title{GoogleCloudAiplatformV1FeatureMonitoringStatsAnomaly Object}
\usage{
GoogleCloudAiplatformV1FeatureMonitoringStatsAnomaly()
}
\value{
GoogleCloudAiplatformV1FeatureMonitoringStatsAnomaly object
}
\description{
GoogleCloudAiplatformV1FeatureMonitoringStatsAnomaly Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of historical SnapshotAnalysis or ImportFeaturesAnalysis stats requested by user, sorted by FeatureStatsAnomaly.start_time descending.
}
\concept{GoogleCloudAiplatformV1FeatureMonitoringStatsAnomaly functions}
