% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1FeatureNoiseSigma}
\alias{GoogleCloudAiplatformV1FeatureNoiseSigma}
\title{GoogleCloudAiplatformV1FeatureNoiseSigma Object}
\usage{
GoogleCloudAiplatformV1FeatureNoiseSigma(noiseSigma = NULL)
}
\arguments{
\item{noiseSigma}{Noise sigma per feature}
}
\value{
GoogleCloudAiplatformV1FeatureNoiseSigma object
}
\description{
GoogleCloudAiplatformV1FeatureNoiseSigma Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Noise sigma by features. Noise sigma represents the standard deviation of the gaussian kernel that will be used to add noise to interpolated inputs prior to computing gradients.
}
\concept{GoogleCloudAiplatformV1FeatureNoiseSigma functions}
