% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1FeatureNoiseSigmaNoiseSigmaForFeature}
\alias{GoogleCloudAiplatformV1FeatureNoiseSigmaNoiseSigmaForFeature}
\title{GoogleCloudAiplatformV1FeatureNoiseSigmaNoiseSigmaForFeature Object}
\usage{
GoogleCloudAiplatformV1FeatureNoiseSigmaNoiseSigmaForFeature(
  name = NULL,
  sigma = NULL
)
}
\arguments{
\item{name}{The name of the input feature for which noise sigma is provided}

\item{sigma}{This represents the standard deviation of the Gaussian kernel that will be used to add noise to the feature prior to computing gradients}
}
\value{
GoogleCloudAiplatformV1FeatureNoiseSigmaNoiseSigmaForFeature object
}
\description{
GoogleCloudAiplatformV1FeatureNoiseSigmaNoiseSigmaForFeature Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Noise sigma for a single feature.
}
\concept{GoogleCloudAiplatformV1FeatureNoiseSigmaNoiseSigmaForFeature functions}
