% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1FeatureSelector}
\alias{GoogleCloudAiplatformV1FeatureSelector}
\title{GoogleCloudAiplatformV1FeatureSelector Object}
\usage{
GoogleCloudAiplatformV1FeatureSelector(idMatcher = NULL)
}
\arguments{
\item{idMatcher}{Required}
}
\value{
GoogleCloudAiplatformV1FeatureSelector object
}
\description{
GoogleCloudAiplatformV1FeatureSelector Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Selector for Features of an EntityType.
}
\concept{GoogleCloudAiplatformV1FeatureSelector functions}
