% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1FeatureStatsAnomaly}
\alias{GoogleCloudAiplatformV1FeatureStatsAnomaly}
\title{GoogleCloudAiplatformV1FeatureStatsAnomaly Object}
\usage{
GoogleCloudAiplatformV1FeatureStatsAnomaly(
  anomalyUri = NULL,
  startTime = NULL,
  score = NULL,
  anomalyDetectionThreshold = NULL,
  distributionDeviation = NULL,
  endTime = NULL,
  statsUri = NULL
)
}
\arguments{
\item{anomalyUri}{Path of the anomaly file for current feature values in Cloud Storage bucket}

\item{startTime}{The start timestamp of window where stats were generated}

\item{score}{Feature importance score, only populated when cross-feature monitoring is enabled}

\item{anomalyDetectionThreshold}{This is the threshold used when detecting anomalies}

\item{distributionDeviation}{Deviation from the current stats to baseline stats}

\item{endTime}{The end timestamp of window where stats were generated}

\item{statsUri}{Path of the stats file for current feature values in Cloud Storage bucket}
}
\value{
GoogleCloudAiplatformV1FeatureStatsAnomaly object
}
\description{
GoogleCloudAiplatformV1FeatureStatsAnomaly Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Stats and Anomaly generated at specific timestamp for specific Feature. The start_time and end_time are used to define the time range of the dataset that current stats belongs to, e.g. prediction traffic is bucketed into prediction datasets by time window. If the Dataset is not defined by time window, start_time = end_time. Timestamp of the stats and anomalies always refers to end_time. Raw stats and anomalies are stored in stats_uri or anomaly_uri in the tensorflow defined protos. Field data_stats contains almost identical information with the raw stats in Vertex AI defined proto, for UI to display.
}
\concept{GoogleCloudAiplatformV1FeatureStatsAnomaly functions}
