% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1Featurestore.labels}
\alias{GoogleCloudAiplatformV1Featurestore.labels}
\title{GoogleCloudAiplatformV1Featurestore.labels Object}
\usage{
GoogleCloudAiplatformV1Featurestore.labels()
}
\value{
GoogleCloudAiplatformV1Featurestore.labels object
}
\description{
GoogleCloudAiplatformV1Featurestore.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Optional. The labels with user-defined metadata to organize your Featurestore. Label keys and values can be no longer than 64 characters (Unicode codepoints), can only contain lowercase letters, numeric characters, underscores and dashes. International characters are allowed. See https://goo.gl/xmQnxf for more information on and examples of labels. No more than 64 user labels can be associated with one Featurestore(System labels are excluded).' System reserved label keys are prefixed with 'aiplatform.googleapis.com/' and are immutable.
}
\seealso{
Other GoogleCloudAiplatformV1Featurestore functions: 
\code{\link{GoogleCloudAiplatformV1Featurestore}()},
\code{\link{projects.locations.featurestores.create}()},
\code{\link{projects.locations.featurestores.patch}()}
}
\concept{GoogleCloudAiplatformV1Featurestore functions}
