% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1Featurestore}
\alias{GoogleCloudAiplatformV1Featurestore}
\title{GoogleCloudAiplatformV1Featurestore Object}
\usage{
GoogleCloudAiplatformV1Featurestore(
  GoogleCloudAiplatformV1Featurestore.labels = NULL,
  labels = NULL,
  encryptionSpec = NULL,
  etag = NULL,
  onlineServingConfig = NULL,
  onlineStorageTtlDays = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1Featurestore.labels}{The \link{GoogleCloudAiplatformV1Featurestore.labels} object or list of objects}

\item{labels}{Optional}

\item{encryptionSpec}{Optional}

\item{etag}{Optional}

\item{onlineServingConfig}{Optional}

\item{onlineStorageTtlDays}{Optional}
}
\value{
GoogleCloudAiplatformV1Featurestore object
}
\description{
GoogleCloudAiplatformV1Featurestore Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Vertex AI Feature Store provides a centralized repository for organizing, storing, and serving ML features. The Featurestore is a top-level container for your features and their values.
}
\seealso{
Other GoogleCloudAiplatformV1Featurestore functions: 
\code{\link{GoogleCloudAiplatformV1Featurestore.labels}()},
\code{\link{projects.locations.featurestores.create}()},
\code{\link{projects.locations.featurestores.patch}()}
}
\concept{GoogleCloudAiplatformV1Featurestore functions}
