% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1FeaturestoreMonitoringConfig}
\alias{GoogleCloudAiplatformV1FeaturestoreMonitoringConfig}
\title{GoogleCloudAiplatformV1FeaturestoreMonitoringConfig Object}
\usage{
GoogleCloudAiplatformV1FeaturestoreMonitoringConfig(
  importFeaturesAnalysis = NULL,
  categoricalThresholdConfig = NULL,
  snapshotAnalysis = NULL,
  numericalThresholdConfig = NULL
)
}
\arguments{
\item{importFeaturesAnalysis}{The config for ImportFeatures Analysis Based Feature Monitoring}

\item{categoricalThresholdConfig}{Threshold for categorical features of anomaly detection}

\item{snapshotAnalysis}{The config for Snapshot Analysis Based Feature Monitoring}

\item{numericalThresholdConfig}{Threshold for numerical features of anomaly detection}
}
\value{
GoogleCloudAiplatformV1FeaturestoreMonitoringConfig object
}
\description{
GoogleCloudAiplatformV1FeaturestoreMonitoringConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration of how features in Featurestore are monitored.
}
\concept{GoogleCloudAiplatformV1FeaturestoreMonitoringConfig functions}
