% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1FeaturestoreMonitoringConfigImportFeaturesAnalysis}
\alias{GoogleCloudAiplatformV1FeaturestoreMonitoringConfigImportFeaturesAnalysis}
\title{GoogleCloudAiplatformV1FeaturestoreMonitoringConfigImportFeaturesAnalysis Object}
\usage{
GoogleCloudAiplatformV1FeaturestoreMonitoringConfigImportFeaturesAnalysis(
  anomalyDetectionBaseline = NULL,
  state = NULL
)
}
\arguments{
\item{anomalyDetectionBaseline}{The baseline used to do anomaly detection for the statistics generated by import features analysis}

\item{state}{Whether to enable / disable / inherite default hebavior for import features analysis}
}
\value{
GoogleCloudAiplatformV1FeaturestoreMonitoringConfigImportFeaturesAnalysis object
}
\description{
GoogleCloudAiplatformV1FeaturestoreMonitoringConfigImportFeaturesAnalysis Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration of the Featurestore's ImportFeature Analysis Based Monitoring. This type of analysis generates statistics for values of each Feature imported by every ImportFeatureValues operation.
}
\concept{GoogleCloudAiplatformV1FeaturestoreMonitoringConfigImportFeaturesAnalysis functions}
