% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1FeaturestoreMonitoringConfigSnapshotAnalysis}
\alias{GoogleCloudAiplatformV1FeaturestoreMonitoringConfigSnapshotAnalysis}
\title{GoogleCloudAiplatformV1FeaturestoreMonitoringConfigSnapshotAnalysis Object}
\usage{
GoogleCloudAiplatformV1FeaturestoreMonitoringConfigSnapshotAnalysis(
  disabled = NULL,
  monitoringIntervalDays = NULL,
  stalenessDays = NULL
)
}
\arguments{
\item{disabled}{The monitoring schedule for snapshot analysis}

\item{monitoringIntervalDays}{Configuration of the snapshot analysis based monitoring pipeline running interval}

\item{stalenessDays}{Customized export features time window for snapshot analysis}
}
\value{
GoogleCloudAiplatformV1FeaturestoreMonitoringConfigSnapshotAnalysis object
}
\description{
GoogleCloudAiplatformV1FeaturestoreMonitoringConfigSnapshotAnalysis Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration of the Featurestore's Snapshot Analysis Based Monitoring. This type of analysis generates statistics for each Feature based on a snapshot of the latest feature value of each entities every monitoring_interval.
}
\concept{GoogleCloudAiplatformV1FeaturestoreMonitoringConfigSnapshotAnalysis functions}
