% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1FeaturestoreMonitoringConfigThresholdConfig}
\alias{GoogleCloudAiplatformV1FeaturestoreMonitoringConfigThresholdConfig}
\title{GoogleCloudAiplatformV1FeaturestoreMonitoringConfigThresholdConfig Object}
\usage{
GoogleCloudAiplatformV1FeaturestoreMonitoringConfigThresholdConfig(
  value = NULL
)
}
\arguments{
\item{value}{Specify a threshold value that can trigger the alert}
}
\value{
GoogleCloudAiplatformV1FeaturestoreMonitoringConfigThresholdConfig object
}
\description{
GoogleCloudAiplatformV1FeaturestoreMonitoringConfigThresholdConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The config for Featurestore Monitoring threshold.
}
\concept{GoogleCloudAiplatformV1FeaturestoreMonitoringConfigThresholdConfig functions}
