% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1FeaturestoreOnlineServingConfig}
\alias{GoogleCloudAiplatformV1FeaturestoreOnlineServingConfig}
\title{GoogleCloudAiplatformV1FeaturestoreOnlineServingConfig Object}
\usage{
GoogleCloudAiplatformV1FeaturestoreOnlineServingConfig(
  scaling = NULL,
  fixedNodeCount = NULL
)
}
\arguments{
\item{scaling}{Online serving scaling configuration}

\item{fixedNodeCount}{The number of nodes for the online store}
}
\value{
GoogleCloudAiplatformV1FeaturestoreOnlineServingConfig object
}
\description{
GoogleCloudAiplatformV1FeaturestoreOnlineServingConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
OnlineServingConfig specifies the details for provisioning online serving resources.
}
\concept{GoogleCloudAiplatformV1FeaturestoreOnlineServingConfig functions}
