% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1FeaturestoreOnlineServingConfigScaling}
\alias{GoogleCloudAiplatformV1FeaturestoreOnlineServingConfigScaling}
\title{GoogleCloudAiplatformV1FeaturestoreOnlineServingConfigScaling Object}
\usage{
GoogleCloudAiplatformV1FeaturestoreOnlineServingConfigScaling(
  cpuUtilizationTarget = NULL,
  minNodeCount = NULL,
  maxNodeCount = NULL
)
}
\arguments{
\item{cpuUtilizationTarget}{Optional}

\item{minNodeCount}{Required}

\item{maxNodeCount}{The maximum number of nodes to scale up to}
}
\value{
GoogleCloudAiplatformV1FeaturestoreOnlineServingConfigScaling object
}
\description{
GoogleCloudAiplatformV1FeaturestoreOnlineServingConfigScaling Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Online serving scaling configuration. If min_node_count and max_node_count are set to the same value, the cluster will be configured with the fixed number of node (no auto-scaling).
}
\concept{GoogleCloudAiplatformV1FeaturestoreOnlineServingConfigScaling functions}
