% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1FeatureValue}
\alias{GoogleCloudAiplatformV1FeatureValue}
\title{GoogleCloudAiplatformV1FeatureValue Object}
\usage{
GoogleCloudAiplatformV1FeatureValue(
  bytesValue = NULL,
  boolValue = NULL,
  stringValue = NULL,
  boolArrayValue = NULL,
  int64Value = NULL,
  int64ArrayValue = NULL,
  doubleArrayValue = NULL,
  metadata = NULL,
  stringArrayValue = NULL,
  doubleValue = NULL
)
}
\arguments{
\item{bytesValue}{Bytes feature value}

\item{boolValue}{Bool type feature value}

\item{stringValue}{String feature value}

\item{boolArrayValue}{A list of bool type feature value}

\item{int64Value}{Int64 feature value}

\item{int64ArrayValue}{A list of int64 type feature value}

\item{doubleArrayValue}{A list of double type feature value}

\item{metadata}{Metadata of feature value}

\item{stringArrayValue}{A list of string type feature value}

\item{doubleValue}{Double type feature value}
}
\value{
GoogleCloudAiplatformV1FeatureValue object
}
\description{
GoogleCloudAiplatformV1FeatureValue Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Value for a feature.
}
\concept{GoogleCloudAiplatformV1FeatureValue functions}
