% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1FeatureValueDestination}
\alias{GoogleCloudAiplatformV1FeatureValueDestination}
\title{GoogleCloudAiplatformV1FeatureValueDestination Object}
\usage{
GoogleCloudAiplatformV1FeatureValueDestination(
  tfrecordDestination = NULL,
  csvDestination = NULL,
  bigqueryDestination = NULL
)
}
\arguments{
\item{tfrecordDestination}{Output in TFRecord format}

\item{csvDestination}{Output in CSV format}

\item{bigqueryDestination}{Output in BigQuery format}
}
\value{
GoogleCloudAiplatformV1FeatureValueDestination object
}
\description{
GoogleCloudAiplatformV1FeatureValueDestination Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A destination location for Feature values and format.
}
\concept{GoogleCloudAiplatformV1FeatureValueDestination functions}
