% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1FeatureValueList}
\alias{GoogleCloudAiplatformV1FeatureValueList}
\title{GoogleCloudAiplatformV1FeatureValueList Object}
\usage{
GoogleCloudAiplatformV1FeatureValueList(values = NULL)
}
\arguments{
\item{values}{A list of feature values}
}
\value{
GoogleCloudAiplatformV1FeatureValueList object
}
\description{
GoogleCloudAiplatformV1FeatureValueList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Container for list of values.
}
\concept{GoogleCloudAiplatformV1FeatureValueList functions}
