% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1FeatureValueMetadata}
\alias{GoogleCloudAiplatformV1FeatureValueMetadata}
\title{GoogleCloudAiplatformV1FeatureValueMetadata Object}
\usage{
GoogleCloudAiplatformV1FeatureValueMetadata(generateTime = NULL)
}
\arguments{
\item{generateTime}{Feature generation timestamp}
}
\value{
GoogleCloudAiplatformV1FeatureValueMetadata object
}
\description{
GoogleCloudAiplatformV1FeatureValueMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata of feature value.
}
\concept{GoogleCloudAiplatformV1FeatureValueMetadata functions}
