% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1FilterSplit}
\alias{GoogleCloudAiplatformV1FilterSplit}
\title{GoogleCloudAiplatformV1FilterSplit Object}
\usage{
GoogleCloudAiplatformV1FilterSplit(
  validationFilter = NULL,
  trainingFilter = NULL,
  testFilter = NULL
)
}
\arguments{
\item{validationFilter}{Required}

\item{trainingFilter}{Required}

\item{testFilter}{Required}
}
\value{
GoogleCloudAiplatformV1FilterSplit object
}
\description{
GoogleCloudAiplatformV1FilterSplit Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Assigns input data to training, validation, and test sets based on the given filters, data pieces not matched by any filter are ignored. Currently only supported for Datasets containing DataItems. If any of the filters in this message are to match nothing, then they can be set as '-' (the minus sign). Supported only for unstructured Datasets.
}
\concept{GoogleCloudAiplatformV1FilterSplit functions}
