% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1FindNeighborsRequest}
\alias{GoogleCloudAiplatformV1FindNeighborsRequest}
\title{GoogleCloudAiplatformV1FindNeighborsRequest Object}
\usage{
GoogleCloudAiplatformV1FindNeighborsRequest(
  returnFullDatapoint = NULL,
  deployedIndexId = NULL,
  queries = NULL
)
}
\arguments{
\item{returnFullDatapoint}{If set to true, the full datapoints (including all vector values and restricts) of the nearest neighbors are returned}

\item{deployedIndexId}{The ID of the DeployedIndex that will serve the request}

\item{queries}{The list of queries}
}
\value{
GoogleCloudAiplatformV1FindNeighborsRequest object
}
\description{
GoogleCloudAiplatformV1FindNeighborsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request message for MatchService.FindNeighbors.
}
\seealso{
Other GoogleCloudAiplatformV1FindNeighborsRequest functions: 
\code{\link{projects.locations.indexEndpoints.findNeighbors}()}
}
\concept{GoogleCloudAiplatformV1FindNeighborsRequest functions}
