% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1FindNeighborsRequestQuery}
\alias{GoogleCloudAiplatformV1FindNeighborsRequestQuery}
\title{GoogleCloudAiplatformV1FindNeighborsRequestQuery Object}
\usage{
GoogleCloudAiplatformV1FindNeighborsRequestQuery(
  approximateNeighborCount = NULL,
  neighborCount = NULL,
  datapoint = NULL,
  fractionLeafNodesToSearchOverride = NULL,
  perCrowdingAttributeNeighborCount = NULL
)
}
\arguments{
\item{approximateNeighborCount}{The number of neighbors to find via approximate search before exact reordering is performed}

\item{neighborCount}{The number of nearest neighbors to be retrieved from database for each query}

\item{datapoint}{Required}

\item{fractionLeafNodesToSearchOverride}{The fraction of the number of leaves to search, set at query time allows user to tune search performance}

\item{perCrowdingAttributeNeighborCount}{Crowding is a constraint on a neighbor list produced by nearest neighbor search requiring that no more than some value k' of the k neighbors returned have the same value of crowding_attribute}
}
\value{
GoogleCloudAiplatformV1FindNeighborsRequestQuery object
}
\description{
GoogleCloudAiplatformV1FindNeighborsRequestQuery Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A query to find a number of the nearest neighbors (most similar vectors) of a vector.
}
\concept{GoogleCloudAiplatformV1FindNeighborsRequestQuery functions}
