% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1FindNeighborsResponse}
\alias{GoogleCloudAiplatformV1FindNeighborsResponse}
\title{GoogleCloudAiplatformV1FindNeighborsResponse Object}
\usage{
GoogleCloudAiplatformV1FindNeighborsResponse(nearestNeighbors = NULL)
}
\arguments{
\item{nearestNeighbors}{The nearest neighbors of the query datapoints}
}
\value{
GoogleCloudAiplatformV1FindNeighborsResponse object
}
\description{
GoogleCloudAiplatformV1FindNeighborsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response message for MatchService.FindNeighbors.
}
\concept{GoogleCloudAiplatformV1FindNeighborsResponse functions}
