% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1FindNeighborsResponseNearestNeighbors}
\alias{GoogleCloudAiplatformV1FindNeighborsResponseNearestNeighbors}
\title{GoogleCloudAiplatformV1FindNeighborsResponseNearestNeighbors Object}
\usage{
GoogleCloudAiplatformV1FindNeighborsResponseNearestNeighbors(
  id = NULL,
  neighbors = NULL
)
}
\arguments{
\item{id}{The ID of the query datapoint}

\item{neighbors}{All its neighbors}
}
\value{
GoogleCloudAiplatformV1FindNeighborsResponseNearestNeighbors object
}
\description{
GoogleCloudAiplatformV1FindNeighborsResponseNearestNeighbors Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Nearest neighbors for one query.
}
\concept{GoogleCloudAiplatformV1FindNeighborsResponseNearestNeighbors functions}
