% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1FindNeighborsResponseNeighbor}
\alias{GoogleCloudAiplatformV1FindNeighborsResponseNeighbor}
\title{GoogleCloudAiplatformV1FindNeighborsResponseNeighbor Object}
\usage{
GoogleCloudAiplatformV1FindNeighborsResponseNeighbor(
  datapoint = NULL,
  distance = NULL
)
}
\arguments{
\item{datapoint}{The datapoint of the neighbor}

\item{distance}{The distance between the neighbor and the query vector}
}
\value{
GoogleCloudAiplatformV1FindNeighborsResponseNeighbor object
}
\description{
GoogleCloudAiplatformV1FindNeighborsResponseNeighbor Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A neighbor of the query vector.
}
\concept{GoogleCloudAiplatformV1FindNeighborsResponseNeighbor functions}
