% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1FractionSplit}
\alias{GoogleCloudAiplatformV1FractionSplit}
\title{GoogleCloudAiplatformV1FractionSplit Object}
\usage{
GoogleCloudAiplatformV1FractionSplit(
  validationFraction = NULL,
  testFraction = NULL,
  trainingFraction = NULL
)
}
\arguments{
\item{validationFraction}{The fraction of the input data that is to be used to validate the Model}

\item{testFraction}{The fraction of the input data that is to be used to evaluate the Model}

\item{trainingFraction}{The fraction of the input data that is to be used to train the Model}
}
\value{
GoogleCloudAiplatformV1FractionSplit object
}
\description{
GoogleCloudAiplatformV1FractionSplit Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Assigns the input data to training, validation, and test sets as per the given fractions. Any of \code{training_fraction}, \code{validation_fraction} and \code{test_fraction} may optionally be provided, they must sum to up to 1. If the provided ones sum to less than 1, the remainder is assigned to sets as decided by Vertex AI. If none of the fractions are set, by default roughly 80\% of data is used for training, 10\% for validation, and 10\% for test.
}
\concept{GoogleCloudAiplatformV1FractionSplit functions}
