% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1GcsDestination}
\alias{GoogleCloudAiplatformV1GcsDestination}
\title{GoogleCloudAiplatformV1GcsDestination Object}
\usage{
GoogleCloudAiplatformV1GcsDestination(outputUriPrefix = NULL)
}
\arguments{
\item{outputUriPrefix}{Required}
}
\value{
GoogleCloudAiplatformV1GcsDestination object
}
\description{
GoogleCloudAiplatformV1GcsDestination Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The Google Cloud Storage location where the output is to be written to.
}
\concept{GoogleCloudAiplatformV1GcsDestination functions}
