% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1GcsSource}
\alias{GoogleCloudAiplatformV1GcsSource}
\title{GoogleCloudAiplatformV1GcsSource Object}
\usage{
GoogleCloudAiplatformV1GcsSource(uris = NULL)
}
\arguments{
\item{uris}{Required}
}
\value{
GoogleCloudAiplatformV1GcsSource object
}
\description{
GoogleCloudAiplatformV1GcsSource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The Google Cloud Storage location for the input content.
}
\concept{GoogleCloudAiplatformV1GcsSource functions}
