% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1IdMatcher}
\alias{GoogleCloudAiplatformV1IdMatcher}
\title{GoogleCloudAiplatformV1IdMatcher Object}
\usage{
GoogleCloudAiplatformV1IdMatcher(ids = NULL)
}
\arguments{
\item{ids}{Required}
}
\value{
GoogleCloudAiplatformV1IdMatcher object
}
\description{
GoogleCloudAiplatformV1IdMatcher Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Matcher for Features of an EntityType by Feature ID.
}
\concept{GoogleCloudAiplatformV1IdMatcher functions}
