% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ImportDataConfig.annotationLabels}
\alias{GoogleCloudAiplatformV1ImportDataConfig.annotationLabels}
\title{GoogleCloudAiplatformV1ImportDataConfig.annotationLabels Object}
\usage{
GoogleCloudAiplatformV1ImportDataConfig.annotationLabels()
}
\value{
GoogleCloudAiplatformV1ImportDataConfig.annotationLabels object
}
\description{
GoogleCloudAiplatformV1ImportDataConfig.annotationLabels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Labels that will be applied to newly imported Annotations. If two Annotations are identical, one of them will be deduped. Two Annotations are considered identical if their payload, payload_schema_uri and all of their labels are the same. These labels will be overridden by Annotation labels specified inside index file referenced by import_schema_uri, e.g. jsonl file.
}
\seealso{
Other GoogleCloudAiplatformV1ImportDataConfig functions: 
\code{\link{GoogleCloudAiplatformV1ImportDataConfig.dataItemLabels}()},
\code{\link{GoogleCloudAiplatformV1ImportDataConfig}()}
}
\concept{GoogleCloudAiplatformV1ImportDataConfig functions}
