% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ImportDataConfig.dataItemLabels}
\alias{GoogleCloudAiplatformV1ImportDataConfig.dataItemLabels}
\title{GoogleCloudAiplatformV1ImportDataConfig.dataItemLabels Object}
\usage{
GoogleCloudAiplatformV1ImportDataConfig.dataItemLabels()
}
\value{
GoogleCloudAiplatformV1ImportDataConfig.dataItemLabels object
}
\description{
GoogleCloudAiplatformV1ImportDataConfig.dataItemLabels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Labels that will be applied to newly imported DataItems. If an identical DataItem as one being imported already exists in the Dataset, then these labels will be appended to these of the already existing one, and if labels with identical key is imported before, the old label value will be overwritten. If two DataItems are identical in the same import data operation, the labels will be combined and if key collision happens in this case, one of the values will be picked randomly. Two DataItems are considered identical if their content bytes are identical (e.g. image bytes or pdf bytes). These labels will be overridden by Annotation labels specified inside index file referenced by import_schema_uri, e.g. jsonl file.
}
\seealso{
Other GoogleCloudAiplatformV1ImportDataConfig functions: 
\code{\link{GoogleCloudAiplatformV1ImportDataConfig.annotationLabels}()},
\code{\link{GoogleCloudAiplatformV1ImportDataConfig}()}
}
\concept{GoogleCloudAiplatformV1ImportDataConfig functions}
