% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ImportDataConfig}
\alias{GoogleCloudAiplatformV1ImportDataConfig}
\title{GoogleCloudAiplatformV1ImportDataConfig Object}
\usage{
GoogleCloudAiplatformV1ImportDataConfig(
  GoogleCloudAiplatformV1ImportDataConfig.annotationLabels = NULL,
  GoogleCloudAiplatformV1ImportDataConfig.dataItemLabels = NULL,
  annotationLabels = NULL,
  importSchemaUri = NULL,
  dataItemLabels = NULL,
  gcsSource = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1ImportDataConfig.annotationLabels}{The \link{GoogleCloudAiplatformV1ImportDataConfig.annotationLabels} object or list of objects}

\item{GoogleCloudAiplatformV1ImportDataConfig.dataItemLabels}{The \link{GoogleCloudAiplatformV1ImportDataConfig.dataItemLabels} object or list of objects}

\item{annotationLabels}{Labels that will be applied to newly imported Annotations}

\item{importSchemaUri}{Required}

\item{dataItemLabels}{Labels that will be applied to newly imported DataItems}

\item{gcsSource}{The Google Cloud Storage location for the input content}
}
\value{
GoogleCloudAiplatformV1ImportDataConfig object
}
\description{
GoogleCloudAiplatformV1ImportDataConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describes the location from where we import data into a Dataset, together with the labels that will be applied to the DataItems and the Annotations.
}
\seealso{
Other GoogleCloudAiplatformV1ImportDataConfig functions: 
\code{\link{GoogleCloudAiplatformV1ImportDataConfig.annotationLabels}()},
\code{\link{GoogleCloudAiplatformV1ImportDataConfig.dataItemLabels}()}
}
\concept{GoogleCloudAiplatformV1ImportDataConfig functions}
