% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ImportDataOperationMetadata}
\alias{GoogleCloudAiplatformV1ImportDataOperationMetadata}
\title{GoogleCloudAiplatformV1ImportDataOperationMetadata Object}
\usage{
GoogleCloudAiplatformV1ImportDataOperationMetadata(genericMetadata = NULL)
}
\arguments{
\item{genericMetadata}{The common part of the operation metadata}
}
\value{
GoogleCloudAiplatformV1ImportDataOperationMetadata object
}
\description{
GoogleCloudAiplatformV1ImportDataOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Runtime operation information for DatasetService.ImportData.
}
\concept{GoogleCloudAiplatformV1ImportDataOperationMetadata functions}
