% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ImportDataRequest}
\alias{GoogleCloudAiplatformV1ImportDataRequest}
\title{GoogleCloudAiplatformV1ImportDataRequest Object}
\usage{
GoogleCloudAiplatformV1ImportDataRequest(importConfigs = NULL)
}
\arguments{
\item{importConfigs}{Required}
}
\value{
GoogleCloudAiplatformV1ImportDataRequest object
}
\description{
GoogleCloudAiplatformV1ImportDataRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for DatasetService.ImportData.
}
\seealso{
Other GoogleCloudAiplatformV1ImportDataRequest functions: 
\code{\link{projects.locations.datasets.import}()}
}
\concept{GoogleCloudAiplatformV1ImportDataRequest functions}
