% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ImportFeatureValuesOperationMetadata}
\alias{GoogleCloudAiplatformV1ImportFeatureValuesOperationMetadata}
\title{GoogleCloudAiplatformV1ImportFeatureValuesOperationMetadata Object}
\usage{
GoogleCloudAiplatformV1ImportFeatureValuesOperationMetadata(
  importedEntityCount = NULL,
  timestampOutsideRetentionRowsCount = NULL,
  genericMetadata = NULL,
  importedFeatureValueCount = NULL,
  sourceUris = NULL,
  invalidRowCount = NULL,
  blockingOperationIds = NULL
)
}
\arguments{
\item{importedEntityCount}{Number of entities that have been imported by the operation}

\item{timestampOutsideRetentionRowsCount}{The number rows that weren't ingested due to having timestamps outside the retention boundary}

\item{genericMetadata}{Operation metadata for Featurestore import Feature values}

\item{importedFeatureValueCount}{Number of Feature values that have been imported by the operation}

\item{sourceUris}{The source URI from where Feature values are imported}

\item{invalidRowCount}{The number of rows in input source that weren't imported due to either * Not having any featureValues}

\item{blockingOperationIds}{List of ImportFeatureValues operations running under a single EntityType that are blocking this operation}
}
\value{
GoogleCloudAiplatformV1ImportFeatureValuesOperationMetadata object
}
\description{
GoogleCloudAiplatformV1ImportFeatureValuesOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details of operations that perform import Feature values.
}
\concept{GoogleCloudAiplatformV1ImportFeatureValuesOperationMetadata functions}
