% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ImportFeatureValuesRequest}
\alias{GoogleCloudAiplatformV1ImportFeatureValuesRequest}
\title{GoogleCloudAiplatformV1ImportFeatureValuesRequest Object}
\usage{
GoogleCloudAiplatformV1ImportFeatureValuesRequest(
  entityIdField = NULL,
  disableIngestionAnalysis = NULL,
  bigquerySource = NULL,
  featureTime = NULL,
  disableOnlineServing = NULL,
  featureTimeField = NULL,
  avroSource = NULL,
  workerCount = NULL,
  featureSpecs = NULL,
  csvSource = NULL
)
}
\arguments{
\item{entityIdField}{Source column that holds entity IDs}

\item{disableIngestionAnalysis}{If true, API doesn't start ingestion analysis pipeline}

\item{bigquerySource}{No description}

\item{featureTime}{Single Feature timestamp for all entities being imported}

\item{disableOnlineServing}{If set, data will not be imported for online serving}

\item{featureTimeField}{Source column that holds the Feature timestamp for all Feature values in each entity}

\item{avroSource}{No description}

\item{workerCount}{Specifies the number of workers that are used to write data to the Featurestore}

\item{featureSpecs}{Required}

\item{csvSource}{No description}
}
\value{
GoogleCloudAiplatformV1ImportFeatureValuesRequest object
}
\description{
GoogleCloudAiplatformV1ImportFeatureValuesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for FeaturestoreService.ImportFeatureValues.
}
\seealso{
Other GoogleCloudAiplatformV1ImportFeatureValuesRequest functions: 
\code{\link{projects.locations.featurestores.entityTypes.importFeatureValues}()}
}
\concept{GoogleCloudAiplatformV1ImportFeatureValuesRequest functions}
