% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ImportFeatureValuesRequestFeatureSpec}
\alias{GoogleCloudAiplatformV1ImportFeatureValuesRequestFeatureSpec}
\title{GoogleCloudAiplatformV1ImportFeatureValuesRequestFeatureSpec Object}
\usage{
GoogleCloudAiplatformV1ImportFeatureValuesRequestFeatureSpec(
  sourceField = NULL,
  id = NULL
)
}
\arguments{
\item{sourceField}{Source column to get the Feature values from}

\item{id}{Required}
}
\value{
GoogleCloudAiplatformV1ImportFeatureValuesRequestFeatureSpec object
}
\description{
GoogleCloudAiplatformV1ImportFeatureValuesRequestFeatureSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Defines the Feature value(s) to import.
}
\concept{GoogleCloudAiplatformV1ImportFeatureValuesRequestFeatureSpec functions}
