% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ImportFeatureValuesResponse}
\alias{GoogleCloudAiplatformV1ImportFeatureValuesResponse}
\title{GoogleCloudAiplatformV1ImportFeatureValuesResponse Object}
\usage{
GoogleCloudAiplatformV1ImportFeatureValuesResponse(
  invalidRowCount = NULL,
  importedEntityCount = NULL,
  importedFeatureValueCount = NULL,
  timestampOutsideRetentionRowsCount = NULL
)
}
\arguments{
\item{invalidRowCount}{The number of rows in input source that weren't imported due to either * Not having any featureValues}

\item{importedEntityCount}{Number of entities that have been imported by the operation}

\item{importedFeatureValueCount}{Number of Feature values that have been imported by the operation}

\item{timestampOutsideRetentionRowsCount}{The number rows that weren't ingested due to having feature timestamps outside the retention boundary}
}
\value{
GoogleCloudAiplatformV1ImportFeatureValuesResponse object
}
\description{
GoogleCloudAiplatformV1ImportFeatureValuesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for FeaturestoreService.ImportFeatureValues.
}
\concept{GoogleCloudAiplatformV1ImportFeatureValuesResponse functions}
