% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ImportModelEvaluationRequest}
\alias{GoogleCloudAiplatformV1ImportModelEvaluationRequest}
\title{GoogleCloudAiplatformV1ImportModelEvaluationRequest Object}
\usage{
GoogleCloudAiplatformV1ImportModelEvaluationRequest(modelEvaluation = NULL)
}
\arguments{
\item{modelEvaluation}{Required}
}
\value{
GoogleCloudAiplatformV1ImportModelEvaluationRequest object
}
\description{
GoogleCloudAiplatformV1ImportModelEvaluationRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for ModelService.ImportModelEvaluation
}
\seealso{
Other GoogleCloudAiplatformV1ImportModelEvaluationRequest functions: 
\code{\link{projects.locations.models.evaluations.import}()}
}
\concept{GoogleCloudAiplatformV1ImportModelEvaluationRequest functions}
