% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1Index}
\alias{GoogleCloudAiplatformV1Index}
\title{GoogleCloudAiplatformV1Index Object}
\usage{
GoogleCloudAiplatformV1Index(
  GoogleCloudAiplatformV1Index.labels = NULL,
  indexUpdateMethod = NULL,
  etag = NULL,
  description = NULL,
  labels = NULL,
  metadata = NULL,
  metadataSchemaUri = NULL,
  displayName = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1Index.labels}{The \link{GoogleCloudAiplatformV1Index.labels} object or list of objects}

\item{indexUpdateMethod}{Immutable}

\item{etag}{Used to perform consistent read-modify-write updates}

\item{description}{The description of the Index}

\item{labels}{The labels with user-defined metadata to organize your Indexes}

\item{metadata}{An additional information about the Index; the schema of the metadata can be found in metadata_schema}

\item{metadataSchemaUri}{Immutable}

\item{displayName}{Required}
}
\value{
GoogleCloudAiplatformV1Index object
}
\description{
GoogleCloudAiplatformV1Index Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A representation of a collection of database items organized in a way that allows for approximate nearest neighbor (a.k.a ANN) algorithms search.
}
\seealso{
Other GoogleCloudAiplatformV1Index functions: 
\code{\link{GoogleCloudAiplatformV1Index.labels}()},
\code{\link{projects.locations.indexes.create}()},
\code{\link{projects.locations.indexes.patch}()}
}
\concept{GoogleCloudAiplatformV1Index functions}
