% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1IndexDatapoint}
\alias{GoogleCloudAiplatformV1IndexDatapoint}
\title{GoogleCloudAiplatformV1IndexDatapoint Object}
\usage{
GoogleCloudAiplatformV1IndexDatapoint(
  featureVector = NULL,
  restricts = NULL,
  crowdingTag = NULL,
  datapointId = NULL
)
}
\arguments{
\item{featureVector}{Required}

\item{restricts}{Optional}

\item{crowdingTag}{Optional}

\item{datapointId}{Required}
}
\value{
GoogleCloudAiplatformV1IndexDatapoint object
}
\description{
GoogleCloudAiplatformV1IndexDatapoint Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A datapoint of Index.
}
\concept{GoogleCloudAiplatformV1IndexDatapoint functions}
