% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1IndexDatapointRestriction}
\alias{GoogleCloudAiplatformV1IndexDatapointRestriction}
\title{GoogleCloudAiplatformV1IndexDatapointRestriction Object}
\usage{
GoogleCloudAiplatformV1IndexDatapointRestriction(
  allowList = NULL,
  namespace = NULL,
  denyList = NULL
)
}
\arguments{
\item{allowList}{The attributes to allow in this namespace}

\item{namespace}{The namespace of this restriction}

\item{denyList}{The attributes to deny in this namespace}
}
\value{
GoogleCloudAiplatformV1IndexDatapointRestriction object
}
\description{
GoogleCloudAiplatformV1IndexDatapointRestriction Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Restriction of a datapoint which describe its attributes(tokens) from each of several attribute categories(namespaces).
}
\concept{GoogleCloudAiplatformV1IndexDatapointRestriction functions}
