% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1IndexEndpoint}
\alias{GoogleCloudAiplatformV1IndexEndpoint}
\title{GoogleCloudAiplatformV1IndexEndpoint Object}
\usage{
GoogleCloudAiplatformV1IndexEndpoint(
  GoogleCloudAiplatformV1IndexEndpoint.labels = NULL,
  displayName = NULL,
  description = NULL,
  labels = NULL,
  network = NULL,
  privateServiceConnectConfig = NULL,
  publicEndpointEnabled = NULL,
  enablePrivateServiceConnect = NULL,
  etag = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1IndexEndpoint.labels}{The \link{GoogleCloudAiplatformV1IndexEndpoint.labels} object or list of objects}

\item{displayName}{Required}

\item{description}{The description of the IndexEndpoint}

\item{labels}{The labels with user-defined metadata to organize your IndexEndpoints}

\item{network}{Optional}

\item{privateServiceConnectConfig}{Optional}

\item{publicEndpointEnabled}{Optional}

\item{enablePrivateServiceConnect}{Optional}

\item{etag}{Used to perform consistent read-modify-write updates}
}
\value{
GoogleCloudAiplatformV1IndexEndpoint object
}
\description{
GoogleCloudAiplatformV1IndexEndpoint Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Indexes are deployed into it. An IndexEndpoint can have multiple DeployedIndexes.
}
\seealso{
Other GoogleCloudAiplatformV1IndexEndpoint functions: 
\code{\link{GoogleCloudAiplatformV1IndexEndpoint.labels}()},
\code{\link{projects.locations.indexEndpoints.create}()},
\code{\link{projects.locations.indexEndpoints.patch}()}
}
\concept{GoogleCloudAiplatformV1IndexEndpoint functions}
