% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1IndexPrivateEndpoints}
\alias{GoogleCloudAiplatformV1IndexPrivateEndpoints}
\title{GoogleCloudAiplatformV1IndexPrivateEndpoints Object}
\usage{
GoogleCloudAiplatformV1IndexPrivateEndpoints()
}
\value{
GoogleCloudAiplatformV1IndexPrivateEndpoints object
}
\description{
GoogleCloudAiplatformV1IndexPrivateEndpoints Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
IndexPrivateEndpoints proto is used to provide paths for users to send requests via private endpoints (e.g. private service access, private service connect). To send request via private service access, use match_grpc_address. To send request via private service connect, use service_attachment.
}
\concept{GoogleCloudAiplatformV1IndexPrivateEndpoints functions}
