% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1InputDataConfig}
\alias{GoogleCloudAiplatformV1InputDataConfig}
\title{GoogleCloudAiplatformV1InputDataConfig Object}
\usage{
GoogleCloudAiplatformV1InputDataConfig(
  savedQueryId = NULL,
  predefinedSplit = NULL,
  filterSplit = NULL,
  annotationsFilter = NULL,
  bigqueryDestination = NULL,
  datasetId = NULL,
  gcsDestination = NULL,
  persistMlUseAssignment = NULL,
  annotationSchemaUri = NULL,
  fractionSplit = NULL,
  stratifiedSplit = NULL,
  timestampSplit = NULL
)
}
\arguments{
\item{savedQueryId}{Only applicable to Datasets that have SavedQueries}

\item{predefinedSplit}{Supported only for tabular Datasets}

\item{filterSplit}{Split based on the provided filters for each set}

\item{annotationsFilter}{Applicable only to Datasets that have DataItems and Annotations}

\item{bigqueryDestination}{Only applicable to custom training with tabular Dataset with BigQuery source}

\item{datasetId}{Required}

\item{gcsDestination}{The Cloud Storage location where the training data is to be written to}

\item{persistMlUseAssignment}{Whether to persist the ML use assignment to data item system labels}

\item{annotationSchemaUri}{Applicable only to custom training with Datasets that have DataItems and Annotations}

\item{fractionSplit}{Split based on fractions defining the size of each set}

\item{stratifiedSplit}{Supported only for tabular Datasets}

\item{timestampSplit}{Supported only for tabular Datasets}
}
\value{
GoogleCloudAiplatformV1InputDataConfig object
}
\description{
GoogleCloudAiplatformV1InputDataConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Specifies Vertex AI owned input data to be used for training, and possibly evaluating, the Model.
}
\concept{GoogleCloudAiplatformV1InputDataConfig functions}
