% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1Int64Array}
\alias{GoogleCloudAiplatformV1Int64Array}
\title{GoogleCloudAiplatformV1Int64Array Object}
\usage{
GoogleCloudAiplatformV1Int64Array(values = NULL)
}
\arguments{
\item{values}{A list of int64 values}
}
\value{
GoogleCloudAiplatformV1Int64Array object
}
\description{
GoogleCloudAiplatformV1Int64Array Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of int64 values.
}
\concept{GoogleCloudAiplatformV1Int64Array functions}
