% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1IntegratedGradientsAttribution}
\alias{GoogleCloudAiplatformV1IntegratedGradientsAttribution}
\title{GoogleCloudAiplatformV1IntegratedGradientsAttribution Object}
\usage{
GoogleCloudAiplatformV1IntegratedGradientsAttribution(
  blurBaselineConfig = NULL,
  stepCount = NULL,
  smoothGradConfig = NULL
)
}
\arguments{
\item{blurBaselineConfig}{Config for IG with blur baseline}

\item{stepCount}{Required}

\item{smoothGradConfig}{Config for SmoothGrad approximation of gradients}
}
\value{
GoogleCloudAiplatformV1IntegratedGradientsAttribution object
}
\description{
GoogleCloudAiplatformV1IntegratedGradientsAttribution Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An attribution method that computes the Aumann-Shapley value taking advantage of the model's fully differentiable structure. Refer to this paper for more details: https://arxiv.org/abs/1703.01365
}
\concept{GoogleCloudAiplatformV1IntegratedGradientsAttribution functions}
