% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1LargeModelReference}
\alias{GoogleCloudAiplatformV1LargeModelReference}
\title{GoogleCloudAiplatformV1LargeModelReference Object}
\usage{
GoogleCloudAiplatformV1LargeModelReference(name = NULL)
}
\arguments{
\item{name}{Required}
}
\value{
GoogleCloudAiplatformV1LargeModelReference object
}
\description{
GoogleCloudAiplatformV1LargeModelReference Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains information about the Large Model.
}
\concept{GoogleCloudAiplatformV1LargeModelReference functions}
