% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1LineageSubgraph}
\alias{GoogleCloudAiplatformV1LineageSubgraph}
\title{GoogleCloudAiplatformV1LineageSubgraph Object}
\usage{
GoogleCloudAiplatformV1LineageSubgraph(
  events = NULL,
  executions = NULL,
  artifacts = NULL
)
}
\arguments{
\item{events}{The Event edges between Artifacts and Executions in the subgraph}

\item{executions}{The Execution nodes in the subgraph}

\item{artifacts}{The Artifact nodes in the subgraph}
}
\value{
GoogleCloudAiplatformV1LineageSubgraph object
}
\description{
GoogleCloudAiplatformV1LineageSubgraph Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A subgraph of the overall lineage graph. Event edges connect Artifact and Execution nodes.
}
\concept{GoogleCloudAiplatformV1LineageSubgraph functions}
