% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ListAnnotationsResponse}
\alias{GoogleCloudAiplatformV1ListAnnotationsResponse}
\title{GoogleCloudAiplatformV1ListAnnotationsResponse Object}
\usage{
GoogleCloudAiplatformV1ListAnnotationsResponse(
  annotations = NULL,
  nextPageToken = NULL
)
}
\arguments{
\item{annotations}{A list of Annotations that matches the specified filter in the request}

\item{nextPageToken}{The standard List next-page token}
}
\value{
GoogleCloudAiplatformV1ListAnnotationsResponse object
}
\description{
GoogleCloudAiplatformV1ListAnnotationsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for DatasetService.ListAnnotations.
}
\concept{GoogleCloudAiplatformV1ListAnnotationsResponse functions}
