% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ListArtifactsResponse}
\alias{GoogleCloudAiplatformV1ListArtifactsResponse}
\title{GoogleCloudAiplatformV1ListArtifactsResponse Object}
\usage{
GoogleCloudAiplatformV1ListArtifactsResponse(
  artifacts = NULL,
  nextPageToken = NULL
)
}
\arguments{
\item{artifacts}{The Artifacts retrieved from the MetadataStore}

\item{nextPageToken}{A token, which can be sent as ListArtifactsRequest}
}
\value{
GoogleCloudAiplatformV1ListArtifactsResponse object
}
\description{
GoogleCloudAiplatformV1ListArtifactsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for MetadataService.ListArtifacts.
}
\concept{GoogleCloudAiplatformV1ListArtifactsResponse functions}
