% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ListBatchPredictionJobsResponse}
\alias{GoogleCloudAiplatformV1ListBatchPredictionJobsResponse}
\title{GoogleCloudAiplatformV1ListBatchPredictionJobsResponse Object}
\usage{
GoogleCloudAiplatformV1ListBatchPredictionJobsResponse(
  batchPredictionJobs = NULL,
  nextPageToken = NULL
)
}
\arguments{
\item{batchPredictionJobs}{List of BatchPredictionJobs in the requested page}

\item{nextPageToken}{A token to retrieve the next page of results}
}
\value{
GoogleCloudAiplatformV1ListBatchPredictionJobsResponse object
}
\description{
GoogleCloudAiplatformV1ListBatchPredictionJobsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for JobService.ListBatchPredictionJobs
}
\concept{GoogleCloudAiplatformV1ListBatchPredictionJobsResponse functions}
