% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ListContextsResponse}
\alias{GoogleCloudAiplatformV1ListContextsResponse}
\title{GoogleCloudAiplatformV1ListContextsResponse Object}
\usage{
GoogleCloudAiplatformV1ListContextsResponse(
  nextPageToken = NULL,
  contexts = NULL
)
}
\arguments{
\item{nextPageToken}{A token, which can be sent as ListContextsRequest}

\item{contexts}{The Contexts retrieved from the MetadataStore}
}
\value{
GoogleCloudAiplatformV1ListContextsResponse object
}
\description{
GoogleCloudAiplatformV1ListContextsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for MetadataService.ListContexts.
}
\concept{GoogleCloudAiplatformV1ListContextsResponse functions}
