% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ListCustomJobsResponse}
\alias{GoogleCloudAiplatformV1ListCustomJobsResponse}
\title{GoogleCloudAiplatformV1ListCustomJobsResponse Object}
\usage{
GoogleCloudAiplatformV1ListCustomJobsResponse(
  nextPageToken = NULL,
  customJobs = NULL
)
}
\arguments{
\item{nextPageToken}{A token to retrieve the next page of results}

\item{customJobs}{List of CustomJobs in the requested page}
}
\value{
GoogleCloudAiplatformV1ListCustomJobsResponse object
}
\description{
GoogleCloudAiplatformV1ListCustomJobsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for JobService.ListCustomJobs
}
\concept{GoogleCloudAiplatformV1ListCustomJobsResponse functions}
