% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ListDataItemsResponse}
\alias{GoogleCloudAiplatformV1ListDataItemsResponse}
\title{GoogleCloudAiplatformV1ListDataItemsResponse Object}
\usage{
GoogleCloudAiplatformV1ListDataItemsResponse(
  dataItems = NULL,
  nextPageToken = NULL
)
}
\arguments{
\item{dataItems}{A list of DataItems that matches the specified filter in the request}

\item{nextPageToken}{The standard List next-page token}
}
\value{
GoogleCloudAiplatformV1ListDataItemsResponse object
}
\description{
GoogleCloudAiplatformV1ListDataItemsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for DatasetService.ListDataItems.
}
\concept{GoogleCloudAiplatformV1ListDataItemsResponse functions}
