% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ListDataLabelingJobsResponse}
\alias{GoogleCloudAiplatformV1ListDataLabelingJobsResponse}
\title{GoogleCloudAiplatformV1ListDataLabelingJobsResponse Object}
\usage{
GoogleCloudAiplatformV1ListDataLabelingJobsResponse(
  nextPageToken = NULL,
  dataLabelingJobs = NULL
)
}
\arguments{
\item{nextPageToken}{The standard List next-page token}

\item{dataLabelingJobs}{A list of DataLabelingJobs that matches the specified filter in the request}
}
\value{
GoogleCloudAiplatformV1ListDataLabelingJobsResponse object
}
\description{
GoogleCloudAiplatformV1ListDataLabelingJobsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for JobService.ListDataLabelingJobs.
}
\concept{GoogleCloudAiplatformV1ListDataLabelingJobsResponse functions}
