% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ListDatasetsResponse}
\alias{GoogleCloudAiplatformV1ListDatasetsResponse}
\title{GoogleCloudAiplatformV1ListDatasetsResponse Object}
\usage{
GoogleCloudAiplatformV1ListDatasetsResponse(
  nextPageToken = NULL,
  datasets = NULL
)
}
\arguments{
\item{nextPageToken}{The standard List next-page token}

\item{datasets}{A list of Datasets that matches the specified filter in the request}
}
\value{
GoogleCloudAiplatformV1ListDatasetsResponse object
}
\description{
GoogleCloudAiplatformV1ListDatasetsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for DatasetService.ListDatasets.
}
\concept{GoogleCloudAiplatformV1ListDatasetsResponse functions}
