% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ListEndpointsResponse}
\alias{GoogleCloudAiplatformV1ListEndpointsResponse}
\title{GoogleCloudAiplatformV1ListEndpointsResponse Object}
\usage{
GoogleCloudAiplatformV1ListEndpointsResponse(
  endpoints = NULL,
  nextPageToken = NULL
)
}
\arguments{
\item{endpoints}{List of Endpoints in the requested page}

\item{nextPageToken}{A token to retrieve the next page of results}
}
\value{
GoogleCloudAiplatformV1ListEndpointsResponse object
}
\description{
GoogleCloudAiplatformV1ListEndpointsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for EndpointService.ListEndpoints.
}
\concept{GoogleCloudAiplatformV1ListEndpointsResponse functions}
