% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ListEntityTypesResponse}
\alias{GoogleCloudAiplatformV1ListEntityTypesResponse}
\title{GoogleCloudAiplatformV1ListEntityTypesResponse Object}
\usage{
GoogleCloudAiplatformV1ListEntityTypesResponse(
  nextPageToken = NULL,
  entityTypes = NULL
)
}
\arguments{
\item{nextPageToken}{A token, which can be sent as ListEntityTypesRequest}

\item{entityTypes}{The EntityTypes matching the request}
}
\value{
GoogleCloudAiplatformV1ListEntityTypesResponse object
}
\description{
GoogleCloudAiplatformV1ListEntityTypesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for FeaturestoreService.ListEntityTypes.
}
\concept{GoogleCloudAiplatformV1ListEntityTypesResponse functions}
