% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ListExecutionsResponse}
\alias{GoogleCloudAiplatformV1ListExecutionsResponse}
\title{GoogleCloudAiplatformV1ListExecutionsResponse Object}
\usage{
GoogleCloudAiplatformV1ListExecutionsResponse(
  nextPageToken = NULL,
  executions = NULL
)
}
\arguments{
\item{nextPageToken}{A token, which can be sent as ListExecutionsRequest}

\item{executions}{The Executions retrieved from the MetadataStore}
}
\value{
GoogleCloudAiplatformV1ListExecutionsResponse object
}
\description{
GoogleCloudAiplatformV1ListExecutionsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for MetadataService.ListExecutions.
}
\concept{GoogleCloudAiplatformV1ListExecutionsResponse functions}
