% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ListFeaturesResponse}
\alias{GoogleCloudAiplatformV1ListFeaturesResponse}
\title{GoogleCloudAiplatformV1ListFeaturesResponse Object}
\usage{
GoogleCloudAiplatformV1ListFeaturesResponse(
  features = NULL,
  nextPageToken = NULL
)
}
\arguments{
\item{features}{The Features matching the request}

\item{nextPageToken}{A token, which can be sent as ListFeaturesRequest}
}
\value{
GoogleCloudAiplatformV1ListFeaturesResponse object
}
\description{
GoogleCloudAiplatformV1ListFeaturesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for FeaturestoreService.ListFeatures. Response message for FeatureRegistryService.ListFeatures.
}
\concept{GoogleCloudAiplatformV1ListFeaturesResponse functions}
