% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ListFeaturestoresResponse}
\alias{GoogleCloudAiplatformV1ListFeaturestoresResponse}
\title{GoogleCloudAiplatformV1ListFeaturestoresResponse Object}
\usage{
GoogleCloudAiplatformV1ListFeaturestoresResponse(
  nextPageToken = NULL,
  featurestores = NULL
)
}
\arguments{
\item{nextPageToken}{A token, which can be sent as ListFeaturestoresRequest}

\item{featurestores}{The Featurestores matching the request}
}
\value{
GoogleCloudAiplatformV1ListFeaturestoresResponse object
}
\description{
GoogleCloudAiplatformV1ListFeaturestoresResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for FeaturestoreService.ListFeaturestores.
}
\concept{GoogleCloudAiplatformV1ListFeaturestoresResponse functions}
