% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ListHyperparameterTuningJobsResponse}
\alias{GoogleCloudAiplatformV1ListHyperparameterTuningJobsResponse}
\title{GoogleCloudAiplatformV1ListHyperparameterTuningJobsResponse Object}
\usage{
GoogleCloudAiplatformV1ListHyperparameterTuningJobsResponse(
  nextPageToken = NULL,
  hyperparameterTuningJobs = NULL
)
}
\arguments{
\item{nextPageToken}{A token to retrieve the next page of results}

\item{hyperparameterTuningJobs}{List of HyperparameterTuningJobs in the requested page}
}
\value{
GoogleCloudAiplatformV1ListHyperparameterTuningJobsResponse object
}
\description{
GoogleCloudAiplatformV1ListHyperparameterTuningJobsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for JobService.ListHyperparameterTuningJobs
}
\concept{GoogleCloudAiplatformV1ListHyperparameterTuningJobsResponse functions}
