% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ListIndexEndpointsResponse}
\alias{GoogleCloudAiplatformV1ListIndexEndpointsResponse}
\title{GoogleCloudAiplatformV1ListIndexEndpointsResponse Object}
\usage{
GoogleCloudAiplatformV1ListIndexEndpointsResponse(
  nextPageToken = NULL,
  indexEndpoints = NULL
)
}
\arguments{
\item{nextPageToken}{A token to retrieve next page of results}

\item{indexEndpoints}{List of IndexEndpoints in the requested page}
}
\value{
GoogleCloudAiplatformV1ListIndexEndpointsResponse object
}
\description{
GoogleCloudAiplatformV1ListIndexEndpointsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for IndexEndpointService.ListIndexEndpoints.
}
\concept{GoogleCloudAiplatformV1ListIndexEndpointsResponse functions}
