% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ListIndexesResponse}
\alias{GoogleCloudAiplatformV1ListIndexesResponse}
\title{GoogleCloudAiplatformV1ListIndexesResponse Object}
\usage{
GoogleCloudAiplatformV1ListIndexesResponse(
  indexes = NULL,
  nextPageToken = NULL
)
}
\arguments{
\item{indexes}{List of indexes in the requested page}

\item{nextPageToken}{A token to retrieve next page of results}
}
\value{
GoogleCloudAiplatformV1ListIndexesResponse object
}
\description{
GoogleCloudAiplatformV1ListIndexesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for IndexService.ListIndexes.
}
\concept{GoogleCloudAiplatformV1ListIndexesResponse functions}
