% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ListMetadataSchemasResponse}
\alias{GoogleCloudAiplatformV1ListMetadataSchemasResponse}
\title{GoogleCloudAiplatformV1ListMetadataSchemasResponse Object}
\usage{
GoogleCloudAiplatformV1ListMetadataSchemasResponse(
  metadataSchemas = NULL,
  nextPageToken = NULL
)
}
\arguments{
\item{metadataSchemas}{The MetadataSchemas found for the MetadataStore}

\item{nextPageToken}{A token, which can be sent as ListMetadataSchemasRequest}
}
\value{
GoogleCloudAiplatformV1ListMetadataSchemasResponse object
}
\description{
GoogleCloudAiplatformV1ListMetadataSchemasResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for MetadataService.ListMetadataSchemas.
}
\concept{GoogleCloudAiplatformV1ListMetadataSchemasResponse functions}
