% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ListMetadataStoresResponse}
\alias{GoogleCloudAiplatformV1ListMetadataStoresResponse}
\title{GoogleCloudAiplatformV1ListMetadataStoresResponse Object}
\usage{
GoogleCloudAiplatformV1ListMetadataStoresResponse(
  metadataStores = NULL,
  nextPageToken = NULL
)
}
\arguments{
\item{metadataStores}{The MetadataStores found for the Location}

\item{nextPageToken}{A token, which can be sent as ListMetadataStoresRequest}
}
\value{
GoogleCloudAiplatformV1ListMetadataStoresResponse object
}
\description{
GoogleCloudAiplatformV1ListMetadataStoresResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for MetadataService.ListMetadataStores.
}
\concept{GoogleCloudAiplatformV1ListMetadataStoresResponse functions}
