% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ListModelDeploymentMonitoringJobsResponse}
\alias{GoogleCloudAiplatformV1ListModelDeploymentMonitoringJobsResponse}
\title{GoogleCloudAiplatformV1ListModelDeploymentMonitoringJobsResponse Object}
\usage{
GoogleCloudAiplatformV1ListModelDeploymentMonitoringJobsResponse(
  modelDeploymentMonitoringJobs = NULL,
  nextPageToken = NULL
)
}
\arguments{
\item{modelDeploymentMonitoringJobs}{A list of ModelDeploymentMonitoringJobs that matches the specified filter in the request}

\item{nextPageToken}{The standard List next-page token}
}
\value{
GoogleCloudAiplatformV1ListModelDeploymentMonitoringJobsResponse object
}
\description{
GoogleCloudAiplatformV1ListModelDeploymentMonitoringJobsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for JobService.ListModelDeploymentMonitoringJobs.
}
\concept{GoogleCloudAiplatformV1ListModelDeploymentMonitoringJobsResponse functions}
