% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ListModelEvaluationSlicesResponse}
\alias{GoogleCloudAiplatformV1ListModelEvaluationSlicesResponse}
\title{GoogleCloudAiplatformV1ListModelEvaluationSlicesResponse Object}
\usage{
GoogleCloudAiplatformV1ListModelEvaluationSlicesResponse(
  modelEvaluationSlices = NULL,
  nextPageToken = NULL
)
}
\arguments{
\item{modelEvaluationSlices}{List of ModelEvaluations in the requested page}

\item{nextPageToken}{A token to retrieve next page of results}
}
\value{
GoogleCloudAiplatformV1ListModelEvaluationSlicesResponse object
}
\description{
GoogleCloudAiplatformV1ListModelEvaluationSlicesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for ModelService.ListModelEvaluationSlices.
}
\concept{GoogleCloudAiplatformV1ListModelEvaluationSlicesResponse functions}
