% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ListModelEvaluationsResponse}
\alias{GoogleCloudAiplatformV1ListModelEvaluationsResponse}
\title{GoogleCloudAiplatformV1ListModelEvaluationsResponse Object}
\usage{
GoogleCloudAiplatformV1ListModelEvaluationsResponse(
  modelEvaluations = NULL,
  nextPageToken = NULL
)
}
\arguments{
\item{modelEvaluations}{List of ModelEvaluations in the requested page}

\item{nextPageToken}{A token to retrieve next page of results}
}
\value{
GoogleCloudAiplatformV1ListModelEvaluationsResponse object
}
\description{
GoogleCloudAiplatformV1ListModelEvaluationsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for ModelService.ListModelEvaluations.
}
\concept{GoogleCloudAiplatformV1ListModelEvaluationsResponse functions}
