% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ListModelsResponse}
\alias{GoogleCloudAiplatformV1ListModelsResponse}
\title{GoogleCloudAiplatformV1ListModelsResponse Object}
\usage{
GoogleCloudAiplatformV1ListModelsResponse(models = NULL, nextPageToken = NULL)
}
\arguments{
\item{models}{List of Models in the requested page}

\item{nextPageToken}{A token to retrieve next page of results}
}
\value{
GoogleCloudAiplatformV1ListModelsResponse object
}
\description{
GoogleCloudAiplatformV1ListModelsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for ModelService.ListModels
}
\concept{GoogleCloudAiplatformV1ListModelsResponse functions}
