% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ListModelVersionsResponse}
\alias{GoogleCloudAiplatformV1ListModelVersionsResponse}
\title{GoogleCloudAiplatformV1ListModelVersionsResponse Object}
\usage{
GoogleCloudAiplatformV1ListModelVersionsResponse(
  models = NULL,
  nextPageToken = NULL
)
}
\arguments{
\item{models}{List of Model versions in the requested page}

\item{nextPageToken}{A token to retrieve the next page of results}
}
\value{
GoogleCloudAiplatformV1ListModelVersionsResponse object
}
\description{
GoogleCloudAiplatformV1ListModelVersionsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for ModelService.ListModelVersions
}
\concept{GoogleCloudAiplatformV1ListModelVersionsResponse functions}
