% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ListNasJobsResponse}
\alias{GoogleCloudAiplatformV1ListNasJobsResponse}
\title{GoogleCloudAiplatformV1ListNasJobsResponse Object}
\usage{
GoogleCloudAiplatformV1ListNasJobsResponse(
  nextPageToken = NULL,
  nasJobs = NULL
)
}
\arguments{
\item{nextPageToken}{A token to retrieve the next page of results}

\item{nasJobs}{List of NasJobs in the requested page}
}
\value{
GoogleCloudAiplatformV1ListNasJobsResponse object
}
\description{
GoogleCloudAiplatformV1ListNasJobsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for JobService.ListNasJobs
}
\concept{GoogleCloudAiplatformV1ListNasJobsResponse functions}
