% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ListNasTrialDetailsResponse}
\alias{GoogleCloudAiplatformV1ListNasTrialDetailsResponse}
\title{GoogleCloudAiplatformV1ListNasTrialDetailsResponse Object}
\usage{
GoogleCloudAiplatformV1ListNasTrialDetailsResponse(
  nasTrialDetails = NULL,
  nextPageToken = NULL
)
}
\arguments{
\item{nasTrialDetails}{List of top NasTrials in the requested page}

\item{nextPageToken}{A token to retrieve the next page of results}
}
\value{
GoogleCloudAiplatformV1ListNasTrialDetailsResponse object
}
\description{
GoogleCloudAiplatformV1ListNasTrialDetailsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for JobService.ListNasTrialDetails
}
\concept{GoogleCloudAiplatformV1ListNasTrialDetailsResponse functions}
