% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ListOptimalTrialsResponse}
\alias{GoogleCloudAiplatformV1ListOptimalTrialsResponse}
\title{GoogleCloudAiplatformV1ListOptimalTrialsResponse Object}
\usage{
GoogleCloudAiplatformV1ListOptimalTrialsResponse(optimalTrials = NULL)
}
\arguments{
\item{optimalTrials}{The pareto-optimal Trials for multiple objective Study or the optimal trial for single objective Study}
}
\value{
GoogleCloudAiplatformV1ListOptimalTrialsResponse object
}
\description{
GoogleCloudAiplatformV1ListOptimalTrialsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for VizierService.ListOptimalTrials.
}
\concept{GoogleCloudAiplatformV1ListOptimalTrialsResponse functions}
