% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ListPipelineJobsResponse}
\alias{GoogleCloudAiplatformV1ListPipelineJobsResponse}
\title{GoogleCloudAiplatformV1ListPipelineJobsResponse Object}
\usage{
GoogleCloudAiplatformV1ListPipelineJobsResponse(
  pipelineJobs = NULL,
  nextPageToken = NULL
)
}
\arguments{
\item{pipelineJobs}{List of PipelineJobs in the requested page}

\item{nextPageToken}{A token to retrieve the next page of results}
}
\value{
GoogleCloudAiplatformV1ListPipelineJobsResponse object
}
\description{
GoogleCloudAiplatformV1ListPipelineJobsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for PipelineService.ListPipelineJobs
}
\concept{GoogleCloudAiplatformV1ListPipelineJobsResponse functions}
