% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ListSavedQueriesResponse}
\alias{GoogleCloudAiplatformV1ListSavedQueriesResponse}
\title{GoogleCloudAiplatformV1ListSavedQueriesResponse Object}
\usage{
GoogleCloudAiplatformV1ListSavedQueriesResponse(
  nextPageToken = NULL,
  savedQueries = NULL
)
}
\arguments{
\item{nextPageToken}{The standard List next-page token}

\item{savedQueries}{A list of SavedQueries that match the specified filter in the request}
}
\value{
GoogleCloudAiplatformV1ListSavedQueriesResponse object
}
\description{
GoogleCloudAiplatformV1ListSavedQueriesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for DatasetService.ListSavedQueries.
}
\concept{GoogleCloudAiplatformV1ListSavedQueriesResponse functions}
