% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ListSchedulesResponse}
\alias{GoogleCloudAiplatformV1ListSchedulesResponse}
\title{GoogleCloudAiplatformV1ListSchedulesResponse Object}
\usage{
GoogleCloudAiplatformV1ListSchedulesResponse(
  schedules = NULL,
  nextPageToken = NULL
)
}
\arguments{
\item{schedules}{List of Schedules in the requested page}

\item{nextPageToken}{A token to retrieve the next page of results}
}
\value{
GoogleCloudAiplatformV1ListSchedulesResponse object
}
\description{
GoogleCloudAiplatformV1ListSchedulesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for ScheduleService.ListSchedules
}
\concept{GoogleCloudAiplatformV1ListSchedulesResponse functions}
