% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ListSpecialistPoolsResponse}
\alias{GoogleCloudAiplatformV1ListSpecialistPoolsResponse}
\title{GoogleCloudAiplatformV1ListSpecialistPoolsResponse Object}
\usage{
GoogleCloudAiplatformV1ListSpecialistPoolsResponse(
  specialistPools = NULL,
  nextPageToken = NULL
)
}
\arguments{
\item{specialistPools}{A list of SpecialistPools that matches the specified filter in the request}

\item{nextPageToken}{The standard List next-page token}
}
\value{
GoogleCloudAiplatformV1ListSpecialistPoolsResponse object
}
\description{
GoogleCloudAiplatformV1ListSpecialistPoolsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for SpecialistPoolService.ListSpecialistPools.
}
\concept{GoogleCloudAiplatformV1ListSpecialistPoolsResponse functions}
