% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ListStudiesResponse}
\alias{GoogleCloudAiplatformV1ListStudiesResponse}
\title{GoogleCloudAiplatformV1ListStudiesResponse Object}
\usage{
GoogleCloudAiplatformV1ListStudiesResponse(
  nextPageToken = NULL,
  studies = NULL
)
}
\arguments{
\item{nextPageToken}{Passes this token as the \code{page_token} field of the request for a subsequent call}

\item{studies}{The studies associated with the project}
}
\value{
GoogleCloudAiplatformV1ListStudiesResponse object
}
\description{
GoogleCloudAiplatformV1ListStudiesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for VizierService.ListStudies.
}
\concept{GoogleCloudAiplatformV1ListStudiesResponse functions}
