% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ListTensorboardExperimentsResponse}
\alias{GoogleCloudAiplatformV1ListTensorboardExperimentsResponse}
\title{GoogleCloudAiplatformV1ListTensorboardExperimentsResponse Object}
\usage{
GoogleCloudAiplatformV1ListTensorboardExperimentsResponse(
  nextPageToken = NULL,
  tensorboardExperiments = NULL
)
}
\arguments{
\item{nextPageToken}{A token, which can be sent as ListTensorboardExperimentsRequest}

\item{tensorboardExperiments}{The TensorboardExperiments mathching the request}
}
\value{
GoogleCloudAiplatformV1ListTensorboardExperimentsResponse object
}
\description{
GoogleCloudAiplatformV1ListTensorboardExperimentsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for TensorboardService.ListTensorboardExperiments.
}
\concept{GoogleCloudAiplatformV1ListTensorboardExperimentsResponse functions}
