% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ListTensorboardRunsResponse}
\alias{GoogleCloudAiplatformV1ListTensorboardRunsResponse}
\title{GoogleCloudAiplatformV1ListTensorboardRunsResponse Object}
\usage{
GoogleCloudAiplatformV1ListTensorboardRunsResponse(
  nextPageToken = NULL,
  tensorboardRuns = NULL
)
}
\arguments{
\item{nextPageToken}{A token, which can be sent as ListTensorboardRunsRequest}

\item{tensorboardRuns}{The TensorboardRuns mathching the request}
}
\value{
GoogleCloudAiplatformV1ListTensorboardRunsResponse object
}
\description{
GoogleCloudAiplatformV1ListTensorboardRunsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for TensorboardService.ListTensorboardRuns.
}
\concept{GoogleCloudAiplatformV1ListTensorboardRunsResponse functions}
