% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ListTensorboardsResponse}
\alias{GoogleCloudAiplatformV1ListTensorboardsResponse}
\title{GoogleCloudAiplatformV1ListTensorboardsResponse Object}
\usage{
GoogleCloudAiplatformV1ListTensorboardsResponse(
  tensorboards = NULL,
  nextPageToken = NULL
)
}
\arguments{
\item{tensorboards}{The Tensorboards mathching the request}

\item{nextPageToken}{A token, which can be sent as ListTensorboardsRequest}
}
\value{
GoogleCloudAiplatformV1ListTensorboardsResponse object
}
\description{
GoogleCloudAiplatformV1ListTensorboardsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for TensorboardService.ListTensorboards.
}
\concept{GoogleCloudAiplatformV1ListTensorboardsResponse functions}
