% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ListTensorboardTimeSeriesResponse}
\alias{GoogleCloudAiplatformV1ListTensorboardTimeSeriesResponse}
\title{GoogleCloudAiplatformV1ListTensorboardTimeSeriesResponse Object}
\usage{
GoogleCloudAiplatformV1ListTensorboardTimeSeriesResponse(
  nextPageToken = NULL,
  tensorboardTimeSeries = NULL
)
}
\arguments{
\item{nextPageToken}{A token, which can be sent as ListTensorboardTimeSeriesRequest}

\item{tensorboardTimeSeries}{The TensorboardTimeSeries mathching the request}
}
\value{
GoogleCloudAiplatformV1ListTensorboardTimeSeriesResponse object
}
\description{
GoogleCloudAiplatformV1ListTensorboardTimeSeriesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for TensorboardService.ListTensorboardTimeSeries.
}
\concept{GoogleCloudAiplatformV1ListTensorboardTimeSeriesResponse functions}
