% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ListTrainingPipelinesResponse}
\alias{GoogleCloudAiplatformV1ListTrainingPipelinesResponse}
\title{GoogleCloudAiplatformV1ListTrainingPipelinesResponse Object}
\usage{
GoogleCloudAiplatformV1ListTrainingPipelinesResponse(
  nextPageToken = NULL,
  trainingPipelines = NULL
)
}
\arguments{
\item{nextPageToken}{A token to retrieve the next page of results}

\item{trainingPipelines}{List of TrainingPipelines in the requested page}
}
\value{
GoogleCloudAiplatformV1ListTrainingPipelinesResponse object
}
\description{
GoogleCloudAiplatformV1ListTrainingPipelinesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for PipelineService.ListTrainingPipelines
}
\concept{GoogleCloudAiplatformV1ListTrainingPipelinesResponse functions}
