% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ListTrialsResponse}
\alias{GoogleCloudAiplatformV1ListTrialsResponse}
\title{GoogleCloudAiplatformV1ListTrialsResponse Object}
\usage{
GoogleCloudAiplatformV1ListTrialsResponse(trials = NULL, nextPageToken = NULL)
}
\arguments{
\item{trials}{The Trials associated with the Study}

\item{nextPageToken}{Pass this token as the \code{page_token} field of the request for a subsequent call}
}
\value{
GoogleCloudAiplatformV1ListTrialsResponse object
}
\description{
GoogleCloudAiplatformV1ListTrialsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for VizierService.ListTrials.
}
\concept{GoogleCloudAiplatformV1ListTrialsResponse functions}
