% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1LookupStudyRequest}
\alias{GoogleCloudAiplatformV1LookupStudyRequest}
\title{GoogleCloudAiplatformV1LookupStudyRequest Object}
\usage{
GoogleCloudAiplatformV1LookupStudyRequest(displayName = NULL)
}
\arguments{
\item{displayName}{Required}
}
\value{
GoogleCloudAiplatformV1LookupStudyRequest object
}
\description{
GoogleCloudAiplatformV1LookupStudyRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for VizierService.LookupStudy.
}
\seealso{
Other GoogleCloudAiplatformV1LookupStudyRequest functions: 
\code{\link{projects.locations.studies.lookup}()}
}
\concept{GoogleCloudAiplatformV1LookupStudyRequest functions}
