% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1MachineSpec}
\alias{GoogleCloudAiplatformV1MachineSpec}
\title{GoogleCloudAiplatformV1MachineSpec Object}
\usage{
GoogleCloudAiplatformV1MachineSpec(
  machineType = NULL,
  acceleratorType = NULL,
  acceleratorCount = NULL
)
}
\arguments{
\item{machineType}{Immutable}

\item{acceleratorType}{Immutable}

\item{acceleratorCount}{The number of accelerators to attach to the machine}
}
\value{
GoogleCloudAiplatformV1MachineSpec object
}
\description{
GoogleCloudAiplatformV1MachineSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Specification of a single machine.
}
\concept{GoogleCloudAiplatformV1MachineSpec functions}
