% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ManualBatchTuningParameters}
\alias{GoogleCloudAiplatformV1ManualBatchTuningParameters}
\title{GoogleCloudAiplatformV1ManualBatchTuningParameters Object}
\usage{
GoogleCloudAiplatformV1ManualBatchTuningParameters(batchSize = NULL)
}
\arguments{
\item{batchSize}{Immutable}
}
\value{
GoogleCloudAiplatformV1ManualBatchTuningParameters object
}
\description{
GoogleCloudAiplatformV1ManualBatchTuningParameters Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Manual batch tuning parameters.
}
\concept{GoogleCloudAiplatformV1ManualBatchTuningParameters functions}
