% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1MergeVersionAliasesRequest}
\alias{GoogleCloudAiplatformV1MergeVersionAliasesRequest}
\title{GoogleCloudAiplatformV1MergeVersionAliasesRequest Object}
\usage{
GoogleCloudAiplatformV1MergeVersionAliasesRequest(versionAliases = NULL)
}
\arguments{
\item{versionAliases}{Required}
}
\value{
GoogleCloudAiplatformV1MergeVersionAliasesRequest object
}
\description{
GoogleCloudAiplatformV1MergeVersionAliasesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for ModelService.MergeVersionAliases.
}
\seealso{
Other GoogleCloudAiplatformV1MergeVersionAliasesRequest functions: 
\code{\link{projects.locations.models.mergeVersionAliases}()}
}
\concept{GoogleCloudAiplatformV1MergeVersionAliasesRequest functions}
