% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1MetadataSchema}
\alias{GoogleCloudAiplatformV1MetadataSchema}
\title{GoogleCloudAiplatformV1MetadataSchema Object}
\usage{
GoogleCloudAiplatformV1MetadataSchema(
  description = NULL,
  schemaVersion = NULL,
  schema = NULL,
  schemaType = NULL
)
}
\arguments{
\item{description}{Description of the Metadata Schema}

\item{schemaVersion}{The version of the MetadataSchema}

\item{schema}{Required}

\item{schemaType}{The type of the MetadataSchema}
}
\value{
GoogleCloudAiplatformV1MetadataSchema object
}
\description{
GoogleCloudAiplatformV1MetadataSchema Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Instance of a general MetadataSchema.
}
\seealso{
Other GoogleCloudAiplatformV1MetadataSchema functions: 
\code{\link{projects.locations.metadataStores.metadataSchemas.create}()}
}
\concept{GoogleCloudAiplatformV1MetadataSchema functions}
