% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1MetadataStore}
\alias{GoogleCloudAiplatformV1MetadataStore}
\title{GoogleCloudAiplatformV1MetadataStore Object}
\usage{
GoogleCloudAiplatformV1MetadataStore(encryptionSpec = NULL, description = NULL)
}
\arguments{
\item{encryptionSpec}{Customer-managed encryption key spec for a Metadata Store}

\item{description}{Description of the MetadataStore}
}
\value{
GoogleCloudAiplatformV1MetadataStore object
}
\description{
GoogleCloudAiplatformV1MetadataStore Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Instance of a metadata store. Contains a set of metadata that can be queried.
}
\seealso{
Other GoogleCloudAiplatformV1MetadataStore functions: 
\code{\link{projects.locations.metadataStores.create}()}
}
\concept{GoogleCloudAiplatformV1MetadataStore functions}
