% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1MetadataStoreMetadataStoreState}
\alias{GoogleCloudAiplatformV1MetadataStoreMetadataStoreState}
\title{GoogleCloudAiplatformV1MetadataStoreMetadataStoreState Object}
\usage{
GoogleCloudAiplatformV1MetadataStoreMetadataStoreState(
  diskUtilizationBytes = NULL
)
}
\arguments{
\item{diskUtilizationBytes}{The disk utilization of the MetadataStore in bytes}
}
\value{
GoogleCloudAiplatformV1MetadataStoreMetadataStoreState object
}
\description{
GoogleCloudAiplatformV1MetadataStoreMetadataStoreState Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents state information for a MetadataStore.
}
\concept{GoogleCloudAiplatformV1MetadataStoreMetadataStoreState functions}
