% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1MigratableResource}
\alias{GoogleCloudAiplatformV1MigratableResource}
\title{GoogleCloudAiplatformV1MigratableResource Object}
\usage{
GoogleCloudAiplatformV1MigratableResource()
}
\value{
GoogleCloudAiplatformV1MigratableResource object
}
\description{
GoogleCloudAiplatformV1MigratableResource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents one resource that exists in automl.googleapis.com, datalabeling.googleapis.com or ml.googleapis.com.
}
\concept{GoogleCloudAiplatformV1MigratableResource functions}
