% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1MigratableResourceAutomlDataset}
\alias{GoogleCloudAiplatformV1MigratableResourceAutomlDataset}
\title{GoogleCloudAiplatformV1MigratableResourceAutomlDataset Object}
\usage{
GoogleCloudAiplatformV1MigratableResourceAutomlDataset(
  datasetDisplayName = NULL,
  dataset = NULL
)
}
\arguments{
\item{datasetDisplayName}{The Dataset's display name in automl}

\item{dataset}{Full resource name of automl Dataset}
}
\value{
GoogleCloudAiplatformV1MigratableResourceAutomlDataset object
}
\description{
GoogleCloudAiplatformV1MigratableResourceAutomlDataset Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents one Dataset in automl.googleapis.com.
}
\concept{GoogleCloudAiplatformV1MigratableResourceAutomlDataset functions}
