% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1MigratableResourceAutomlModel}
\alias{GoogleCloudAiplatformV1MigratableResourceAutomlModel}
\title{GoogleCloudAiplatformV1MigratableResourceAutomlModel Object}
\usage{
GoogleCloudAiplatformV1MigratableResourceAutomlModel(
  model = NULL,
  modelDisplayName = NULL
)
}
\arguments{
\item{model}{Full resource name of automl Model}

\item{modelDisplayName}{The Model's display name in automl}
}
\value{
GoogleCloudAiplatformV1MigratableResourceAutomlModel object
}
\description{
GoogleCloudAiplatformV1MigratableResourceAutomlModel Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents one Model in automl.googleapis.com.
}
\concept{GoogleCloudAiplatformV1MigratableResourceAutomlModel functions}
