% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1MigratableResourceDataLabelingDataset}
\alias{GoogleCloudAiplatformV1MigratableResourceDataLabelingDataset}
\title{GoogleCloudAiplatformV1MigratableResourceDataLabelingDataset Object}
\usage{
GoogleCloudAiplatformV1MigratableResourceDataLabelingDataset(
  datasetDisplayName = NULL,
  dataset = NULL,
  dataLabelingAnnotatedDatasets = NULL
)
}
\arguments{
\item{datasetDisplayName}{The Dataset's display name in datalabeling}

\item{dataset}{Full resource name of data labeling Dataset}

\item{dataLabelingAnnotatedDatasets}{The migratable AnnotatedDataset in datalabeling}
}
\value{
GoogleCloudAiplatformV1MigratableResourceDataLabelingDataset object
}
\description{
GoogleCloudAiplatformV1MigratableResourceDataLabelingDataset Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents one Dataset in datalabeling.googleapis.com.
}
\concept{GoogleCloudAiplatformV1MigratableResourceDataLabelingDataset functions}
