% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1MigratableResourceDataLabelingDatasetDataLabelingAnnotatedDataset}
\alias{GoogleCloudAiplatformV1MigratableResourceDataLabelingDatasetDataLabelingAnnotatedDataset}
\title{GoogleCloudAiplatformV1MigratableResourceDataLabelingDatasetDataLabelingAnnotatedDataset Object}
\usage{
GoogleCloudAiplatformV1MigratableResourceDataLabelingDatasetDataLabelingAnnotatedDataset(
  annotatedDatasetDisplayName = NULL,
  annotatedDataset = NULL
)
}
\arguments{
\item{annotatedDatasetDisplayName}{The AnnotatedDataset's display name in datalabeling}

\item{annotatedDataset}{Full resource name of data labeling AnnotatedDataset}
}
\value{
GoogleCloudAiplatformV1MigratableResourceDataLabelingDatasetDataLabelingAnnotatedDataset object
}
\description{
GoogleCloudAiplatformV1MigratableResourceDataLabelingDatasetDataLabelingAnnotatedDataset Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents one AnnotatedDataset in datalabeling.googleapis.com.
}
\concept{GoogleCloudAiplatformV1MigratableResourceDataLabelingDatasetDataLabelingAnnotatedDataset functions}
