% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1MigratableResourceMlEngineModelVersion}
\alias{GoogleCloudAiplatformV1MigratableResourceMlEngineModelVersion}
\title{GoogleCloudAiplatformV1MigratableResourceMlEngineModelVersion Object}
\usage{
GoogleCloudAiplatformV1MigratableResourceMlEngineModelVersion(
  version = NULL,
  endpoint = NULL
)
}
\arguments{
\item{version}{Full resource name of ml engine model Version}

\item{endpoint}{The ml}
}
\value{
GoogleCloudAiplatformV1MigratableResourceMlEngineModelVersion object
}
\description{
GoogleCloudAiplatformV1MigratableResourceMlEngineModelVersion Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents one model Version in ml.googleapis.com.
}
\concept{GoogleCloudAiplatformV1MigratableResourceMlEngineModelVersion functions}
