% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1MigrateResourceRequest}
\alias{GoogleCloudAiplatformV1MigrateResourceRequest}
\title{GoogleCloudAiplatformV1MigrateResourceRequest Object}
\usage{
GoogleCloudAiplatformV1MigrateResourceRequest(
  migrateMlEngineModelVersionConfig = NULL,
  migrateAutomlDatasetConfig = NULL,
  migrateDataLabelingDatasetConfig = NULL,
  migrateAutomlModelConfig = NULL
)
}
\arguments{
\item{migrateMlEngineModelVersionConfig}{Config for migrating Version in ml}

\item{migrateAutomlDatasetConfig}{Config for migrating Dataset in automl}

\item{migrateDataLabelingDatasetConfig}{Config for migrating Dataset in datalabeling}

\item{migrateAutomlModelConfig}{Config for migrating Model in automl}
}
\value{
GoogleCloudAiplatformV1MigrateResourceRequest object
}
\description{
GoogleCloudAiplatformV1MigrateResourceRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Config of migrating one resource from automl.googleapis.com, datalabeling.googleapis.com and ml.googleapis.com to Vertex AI.
}
\concept{GoogleCloudAiplatformV1MigrateResourceRequest functions}
