% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1MigrateResourceRequestMigrateAutomlDatasetConfig}
\alias{GoogleCloudAiplatformV1MigrateResourceRequestMigrateAutomlDatasetConfig}
\title{GoogleCloudAiplatformV1MigrateResourceRequestMigrateAutomlDatasetConfig Object}
\usage{
GoogleCloudAiplatformV1MigrateResourceRequestMigrateAutomlDatasetConfig(
  dataset = NULL,
  datasetDisplayName = NULL
)
}
\arguments{
\item{dataset}{Required}

\item{datasetDisplayName}{Required}
}
\value{
GoogleCloudAiplatformV1MigrateResourceRequestMigrateAutomlDatasetConfig object
}
\description{
GoogleCloudAiplatformV1MigrateResourceRequestMigrateAutomlDatasetConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Config for migrating Dataset in automl.googleapis.com to Vertex AI's Dataset.
}
\concept{GoogleCloudAiplatformV1MigrateResourceRequestMigrateAutomlDatasetConfig functions}
