% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1MigrateResourceRequestMigrateAutomlModelConfig}
\alias{GoogleCloudAiplatformV1MigrateResourceRequestMigrateAutomlModelConfig}
\title{GoogleCloudAiplatformV1MigrateResourceRequestMigrateAutomlModelConfig Object}
\usage{
GoogleCloudAiplatformV1MigrateResourceRequestMigrateAutomlModelConfig(
  modelDisplayName = NULL,
  model = NULL
)
}
\arguments{
\item{modelDisplayName}{Optional}

\item{model}{Required}
}
\value{
GoogleCloudAiplatformV1MigrateResourceRequestMigrateAutomlModelConfig object
}
\description{
GoogleCloudAiplatformV1MigrateResourceRequestMigrateAutomlModelConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Config for migrating Model in automl.googleapis.com to Vertex AI's Model.
}
\concept{GoogleCloudAiplatformV1MigrateResourceRequestMigrateAutomlModelConfig functions}
