% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1MigrateResourceRequestMigrateDataLabelingDatasetConfig}
\alias{GoogleCloudAiplatformV1MigrateResourceRequestMigrateDataLabelingDatasetConfig}
\title{GoogleCloudAiplatformV1MigrateResourceRequestMigrateDataLabelingDatasetConfig Object}
\usage{
GoogleCloudAiplatformV1MigrateResourceRequestMigrateDataLabelingDatasetConfig(
  dataset = NULL,
  migrateDataLabelingAnnotatedDatasetConfigs = NULL,
  datasetDisplayName = NULL
)
}
\arguments{
\item{dataset}{Required}

\item{migrateDataLabelingAnnotatedDatasetConfigs}{Optional}

\item{datasetDisplayName}{Optional}
}
\value{
GoogleCloudAiplatformV1MigrateResourceRequestMigrateDataLabelingDatasetConfig object
}
\description{
GoogleCloudAiplatformV1MigrateResourceRequestMigrateDataLabelingDatasetConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Config for migrating Dataset in datalabeling.googleapis.com to Vertex AI's Dataset.
}
\concept{GoogleCloudAiplatformV1MigrateResourceRequestMigrateDataLabelingDatasetConfig functions}
