% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1MigrateResourceRequestMigrateMlEngineModelVersionConfig}
\alias{GoogleCloudAiplatformV1MigrateResourceRequestMigrateMlEngineModelVersionConfig}
\title{GoogleCloudAiplatformV1MigrateResourceRequestMigrateMlEngineModelVersionConfig Object}
\usage{
GoogleCloudAiplatformV1MigrateResourceRequestMigrateMlEngineModelVersionConfig(
  modelDisplayName = NULL,
  endpoint = NULL,
  modelVersion = NULL
)
}
\arguments{
\item{modelDisplayName}{Required}

\item{endpoint}{Required}

\item{modelVersion}{Required}
}
\value{
GoogleCloudAiplatformV1MigrateResourceRequestMigrateMlEngineModelVersionConfig object
}
\description{
GoogleCloudAiplatformV1MigrateResourceRequestMigrateMlEngineModelVersionConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Config for migrating version in ml.googleapis.com to Vertex AI's Model.
}
\concept{GoogleCloudAiplatformV1MigrateResourceRequestMigrateMlEngineModelVersionConfig functions}
