% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1MigrateResourceResponse}
\alias{GoogleCloudAiplatformV1MigrateResourceResponse}
\title{GoogleCloudAiplatformV1MigrateResourceResponse Object}
\usage{
GoogleCloudAiplatformV1MigrateResourceResponse(
  model = NULL,
  dataset = NULL,
  migratableResource = NULL
)
}
\arguments{
\item{model}{Migrated Model's resource name}

\item{dataset}{Migrated Dataset's resource name}

\item{migratableResource}{Before migration, the identifier in ml}
}
\value{
GoogleCloudAiplatformV1MigrateResourceResponse object
}
\description{
GoogleCloudAiplatformV1MigrateResourceResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describes a successfully migrated resource.
}
\concept{GoogleCloudAiplatformV1MigrateResourceResponse functions}
