% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1Model}
\alias{GoogleCloudAiplatformV1Model}
\title{GoogleCloudAiplatformV1Model Object}
\usage{
GoogleCloudAiplatformV1Model(
  GoogleCloudAiplatformV1Model.labels = NULL,
  encryptionSpec = NULL,
  explanationSpec = NULL,
  displayName = NULL,
  metadata = NULL,
  containerSpec = NULL,
  metadataSchemaUri = NULL,
  name = NULL,
  artifactUri = NULL,
  pipelineJob = NULL,
  etag = NULL,
  predictSchemata = NULL,
  labels = NULL,
  versionDescription = NULL,
  description = NULL,
  versionAliases = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1Model.labels}{The \link{GoogleCloudAiplatformV1Model.labels} object or list of objects}

\item{encryptionSpec}{Customer-managed encryption key spec for a Model}

\item{explanationSpec}{The default explanation specification for this Model}

\item{displayName}{Required}

\item{metadata}{Immutable}

\item{containerSpec}{Input only}

\item{metadataSchemaUri}{Immutable}

\item{name}{The resource name of the Model}

\item{artifactUri}{Immutable}

\item{pipelineJob}{Optional}

\item{etag}{Used to perform consistent read-modify-write updates}

\item{predictSchemata}{The schemata that describe formats of the Model's predictions and explanations as given and returned via PredictionService}

\item{labels}{The labels with user-defined metadata to organize your Models}

\item{versionDescription}{The description of this version}

\item{description}{The description of the Model}

\item{versionAliases}{User provided version aliases so that a model version can be referenced via alias (i}
}
\value{
GoogleCloudAiplatformV1Model object
}
\description{
GoogleCloudAiplatformV1Model Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A trained machine learning Model.
}
\seealso{
Other GoogleCloudAiplatformV1Model functions: 
\code{\link{GoogleCloudAiplatformV1Model.labels}()},
\code{\link{projects.locations.models.patch}()}
}
\concept{GoogleCloudAiplatformV1Model functions}
