% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ModelContainerSpec}
\alias{GoogleCloudAiplatformV1ModelContainerSpec}
\title{GoogleCloudAiplatformV1ModelContainerSpec Object}
\usage{
GoogleCloudAiplatformV1ModelContainerSpec(
  healthRoute = NULL,
  command = NULL,
  env = NULL,
  predictRoute = NULL,
  imageUri = NULL,
  ports = NULL,
  args = NULL
)
}
\arguments{
\item{healthRoute}{Immutable}

\item{command}{Immutable}

\item{env}{Immutable}

\item{predictRoute}{Immutable}

\item{imageUri}{Required}

\item{ports}{Immutable}

\item{args}{Immutable}
}
\value{
GoogleCloudAiplatformV1ModelContainerSpec object
}
\description{
GoogleCloudAiplatformV1ModelContainerSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Specification of a container for serving predictions. Some fields in this message correspond to fields in the \href{https://kubernetes.io/docs/reference/generated/kubernetes-api/v1.23/#container-v1-core}{Kubernetes Container v1 core specification}.
}
\concept{GoogleCloudAiplatformV1ModelContainerSpec functions}
