% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ModelDeploymentMonitoringBigQueryTable}
\alias{GoogleCloudAiplatformV1ModelDeploymentMonitoringBigQueryTable}
\title{GoogleCloudAiplatformV1ModelDeploymentMonitoringBigQueryTable Object}
\usage{
GoogleCloudAiplatformV1ModelDeploymentMonitoringBigQueryTable(
  bigqueryTablePath = NULL,
  logType = NULL,
  logSource = NULL
)
}
\arguments{
\item{bigqueryTablePath}{The created BigQuery table to store logs}

\item{logType}{The type of log}

\item{logSource}{The source of log}
}
\value{
GoogleCloudAiplatformV1ModelDeploymentMonitoringBigQueryTable object
}
\description{
GoogleCloudAiplatformV1ModelDeploymentMonitoringBigQueryTable Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
ModelDeploymentMonitoringBigQueryTable specifies the BigQuery table name as well as some information of the logs stored in this table.
}
\concept{GoogleCloudAiplatformV1ModelDeploymentMonitoringBigQueryTable functions}
