% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ModelDeploymentMonitoringObjectiveConfig}
\alias{GoogleCloudAiplatformV1ModelDeploymentMonitoringObjectiveConfig}
\title{GoogleCloudAiplatformV1ModelDeploymentMonitoringObjectiveConfig Object}
\usage{
GoogleCloudAiplatformV1ModelDeploymentMonitoringObjectiveConfig(
  deployedModelId = NULL,
  objectiveConfig = NULL
)
}
\arguments{
\item{deployedModelId}{The DeployedModel ID of the objective config}

\item{objectiveConfig}{The objective config of for the modelmonitoring job of this deployed model}
}
\value{
GoogleCloudAiplatformV1ModelDeploymentMonitoringObjectiveConfig object
}
\description{
GoogleCloudAiplatformV1ModelDeploymentMonitoringObjectiveConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
ModelDeploymentMonitoringObjectiveConfig contains the pair of deployed_model_id to ModelMonitoringObjectiveConfig.
}
\concept{GoogleCloudAiplatformV1ModelDeploymentMonitoringObjectiveConfig functions}
