% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ModelDeploymentMonitoringScheduleConfig}
\alias{GoogleCloudAiplatformV1ModelDeploymentMonitoringScheduleConfig}
\title{GoogleCloudAiplatformV1ModelDeploymentMonitoringScheduleConfig Object}
\usage{
GoogleCloudAiplatformV1ModelDeploymentMonitoringScheduleConfig(
  monitorInterval = NULL,
  monitorWindow = NULL
)
}
\arguments{
\item{monitorInterval}{Required}

\item{monitorWindow}{The time window of the prediction data being included in each prediction dataset}
}
\value{
GoogleCloudAiplatformV1ModelDeploymentMonitoringScheduleConfig object
}
\description{
GoogleCloudAiplatformV1ModelDeploymentMonitoringScheduleConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The config for scheduling monitoring job.
}
\concept{GoogleCloudAiplatformV1ModelDeploymentMonitoringScheduleConfig functions}
