% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ModelEvaluationModelEvaluationExplanationSpec}
\alias{GoogleCloudAiplatformV1ModelEvaluationModelEvaluationExplanationSpec}
\title{GoogleCloudAiplatformV1ModelEvaluationModelEvaluationExplanationSpec Object}
\usage{
GoogleCloudAiplatformV1ModelEvaluationModelEvaluationExplanationSpec(
  explanationType = NULL,
  explanationSpec = NULL
)
}
\arguments{
\item{explanationType}{Explanation type}

\item{explanationSpec}{Explanation spec details}
}
\value{
GoogleCloudAiplatformV1ModelEvaluationModelEvaluationExplanationSpec object
}
\description{
GoogleCloudAiplatformV1ModelEvaluationModelEvaluationExplanationSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\concept{GoogleCloudAiplatformV1ModelEvaluationModelEvaluationExplanationSpec functions}
