% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ModelEvaluationSlice}
\alias{GoogleCloudAiplatformV1ModelEvaluationSlice}
\title{GoogleCloudAiplatformV1ModelEvaluationSlice Object}
\usage{
GoogleCloudAiplatformV1ModelEvaluationSlice()
}
\value{
GoogleCloudAiplatformV1ModelEvaluationSlice object
}
\description{
GoogleCloudAiplatformV1ModelEvaluationSlice Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A collection of metrics calculated by comparing Model's predictions on a slice of the test data against ground truth annotations.
}
\concept{GoogleCloudAiplatformV1ModelEvaluationSlice functions}
