% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ModelEvaluationSliceSlice}
\alias{GoogleCloudAiplatformV1ModelEvaluationSliceSlice}
\title{GoogleCloudAiplatformV1ModelEvaluationSliceSlice Object}
\usage{
GoogleCloudAiplatformV1ModelEvaluationSliceSlice()
}
\value{
GoogleCloudAiplatformV1ModelEvaluationSliceSlice object
}
\description{
GoogleCloudAiplatformV1ModelEvaluationSliceSlice Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Definition of a slice.
}
\concept{GoogleCloudAiplatformV1ModelEvaluationSliceSlice functions}
