% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ModelEvaluationSliceSliceSliceSpec.configs}
\alias{GoogleCloudAiplatformV1ModelEvaluationSliceSliceSliceSpec.configs}
\title{GoogleCloudAiplatformV1ModelEvaluationSliceSliceSliceSpec.configs Object}
\usage{
GoogleCloudAiplatformV1ModelEvaluationSliceSliceSliceSpec.configs()
}
\value{
GoogleCloudAiplatformV1ModelEvaluationSliceSliceSliceSpec.configs object
}
\description{
GoogleCloudAiplatformV1ModelEvaluationSliceSliceSliceSpec.configs Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Mapping configuration for this SliceSpec. The key is the name of the feature. By default, the key will be prefixed by 'instance' as a dictionary prefix for Vertex Batch Predictions output format.
}
\seealso{
Other GoogleCloudAiplatformV1ModelEvaluationSliceSliceSliceSpec functions: 
\code{\link{GoogleCloudAiplatformV1ModelEvaluationSliceSliceSliceSpec}()}
}
\concept{GoogleCloudAiplatformV1ModelEvaluationSliceSliceSliceSpec functions}
