% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ModelEvaluationSliceSliceSliceSpec}
\alias{GoogleCloudAiplatformV1ModelEvaluationSliceSliceSliceSpec}
\title{GoogleCloudAiplatformV1ModelEvaluationSliceSliceSliceSpec Object}
\usage{
GoogleCloudAiplatformV1ModelEvaluationSliceSliceSliceSpec(
  GoogleCloudAiplatformV1ModelEvaluationSliceSliceSliceSpec.configs = NULL,
  configs = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1ModelEvaluationSliceSliceSliceSpec.configs}{The \link{GoogleCloudAiplatformV1ModelEvaluationSliceSliceSliceSpec.configs} object or list of objects}

\item{configs}{Mapping configuration for this SliceSpec}
}
\value{
GoogleCloudAiplatformV1ModelEvaluationSliceSliceSliceSpec object
}
\description{
GoogleCloudAiplatformV1ModelEvaluationSliceSliceSliceSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Specification for how the data should be sliced.
}
\seealso{
Other GoogleCloudAiplatformV1ModelEvaluationSliceSliceSliceSpec functions: 
\code{\link{GoogleCloudAiplatformV1ModelEvaluationSliceSliceSliceSpec.configs}()}
}
\concept{GoogleCloudAiplatformV1ModelEvaluationSliceSliceSliceSpec functions}
