% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ModelEvaluationSliceSliceSliceSpecSliceConfig}
\alias{GoogleCloudAiplatformV1ModelEvaluationSliceSliceSliceSpecSliceConfig}
\title{GoogleCloudAiplatformV1ModelEvaluationSliceSliceSliceSpecSliceConfig Object}
\usage{
GoogleCloudAiplatformV1ModelEvaluationSliceSliceSliceSpecSliceConfig(
  value = NULL,
  range = NULL,
  allValues = NULL
)
}
\arguments{
\item{value}{A unique specific value for a given feature}

\item{range}{A range of values for a numerical feature}

\item{allValues}{If all_values is set to true, then all possible labels of the keyed feature will have another slice computed}
}
\value{
GoogleCloudAiplatformV1ModelEvaluationSliceSliceSliceSpecSliceConfig object
}
\description{
GoogleCloudAiplatformV1ModelEvaluationSliceSliceSliceSpecSliceConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Specification message containing the config for this SliceSpec. When \code{kind} is selected as \code{value} and/or \code{range}, only a single slice will be computed. When \code{all_values} is present, a separate slice will be computed for each possible label/value for the corresponding key in \code{config}. Examples, with feature zip_code with values 12345, 23334, 88888 and feature country with values 'US', 'Canada', 'Mexico' in the dataset: Example 1: { 'zip_code': { 'value': { 'float_value': 12345.0 } } } A single slice for any data with zip_code 12345 in the dataset. Example 2: { 'zip_code': { 'range': { 'low': 12345, 'high': 20000 } } } A single slice containing data where the zip_codes between 12345 and 20000 For this example, data with the zip_code of 12345 will be in this slice. Example 3: { 'zip_code': { 'range': { 'low': 10000, 'high': 20000 } }, 'country': { 'value': { 'string_value': 'US' } } } A single slice containing data where the zip_codes between 10000 and 20000 has the country 'US'. For this example, data with the zip_code of 12345 and country 'US' will be in this slice. Example 4: { 'country': {'all_values': { 'value': true } } } Three slices are computed, one for each unique country in the dataset. Example 5: { 'country': { 'all_values': { 'value': true } }, 'zip_code': { 'value': { 'float_value': 12345.0 } } } Three slices are computed, one for each unique country in the dataset where the zip_code is also 12345. For this example, data with zip_code 12345 and country 'US' will be in one slice, zip_code 12345 and country 'Canada' in another slice, and zip_code 12345 and country 'Mexico' in another slice, totaling 3 slices.
}
\concept{GoogleCloudAiplatformV1ModelEvaluationSliceSliceSliceSpecSliceConfig functions}
