% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ModelEvaluationSliceSliceSliceSpecValue}
\alias{GoogleCloudAiplatformV1ModelEvaluationSliceSliceSliceSpecValue}
\title{GoogleCloudAiplatformV1ModelEvaluationSliceSliceSliceSpecValue Object}
\usage{
GoogleCloudAiplatformV1ModelEvaluationSliceSliceSliceSpecValue(
  floatValue = NULL,
  stringValue = NULL
)
}
\arguments{
\item{floatValue}{Float type}

\item{stringValue}{String type}
}
\value{
GoogleCloudAiplatformV1ModelEvaluationSliceSliceSliceSpecValue object
}
\description{
GoogleCloudAiplatformV1ModelEvaluationSliceSliceSliceSpecValue Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Single value that supports strings and floats.
}
\concept{GoogleCloudAiplatformV1ModelEvaluationSliceSliceSliceSpecValue functions}
