% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ModelExplanation}
\alias{GoogleCloudAiplatformV1ModelExplanation}
\title{GoogleCloudAiplatformV1ModelExplanation Object}
\usage{
GoogleCloudAiplatformV1ModelExplanation()
}
\value{
GoogleCloudAiplatformV1ModelExplanation object
}
\description{
GoogleCloudAiplatformV1ModelExplanation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Aggregated explanation metrics for a Model over a set of instances.
}
\concept{GoogleCloudAiplatformV1ModelExplanation functions}
