% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ModelExportFormat}
\alias{GoogleCloudAiplatformV1ModelExportFormat}
\title{GoogleCloudAiplatformV1ModelExportFormat Object}
\usage{
GoogleCloudAiplatformV1ModelExportFormat()
}
\value{
GoogleCloudAiplatformV1ModelExportFormat object
}
\description{
GoogleCloudAiplatformV1ModelExportFormat Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents export format supported by the Model. All formats export to Google Cloud Storage.
}
\concept{GoogleCloudAiplatformV1ModelExportFormat functions}
