% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ModelMonitoringAlertConfig}
\alias{GoogleCloudAiplatformV1ModelMonitoringAlertConfig}
\title{GoogleCloudAiplatformV1ModelMonitoringAlertConfig Object}
\usage{
GoogleCloudAiplatformV1ModelMonitoringAlertConfig(
  emailAlertConfig = NULL,
  enableLogging = NULL
)
}
\arguments{
\item{emailAlertConfig}{Email alert config}

\item{enableLogging}{Dump the anomalies to Cloud Logging}
}
\value{
GoogleCloudAiplatformV1ModelMonitoringAlertConfig object
}
\description{
GoogleCloudAiplatformV1ModelMonitoringAlertConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\concept{GoogleCloudAiplatformV1ModelMonitoringAlertConfig functions}
