% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ModelMonitoringAlertConfigEmailAlertConfig}
\alias{GoogleCloudAiplatformV1ModelMonitoringAlertConfigEmailAlertConfig}
\title{GoogleCloudAiplatformV1ModelMonitoringAlertConfigEmailAlertConfig Object}
\usage{
GoogleCloudAiplatformV1ModelMonitoringAlertConfigEmailAlertConfig(
  userEmails = NULL
)
}
\arguments{
\item{userEmails}{The email addresses to send the alert}
}
\value{
GoogleCloudAiplatformV1ModelMonitoringAlertConfigEmailAlertConfig object
}
\description{
GoogleCloudAiplatformV1ModelMonitoringAlertConfigEmailAlertConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The config for email alert.
}
\concept{GoogleCloudAiplatformV1ModelMonitoringAlertConfigEmailAlertConfig functions}
