% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfig}
\alias{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfig}
\title{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfig Object}
\usage{
GoogleCloudAiplatformV1ModelMonitoringObjectiveConfig(
  trainingPredictionSkewDetectionConfig = NULL,
  predictionDriftDetectionConfig = NULL,
  trainingDataset = NULL,
  explanationConfig = NULL
)
}
\arguments{
\item{trainingPredictionSkewDetectionConfig}{The config for skew between training data and prediction data}

\item{predictionDriftDetectionConfig}{The config for drift of prediction data}

\item{trainingDataset}{Training dataset for models}

\item{explanationConfig}{The config for integrating with Vertex Explainable AI}
}
\value{
GoogleCloudAiplatformV1ModelMonitoringObjectiveConfig object
}
\description{
GoogleCloudAiplatformV1ModelMonitoringObjectiveConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The objective configuration for model monitoring, including the information needed to detect anomalies for one particular model.
}
\concept{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfig functions}
