% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigExplanationConfig}
\alias{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigExplanationConfig}
\title{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigExplanationConfig Object}
\usage{
GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigExplanationConfig(
  explanationBaseline = NULL,
  enableFeatureAttributes = NULL
)
}
\arguments{
\item{explanationBaseline}{Predictions generated by the BatchPredictionJob using baseline dataset}

\item{enableFeatureAttributes}{If want to analyze the Vertex Explainable AI feature attribute scores or not}
}
\value{
GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigExplanationConfig object
}
\description{
GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigExplanationConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The config for integrating with Vertex Explainable AI. Only applicable if the Model has explanation_spec populated.
}
\concept{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigExplanationConfig functions}
