% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigExplanationConfigExplanationBaseline}
\alias{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigExplanationConfigExplanationBaseline}
\title{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigExplanationConfigExplanationBaseline Object}
\usage{
GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigExplanationConfigExplanationBaseline(
  predictionFormat = NULL,
  bigquery = NULL,
  gcs = NULL
)
}
\arguments{
\item{predictionFormat}{The storage format of the predictions generated BatchPrediction job}

\item{bigquery}{BigQuery location for BatchExplain output}

\item{gcs}{Cloud Storage location for BatchExplain output}
}
\value{
GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigExplanationConfigExplanationBaseline object
}
\description{
GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigExplanationConfigExplanationBaseline Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Output from BatchPredictionJob for Model Monitoring baseline dataset, which can be used to generate baseline attribution scores.
}
\concept{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigExplanationConfigExplanationBaseline functions}
