% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigPredictionDriftDetectionConfig.attributionScoreDriftThresholds}
\alias{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigPredictionDriftDetectionConfig.attributionScoreDriftThresholds}
\title{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigPredictionDriftDetectionConfig.attributionScoreDriftThresholds Object}
\usage{
GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigPredictionDriftDetectionConfig.attributionScoreDriftThresholds(
  
)
}
\value{
GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigPredictionDriftDetectionConfig.attributionScoreDriftThresholds object
}
\description{
GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigPredictionDriftDetectionConfig.attributionScoreDriftThresholds Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Key is the feature name and value is the threshold. The threshold here is against attribution score distance between different time windows.
}
\seealso{
Other GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigPredictionDriftDetectionConfig functions: 
\code{\link{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigPredictionDriftDetectionConfig.driftThresholds}()},
\code{\link{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigPredictionDriftDetectionConfig}()}
}
\concept{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigPredictionDriftDetectionConfig functions}
