% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigPredictionDriftDetectionConfig.driftThresholds}
\alias{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigPredictionDriftDetectionConfig.driftThresholds}
\title{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigPredictionDriftDetectionConfig.driftThresholds Object}
\usage{
GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigPredictionDriftDetectionConfig.driftThresholds(
  
)
}
\value{
GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigPredictionDriftDetectionConfig.driftThresholds object
}
\description{
GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigPredictionDriftDetectionConfig.driftThresholds Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Key is the feature name and value is the threshold. If a feature needs to be monitored for drift, a value threshold must be configured for that feature. The threshold here is against feature distribution distance between different time windws.
}
\seealso{
Other GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigPredictionDriftDetectionConfig functions: 
\code{\link{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigPredictionDriftDetectionConfig.attributionScoreDriftThresholds}()},
\code{\link{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigPredictionDriftDetectionConfig}()}
}
\concept{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigPredictionDriftDetectionConfig functions}
