% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigTrainingDataset}
\alias{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigTrainingDataset}
\title{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigTrainingDataset Object}
\usage{
GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigTrainingDataset(
  dataFormat = NULL,
  targetField = NULL,
  bigquerySource = NULL,
  dataset = NULL,
  loggingSamplingStrategy = NULL,
  gcsSource = NULL
)
}
\arguments{
\item{dataFormat}{Data format of the dataset, only applicable if the input is from Google Cloud Storage}

\item{targetField}{The target field name the model is to predict}

\item{bigquerySource}{The BigQuery table of the unmanaged Dataset used to train this Model}

\item{dataset}{The resource name of the Dataset used to train this Model}

\item{loggingSamplingStrategy}{Strategy to sample data from Training Dataset}

\item{gcsSource}{The Google Cloud Storage uri of the unmanaged Dataset used to train this Model}
}
\value{
GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigTrainingDataset object
}
\description{
GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigTrainingDataset Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Training Dataset information.
}
\concept{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigTrainingDataset functions}
