% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigTrainingPredictionSkewDetectionConfig.attributionScoreSkewThresholds}
\alias{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigTrainingPredictionSkewDetectionConfig.attributionScoreSkewThresholds}
\title{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigTrainingPredictionSkewDetectionConfig.attributionScoreSkewThresholds Object}
\usage{
GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigTrainingPredictionSkewDetectionConfig.attributionScoreSkewThresholds(
  
)
}
\value{
GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigTrainingPredictionSkewDetectionConfig.attributionScoreSkewThresholds object
}
\description{
GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigTrainingPredictionSkewDetectionConfig.attributionScoreSkewThresholds Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Key is the feature name and value is the threshold. The threshold here is against attribution score distance between the training and prediction feature.
}
\seealso{
Other GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigTrainingPredictionSkewDetectionConfig functions: 
\code{\link{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigTrainingPredictionSkewDetectionConfig.skewThresholds}()},
\code{\link{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigTrainingPredictionSkewDetectionConfig}()}
}
\concept{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigTrainingPredictionSkewDetectionConfig functions}
