% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigTrainingPredictionSkewDetectionConfig}
\alias{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigTrainingPredictionSkewDetectionConfig}
\title{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigTrainingPredictionSkewDetectionConfig Object}
\usage{
GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigTrainingPredictionSkewDetectionConfig(
 
    GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigTrainingPredictionSkewDetectionConfig.attributionScoreSkewThresholds
    = NULL,
 
    GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigTrainingPredictionSkewDetectionConfig.skewThresholds
    = NULL,
  attributionScoreSkewThresholds = NULL,
  defaultSkewThreshold = NULL,
  skewThresholds = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigTrainingPredictionSkewDetectionConfig.attributionScoreSkewThresholds}{The \link{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigTrainingPredictionSkewDetectionConfig.attributionScoreSkewThresholds} object or list of objects}

\item{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigTrainingPredictionSkewDetectionConfig.skewThresholds}{The \link{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigTrainingPredictionSkewDetectionConfig.skewThresholds} object or list of objects}

\item{attributionScoreSkewThresholds}{Key is the feature name and value is the threshold}

\item{defaultSkewThreshold}{Skew anomaly detection threshold used by all features}

\item{skewThresholds}{Key is the feature name and value is the threshold}
}
\value{
GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigTrainingPredictionSkewDetectionConfig object
}
\description{
GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigTrainingPredictionSkewDetectionConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The config for Training & Prediction data skew detection. It specifies the training dataset sources and the skew detection parameters.
}
\seealso{
Other GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigTrainingPredictionSkewDetectionConfig functions: 
\code{\link{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigTrainingPredictionSkewDetectionConfig.attributionScoreSkewThresholds}()},
\code{\link{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigTrainingPredictionSkewDetectionConfig.skewThresholds}()}
}
\concept{GoogleCloudAiplatformV1ModelMonitoringObjectiveConfigTrainingPredictionSkewDetectionConfig functions}
