% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ModelMonitoringStatsAnomalies}
\alias{GoogleCloudAiplatformV1ModelMonitoringStatsAnomalies}
\title{GoogleCloudAiplatformV1ModelMonitoringStatsAnomalies Object}
\usage{
GoogleCloudAiplatformV1ModelMonitoringStatsAnomalies(
  objective = NULL,
  deployedModelId = NULL,
  featureStats = NULL,
  anomalyCount = NULL
)
}
\arguments{
\item{objective}{Model Monitoring Objective those stats and anomalies belonging to}

\item{deployedModelId}{Deployed Model ID}

\item{featureStats}{A list of historical Stats and Anomalies generated for all Features}

\item{anomalyCount}{Number of anomalies within all stats}
}
\value{
GoogleCloudAiplatformV1ModelMonitoringStatsAnomalies object
}
\description{
GoogleCloudAiplatformV1ModelMonitoringStatsAnomalies Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Statistics and anomalies generated by Model Monitoring.
}
\concept{GoogleCloudAiplatformV1ModelMonitoringStatsAnomalies functions}
