% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ModelMonitoringStatsAnomaliesFeatureHistoricStatsAnomalies}
\alias{GoogleCloudAiplatformV1ModelMonitoringStatsAnomaliesFeatureHistoricStatsAnomalies}
\title{GoogleCloudAiplatformV1ModelMonitoringStatsAnomaliesFeatureHistoricStatsAnomalies Object}
\usage{
GoogleCloudAiplatformV1ModelMonitoringStatsAnomaliesFeatureHistoricStatsAnomalies(
  threshold = NULL,
  trainingStats = NULL,
  predictionStats = NULL,
  featureDisplayName = NULL
)
}
\arguments{
\item{threshold}{Threshold for anomaly detection}

\item{trainingStats}{Stats calculated for the Training Dataset}

\item{predictionStats}{A list of historical stats generated by different time window's Prediction Dataset}

\item{featureDisplayName}{Display Name of the Feature}
}
\value{
GoogleCloudAiplatformV1ModelMonitoringStatsAnomaliesFeatureHistoricStatsAnomalies object
}
\description{
GoogleCloudAiplatformV1ModelMonitoringStatsAnomaliesFeatureHistoricStatsAnomalies Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Historical Stats (and Anomalies) for a specific Feature.
}
\concept{GoogleCloudAiplatformV1ModelMonitoringStatsAnomaliesFeatureHistoricStatsAnomalies functions}
