% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ModelOriginalModelInfo}
\alias{GoogleCloudAiplatformV1ModelOriginalModelInfo}
\title{GoogleCloudAiplatformV1ModelOriginalModelInfo Object}
\usage{
GoogleCloudAiplatformV1ModelOriginalModelInfo()
}
\value{
GoogleCloudAiplatformV1ModelOriginalModelInfo object
}
\description{
GoogleCloudAiplatformV1ModelOriginalModelInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains information about the original Model if this Model is a copy.
}
\concept{GoogleCloudAiplatformV1ModelOriginalModelInfo functions}
