% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ModelSourceInfo}
\alias{GoogleCloudAiplatformV1ModelSourceInfo}
\title{GoogleCloudAiplatformV1ModelSourceInfo Object}
\usage{
GoogleCloudAiplatformV1ModelSourceInfo(sourceType = NULL, copy = NULL)
}
\arguments{
\item{sourceType}{Type of the model source}

\item{copy}{If this Model is copy of another Model}
}
\value{
GoogleCloudAiplatformV1ModelSourceInfo object
}
\description{
GoogleCloudAiplatformV1ModelSourceInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Detail description of the source information of the model.
}
\concept{GoogleCloudAiplatformV1ModelSourceInfo functions}
