% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1MutateDeployedIndexOperationMetadata}
\alias{GoogleCloudAiplatformV1MutateDeployedIndexOperationMetadata}
\title{GoogleCloudAiplatformV1MutateDeployedIndexOperationMetadata Object}
\usage{
GoogleCloudAiplatformV1MutateDeployedIndexOperationMetadata(
  genericMetadata = NULL,
  deployedIndexId = NULL
)
}
\arguments{
\item{genericMetadata}{The operation generic information}

\item{deployedIndexId}{The unique index id specified by user}
}
\value{
GoogleCloudAiplatformV1MutateDeployedIndexOperationMetadata object
}
\description{
GoogleCloudAiplatformV1MutateDeployedIndexOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Runtime operation information for IndexEndpointService.MutateDeployedIndex.
}
\concept{GoogleCloudAiplatformV1MutateDeployedIndexOperationMetadata functions}
