% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1MutateDeployedIndexResponse}
\alias{GoogleCloudAiplatformV1MutateDeployedIndexResponse}
\title{GoogleCloudAiplatformV1MutateDeployedIndexResponse Object}
\usage{
GoogleCloudAiplatformV1MutateDeployedIndexResponse(deployedIndex = NULL)
}
\arguments{
\item{deployedIndex}{The DeployedIndex that had been updated in the IndexEndpoint}
}
\value{
GoogleCloudAiplatformV1MutateDeployedIndexResponse object
}
\description{
GoogleCloudAiplatformV1MutateDeployedIndexResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for IndexEndpointService.MutateDeployedIndex.
}
\concept{GoogleCloudAiplatformV1MutateDeployedIndexResponse functions}
