% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1MutateDeployedModelOperationMetadata}
\alias{GoogleCloudAiplatformV1MutateDeployedModelOperationMetadata}
\title{GoogleCloudAiplatformV1MutateDeployedModelOperationMetadata Object}
\usage{
GoogleCloudAiplatformV1MutateDeployedModelOperationMetadata(
  genericMetadata = NULL
)
}
\arguments{
\item{genericMetadata}{The operation generic information}
}
\value{
GoogleCloudAiplatformV1MutateDeployedModelOperationMetadata object
}
\description{
GoogleCloudAiplatformV1MutateDeployedModelOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Runtime operation information for EndpointService.MutateDeployedModel.
}
\concept{GoogleCloudAiplatformV1MutateDeployedModelOperationMetadata functions}
