% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1MutateDeployedModelRequest}
\alias{GoogleCloudAiplatformV1MutateDeployedModelRequest}
\title{GoogleCloudAiplatformV1MutateDeployedModelRequest Object}
\usage{
GoogleCloudAiplatformV1MutateDeployedModelRequest(
  deployedModel = NULL,
  updateMask = NULL
)
}
\arguments{
\item{deployedModel}{Required}

\item{updateMask}{Required}
}
\value{
GoogleCloudAiplatformV1MutateDeployedModelRequest object
}
\description{
GoogleCloudAiplatformV1MutateDeployedModelRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for EndpointService.MutateDeployedModel.
}
\seealso{
Other GoogleCloudAiplatformV1MutateDeployedModelRequest functions: 
\code{\link{projects.locations.endpoints.mutateDeployedModel}()}
}
\concept{GoogleCloudAiplatformV1MutateDeployedModelRequest functions}
