% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1MutateDeployedModelResponse}
\alias{GoogleCloudAiplatformV1MutateDeployedModelResponse}
\title{GoogleCloudAiplatformV1MutateDeployedModelResponse Object}
\usage{
GoogleCloudAiplatformV1MutateDeployedModelResponse(deployedModel = NULL)
}
\arguments{
\item{deployedModel}{The DeployedModel that's being mutated}
}
\value{
GoogleCloudAiplatformV1MutateDeployedModelResponse object
}
\description{
GoogleCloudAiplatformV1MutateDeployedModelResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for EndpointService.MutateDeployedModel.
}
\concept{GoogleCloudAiplatformV1MutateDeployedModelResponse functions}
