% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1NasJob}
\alias{GoogleCloudAiplatformV1NasJob}
\title{GoogleCloudAiplatformV1NasJob Object}
\usage{
GoogleCloudAiplatformV1NasJob(
  GoogleCloudAiplatformV1NasJob.labels = NULL,
  labels = NULL,
  nasJobSpec = NULL,
  encryptionSpec = NULL,
  displayName = NULL,
  enableRestrictedImageTraining = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1NasJob.labels}{The \link{GoogleCloudAiplatformV1NasJob.labels} object or list of objects}

\item{labels}{The labels with user-defined metadata to organize NasJobs}

\item{nasJobSpec}{Required}

\item{encryptionSpec}{Customer-managed encryption key options for a NasJob}

\item{displayName}{Required}

\item{enableRestrictedImageTraining}{Optional}
}
\value{
GoogleCloudAiplatformV1NasJob object
}
\description{
GoogleCloudAiplatformV1NasJob Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a Neural Architecture Search (NAS) job.
}
\seealso{
Other GoogleCloudAiplatformV1NasJob functions: 
\code{\link{GoogleCloudAiplatformV1NasJob.labels}()},
\code{\link{projects.locations.nasJobs.create}()}
}
\concept{GoogleCloudAiplatformV1NasJob functions}
