% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1NasJobSpec}
\alias{GoogleCloudAiplatformV1NasJobSpec}
\title{GoogleCloudAiplatformV1NasJobSpec Object}
\usage{
GoogleCloudAiplatformV1NasJobSpec(
  multiTrialAlgorithmSpec = NULL,
  resumeNasJobId = NULL,
  searchSpaceSpec = NULL
)
}
\arguments{
\item{multiTrialAlgorithmSpec}{The spec of multi-trial algorithms}

\item{resumeNasJobId}{The ID of the existing NasJob in the same Project and Location which will be used to resume search}

\item{searchSpaceSpec}{It defines the search space for Neural Architecture Search (NAS)}
}
\value{
GoogleCloudAiplatformV1NasJobSpec object
}
\description{
GoogleCloudAiplatformV1NasJobSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents the spec of a NasJob.
}
\concept{GoogleCloudAiplatformV1NasJobSpec functions}
