% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1NasJobSpecMultiTrialAlgorithmSpec}
\alias{GoogleCloudAiplatformV1NasJobSpecMultiTrialAlgorithmSpec}
\title{GoogleCloudAiplatformV1NasJobSpecMultiTrialAlgorithmSpec Object}
\usage{
GoogleCloudAiplatformV1NasJobSpecMultiTrialAlgorithmSpec(
  searchTrialSpec = NULL,
  metric = NULL,
  trainTrialSpec = NULL,
  multiTrialAlgorithm = NULL
)
}
\arguments{
\item{searchTrialSpec}{Required}

\item{metric}{Metric specs for the NAS job}

\item{trainTrialSpec}{Spec for train trials}

\item{multiTrialAlgorithm}{The multi-trial Neural Architecture Search (NAS) algorithm type}
}
\value{
GoogleCloudAiplatformV1NasJobSpecMultiTrialAlgorithmSpec object
}
\description{
GoogleCloudAiplatformV1NasJobSpecMultiTrialAlgorithmSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The spec of multi-trial Neural Architecture Search (NAS).
}
\concept{GoogleCloudAiplatformV1NasJobSpecMultiTrialAlgorithmSpec functions}
