% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1NasJobSpecMultiTrialAlgorithmSpecMetricSpec}
\alias{GoogleCloudAiplatformV1NasJobSpecMultiTrialAlgorithmSpecMetricSpec}
\title{GoogleCloudAiplatformV1NasJobSpecMultiTrialAlgorithmSpecMetricSpec Object}
\usage{
GoogleCloudAiplatformV1NasJobSpecMultiTrialAlgorithmSpecMetricSpec(
  goal = NULL,
  metricId = NULL
)
}
\arguments{
\item{goal}{Required}

\item{metricId}{Required}
}
\value{
GoogleCloudAiplatformV1NasJobSpecMultiTrialAlgorithmSpecMetricSpec object
}
\description{
GoogleCloudAiplatformV1NasJobSpecMultiTrialAlgorithmSpecMetricSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a metric to optimize.
}
\concept{GoogleCloudAiplatformV1NasJobSpecMultiTrialAlgorithmSpecMetricSpec functions}
