% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1NasJobSpecMultiTrialAlgorithmSpecSearchTrialSpec}
\alias{GoogleCloudAiplatformV1NasJobSpecMultiTrialAlgorithmSpecSearchTrialSpec}
\title{GoogleCloudAiplatformV1NasJobSpecMultiTrialAlgorithmSpecSearchTrialSpec Object}
\usage{
GoogleCloudAiplatformV1NasJobSpecMultiTrialAlgorithmSpecSearchTrialSpec(
  maxFailedTrialCount = NULL,
  maxParallelTrialCount = NULL,
  searchTrialJobSpec = NULL,
  maxTrialCount = NULL
)
}
\arguments{
\item{maxFailedTrialCount}{The number of failed trials that need to be seen before failing the NasJob}

\item{maxParallelTrialCount}{Required}

\item{searchTrialJobSpec}{Required}

\item{maxTrialCount}{Required}
}
\value{
GoogleCloudAiplatformV1NasJobSpecMultiTrialAlgorithmSpecSearchTrialSpec object
}
\description{
GoogleCloudAiplatformV1NasJobSpecMultiTrialAlgorithmSpecSearchTrialSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represent spec for search trials.
}
\concept{GoogleCloudAiplatformV1NasJobSpecMultiTrialAlgorithmSpecSearchTrialSpec functions}
