% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1NasTrial}
\alias{GoogleCloudAiplatformV1NasTrial}
\title{GoogleCloudAiplatformV1NasTrial Object}
\usage{
GoogleCloudAiplatformV1NasTrial()
}
\value{
GoogleCloudAiplatformV1NasTrial object
}
\description{
GoogleCloudAiplatformV1NasTrial Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a uCAIP NasJob trial.
}
\concept{GoogleCloudAiplatformV1NasTrial functions}
