% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1NasTrialDetail}
\alias{GoogleCloudAiplatformV1NasTrialDetail}
\title{GoogleCloudAiplatformV1NasTrialDetail Object}
\usage{
GoogleCloudAiplatformV1NasTrialDetail(
  trainTrial = NULL,
  parameters = NULL,
  searchTrial = NULL
)
}
\arguments{
\item{trainTrial}{The train NasTrial corresponding to search_trial}

\item{parameters}{The parameters for the NasJob NasTrial}

\item{searchTrial}{The requested search NasTrial}
}
\value{
GoogleCloudAiplatformV1NasTrialDetail object
}
\description{
GoogleCloudAiplatformV1NasTrialDetail Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a NasTrial details along with its parameters. If there is a corresponding train NasTrial, the train NasTrial is also returned.
}
\concept{GoogleCloudAiplatformV1NasTrialDetail functions}
