% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1NearestNeighborSearchOperationMetadata}
\alias{GoogleCloudAiplatformV1NearestNeighborSearchOperationMetadata}
\title{GoogleCloudAiplatformV1NearestNeighborSearchOperationMetadata Object}
\usage{
GoogleCloudAiplatformV1NearestNeighborSearchOperationMetadata(
  dataBytesCount = NULL,
  contentValidationStats = NULL
)
}
\arguments{
\item{dataBytesCount}{The ingested data size in bytes}

\item{contentValidationStats}{The validation stats of the content (per file) to be inserted or updated on the Matching Engine Index resource}
}
\value{
GoogleCloudAiplatformV1NearestNeighborSearchOperationMetadata object
}
\description{
GoogleCloudAiplatformV1NearestNeighborSearchOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Runtime operation metadata with regard to Matching Engine Index.
}
\concept{GoogleCloudAiplatformV1NearestNeighborSearchOperationMetadata functions}
