% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1NearestNeighborSearchOperationMetadataContentValidationStats}
\alias{GoogleCloudAiplatformV1NearestNeighborSearchOperationMetadataContentValidationStats}
\title{GoogleCloudAiplatformV1NearestNeighborSearchOperationMetadataContentValidationStats Object}
\usage{
GoogleCloudAiplatformV1NearestNeighborSearchOperationMetadataContentValidationStats(
  sourceGcsUri = NULL,
  validRecordCount = NULL,
  invalidRecordCount = NULL,
  partialErrors = NULL
)
}
\arguments{
\item{sourceGcsUri}{Cloud Storage URI pointing to the original file in user's bucket}

\item{validRecordCount}{Number of records in this file that were successfully processed}

\item{invalidRecordCount}{Number of records in this file we skipped due to validate errors}

\item{partialErrors}{The detail information of the partial failures encountered for those invalid records that couldn't be parsed}
}
\value{
GoogleCloudAiplatformV1NearestNeighborSearchOperationMetadataContentValidationStats object
}
\description{
GoogleCloudAiplatformV1NearestNeighborSearchOperationMetadataContentValidationStats Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\concept{GoogleCloudAiplatformV1NearestNeighborSearchOperationMetadataContentValidationStats functions}
