% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1NearestNeighborSearchOperationMetadataRecordError}
\alias{GoogleCloudAiplatformV1NearestNeighborSearchOperationMetadataRecordError}
\title{GoogleCloudAiplatformV1NearestNeighborSearchOperationMetadataRecordError Object}
\usage{
GoogleCloudAiplatformV1NearestNeighborSearchOperationMetadataRecordError(
  rawRecord = NULL,
  embeddingId = NULL,
  errorMessage = NULL,
  sourceGcsUri = NULL,
  errorType = NULL
)
}
\arguments{
\item{rawRecord}{The original content of this record}

\item{embeddingId}{Empty if the embedding id is failed to parse}

\item{errorMessage}{A human-readable message that is shown to the user to help them fix the error}

\item{sourceGcsUri}{Cloud Storage URI pointing to the original file in user's bucket}

\item{errorType}{The error type of this record}
}
\value{
GoogleCloudAiplatformV1NearestNeighborSearchOperationMetadataRecordError object
}
\description{
GoogleCloudAiplatformV1NearestNeighborSearchOperationMetadataRecordError Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\concept{GoogleCloudAiplatformV1NearestNeighborSearchOperationMetadataRecordError functions}
