% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1NfsMount}
\alias{GoogleCloudAiplatformV1NfsMount}
\title{GoogleCloudAiplatformV1NfsMount Object}
\usage{
GoogleCloudAiplatformV1NfsMount(mountPoint = NULL, server = NULL, path = NULL)
}
\arguments{
\item{mountPoint}{Required}

\item{server}{Required}

\item{path}{Required}
}
\value{
GoogleCloudAiplatformV1NfsMount object
}
\description{
GoogleCloudAiplatformV1NfsMount Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a mount configuration for Network File System (NFS) to mount.
}
\concept{GoogleCloudAiplatformV1NfsMount functions}
