% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1PipelineJob.labels}
\alias{GoogleCloudAiplatformV1PipelineJob.labels}
\title{GoogleCloudAiplatformV1PipelineJob.labels Object}
\usage{
GoogleCloudAiplatformV1PipelineJob.labels()
}
\value{
GoogleCloudAiplatformV1PipelineJob.labels object
}
\description{
GoogleCloudAiplatformV1PipelineJob.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The labels with user-defined metadata to organize PipelineJob. Label keys and values can be no longer than 64 characters (Unicode codepoints), can only contain lowercase letters, numeric characters, underscores and dashes. International characters are allowed. See https://goo.gl/xmQnxf for more information and examples of labels. Note there is some reserved label key for Vertex AI Pipelines. - \code{vertex-ai-pipelines-run-billing-id}, user set value will get overrided.
}
\seealso{
Other GoogleCloudAiplatformV1PipelineJob functions: 
\code{\link{GoogleCloudAiplatformV1PipelineJob.pipelineSpec}()},
\code{\link{GoogleCloudAiplatformV1PipelineJob}()},
\code{\link{projects.locations.pipelineJobs.create}()}
}
\concept{GoogleCloudAiplatformV1PipelineJob functions}
