% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1PipelineJob.pipelineSpec}
\alias{GoogleCloudAiplatformV1PipelineJob.pipelineSpec}
\title{GoogleCloudAiplatformV1PipelineJob.pipelineSpec Object}
\usage{
GoogleCloudAiplatformV1PipelineJob.pipelineSpec()
}
\value{
GoogleCloudAiplatformV1PipelineJob.pipelineSpec object
}
\description{
GoogleCloudAiplatformV1PipelineJob.pipelineSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The spec of the pipeline.
}
\seealso{
Other GoogleCloudAiplatformV1PipelineJob functions: 
\code{\link{GoogleCloudAiplatformV1PipelineJob.labels}()},
\code{\link{GoogleCloudAiplatformV1PipelineJob}()},
\code{\link{projects.locations.pipelineJobs.create}()}
}
\concept{GoogleCloudAiplatformV1PipelineJob functions}
