% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1PipelineJob}
\alias{GoogleCloudAiplatformV1PipelineJob}
\title{GoogleCloudAiplatformV1PipelineJob Object}
\usage{
GoogleCloudAiplatformV1PipelineJob(
  GoogleCloudAiplatformV1PipelineJob.labels = NULL,
  GoogleCloudAiplatformV1PipelineJob.pipelineSpec = NULL,
  runtimeConfig = NULL,
  templateUri = NULL,
  serviceAccount = NULL,
  displayName = NULL,
  labels = NULL,
  reservedIpRanges = NULL,
  encryptionSpec = NULL,
  pipelineSpec = NULL,
  network = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1PipelineJob.labels}{The \link{GoogleCloudAiplatformV1PipelineJob.labels} object or list of objects}

\item{GoogleCloudAiplatformV1PipelineJob.pipelineSpec}{The \link{GoogleCloudAiplatformV1PipelineJob.pipelineSpec} object or list of objects}

\item{runtimeConfig}{Runtime config of the pipeline}

\item{templateUri}{A template uri from where the PipelineJob}

\item{serviceAccount}{The service account that the pipeline workload runs as}

\item{displayName}{The display name of the Pipeline}

\item{labels}{The labels with user-defined metadata to organize PipelineJob}

\item{reservedIpRanges}{A list of names for the reserved ip ranges under the VPC network that can be used for this Pipeline Job's workload}

\item{encryptionSpec}{Customer-managed encryption key spec for a pipelineJob}

\item{pipelineSpec}{The spec of the pipeline}

\item{network}{The full name of the Compute Engine \href{/compute/docs/networks-and-firewalls#networks}{network} to which the Pipeline Job's workload should be peered}
}
\value{
GoogleCloudAiplatformV1PipelineJob object
}
\description{
GoogleCloudAiplatformV1PipelineJob Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An instance of a machine learning PipelineJob.
}
\seealso{
Other GoogleCloudAiplatformV1PipelineJob functions: 
\code{\link{GoogleCloudAiplatformV1PipelineJob.labels}()},
\code{\link{GoogleCloudAiplatformV1PipelineJob.pipelineSpec}()},
\code{\link{projects.locations.pipelineJobs.create}()}
}
\concept{GoogleCloudAiplatformV1PipelineJob functions}
