% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1PipelineJobRuntimeConfig.inputArtifacts}
\alias{GoogleCloudAiplatformV1PipelineJobRuntimeConfig.inputArtifacts}
\title{GoogleCloudAiplatformV1PipelineJobRuntimeConfig.inputArtifacts Object}
\usage{
GoogleCloudAiplatformV1PipelineJobRuntimeConfig.inputArtifacts()
}
\value{
GoogleCloudAiplatformV1PipelineJobRuntimeConfig.inputArtifacts object
}
\description{
GoogleCloudAiplatformV1PipelineJobRuntimeConfig.inputArtifacts Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The runtime artifacts of the PipelineJob. The key will be the input artifact name and the value would be one of the InputArtifact.
}
\seealso{
Other GoogleCloudAiplatformV1PipelineJobRuntimeConfig functions: 
\code{\link{GoogleCloudAiplatformV1PipelineJobRuntimeConfig.parameterValues}()},
\code{\link{GoogleCloudAiplatformV1PipelineJobRuntimeConfig.parameters}()},
\code{\link{GoogleCloudAiplatformV1PipelineJobRuntimeConfig}()}
}
\concept{GoogleCloudAiplatformV1PipelineJobRuntimeConfig functions}
