% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1PipelineJobRuntimeConfig.parameters}
\alias{GoogleCloudAiplatformV1PipelineJobRuntimeConfig.parameters}
\title{GoogleCloudAiplatformV1PipelineJobRuntimeConfig.parameters Object}
\usage{
GoogleCloudAiplatformV1PipelineJobRuntimeConfig.parameters()
}
\value{
GoogleCloudAiplatformV1PipelineJobRuntimeConfig.parameters object
}
\description{
GoogleCloudAiplatformV1PipelineJobRuntimeConfig.parameters Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Deprecated. Use RuntimeConfig.parameter_values instead. The runtime parameters of the PipelineJob. The parameters will be passed into PipelineJob.pipeline_spec to replace the placeholders at runtime. This field is used by pipelines built using \code{PipelineJob.pipeline_spec.schema_version} 2.0.0 or lower, such as pipelines built using Kubeflow Pipelines SDK 1.8 or lower.
}
\seealso{
Other GoogleCloudAiplatformV1PipelineJobRuntimeConfig functions: 
\code{\link{GoogleCloudAiplatformV1PipelineJobRuntimeConfig.inputArtifacts}()},
\code{\link{GoogleCloudAiplatformV1PipelineJobRuntimeConfig.parameterValues}()},
\code{\link{GoogleCloudAiplatformV1PipelineJobRuntimeConfig}()}
}
\concept{GoogleCloudAiplatformV1PipelineJobRuntimeConfig functions}
