% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1PipelineJobRuntimeConfig}
\alias{GoogleCloudAiplatformV1PipelineJobRuntimeConfig}
\title{GoogleCloudAiplatformV1PipelineJobRuntimeConfig Object}
\usage{
GoogleCloudAiplatformV1PipelineJobRuntimeConfig(
  GoogleCloudAiplatformV1PipelineJobRuntimeConfig.parameterValues = NULL,
  GoogleCloudAiplatformV1PipelineJobRuntimeConfig.inputArtifacts = NULL,
  GoogleCloudAiplatformV1PipelineJobRuntimeConfig.parameters = NULL,
  gcsOutputDirectory = NULL,
  parameterValues = NULL,
  inputArtifacts = NULL,
  parameters = NULL,
  failurePolicy = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1PipelineJobRuntimeConfig.parameterValues}{The \link{GoogleCloudAiplatformV1PipelineJobRuntimeConfig.parameterValues} object or list of objects}

\item{GoogleCloudAiplatformV1PipelineJobRuntimeConfig.inputArtifacts}{The \link{GoogleCloudAiplatformV1PipelineJobRuntimeConfig.inputArtifacts} object or list of objects}

\item{GoogleCloudAiplatformV1PipelineJobRuntimeConfig.parameters}{The \link{GoogleCloudAiplatformV1PipelineJobRuntimeConfig.parameters} object or list of objects}

\item{gcsOutputDirectory}{Required}

\item{parameterValues}{The runtime parameters of the PipelineJob}

\item{inputArtifacts}{The runtime artifacts of the PipelineJob}

\item{parameters}{Deprecated}

\item{failurePolicy}{Represents the failure policy of a pipeline}
}
\value{
GoogleCloudAiplatformV1PipelineJobRuntimeConfig object
}
\description{
GoogleCloudAiplatformV1PipelineJobRuntimeConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The runtime config of a PipelineJob.
}
\seealso{
Other GoogleCloudAiplatformV1PipelineJobRuntimeConfig functions: 
\code{\link{GoogleCloudAiplatformV1PipelineJobRuntimeConfig.inputArtifacts}()},
\code{\link{GoogleCloudAiplatformV1PipelineJobRuntimeConfig.parameterValues}()},
\code{\link{GoogleCloudAiplatformV1PipelineJobRuntimeConfig.parameters}()}
}
\concept{GoogleCloudAiplatformV1PipelineJobRuntimeConfig functions}
