% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1PipelineJobRuntimeConfigInputArtifact}
\alias{GoogleCloudAiplatformV1PipelineJobRuntimeConfigInputArtifact}
\title{GoogleCloudAiplatformV1PipelineJobRuntimeConfigInputArtifact Object}
\usage{
GoogleCloudAiplatformV1PipelineJobRuntimeConfigInputArtifact(artifactId = NULL)
}
\arguments{
\item{artifactId}{Artifact resource id from MLMD}
}
\value{
GoogleCloudAiplatformV1PipelineJobRuntimeConfigInputArtifact object
}
\description{
GoogleCloudAiplatformV1PipelineJobRuntimeConfigInputArtifact Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The type of an input artifact.
}
\concept{GoogleCloudAiplatformV1PipelineJobRuntimeConfigInputArtifact functions}
