% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1PipelineTaskDetail.inputs}
\alias{GoogleCloudAiplatformV1PipelineTaskDetail.inputs}
\title{GoogleCloudAiplatformV1PipelineTaskDetail.inputs Object}
\usage{
GoogleCloudAiplatformV1PipelineTaskDetail.inputs()
}
\value{
GoogleCloudAiplatformV1PipelineTaskDetail.inputs object
}
\description{
GoogleCloudAiplatformV1PipelineTaskDetail.inputs Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Output only. The runtime input artifacts of the task.
}
\seealso{
Other GoogleCloudAiplatformV1PipelineTaskDetail functions: 
\code{\link{GoogleCloudAiplatformV1PipelineTaskDetail.outputs}()},
\code{\link{GoogleCloudAiplatformV1PipelineTaskDetail}()}
}
\concept{GoogleCloudAiplatformV1PipelineTaskDetail functions}
