% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1PipelineTaskDetail.outputs}
\alias{GoogleCloudAiplatformV1PipelineTaskDetail.outputs}
\title{GoogleCloudAiplatformV1PipelineTaskDetail.outputs Object}
\usage{
GoogleCloudAiplatformV1PipelineTaskDetail.outputs()
}
\value{
GoogleCloudAiplatformV1PipelineTaskDetail.outputs object
}
\description{
GoogleCloudAiplatformV1PipelineTaskDetail.outputs Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Output only. The runtime output artifacts of the task.
}
\seealso{
Other GoogleCloudAiplatformV1PipelineTaskDetail functions: 
\code{\link{GoogleCloudAiplatformV1PipelineTaskDetail.inputs}()},
\code{\link{GoogleCloudAiplatformV1PipelineTaskDetail}()}
}
\concept{GoogleCloudAiplatformV1PipelineTaskDetail functions}
