% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1PipelineTaskDetail}
\alias{GoogleCloudAiplatformV1PipelineTaskDetail}
\title{GoogleCloudAiplatformV1PipelineTaskDetail Object}
\usage{
GoogleCloudAiplatformV1PipelineTaskDetail(
  GoogleCloudAiplatformV1PipelineTaskDetail.inputs = NULL,
  GoogleCloudAiplatformV1PipelineTaskDetail.outputs = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1PipelineTaskDetail.inputs}{The \link{GoogleCloudAiplatformV1PipelineTaskDetail.inputs} object or list of objects}

\item{GoogleCloudAiplatformV1PipelineTaskDetail.outputs}{The \link{GoogleCloudAiplatformV1PipelineTaskDetail.outputs} object or list of objects}
}
\value{
GoogleCloudAiplatformV1PipelineTaskDetail object
}
\description{
GoogleCloudAiplatformV1PipelineTaskDetail Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The runtime detail of a task execution.
}
\seealso{
Other GoogleCloudAiplatformV1PipelineTaskDetail functions: 
\code{\link{GoogleCloudAiplatformV1PipelineTaskDetail.inputs}()},
\code{\link{GoogleCloudAiplatformV1PipelineTaskDetail.outputs}()}
}
\concept{GoogleCloudAiplatformV1PipelineTaskDetail functions}
