% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1PipelineTaskDetailArtifactList}
\alias{GoogleCloudAiplatformV1PipelineTaskDetailArtifactList}
\title{GoogleCloudAiplatformV1PipelineTaskDetailArtifactList Object}
\usage{
GoogleCloudAiplatformV1PipelineTaskDetailArtifactList()
}
\value{
GoogleCloudAiplatformV1PipelineTaskDetailArtifactList object
}
\description{
GoogleCloudAiplatformV1PipelineTaskDetailArtifactList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of artifact metadata.
}
\concept{GoogleCloudAiplatformV1PipelineTaskDetailArtifactList functions}
