% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1PipelineTaskExecutorDetailContainerDetail}
\alias{GoogleCloudAiplatformV1PipelineTaskExecutorDetailContainerDetail}
\title{GoogleCloudAiplatformV1PipelineTaskExecutorDetailContainerDetail Object}
\usage{
GoogleCloudAiplatformV1PipelineTaskExecutorDetailContainerDetail()
}
\value{
GoogleCloudAiplatformV1PipelineTaskExecutorDetailContainerDetail object
}
\description{
GoogleCloudAiplatformV1PipelineTaskExecutorDetailContainerDetail Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The detail of a container execution. It contains the job names of the lifecycle of a container execution.
}
\concept{GoogleCloudAiplatformV1PipelineTaskExecutorDetailContainerDetail functions}
