% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1PipelineTaskExecutorDetailCustomJobDetail}
\alias{GoogleCloudAiplatformV1PipelineTaskExecutorDetailCustomJobDetail}
\title{GoogleCloudAiplatformV1PipelineTaskExecutorDetailCustomJobDetail Object}
\usage{
GoogleCloudAiplatformV1PipelineTaskExecutorDetailCustomJobDetail()
}
\value{
GoogleCloudAiplatformV1PipelineTaskExecutorDetailCustomJobDetail object
}
\description{
GoogleCloudAiplatformV1PipelineTaskExecutorDetailCustomJobDetail Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The detailed info for a custom job executor.
}
\concept{GoogleCloudAiplatformV1PipelineTaskExecutorDetailCustomJobDetail functions}
