% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1PipelineTemplateMetadata}
\alias{GoogleCloudAiplatformV1PipelineTemplateMetadata}
\title{GoogleCloudAiplatformV1PipelineTemplateMetadata Object}
\usage{
GoogleCloudAiplatformV1PipelineTemplateMetadata(version = NULL)
}
\arguments{
\item{version}{The version_name in artifact registry}
}
\value{
GoogleCloudAiplatformV1PipelineTemplateMetadata object
}
\description{
GoogleCloudAiplatformV1PipelineTemplateMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Pipeline template metadata if PipelineJob.template_uri is from supported template registry. Currently, the only supported registry is Artifact Registry.
}
\concept{GoogleCloudAiplatformV1PipelineTemplateMetadata functions}
