% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1Port}
\alias{GoogleCloudAiplatformV1Port}
\title{GoogleCloudAiplatformV1Port Object}
\usage{
GoogleCloudAiplatformV1Port(containerPort = NULL)
}
\arguments{
\item{containerPort}{The number of the port to expose on the pod's IP address}
}
\value{
GoogleCloudAiplatformV1Port object
}
\description{
GoogleCloudAiplatformV1Port Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a network port in a container.
}
\concept{GoogleCloudAiplatformV1Port functions}
