% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1PredefinedSplit}
\alias{GoogleCloudAiplatformV1PredefinedSplit}
\title{GoogleCloudAiplatformV1PredefinedSplit Object}
\usage{
GoogleCloudAiplatformV1PredefinedSplit(key = NULL)
}
\arguments{
\item{key}{Required}
}
\value{
GoogleCloudAiplatformV1PredefinedSplit object
}
\description{
GoogleCloudAiplatformV1PredefinedSplit Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Assigns input data to training, validation, and test sets based on the value of a provided key. Supported only for tabular Datasets.
}
\concept{GoogleCloudAiplatformV1PredefinedSplit functions}
