% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1PredictRequest}
\alias{GoogleCloudAiplatformV1PredictRequest}
\title{GoogleCloudAiplatformV1PredictRequest Object}
\usage{
GoogleCloudAiplatformV1PredictRequest(instances = NULL, parameters = NULL)
}
\arguments{
\item{instances}{Required}

\item{parameters}{The parameters that govern the prediction}
}
\value{
GoogleCloudAiplatformV1PredictRequest object
}
\description{
GoogleCloudAiplatformV1PredictRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for PredictionService.Predict.
}
\seealso{
Other GoogleCloudAiplatformV1PredictRequest functions: 
\code{\link{projects.locations.endpoints.predict}()},
\code{\link{projects.locations.publishers.models.predict}()}
}
\concept{GoogleCloudAiplatformV1PredictRequest functions}
