% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1PredictRequestResponseLoggingConfig}
\alias{GoogleCloudAiplatformV1PredictRequestResponseLoggingConfig}
\title{GoogleCloudAiplatformV1PredictRequestResponseLoggingConfig Object}
\usage{
GoogleCloudAiplatformV1PredictRequestResponseLoggingConfig(
  samplingRate = NULL,
  bigqueryDestination = NULL,
  enabled = NULL
)
}
\arguments{
\item{samplingRate}{Percentage of requests to be logged, expressed as a fraction in range(0,1]}

\item{bigqueryDestination}{BigQuery table for logging}

\item{enabled}{If logging is enabled or not}
}
\value{
GoogleCloudAiplatformV1PredictRequestResponseLoggingConfig object
}
\description{
GoogleCloudAiplatformV1PredictRequestResponseLoggingConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration for logging request-response to a BigQuery table.
}
\concept{GoogleCloudAiplatformV1PredictRequestResponseLoggingConfig functions}
