% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1PredictResponse}
\alias{GoogleCloudAiplatformV1PredictResponse}
\title{GoogleCloudAiplatformV1PredictResponse Object}
\usage{
GoogleCloudAiplatformV1PredictResponse(
  deployedModelId = NULL,
  predictions = NULL
)
}
\arguments{
\item{deployedModelId}{ID of the Endpoint's DeployedModel that served this prediction}

\item{predictions}{The predictions that are the output of the predictions call}
}
\value{
GoogleCloudAiplatformV1PredictResponse object
}
\description{
GoogleCloudAiplatformV1PredictResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for PredictionService.Predict.
}
\concept{GoogleCloudAiplatformV1PredictResponse functions}
