% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1PredictSchemata}
\alias{GoogleCloudAiplatformV1PredictSchemata}
\title{GoogleCloudAiplatformV1PredictSchemata Object}
\usage{
GoogleCloudAiplatformV1PredictSchemata(
  parametersSchemaUri = NULL,
  predictionSchemaUri = NULL,
  instanceSchemaUri = NULL
)
}
\arguments{
\item{parametersSchemaUri}{Immutable}

\item{predictionSchemaUri}{Immutable}

\item{instanceSchemaUri}{Immutable}
}
\value{
GoogleCloudAiplatformV1PredictSchemata object
}
\description{
GoogleCloudAiplatformV1PredictSchemata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains the schemata used in Model's predictions and explanations via PredictionService.Predict, PredictionService.Explain and BatchPredictionJob.
}
\concept{GoogleCloudAiplatformV1PredictSchemata functions}
