% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1Presets}
\alias{GoogleCloudAiplatformV1Presets}
\title{GoogleCloudAiplatformV1Presets Object}
\usage{
GoogleCloudAiplatformV1Presets(query = NULL, modality = NULL)
}
\arguments{
\item{query}{Preset option controlling parameters for speed-precision trade-off when querying for examples}

\item{modality}{The modality of the uploaded model, which automatically configures the distance measurement and feature normalization for the underlying example index and queries}
}
\value{
GoogleCloudAiplatformV1Presets object
}
\description{
GoogleCloudAiplatformV1Presets Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Preset configuration for example-based explanations
}
\concept{GoogleCloudAiplatformV1Presets functions}
