% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1PrivateEndpoints}
\alias{GoogleCloudAiplatformV1PrivateEndpoints}
\title{GoogleCloudAiplatformV1PrivateEndpoints Object}
\usage{
GoogleCloudAiplatformV1PrivateEndpoints()
}
\value{
GoogleCloudAiplatformV1PrivateEndpoints object
}
\description{
GoogleCloudAiplatformV1PrivateEndpoints Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
PrivateEndpoints proto is used to provide paths for users to send requests privately. To send request via private service access, use predict_http_uri, explain_http_uri or health_http_uri. To send request via private service connect, use service_attachment.
}
\concept{GoogleCloudAiplatformV1PrivateEndpoints functions}
