% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1PrivateServiceConnectConfig}
\alias{GoogleCloudAiplatformV1PrivateServiceConnectConfig}
\title{GoogleCloudAiplatformV1PrivateServiceConnectConfig Object}
\usage{
GoogleCloudAiplatformV1PrivateServiceConnectConfig(
  projectAllowlist = NULL,
  enablePrivateServiceConnect = NULL
)
}
\arguments{
\item{projectAllowlist}{A list of Projects from which the forwarding rule will target the service attachment}

\item{enablePrivateServiceConnect}{Required}
}
\value{
GoogleCloudAiplatformV1PrivateServiceConnectConfig object
}
\description{
GoogleCloudAiplatformV1PrivateServiceConnectConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents configuration for private service connect.
}
\concept{GoogleCloudAiplatformV1PrivateServiceConnectConfig functions}
