% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1PublisherModel}
\alias{GoogleCloudAiplatformV1PublisherModel}
\title{GoogleCloudAiplatformV1PublisherModel Object}
\usage{
GoogleCloudAiplatformV1PublisherModel(
  openSourceCategory = NULL,
  frameworks = NULL,
  predictSchemata = NULL,
  supportedActions = NULL,
  launchStage = NULL
)
}
\arguments{
\item{openSourceCategory}{Required}

\item{frameworks}{Optional}

\item{predictSchemata}{Optional}

\item{supportedActions}{Optional}

\item{launchStage}{Optional}
}
\value{
GoogleCloudAiplatformV1PublisherModel object
}
\description{
GoogleCloudAiplatformV1PublisherModel Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Model Garden Publisher Model.
}
\concept{GoogleCloudAiplatformV1PublisherModel functions}
