% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1PublisherModelCallToAction}
\alias{GoogleCloudAiplatformV1PublisherModelCallToAction}
\title{GoogleCloudAiplatformV1PublisherModelCallToAction Object}
\usage{
GoogleCloudAiplatformV1PublisherModelCallToAction(
  openGenie = NULL,
  openGenerationAiStudio = NULL,
  openNotebook = NULL,
  requestAccess = NULL,
  deploy = NULL,
  openPromptTuningPipeline = NULL,
  viewRestApi = NULL,
  createApplication = NULL,
  openFineTuningPipeline = NULL
)
}
\arguments{
\item{openGenie}{Optional}

\item{openGenerationAiStudio}{Optional}

\item{openNotebook}{Optional}

\item{requestAccess}{Optional}

\item{deploy}{Optional}

\item{openPromptTuningPipeline}{Optional}

\item{viewRestApi}{Optional}

\item{createApplication}{Optional}

\item{openFineTuningPipeline}{Optional}
}
\value{
GoogleCloudAiplatformV1PublisherModelCallToAction object
}
\description{
GoogleCloudAiplatformV1PublisherModelCallToAction Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Actions could take on this Publisher Model.
}
\concept{GoogleCloudAiplatformV1PublisherModelCallToAction functions}
