% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1PublisherModelCallToActionDeploy}
\alias{GoogleCloudAiplatformV1PublisherModelCallToActionDeploy}
\title{GoogleCloudAiplatformV1PublisherModelCallToActionDeploy Object}
\usage{
GoogleCloudAiplatformV1PublisherModelCallToActionDeploy(
  artifactUri = NULL,
  dedicatedResources = NULL,
  modelDisplayName = NULL,
  largeModelReference = NULL,
  title = NULL,
  containerSpec = NULL,
  sharedResources = NULL,
  automaticResources = NULL
)
}
\arguments{
\item{artifactUri}{Optional}

\item{dedicatedResources}{A description of resources that are dedicated to the DeployedModel, and that need a higher degree of manual configuration}

\item{modelDisplayName}{Optional}

\item{largeModelReference}{Optional}

\item{title}{Required}

\item{containerSpec}{Optional}

\item{sharedResources}{The resource name of the shared DeploymentResourcePool to deploy on}

\item{automaticResources}{A description of resources that to large degree are decided by Vertex AI, and require only a modest additional configuration}
}
\value{
GoogleCloudAiplatformV1PublisherModelCallToActionDeploy object
}
\description{
GoogleCloudAiplatformV1PublisherModelCallToActionDeploy Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Model metadata that is needed for UploadModel or DeployModel/CreateEndpoint requests.
}
\concept{GoogleCloudAiplatformV1PublisherModelCallToActionDeploy functions}
