% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1PublisherModelCallToActionRegionalResourceReferences}
\alias{GoogleCloudAiplatformV1PublisherModelCallToActionRegionalResourceReferences}
\title{GoogleCloudAiplatformV1PublisherModelCallToActionRegionalResourceReferences Object}
\usage{
GoogleCloudAiplatformV1PublisherModelCallToActionRegionalResourceReferences(
  GoogleCloudAiplatformV1PublisherModelCallToActionRegionalResourceReferences.references
    = NULL,
  references = NULL,
  title = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1PublisherModelCallToActionRegionalResourceReferences.references}{The \link{GoogleCloudAiplatformV1PublisherModelCallToActionRegionalResourceReferences.references} object or list of objects}

\item{references}{Required}

\item{title}{Required}
}
\value{
GoogleCloudAiplatformV1PublisherModelCallToActionRegionalResourceReferences object
}
\description{
GoogleCloudAiplatformV1PublisherModelCallToActionRegionalResourceReferences Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The regional resource name or the URI. Key is region, e.g., us-central1, europe-west2, global, etc..
}
\seealso{
Other GoogleCloudAiplatformV1PublisherModelCallToActionRegionalResourceReferences functions: 
\code{\link{GoogleCloudAiplatformV1PublisherModelCallToActionRegionalResourceReferences.references}()}
}
\concept{GoogleCloudAiplatformV1PublisherModelCallToActionRegionalResourceReferences functions}
