% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1PublisherModelCallToActionRegionalResourceReferences.references}
\alias{GoogleCloudAiplatformV1PublisherModelCallToActionRegionalResourceReferences.references}
\title{GoogleCloudAiplatformV1PublisherModelCallToActionRegionalResourceReferences.references Object}
\usage{
GoogleCloudAiplatformV1PublisherModelCallToActionRegionalResourceReferences.references(
  
)
}
\value{
GoogleCloudAiplatformV1PublisherModelCallToActionRegionalResourceReferences.references object
}
\description{
GoogleCloudAiplatformV1PublisherModelCallToActionRegionalResourceReferences.references Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Required.
}
\seealso{
Other GoogleCloudAiplatformV1PublisherModelCallToActionRegionalResourceReferences functions: 
\code{\link{GoogleCloudAiplatformV1PublisherModelCallToActionRegionalResourceReferences}()}
}
\concept{GoogleCloudAiplatformV1PublisherModelCallToActionRegionalResourceReferences functions}
