% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1PublisherModelCallToActionViewRestApi}
\alias{GoogleCloudAiplatformV1PublisherModelCallToActionViewRestApi}
\title{GoogleCloudAiplatformV1PublisherModelCallToActionViewRestApi Object}
\usage{
GoogleCloudAiplatformV1PublisherModelCallToActionViewRestApi(
  documentations = NULL,
  title = NULL
)
}
\arguments{
\item{documentations}{Required}

\item{title}{Required}
}
\value{
GoogleCloudAiplatformV1PublisherModelCallToActionViewRestApi object
}
\description{
GoogleCloudAiplatformV1PublisherModelCallToActionViewRestApi Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Rest API docs.
}
\concept{GoogleCloudAiplatformV1PublisherModelCallToActionViewRestApi functions}
