% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1PublisherModelDocumentation}
\alias{GoogleCloudAiplatformV1PublisherModelDocumentation}
\title{GoogleCloudAiplatformV1PublisherModelDocumentation Object}
\usage{
GoogleCloudAiplatformV1PublisherModelDocumentation(
  title = NULL,
  content = NULL
)
}
\arguments{
\item{title}{Required}

\item{content}{Required}
}
\value{
GoogleCloudAiplatformV1PublisherModelDocumentation object
}
\description{
GoogleCloudAiplatformV1PublisherModelDocumentation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A named piece of documentation.
}
\concept{GoogleCloudAiplatformV1PublisherModelDocumentation functions}
