% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1PublisherModelResourceReference}
\alias{GoogleCloudAiplatformV1PublisherModelResourceReference}
\title{GoogleCloudAiplatformV1PublisherModelResourceReference Object}
\usage{
GoogleCloudAiplatformV1PublisherModelResourceReference(
  resourceName = NULL,
  uri = NULL
)
}
\arguments{
\item{resourceName}{The resource name of the Google Cloud resource}

\item{uri}{The URI of the resource}
}
\value{
GoogleCloudAiplatformV1PublisherModelResourceReference object
}
\description{
GoogleCloudAiplatformV1PublisherModelResourceReference Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Reference to a resource.
}
\concept{GoogleCloudAiplatformV1PublisherModelResourceReference functions}
