% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1PurgeArtifactsMetadata}
\alias{GoogleCloudAiplatformV1PurgeArtifactsMetadata}
\title{GoogleCloudAiplatformV1PurgeArtifactsMetadata Object}
\usage{
GoogleCloudAiplatformV1PurgeArtifactsMetadata(genericMetadata = NULL)
}
\arguments{
\item{genericMetadata}{Operation metadata for purging Artifacts}
}
\value{
GoogleCloudAiplatformV1PurgeArtifactsMetadata object
}
\description{
GoogleCloudAiplatformV1PurgeArtifactsMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details of operations that perform MetadataService.PurgeArtifacts.
}
\concept{GoogleCloudAiplatformV1PurgeArtifactsMetadata functions}
