% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1PurgeArtifactsRequest}
\alias{GoogleCloudAiplatformV1PurgeArtifactsRequest}
\title{GoogleCloudAiplatformV1PurgeArtifactsRequest Object}
\usage{
GoogleCloudAiplatformV1PurgeArtifactsRequest(force = NULL, filter = NULL)
}
\arguments{
\item{force}{Optional}

\item{filter}{Required}
}
\value{
GoogleCloudAiplatformV1PurgeArtifactsRequest object
}
\description{
GoogleCloudAiplatformV1PurgeArtifactsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for MetadataService.PurgeArtifacts.
}
\seealso{
Other GoogleCloudAiplatformV1PurgeArtifactsRequest functions: 
\code{\link{projects.locations.metadataStores.artifacts.purge}()}
}
\concept{GoogleCloudAiplatformV1PurgeArtifactsRequest functions}
