% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1PurgeArtifactsResponse}
\alias{GoogleCloudAiplatformV1PurgeArtifactsResponse}
\title{GoogleCloudAiplatformV1PurgeArtifactsResponse Object}
\usage{
GoogleCloudAiplatformV1PurgeArtifactsResponse(
  purgeSample = NULL,
  purgeCount = NULL
)
}
\arguments{
\item{purgeSample}{A sample of the Artifact names that will be deleted}

\item{purgeCount}{The number of Artifacts that this request deleted (or, if \code{force} is false, the number of Artifacts that will be deleted)}
}
\value{
GoogleCloudAiplatformV1PurgeArtifactsResponse object
}
\description{
GoogleCloudAiplatformV1PurgeArtifactsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for MetadataService.PurgeArtifacts.
}
\concept{GoogleCloudAiplatformV1PurgeArtifactsResponse functions}
