% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1PurgeContextsMetadata}
\alias{GoogleCloudAiplatformV1PurgeContextsMetadata}
\title{GoogleCloudAiplatformV1PurgeContextsMetadata Object}
\usage{
GoogleCloudAiplatformV1PurgeContextsMetadata(genericMetadata = NULL)
}
\arguments{
\item{genericMetadata}{Operation metadata for purging Contexts}
}
\value{
GoogleCloudAiplatformV1PurgeContextsMetadata object
}
\description{
GoogleCloudAiplatformV1PurgeContextsMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details of operations that perform MetadataService.PurgeContexts.
}
\concept{GoogleCloudAiplatformV1PurgeContextsMetadata functions}
