% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1PurgeContextsRequest}
\alias{GoogleCloudAiplatformV1PurgeContextsRequest}
\title{GoogleCloudAiplatformV1PurgeContextsRequest Object}
\usage{
GoogleCloudAiplatformV1PurgeContextsRequest(filter = NULL, force = NULL)
}
\arguments{
\item{filter}{Required}

\item{force}{Optional}
}
\value{
GoogleCloudAiplatformV1PurgeContextsRequest object
}
\description{
GoogleCloudAiplatformV1PurgeContextsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for MetadataService.PurgeContexts.
}
\seealso{
Other GoogleCloudAiplatformV1PurgeContextsRequest functions: 
\code{\link{projects.locations.metadataStores.contexts.purge}()}
}
\concept{GoogleCloudAiplatformV1PurgeContextsRequest functions}
