% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1PurgeContextsResponse}
\alias{GoogleCloudAiplatformV1PurgeContextsResponse}
\title{GoogleCloudAiplatformV1PurgeContextsResponse Object}
\usage{
GoogleCloudAiplatformV1PurgeContextsResponse(
  purgeSample = NULL,
  purgeCount = NULL
)
}
\arguments{
\item{purgeSample}{A sample of the Context names that will be deleted}

\item{purgeCount}{The number of Contexts that this request deleted (or, if \code{force} is false, the number of Contexts that will be deleted)}
}
\value{
GoogleCloudAiplatformV1PurgeContextsResponse object
}
\description{
GoogleCloudAiplatformV1PurgeContextsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for MetadataService.PurgeContexts.
}
\concept{GoogleCloudAiplatformV1PurgeContextsResponse functions}
