% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1PurgeExecutionsMetadata}
\alias{GoogleCloudAiplatformV1PurgeExecutionsMetadata}
\title{GoogleCloudAiplatformV1PurgeExecutionsMetadata Object}
\usage{
GoogleCloudAiplatformV1PurgeExecutionsMetadata(genericMetadata = NULL)
}
\arguments{
\item{genericMetadata}{Operation metadata for purging Executions}
}
\value{
GoogleCloudAiplatformV1PurgeExecutionsMetadata object
}
\description{
GoogleCloudAiplatformV1PurgeExecutionsMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details of operations that perform MetadataService.PurgeExecutions.
}
\concept{GoogleCloudAiplatformV1PurgeExecutionsMetadata functions}
