% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1PurgeExecutionsResponse}
\alias{GoogleCloudAiplatformV1PurgeExecutionsResponse}
\title{GoogleCloudAiplatformV1PurgeExecutionsResponse Object}
\usage{
GoogleCloudAiplatformV1PurgeExecutionsResponse(
  purgeCount = NULL,
  purgeSample = NULL
)
}
\arguments{
\item{purgeCount}{The number of Executions that this request deleted (or, if \code{force} is false, the number of Executions that will be deleted)}

\item{purgeSample}{A sample of the Execution names that will be deleted}
}
\value{
GoogleCloudAiplatformV1PurgeExecutionsResponse object
}
\description{
GoogleCloudAiplatformV1PurgeExecutionsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for MetadataService.PurgeExecutions.
}
\concept{GoogleCloudAiplatformV1PurgeExecutionsResponse functions}
