% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1PythonPackageSpec}
\alias{GoogleCloudAiplatformV1PythonPackageSpec}
\title{GoogleCloudAiplatformV1PythonPackageSpec Object}
\usage{
GoogleCloudAiplatformV1PythonPackageSpec(
  pythonModule = NULL,
  packageUris = NULL,
  executorImageUri = NULL,
  args = NULL,
  env = NULL
)
}
\arguments{
\item{pythonModule}{Required}

\item{packageUris}{Required}

\item{executorImageUri}{Required}

\item{args}{Command line arguments to be passed to the Python task}

\item{env}{Environment variables to be passed to the python module}
}
\value{
GoogleCloudAiplatformV1PythonPackageSpec object
}
\description{
GoogleCloudAiplatformV1PythonPackageSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The spec of a Python packaged code.
}
\concept{GoogleCloudAiplatformV1PythonPackageSpec functions}
