% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1RawPredictRequest}
\alias{GoogleCloudAiplatformV1RawPredictRequest}
\title{GoogleCloudAiplatformV1RawPredictRequest Object}
\usage{
GoogleCloudAiplatformV1RawPredictRequest(httpBody = NULL)
}
\arguments{
\item{httpBody}{The prediction input}
}
\value{
GoogleCloudAiplatformV1RawPredictRequest object
}
\description{
GoogleCloudAiplatformV1RawPredictRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for PredictionService.RawPredict.
}
\seealso{
Other GoogleCloudAiplatformV1RawPredictRequest functions: 
\code{\link{projects.locations.endpoints.rawPredict}()},
\code{\link{projects.locations.publishers.models.rawPredict}()}
}
\concept{GoogleCloudAiplatformV1RawPredictRequest functions}
