% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ReadFeatureValuesRequest}
\alias{GoogleCloudAiplatformV1ReadFeatureValuesRequest}
\title{GoogleCloudAiplatformV1ReadFeatureValuesRequest Object}
\usage{
GoogleCloudAiplatformV1ReadFeatureValuesRequest(
  entityId = NULL,
  featureSelector = NULL
)
}
\arguments{
\item{entityId}{Required}

\item{featureSelector}{Required}
}
\value{
GoogleCloudAiplatformV1ReadFeatureValuesRequest object
}
\description{
GoogleCloudAiplatformV1ReadFeatureValuesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for FeaturestoreOnlineServingService.ReadFeatureValues.
}
\seealso{
Other GoogleCloudAiplatformV1ReadFeatureValuesRequest functions: 
\code{\link{projects.locations.featurestores.entityTypes.readFeatureValues}()}
}
\concept{GoogleCloudAiplatformV1ReadFeatureValuesRequest functions}
