% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ReadFeatureValuesResponse}
\alias{GoogleCloudAiplatformV1ReadFeatureValuesResponse}
\title{GoogleCloudAiplatformV1ReadFeatureValuesResponse Object}
\usage{
GoogleCloudAiplatformV1ReadFeatureValuesResponse(
  header = NULL,
  entityView = NULL
)
}
\arguments{
\item{header}{Response header}

\item{entityView}{Entity view with Feature values}
}
\value{
GoogleCloudAiplatformV1ReadFeatureValuesResponse object
}
\description{
GoogleCloudAiplatformV1ReadFeatureValuesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for FeaturestoreOnlineServingService.ReadFeatureValues.
}
\concept{GoogleCloudAiplatformV1ReadFeatureValuesResponse functions}
