% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ReadFeatureValuesResponseEntityView}
\alias{GoogleCloudAiplatformV1ReadFeatureValuesResponseEntityView}
\title{GoogleCloudAiplatformV1ReadFeatureValuesResponseEntityView Object}
\usage{
GoogleCloudAiplatformV1ReadFeatureValuesResponseEntityView(
  entityId = NULL,
  data = NULL
)
}
\arguments{
\item{entityId}{ID of the requested entity}

\item{data}{Each piece of data holds the k requested values for one requested Feature}
}
\value{
GoogleCloudAiplatformV1ReadFeatureValuesResponseEntityView object
}
\description{
GoogleCloudAiplatformV1ReadFeatureValuesResponseEntityView Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Entity view with Feature values.
}
\concept{GoogleCloudAiplatformV1ReadFeatureValuesResponseEntityView functions}
