% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ReadFeatureValuesResponseEntityViewData}
\alias{GoogleCloudAiplatformV1ReadFeatureValuesResponseEntityViewData}
\title{GoogleCloudAiplatformV1ReadFeatureValuesResponseEntityViewData Object}
\usage{
GoogleCloudAiplatformV1ReadFeatureValuesResponseEntityViewData(
  values = NULL,
  value = NULL
)
}
\arguments{
\item{values}{Feature values list if values, successive in time, are requested}

\item{value}{Feature value if a single value is requested}
}
\value{
GoogleCloudAiplatformV1ReadFeatureValuesResponseEntityViewData object
}
\description{
GoogleCloudAiplatformV1ReadFeatureValuesResponseEntityViewData Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Container to hold value(s), successive in time, for one Feature from the request.
}
\concept{GoogleCloudAiplatformV1ReadFeatureValuesResponseEntityViewData functions}
