% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ReadFeatureValuesResponseHeader}
\alias{GoogleCloudAiplatformV1ReadFeatureValuesResponseHeader}
\title{GoogleCloudAiplatformV1ReadFeatureValuesResponseHeader Object}
\usage{
GoogleCloudAiplatformV1ReadFeatureValuesResponseHeader(
  entityType = NULL,
  featureDescriptors = NULL
)
}
\arguments{
\item{entityType}{The resource name of the EntityType from the ReadFeatureValuesRequest}

\item{featureDescriptors}{List of Feature metadata corresponding to each piece of ReadFeatureValuesResponse}
}
\value{
GoogleCloudAiplatformV1ReadFeatureValuesResponseHeader object
}
\description{
GoogleCloudAiplatformV1ReadFeatureValuesResponseHeader Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response header with metadata for the requested ReadFeatureValuesRequest.entity_type and Features.
}
\concept{GoogleCloudAiplatformV1ReadFeatureValuesResponseHeader functions}
