% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ReadIndexDatapointsRequest}
\alias{GoogleCloudAiplatformV1ReadIndexDatapointsRequest}
\title{GoogleCloudAiplatformV1ReadIndexDatapointsRequest Object}
\usage{
GoogleCloudAiplatformV1ReadIndexDatapointsRequest(
  deployedIndexId = NULL,
  ids = NULL
)
}
\arguments{
\item{deployedIndexId}{The ID of the DeployedIndex that will serve the request}

\item{ids}{IDs of the datapoints to be searched for}
}
\value{
GoogleCloudAiplatformV1ReadIndexDatapointsRequest object
}
\description{
GoogleCloudAiplatformV1ReadIndexDatapointsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request message for MatchService.ReadIndexDatapoints.
}
\seealso{
Other GoogleCloudAiplatformV1ReadIndexDatapointsRequest functions: 
\code{\link{projects.locations.indexEndpoints.readIndexDatapoints}()}
}
\concept{GoogleCloudAiplatformV1ReadIndexDatapointsRequest functions}
