% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ReadIndexDatapointsResponse}
\alias{GoogleCloudAiplatformV1ReadIndexDatapointsResponse}
\title{GoogleCloudAiplatformV1ReadIndexDatapointsResponse Object}
\usage{
GoogleCloudAiplatformV1ReadIndexDatapointsResponse(datapoints = NULL)
}
\arguments{
\item{datapoints}{The result list of datapoints}
}
\value{
GoogleCloudAiplatformV1ReadIndexDatapointsResponse object
}
\description{
GoogleCloudAiplatformV1ReadIndexDatapointsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response message for MatchService.ReadIndexDatapoints.
}
\concept{GoogleCloudAiplatformV1ReadIndexDatapointsResponse functions}
