% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ReadTensorboardBlobDataResponse}
\alias{GoogleCloudAiplatformV1ReadTensorboardBlobDataResponse}
\title{GoogleCloudAiplatformV1ReadTensorboardBlobDataResponse Object}
\usage{
GoogleCloudAiplatformV1ReadTensorboardBlobDataResponse(blobs = NULL)
}
\arguments{
\item{blobs}{Blob messages containing blob bytes}
}
\value{
GoogleCloudAiplatformV1ReadTensorboardBlobDataResponse object
}
\description{
GoogleCloudAiplatformV1ReadTensorboardBlobDataResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for TensorboardService.ReadTensorboardBlobData.
}
\concept{GoogleCloudAiplatformV1ReadTensorboardBlobDataResponse functions}
