% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ReadTensorboardSizeResponse}
\alias{GoogleCloudAiplatformV1ReadTensorboardSizeResponse}
\title{GoogleCloudAiplatformV1ReadTensorboardSizeResponse Object}
\usage{
GoogleCloudAiplatformV1ReadTensorboardSizeResponse(storageSizeByte = NULL)
}
\arguments{
\item{storageSizeByte}{Payload storage size for the TensorBoard}
}
\value{
GoogleCloudAiplatformV1ReadTensorboardSizeResponse object
}
\description{
GoogleCloudAiplatformV1ReadTensorboardSizeResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for TensorboardService.ReadTensorboardSize.
}
\concept{GoogleCloudAiplatformV1ReadTensorboardSizeResponse functions}
