% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ReadTensorboardTimeSeriesDataResponse}
\alias{GoogleCloudAiplatformV1ReadTensorboardTimeSeriesDataResponse}
\title{GoogleCloudAiplatformV1ReadTensorboardTimeSeriesDataResponse Object}
\usage{
GoogleCloudAiplatformV1ReadTensorboardTimeSeriesDataResponse(
  timeSeriesData = NULL
)
}
\arguments{
\item{timeSeriesData}{The returned time series data}
}
\value{
GoogleCloudAiplatformV1ReadTensorboardTimeSeriesDataResponse object
}
\description{
GoogleCloudAiplatformV1ReadTensorboardTimeSeriesDataResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for TensorboardService.ReadTensorboardTimeSeriesData.
}
\concept{GoogleCloudAiplatformV1ReadTensorboardTimeSeriesDataResponse functions}
