% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ReadTensorboardUsageResponse}
\alias{GoogleCloudAiplatformV1ReadTensorboardUsageResponse}
\title{GoogleCloudAiplatformV1ReadTensorboardUsageResponse Object}
\usage{
GoogleCloudAiplatformV1ReadTensorboardUsageResponse(
  GoogleCloudAiplatformV1ReadTensorboardUsageResponse.monthlyUsageData = NULL,
  monthlyUsageData = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1ReadTensorboardUsageResponse.monthlyUsageData}{The \link{GoogleCloudAiplatformV1ReadTensorboardUsageResponse.monthlyUsageData} object or list of objects}

\item{monthlyUsageData}{Maps year-month (YYYYMM) string to per month usage data}
}
\value{
GoogleCloudAiplatformV1ReadTensorboardUsageResponse object
}
\description{
GoogleCloudAiplatformV1ReadTensorboardUsageResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for TensorboardService.ReadTensorboardUsage.
}
\seealso{
Other GoogleCloudAiplatformV1ReadTensorboardUsageResponse functions: 
\code{\link{GoogleCloudAiplatformV1ReadTensorboardUsageResponse.monthlyUsageData}()}
}
\concept{GoogleCloudAiplatformV1ReadTensorboardUsageResponse functions}
