% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ReadTensorboardUsageResponsePerMonthUsageData}
\alias{GoogleCloudAiplatformV1ReadTensorboardUsageResponsePerMonthUsageData}
\title{GoogleCloudAiplatformV1ReadTensorboardUsageResponsePerMonthUsageData Object}
\usage{
GoogleCloudAiplatformV1ReadTensorboardUsageResponsePerMonthUsageData(
  userUsageData = NULL
)
}
\arguments{
\item{userUsageData}{Usage data for each user in the given month}
}
\value{
GoogleCloudAiplatformV1ReadTensorboardUsageResponsePerMonthUsageData object
}
\description{
GoogleCloudAiplatformV1ReadTensorboardUsageResponsePerMonthUsageData Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Per month usage data
}
\concept{GoogleCloudAiplatformV1ReadTensorboardUsageResponsePerMonthUsageData functions}
