% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ReadTensorboardUsageResponsePerUserUsageData}
\alias{GoogleCloudAiplatformV1ReadTensorboardUsageResponsePerUserUsageData}
\title{GoogleCloudAiplatformV1ReadTensorboardUsageResponsePerUserUsageData Object}
\usage{
GoogleCloudAiplatformV1ReadTensorboardUsageResponsePerUserUsageData(
  viewCount = NULL,
  username = NULL
)
}
\arguments{
\item{viewCount}{Number of times the user has read data within the Tensorboard}

\item{username}{User's username}
}
\value{
GoogleCloudAiplatformV1ReadTensorboardUsageResponsePerUserUsageData object
}
\description{
GoogleCloudAiplatformV1ReadTensorboardUsageResponsePerUserUsageData Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Per user usage data.
}
\concept{GoogleCloudAiplatformV1ReadTensorboardUsageResponsePerUserUsageData functions}
