% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1RemoveContextChildrenRequest}
\alias{GoogleCloudAiplatformV1RemoveContextChildrenRequest}
\title{GoogleCloudAiplatformV1RemoveContextChildrenRequest Object}
\usage{
GoogleCloudAiplatformV1RemoveContextChildrenRequest(childContexts = NULL)
}
\arguments{
\item{childContexts}{The resource names of the child Contexts}
}
\value{
GoogleCloudAiplatformV1RemoveContextChildrenRequest object
}
\description{
GoogleCloudAiplatformV1RemoveContextChildrenRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for MetadataService.DeleteContextChildrenRequest.
}
\seealso{
Other GoogleCloudAiplatformV1RemoveContextChildrenRequest functions: 
\code{\link{projects.locations.metadataStores.contexts.removeContextChildren}()}
}
\concept{GoogleCloudAiplatformV1RemoveContextChildrenRequest functions}
