% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1RemoveDatapointsRequest}
\alias{GoogleCloudAiplatformV1RemoveDatapointsRequest}
\title{GoogleCloudAiplatformV1RemoveDatapointsRequest Object}
\usage{
GoogleCloudAiplatformV1RemoveDatapointsRequest(datapointIds = NULL)
}
\arguments{
\item{datapointIds}{A list of datapoint ids to be deleted}
}
\value{
GoogleCloudAiplatformV1RemoveDatapointsRequest object
}
\description{
GoogleCloudAiplatformV1RemoveDatapointsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for IndexService.RemoveDatapoints
}
\seealso{
Other GoogleCloudAiplatformV1RemoveDatapointsRequest functions: 
\code{\link{projects.locations.indexes.removeDatapoints}()}
}
\concept{GoogleCloudAiplatformV1RemoveDatapointsRequest functions}
