% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ResourcesConsumed}
\alias{GoogleCloudAiplatformV1ResourcesConsumed}
\title{GoogleCloudAiplatformV1ResourcesConsumed Object}
\usage{
GoogleCloudAiplatformV1ResourcesConsumed()
}
\value{
GoogleCloudAiplatformV1ResourcesConsumed object
}
\description{
GoogleCloudAiplatformV1ResourcesConsumed Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Statistics information about resource consumption.
}
\concept{GoogleCloudAiplatformV1ResourcesConsumed functions}
