% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1RestoreDatasetVersionOperationMetadata}
\alias{GoogleCloudAiplatformV1RestoreDatasetVersionOperationMetadata}
\title{GoogleCloudAiplatformV1RestoreDatasetVersionOperationMetadata Object}
\usage{
GoogleCloudAiplatformV1RestoreDatasetVersionOperationMetadata(
  genericMetadata = NULL
)
}
\arguments{
\item{genericMetadata}{The common part of the operation metadata}
}
\value{
GoogleCloudAiplatformV1RestoreDatasetVersionOperationMetadata object
}
\description{
GoogleCloudAiplatformV1RestoreDatasetVersionOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Runtime operation information for DatasetService.RestoreDatasetVersion.
}
\concept{GoogleCloudAiplatformV1RestoreDatasetVersionOperationMetadata functions}
