% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ResumeModelDeploymentMonitoringJobRequest}
\alias{GoogleCloudAiplatformV1ResumeModelDeploymentMonitoringJobRequest}
\title{GoogleCloudAiplatformV1ResumeModelDeploymentMonitoringJobRequest Object}
\usage{
GoogleCloudAiplatformV1ResumeModelDeploymentMonitoringJobRequest()
}
\value{
GoogleCloudAiplatformV1ResumeModelDeploymentMonitoringJobRequest object
}
\description{
GoogleCloudAiplatformV1ResumeModelDeploymentMonitoringJobRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for JobService.ResumeModelDeploymentMonitoringJob.
}
\seealso{
Other GoogleCloudAiplatformV1ResumeModelDeploymentMonitoringJobRequest functions: 
\code{\link{projects.locations.modelDeploymentMonitoringJobs.resume}()}
}
\concept{GoogleCloudAiplatformV1ResumeModelDeploymentMonitoringJobRequest functions}
