% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SampleConfig}
\alias{GoogleCloudAiplatformV1SampleConfig}
\title{GoogleCloudAiplatformV1SampleConfig Object}
\usage{
GoogleCloudAiplatformV1SampleConfig(
  sampleStrategy = NULL,
  initialBatchSamplePercentage = NULL,
  followingBatchSamplePercentage = NULL
)
}
\arguments{
\item{sampleStrategy}{Field to choose sampling strategy}

\item{initialBatchSamplePercentage}{The percentage of data needed to be labeled in the first batch}

\item{followingBatchSamplePercentage}{The percentage of data needed to be labeled in each following batch (except the first batch)}
}
\value{
GoogleCloudAiplatformV1SampleConfig object
}
\description{
GoogleCloudAiplatformV1SampleConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Active learning data sampling config. For every active learning labeling iteration, it will select a batch of data based on the sampling strategy.
}
\concept{GoogleCloudAiplatformV1SampleConfig functions}
