% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SamplingStrategy}
\alias{GoogleCloudAiplatformV1SamplingStrategy}
\title{GoogleCloudAiplatformV1SamplingStrategy Object}
\usage{
GoogleCloudAiplatformV1SamplingStrategy(randomSampleConfig = NULL)
}
\arguments{
\item{randomSampleConfig}{Random sample config}
}
\value{
GoogleCloudAiplatformV1SamplingStrategy object
}
\description{
GoogleCloudAiplatformV1SamplingStrategy Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Sampling Strategy for logging, can be for both training and prediction dataset.
}
\concept{GoogleCloudAiplatformV1SamplingStrategy functions}
