% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SavedQuery}
\alias{GoogleCloudAiplatformV1SavedQuery}
\title{GoogleCloudAiplatformV1SavedQuery Object}
\usage{
GoogleCloudAiplatformV1SavedQuery(
  metadata = NULL,
  displayName = NULL,
  problemType = NULL,
  etag = NULL
)
}
\arguments{
\item{metadata}{Some additional information about the SavedQuery}

\item{displayName}{Required}

\item{problemType}{Required}

\item{etag}{Used to perform a consistent read-modify-write update}
}
\value{
GoogleCloudAiplatformV1SavedQuery object
}
\description{
GoogleCloudAiplatformV1SavedQuery Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A SavedQuery is a view of the dataset. It references a subset of annotations by problem type and filters.
}
\concept{GoogleCloudAiplatformV1SavedQuery functions}
