% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1Scalar}
\alias{GoogleCloudAiplatformV1Scalar}
\title{GoogleCloudAiplatformV1Scalar Object}
\usage{
GoogleCloudAiplatformV1Scalar(value = NULL)
}
\arguments{
\item{value}{Value of the point at this step / timestamp}
}
\value{
GoogleCloudAiplatformV1Scalar object
}
\description{
GoogleCloudAiplatformV1Scalar Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
One point viewable on a scalar metric plot.
}
\concept{GoogleCloudAiplatformV1Scalar functions}
