% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1Schedule}
\alias{GoogleCloudAiplatformV1Schedule}
\title{GoogleCloudAiplatformV1Schedule Object}
\usage{
GoogleCloudAiplatformV1Schedule(
  cron = NULL,
  displayName = NULL,
  createPipelineJobRequest = NULL,
  endTime = NULL,
  allowQueueing = NULL,
  maxConcurrentRunCount = NULL,
  name = NULL,
  maxRunCount = NULL,
  startTime = NULL
)
}
\arguments{
\item{cron}{Cron schedule (https://en}

\item{displayName}{Required}

\item{createPipelineJobRequest}{Request for PipelineService}

\item{endTime}{Optional}

\item{allowQueueing}{Optional}

\item{maxConcurrentRunCount}{Required}

\item{name}{Immutable}

\item{maxRunCount}{Optional}

\item{startTime}{Optional}
}
\value{
GoogleCloudAiplatformV1Schedule object
}
\description{
GoogleCloudAiplatformV1Schedule Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An instance of a Schedule periodically schedules runs to make API calls based on user specified time specification and API request type.
}
\seealso{
Other GoogleCloudAiplatformV1Schedule functions: 
\code{\link{projects.locations.schedules.create}()},
\code{\link{projects.locations.schedules.patch}()}
}
\concept{GoogleCloudAiplatformV1Schedule functions}
