% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ScheduleRunResponse}
\alias{GoogleCloudAiplatformV1ScheduleRunResponse}
\title{GoogleCloudAiplatformV1ScheduleRunResponse Object}
\usage{
GoogleCloudAiplatformV1ScheduleRunResponse(
  scheduledRunTime = NULL,
  runResponse = NULL
)
}
\arguments{
\item{scheduledRunTime}{The scheduled run time based on the user-specified schedule}

\item{runResponse}{The response of the scheduled run}
}
\value{
GoogleCloudAiplatformV1ScheduleRunResponse object
}
\description{
GoogleCloudAiplatformV1ScheduleRunResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Status of a scheduled run.
}
\concept{GoogleCloudAiplatformV1ScheduleRunResponse functions}
