% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1Scheduling}
\alias{GoogleCloudAiplatformV1Scheduling}
\title{GoogleCloudAiplatformV1Scheduling Object}
\usage{
GoogleCloudAiplatformV1Scheduling(
  restartJobOnWorkerRestart = NULL,
  timeout = NULL
)
}
\arguments{
\item{restartJobOnWorkerRestart}{Restarts the entire CustomJob if a worker gets restarted}

\item{timeout}{The maximum job running time}
}
\value{
GoogleCloudAiplatformV1Scheduling object
}
\description{
GoogleCloudAiplatformV1Scheduling Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
All parameters related to queuing and scheduling of custom jobs.
}
\concept{GoogleCloudAiplatformV1Scheduling functions}
