% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaAnnotationSpecColor}
\alias{GoogleCloudAiplatformV1SchemaAnnotationSpecColor}
\title{GoogleCloudAiplatformV1SchemaAnnotationSpecColor Object}
\usage{
GoogleCloudAiplatformV1SchemaAnnotationSpecColor(
  color = NULL,
  displayName = NULL,
  id = NULL
)
}
\arguments{
\item{color}{The color of the AnnotationSpec in a segmentation mask}

\item{displayName}{The display name of the AnnotationSpec represented by the color in the segmentation mask}

\item{id}{The ID of the AnnotationSpec represented by the color in the segmentation mask}
}
\value{
GoogleCloudAiplatformV1SchemaAnnotationSpecColor object
}
\description{
GoogleCloudAiplatformV1SchemaAnnotationSpecColor Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An entry of mapping between color and AnnotationSpec. The mapping is used in segmentation mask.
}
\concept{GoogleCloudAiplatformV1SchemaAnnotationSpecColor functions}
