% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaImageBoundingBoxAnnotation}
\alias{GoogleCloudAiplatformV1SchemaImageBoundingBoxAnnotation}
\title{GoogleCloudAiplatformV1SchemaImageBoundingBoxAnnotation Object}
\usage{
GoogleCloudAiplatformV1SchemaImageBoundingBoxAnnotation(
  xMin = NULL,
  yMax = NULL,
  annotationSpecId = NULL,
  xMax = NULL,
  displayName = NULL,
  yMin = NULL
)
}
\arguments{
\item{xMin}{The leftmost coordinate of the bounding box}

\item{yMax}{The bottommost coordinate of the bounding box}

\item{annotationSpecId}{The resource Id of the AnnotationSpec that this Annotation pertains to}

\item{xMax}{The rightmost coordinate of the bounding box}

\item{displayName}{The display name of the AnnotationSpec that this Annotation pertains to}

\item{yMin}{The topmost coordinate of the bounding box}
}
\value{
GoogleCloudAiplatformV1SchemaImageBoundingBoxAnnotation object
}
\description{
GoogleCloudAiplatformV1SchemaImageBoundingBoxAnnotation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Annotation details specific to image object detection.
}
\concept{GoogleCloudAiplatformV1SchemaImageBoundingBoxAnnotation functions}
