% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaImageClassificationAnnotation}
\alias{GoogleCloudAiplatformV1SchemaImageClassificationAnnotation}
\title{GoogleCloudAiplatformV1SchemaImageClassificationAnnotation Object}
\usage{
GoogleCloudAiplatformV1SchemaImageClassificationAnnotation(
  displayName = NULL,
  annotationSpecId = NULL
)
}
\arguments{
\item{displayName}{The display name of the AnnotationSpec that this Annotation pertains to}

\item{annotationSpecId}{The resource Id of the AnnotationSpec that this Annotation pertains to}
}
\value{
GoogleCloudAiplatformV1SchemaImageClassificationAnnotation object
}
\description{
GoogleCloudAiplatformV1SchemaImageClassificationAnnotation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Annotation details specific to image classification.
}
\concept{GoogleCloudAiplatformV1SchemaImageClassificationAnnotation functions}
