% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaImageDataItem}
\alias{GoogleCloudAiplatformV1SchemaImageDataItem}
\title{GoogleCloudAiplatformV1SchemaImageDataItem Object}
\usage{
GoogleCloudAiplatformV1SchemaImageDataItem(gcsUri = NULL)
}
\arguments{
\item{gcsUri}{Required}
}
\value{
GoogleCloudAiplatformV1SchemaImageDataItem object
}
\description{
GoogleCloudAiplatformV1SchemaImageDataItem Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Payload of Image DataItem.
}
\concept{GoogleCloudAiplatformV1SchemaImageDataItem functions}
