% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaImageDatasetMetadata}
\alias{GoogleCloudAiplatformV1SchemaImageDatasetMetadata}
\title{GoogleCloudAiplatformV1SchemaImageDatasetMetadata Object}
\usage{
GoogleCloudAiplatformV1SchemaImageDatasetMetadata(
  dataItemSchemaUri = NULL,
  gcsBucket = NULL
)
}
\arguments{
\item{dataItemSchemaUri}{Points to a YAML file stored on Google Cloud Storage describing payload of the Image DataItems that belong to this Dataset}

\item{gcsBucket}{Google Cloud Storage Bucket name that contains the blob data of this Dataset}
}
\value{
GoogleCloudAiplatformV1SchemaImageDatasetMetadata object
}
\description{
GoogleCloudAiplatformV1SchemaImageDatasetMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The metadata of Datasets that contain Image DataItems.
}
\concept{GoogleCloudAiplatformV1SchemaImageDatasetMetadata functions}
