% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaImageSegmentationAnnotation}
\alias{GoogleCloudAiplatformV1SchemaImageSegmentationAnnotation}
\title{GoogleCloudAiplatformV1SchemaImageSegmentationAnnotation Object}
\usage{
GoogleCloudAiplatformV1SchemaImageSegmentationAnnotation(
  maskAnnotation = NULL,
  polylineAnnotation = NULL,
  polygonAnnotation = NULL
)
}
\arguments{
\item{maskAnnotation}{Mask based segmentation annotation}

\item{polylineAnnotation}{Polyline annotation}

\item{polygonAnnotation}{Polygon annotation}
}
\value{
GoogleCloudAiplatformV1SchemaImageSegmentationAnnotation object
}
\description{
GoogleCloudAiplatformV1SchemaImageSegmentationAnnotation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Annotation details specific to image segmentation.
}
\concept{GoogleCloudAiplatformV1SchemaImageSegmentationAnnotation functions}
