% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaImageSegmentationAnnotationMaskAnnotation}
\alias{GoogleCloudAiplatformV1SchemaImageSegmentationAnnotationMaskAnnotation}
\title{GoogleCloudAiplatformV1SchemaImageSegmentationAnnotationMaskAnnotation Object}
\usage{
GoogleCloudAiplatformV1SchemaImageSegmentationAnnotationMaskAnnotation(
  maskGcsUri = NULL,
  annotationSpecColors = NULL
)
}
\arguments{
\item{maskGcsUri}{Google Cloud Storage URI that points to the mask image}

\item{annotationSpecColors}{The mapping between color and AnnotationSpec for this Annotation}
}
\value{
GoogleCloudAiplatformV1SchemaImageSegmentationAnnotationMaskAnnotation object
}
\description{
GoogleCloudAiplatformV1SchemaImageSegmentationAnnotationMaskAnnotation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The mask based segmentation annotation.
}
\concept{GoogleCloudAiplatformV1SchemaImageSegmentationAnnotationMaskAnnotation functions}
