% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaImageSegmentationAnnotationPolygonAnnotation}
\alias{GoogleCloudAiplatformV1SchemaImageSegmentationAnnotationPolygonAnnotation}
\title{GoogleCloudAiplatformV1SchemaImageSegmentationAnnotationPolygonAnnotation Object}
\usage{
GoogleCloudAiplatformV1SchemaImageSegmentationAnnotationPolygonAnnotation(
  displayName = NULL,
  annotationSpecId = NULL,
  vertexes = NULL
)
}
\arguments{
\item{displayName}{The display name of the AnnotationSpec that this Annotation pertains to}

\item{annotationSpecId}{The resource Id of the AnnotationSpec that this Annotation pertains to}

\item{vertexes}{The vertexes are connected one by one and the last vertex is connected to the first one to represent a polygon}
}
\value{
GoogleCloudAiplatformV1SchemaImageSegmentationAnnotationPolygonAnnotation object
}
\description{
GoogleCloudAiplatformV1SchemaImageSegmentationAnnotationPolygonAnnotation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a polygon in image.
}
\concept{GoogleCloudAiplatformV1SchemaImageSegmentationAnnotationPolygonAnnotation functions}
