% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaImageSegmentationAnnotationPolylineAnnotation}
\alias{GoogleCloudAiplatformV1SchemaImageSegmentationAnnotationPolylineAnnotation}
\title{GoogleCloudAiplatformV1SchemaImageSegmentationAnnotationPolylineAnnotation Object}
\usage{
GoogleCloudAiplatformV1SchemaImageSegmentationAnnotationPolylineAnnotation(
  vertexes = NULL,
  annotationSpecId = NULL,
  displayName = NULL
)
}
\arguments{
\item{vertexes}{The vertexes are connected one by one and the last vertex in not connected to the first one}

\item{annotationSpecId}{The resource Id of the AnnotationSpec that this Annotation pertains to}

\item{displayName}{The display name of the AnnotationSpec that this Annotation pertains to}
}
\value{
GoogleCloudAiplatformV1SchemaImageSegmentationAnnotationPolylineAnnotation object
}
\description{
GoogleCloudAiplatformV1SchemaImageSegmentationAnnotationPolylineAnnotation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a polyline in image.
}
\concept{GoogleCloudAiplatformV1SchemaImageSegmentationAnnotationPolylineAnnotation functions}
