% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaModelevaluationMetricsBoundingBoxMetrics}
\alias{GoogleCloudAiplatformV1SchemaModelevaluationMetricsBoundingBoxMetrics}
\title{GoogleCloudAiplatformV1SchemaModelevaluationMetricsBoundingBoxMetrics Object}
\usage{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsBoundingBoxMetrics(
  confidenceMetrics = NULL,
  meanAveragePrecision = NULL,
  iouThreshold = NULL
)
}
\arguments{
\item{confidenceMetrics}{Metrics for each label-match confidence_threshold from 0}

\item{meanAveragePrecision}{The mean average precision, most often close to \code{auPrc}}

\item{iouThreshold}{The intersection-over-union threshold value used to compute this metrics entry}
}
\value{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsBoundingBoxMetrics object
}
\description{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsBoundingBoxMetrics Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Bounding box matching model metrics for a single intersection-over-union threshold and multiple label match confidence thresholds.
}
\concept{GoogleCloudAiplatformV1SchemaModelevaluationMetricsBoundingBoxMetrics functions}
