% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaModelevaluationMetricsClassificationEvaluationMetrics}
\alias{GoogleCloudAiplatformV1SchemaModelevaluationMetricsClassificationEvaluationMetrics}
\title{GoogleCloudAiplatformV1SchemaModelevaluationMetricsClassificationEvaluationMetrics Object}
\usage{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsClassificationEvaluationMetrics(
  confidenceMetrics = NULL,
  auRoc = NULL,
  confusionMatrix = NULL,
  auPrc = NULL,
  logLoss = NULL
)
}
\arguments{
\item{confidenceMetrics}{Metrics for each \code{confidenceThreshold} in 0}

\item{auRoc}{The Area Under Receiver Operating Characteristic curve metric}

\item{confusionMatrix}{Confusion matrix of the evaluation}

\item{auPrc}{The Area Under Precision-Recall Curve metric}

\item{logLoss}{The Log Loss metric}
}
\value{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsClassificationEvaluationMetrics object
}
\description{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsClassificationEvaluationMetrics Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metrics for classification evaluation results.
}
\concept{GoogleCloudAiplatformV1SchemaModelevaluationMetricsClassificationEvaluationMetrics functions}
