% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaModelevaluationMetricsConfusionMatrix}
\alias{GoogleCloudAiplatformV1SchemaModelevaluationMetricsConfusionMatrix}
\title{GoogleCloudAiplatformV1SchemaModelevaluationMetricsConfusionMatrix Object}
\usage{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsConfusionMatrix(
  rows = NULL,
  annotationSpecs = NULL
)
}
\arguments{
\item{rows}{Rows in the confusion matrix}

\item{annotationSpecs}{AnnotationSpecs used in the confusion matrix}
}
\value{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsConfusionMatrix object
}
\description{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsConfusionMatrix Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\concept{GoogleCloudAiplatformV1SchemaModelevaluationMetricsConfusionMatrix functions}
