% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaModelevaluationMetricsForecastingEvaluationMetrics}
\alias{GoogleCloudAiplatformV1SchemaModelevaluationMetricsForecastingEvaluationMetrics}
\title{GoogleCloudAiplatformV1SchemaModelevaluationMetricsForecastingEvaluationMetrics Object}
\usage{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsForecastingEvaluationMetrics(
  weightedAbsolutePercentageError = NULL,
  quantileMetrics = NULL,
  rootMeanSquaredError = NULL,
  meanAbsoluteError = NULL,
  rootMeanSquaredLogError = NULL,
  meanAbsolutePercentageError = NULL,
  rSquared = NULL,
  rootMeanSquaredPercentageError = NULL
)
}
\arguments{
\item{weightedAbsolutePercentageError}{Weighted Absolute Percentage Error}

\item{quantileMetrics}{The quantile metrics entries for each quantile}

\item{rootMeanSquaredError}{Root Mean Squared Error (RMSE)}

\item{meanAbsoluteError}{Mean Absolute Error (MAE)}

\item{rootMeanSquaredLogError}{Root mean squared log error}

\item{meanAbsolutePercentageError}{Mean absolute percentage error}

\item{rSquared}{Coefficient of determination as Pearson correlation coefficient}

\item{rootMeanSquaredPercentageError}{Root Mean Square Percentage Error}
}
\value{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsForecastingEvaluationMetrics object
}
\description{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsForecastingEvaluationMetrics Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metrics for forecasting evaluation results.
}
\concept{GoogleCloudAiplatformV1SchemaModelevaluationMetricsForecastingEvaluationMetrics functions}
