% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaModelevaluationMetricsForecastingEvaluationMetricsQuantileMetricsEntry}
\alias{GoogleCloudAiplatformV1SchemaModelevaluationMetricsForecastingEvaluationMetricsQuantileMetricsEntry}
\title{GoogleCloudAiplatformV1SchemaModelevaluationMetricsForecastingEvaluationMetricsQuantileMetricsEntry Object}
\usage{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsForecastingEvaluationMetricsQuantileMetricsEntry(
  scaledPinballLoss = NULL,
  quantile = NULL,
  observedQuantile = NULL
)
}
\arguments{
\item{scaledPinballLoss}{The scaled pinball loss of this quantile}

\item{quantile}{The quantile for this entry}

\item{observedQuantile}{This is a custom metric that calculates the percentage of true values that were less than the predicted value for that quantile}
}
\value{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsForecastingEvaluationMetricsQuantileMetricsEntry object
}
\description{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsForecastingEvaluationMetricsQuantileMetricsEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Entry for the Quantiles loss type optimization objective.
}
\concept{GoogleCloudAiplatformV1SchemaModelevaluationMetricsForecastingEvaluationMetricsQuantileMetricsEntry functions}
