% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaModelevaluationMetricsImageObjectDetectionEvaluationMetrics}
\alias{GoogleCloudAiplatformV1SchemaModelevaluationMetricsImageObjectDetectionEvaluationMetrics}
\title{GoogleCloudAiplatformV1SchemaModelevaluationMetricsImageObjectDetectionEvaluationMetrics Object}
\usage{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsImageObjectDetectionEvaluationMetrics(
  evaluatedBoundingBoxCount = NULL,
  boundingBoxMetrics = NULL,
  boundingBoxMeanAveragePrecision = NULL
)
}
\arguments{
\item{evaluatedBoundingBoxCount}{The total number of bounding boxes (i}

\item{boundingBoxMetrics}{The bounding boxes match metrics for each intersection-over-union threshold 0}

\item{boundingBoxMeanAveragePrecision}{The single metric for bounding boxes evaluation: the \code{meanAveragePrecision} averaged over all \code{boundingBoxMetricsEntries}}
}
\value{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsImageObjectDetectionEvaluationMetrics object
}
\description{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsImageObjectDetectionEvaluationMetrics Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metrics for image object detection evaluation results.
}
\concept{GoogleCloudAiplatformV1SchemaModelevaluationMetricsImageObjectDetectionEvaluationMetrics functions}
