% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaModelevaluationMetricsImageSegmentationEvaluationMetrics}
\alias{GoogleCloudAiplatformV1SchemaModelevaluationMetricsImageSegmentationEvaluationMetrics}
\title{GoogleCloudAiplatformV1SchemaModelevaluationMetricsImageSegmentationEvaluationMetrics Object}
\usage{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsImageSegmentationEvaluationMetrics(
  confidenceMetricsEntries = NULL
)
}
\arguments{
\item{confidenceMetricsEntries}{Metrics for each confidenceThreshold in 0}
}
\value{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsImageSegmentationEvaluationMetrics object
}
\description{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsImageSegmentationEvaluationMetrics Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metrics for image segmentation evaluation results.
}
\concept{GoogleCloudAiplatformV1SchemaModelevaluationMetricsImageSegmentationEvaluationMetrics functions}
