% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaModelevaluationMetricsImageSegmentationEvaluationMetricsConfidenceMetricsEntry}
\alias{GoogleCloudAiplatformV1SchemaModelevaluationMetricsImageSegmentationEvaluationMetricsConfidenceMetricsEntry}
\title{GoogleCloudAiplatformV1SchemaModelevaluationMetricsImageSegmentationEvaluationMetricsConfidenceMetricsEntry Object}
\usage{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsImageSegmentationEvaluationMetricsConfidenceMetricsEntry(
  confidenceThreshold = NULL,
  iouScore = NULL,
  recall = NULL,
  precision = NULL,
  confusionMatrix = NULL,
  diceScoreCoefficient = NULL
)
}
\arguments{
\item{confidenceThreshold}{Metrics are computed with an assumption that the model never returns predictions with score lower than this value}

\item{iouScore}{The intersection-over-union score}

\item{recall}{Recall (True Positive Rate) for the given confidence threshold}

\item{precision}{Precision for the given confidence threshold}

\item{confusionMatrix}{Confusion matrix for the given confidence threshold}

\item{diceScoreCoefficient}{DSC or the F1 score, The harmonic mean of recall and precision}
}
\value{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsImageSegmentationEvaluationMetricsConfidenceMetricsEntry object
}
\description{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsImageSegmentationEvaluationMetricsConfidenceMetricsEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\concept{GoogleCloudAiplatformV1SchemaModelevaluationMetricsImageSegmentationEvaluationMetricsConfidenceMetricsEntry functions}
