% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaModelevaluationMetricsRegressionEvaluationMetrics}
\alias{GoogleCloudAiplatformV1SchemaModelevaluationMetricsRegressionEvaluationMetrics}
\title{GoogleCloudAiplatformV1SchemaModelevaluationMetricsRegressionEvaluationMetrics Object}
\usage{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsRegressionEvaluationMetrics(
  rootMeanSquaredLogError = NULL,
  meanAbsoluteError = NULL,
  meanAbsolutePercentageError = NULL,
  rSquared = NULL,
  rootMeanSquaredError = NULL
)
}
\arguments{
\item{rootMeanSquaredLogError}{Root mean squared log error}

\item{meanAbsoluteError}{Mean Absolute Error (MAE)}

\item{meanAbsolutePercentageError}{Mean absolute percentage error}

\item{rSquared}{Coefficient of determination as Pearson correlation coefficient}

\item{rootMeanSquaredError}{Root Mean Squared Error (RMSE)}
}
\value{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsRegressionEvaluationMetrics object
}
\description{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsRegressionEvaluationMetrics Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metrics for regression evaluation results.
}
\concept{GoogleCloudAiplatformV1SchemaModelevaluationMetricsRegressionEvaluationMetrics functions}
