% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaModelevaluationMetricsSummarizationEvaluationMetrics}
\alias{GoogleCloudAiplatformV1SchemaModelevaluationMetricsSummarizationEvaluationMetrics}
\title{GoogleCloudAiplatformV1SchemaModelevaluationMetricsSummarizationEvaluationMetrics Object}
\usage{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsSummarizationEvaluationMetrics(
  rougeLSum = NULL
)
}
\arguments{
\item{rougeLSum}{ROUGE-L (Longest Common Subsequence) scoring at summary level}
}
\value{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsSummarizationEvaluationMetrics object
}
\description{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsSummarizationEvaluationMetrics Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\concept{GoogleCloudAiplatformV1SchemaModelevaluationMetricsSummarizationEvaluationMetrics functions}
