% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaModelevaluationMetricsTextExtractionEvaluationMetrics}
\alias{GoogleCloudAiplatformV1SchemaModelevaluationMetricsTextExtractionEvaluationMetrics}
\title{GoogleCloudAiplatformV1SchemaModelevaluationMetricsTextExtractionEvaluationMetrics Object}
\usage{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsTextExtractionEvaluationMetrics(
  confusionMatrix = NULL,
  confidenceMetrics = NULL
)
}
\arguments{
\item{confusionMatrix}{Confusion matrix of the evaluation}

\item{confidenceMetrics}{Metrics that have confidence thresholds}
}
\value{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsTextExtractionEvaluationMetrics object
}
\description{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsTextExtractionEvaluationMetrics Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metrics for text extraction evaluation results.
}
\concept{GoogleCloudAiplatformV1SchemaModelevaluationMetricsTextExtractionEvaluationMetrics functions}
