% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaModelevaluationMetricsTextExtractionEvaluationMetricsConfidenceMetrics}
\alias{GoogleCloudAiplatformV1SchemaModelevaluationMetricsTextExtractionEvaluationMetricsConfidenceMetrics}
\title{GoogleCloudAiplatformV1SchemaModelevaluationMetricsTextExtractionEvaluationMetricsConfidenceMetrics Object}
\usage{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsTextExtractionEvaluationMetricsConfidenceMetrics(
  precision = NULL,
  confidenceThreshold = NULL,
  recall = NULL,
  f1Score = NULL
)
}
\arguments{
\item{precision}{Precision for the given confidence threshold}

\item{confidenceThreshold}{Metrics are computed with an assumption that the Model never returns predictions with score lower than this value}

\item{recall}{Recall (True Positive Rate) for the given confidence threshold}

\item{f1Score}{The harmonic mean of recall and precision}
}
\value{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsTextExtractionEvaluationMetricsConfidenceMetrics object
}
\description{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsTextExtractionEvaluationMetricsConfidenceMetrics Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\concept{GoogleCloudAiplatformV1SchemaModelevaluationMetricsTextExtractionEvaluationMetricsConfidenceMetrics functions}
