% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaModelevaluationMetricsTextSentimentEvaluationMetrics}
\alias{GoogleCloudAiplatformV1SchemaModelevaluationMetricsTextSentimentEvaluationMetrics}
\title{GoogleCloudAiplatformV1SchemaModelevaluationMetricsTextSentimentEvaluationMetrics Object}
\usage{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsTextSentimentEvaluationMetrics(
  meanAbsoluteError = NULL,
  linearKappa = NULL,
  quadraticKappa = NULL,
  precision = NULL,
  f1Score = NULL,
  meanSquaredError = NULL,
  confusionMatrix = NULL,
  recall = NULL
)
}
\arguments{
\item{meanAbsoluteError}{Mean absolute error}

\item{linearKappa}{Linear weighted kappa}

\item{quadraticKappa}{Quadratic weighted kappa}

\item{precision}{Precision}

\item{f1Score}{The harmonic mean of recall and precision}

\item{meanSquaredError}{Mean squared error}

\item{confusionMatrix}{Confusion matrix of the evaluation}

\item{recall}{Recall}
}
\value{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsTextSentimentEvaluationMetrics object
}
\description{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsTextSentimentEvaluationMetrics Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Model evaluation metrics for text sentiment problems.
}
\concept{GoogleCloudAiplatformV1SchemaModelevaluationMetricsTextSentimentEvaluationMetrics functions}
