% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaModelevaluationMetricsTrackMetricsConfidenceMetrics}
\alias{GoogleCloudAiplatformV1SchemaModelevaluationMetricsTrackMetricsConfidenceMetrics}
\title{GoogleCloudAiplatformV1SchemaModelevaluationMetricsTrackMetricsConfidenceMetrics Object}
\usage{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsTrackMetricsConfidenceMetrics(
  confidenceThreshold = NULL,
  trackingRecall = NULL,
  boundingBoxIou = NULL,
  trackingPrecision = NULL,
  mismatchRate = NULL
)
}
\arguments{
\item{confidenceThreshold}{The confidence threshold value used to compute the metrics}

\item{trackingRecall}{Tracking recall}

\item{boundingBoxIou}{Bounding box intersection-over-union precision}

\item{trackingPrecision}{Tracking precision}

\item{mismatchRate}{Mismatch rate, which measures the tracking consistency, i}
}
\value{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsTrackMetricsConfidenceMetrics object
}
\description{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsTrackMetricsConfidenceMetrics Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metrics for a single confidence threshold.
}
\concept{GoogleCloudAiplatformV1SchemaModelevaluationMetricsTrackMetricsConfidenceMetrics functions}
