% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaModelevaluationMetricsVideoActionMetrics}
\alias{GoogleCloudAiplatformV1SchemaModelevaluationMetricsVideoActionMetrics}
\title{GoogleCloudAiplatformV1SchemaModelevaluationMetricsVideoActionMetrics Object}
\usage{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsVideoActionMetrics(
  confidenceMetrics = NULL,
  meanAveragePrecision = NULL,
  precisionWindowLength = NULL
)
}
\arguments{
\item{confidenceMetrics}{Metrics for each label-match confidence_threshold from 0}

\item{meanAveragePrecision}{The mean average precision}

\item{precisionWindowLength}{This VideoActionMetrics is calculated based on this prediction window length}
}
\value{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsVideoActionMetrics object
}
\description{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsVideoActionMetrics Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The Evaluation metrics given a specific precision_window_length.
}
\concept{GoogleCloudAiplatformV1SchemaModelevaluationMetricsVideoActionMetrics functions}
