% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaModelevaluationMetricsVideoActionRecognitionMetrics}
\alias{GoogleCloudAiplatformV1SchemaModelevaluationMetricsVideoActionRecognitionMetrics}
\title{GoogleCloudAiplatformV1SchemaModelevaluationMetricsVideoActionRecognitionMetrics Object}
\usage{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsVideoActionRecognitionMetrics(
  evaluatedActionCount = NULL,
  videoActionMetrics = NULL
)
}
\arguments{
\item{evaluatedActionCount}{The number of ground truth actions used to create this evaluation}

\item{videoActionMetrics}{The metric entries for precision window lengths: 1s,2s,3s}
}
\value{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsVideoActionRecognitionMetrics object
}
\description{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsVideoActionRecognitionMetrics Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Model evaluation metrics for video action recognition.
}
\concept{GoogleCloudAiplatformV1SchemaModelevaluationMetricsVideoActionRecognitionMetrics functions}
