% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaModelevaluationMetricsVideoObjectTrackingMetrics}
\alias{GoogleCloudAiplatformV1SchemaModelevaluationMetricsVideoObjectTrackingMetrics}
\title{GoogleCloudAiplatformV1SchemaModelevaluationMetricsVideoObjectTrackingMetrics Object}
\usage{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsVideoObjectTrackingMetrics(
  trackMetrics = NULL,
  trackMeanMismatchRate = NULL,
  trackMeanAveragePrecision = NULL,
  evaluatedTrackCount = NULL,
  evaluatedFrameCount = NULL,
  boundingBoxMetrics = NULL,
  evaluatedBoundingBoxCount = NULL,
  boundingBoxMeanAveragePrecision = NULL,
  trackMeanBoundingBoxIou = NULL
)
}
\arguments{
\item{trackMetrics}{UNIMPLEMENTED}

\item{trackMeanMismatchRate}{UNIMPLEMENTED}

\item{trackMeanAveragePrecision}{UNIMPLEMENTED}

\item{evaluatedTrackCount}{UNIMPLEMENTED}

\item{evaluatedFrameCount}{UNIMPLEMENTED}

\item{boundingBoxMetrics}{The bounding boxes match metrics for each intersection-over-union threshold 0}

\item{evaluatedBoundingBoxCount}{UNIMPLEMENTED}

\item{boundingBoxMeanAveragePrecision}{The single metric for bounding boxes evaluation: the \code{meanAveragePrecision} averaged over all \code{boundingBoxMetrics}}

\item{trackMeanBoundingBoxIou}{UNIMPLEMENTED}
}
\value{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsVideoObjectTrackingMetrics object
}
\description{
GoogleCloudAiplatformV1SchemaModelevaluationMetricsVideoObjectTrackingMetrics Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Model evaluation metrics for video object tracking problems. Evaluates prediction quality of both labeled bounding boxes and labeled tracks (i.e. series of bounding boxes sharing same label and instance ID).
}
\concept{GoogleCloudAiplatformV1SchemaModelevaluationMetricsVideoObjectTrackingMetrics functions}
