% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaPredictInstanceImageClassificationPredictionInstance}
\alias{GoogleCloudAiplatformV1SchemaPredictInstanceImageClassificationPredictionInstance}
\title{GoogleCloudAiplatformV1SchemaPredictInstanceImageClassificationPredictionInstance Object}
\usage{
GoogleCloudAiplatformV1SchemaPredictInstanceImageClassificationPredictionInstance(
  content = NULL,
  mimeType = NULL
)
}
\arguments{
\item{content}{The image bytes or Cloud Storage URI to make the prediction on}

\item{mimeType}{The MIME type of the content of the image}
}
\value{
GoogleCloudAiplatformV1SchemaPredictInstanceImageClassificationPredictionInstance object
}
\description{
GoogleCloudAiplatformV1SchemaPredictInstanceImageClassificationPredictionInstance Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Prediction input format for Image Classification.
}
\concept{GoogleCloudAiplatformV1SchemaPredictInstanceImageClassificationPredictionInstance functions}
