% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaPredictInstanceImageObjectDetectionPredictionInstance}
\alias{GoogleCloudAiplatformV1SchemaPredictInstanceImageObjectDetectionPredictionInstance}
\title{GoogleCloudAiplatformV1SchemaPredictInstanceImageObjectDetectionPredictionInstance Object}
\usage{
GoogleCloudAiplatformV1SchemaPredictInstanceImageObjectDetectionPredictionInstance(
  content = NULL,
  mimeType = NULL
)
}
\arguments{
\item{content}{The image bytes or Cloud Storage URI to make the prediction on}

\item{mimeType}{The MIME type of the content of the image}
}
\value{
GoogleCloudAiplatformV1SchemaPredictInstanceImageObjectDetectionPredictionInstance object
}
\description{
GoogleCloudAiplatformV1SchemaPredictInstanceImageObjectDetectionPredictionInstance Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Prediction input format for Image Object Detection.
}
\concept{GoogleCloudAiplatformV1SchemaPredictInstanceImageObjectDetectionPredictionInstance functions}
