% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaPredictInstanceImageSegmentationPredictionInstance}
\alias{GoogleCloudAiplatformV1SchemaPredictInstanceImageSegmentationPredictionInstance}
\title{GoogleCloudAiplatformV1SchemaPredictInstanceImageSegmentationPredictionInstance Object}
\usage{
GoogleCloudAiplatformV1SchemaPredictInstanceImageSegmentationPredictionInstance(
  mimeType = NULL,
  content = NULL
)
}
\arguments{
\item{mimeType}{The MIME type of the content of the image}

\item{content}{The image bytes to make the predictions on}
}
\value{
GoogleCloudAiplatformV1SchemaPredictInstanceImageSegmentationPredictionInstance object
}
\description{
GoogleCloudAiplatformV1SchemaPredictInstanceImageSegmentationPredictionInstance Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Prediction input format for Image Segmentation.
}
\concept{GoogleCloudAiplatformV1SchemaPredictInstanceImageSegmentationPredictionInstance functions}
