% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaPredictInstanceTextClassificationPredictionInstance}
\alias{GoogleCloudAiplatformV1SchemaPredictInstanceTextClassificationPredictionInstance}
\title{GoogleCloudAiplatformV1SchemaPredictInstanceTextClassificationPredictionInstance Object}
\usage{
GoogleCloudAiplatformV1SchemaPredictInstanceTextClassificationPredictionInstance(
  mimeType = NULL,
  content = NULL
)
}
\arguments{
\item{mimeType}{The MIME type of the text snippet}

\item{content}{The text snippet to make the predictions on}
}
\value{
GoogleCloudAiplatformV1SchemaPredictInstanceTextClassificationPredictionInstance object
}
\description{
GoogleCloudAiplatformV1SchemaPredictInstanceTextClassificationPredictionInstance Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Prediction input format for Text Classification.
}
\concept{GoogleCloudAiplatformV1SchemaPredictInstanceTextClassificationPredictionInstance functions}
