% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaPredictInstanceTextExtractionPredictionInstance}
\alias{GoogleCloudAiplatformV1SchemaPredictInstanceTextExtractionPredictionInstance}
\title{GoogleCloudAiplatformV1SchemaPredictInstanceTextExtractionPredictionInstance Object}
\usage{
GoogleCloudAiplatformV1SchemaPredictInstanceTextExtractionPredictionInstance(
  content = NULL,
  mimeType = NULL,
  key = NULL
)
}
\arguments{
\item{content}{The text snippet to make the predictions on}

\item{mimeType}{The MIME type of the text snippet}

\item{key}{This field is only used for batch prediction}
}
\value{
GoogleCloudAiplatformV1SchemaPredictInstanceTextExtractionPredictionInstance object
}
\description{
GoogleCloudAiplatformV1SchemaPredictInstanceTextExtractionPredictionInstance Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Prediction input format for Text Extraction.
}
\concept{GoogleCloudAiplatformV1SchemaPredictInstanceTextExtractionPredictionInstance functions}
