% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaPredictInstanceTextSentimentPredictionInstance}
\alias{GoogleCloudAiplatformV1SchemaPredictInstanceTextSentimentPredictionInstance}
\title{GoogleCloudAiplatformV1SchemaPredictInstanceTextSentimentPredictionInstance Object}
\usage{
GoogleCloudAiplatformV1SchemaPredictInstanceTextSentimentPredictionInstance(
  content = NULL,
  mimeType = NULL
)
}
\arguments{
\item{content}{The text snippet to make the predictions on}

\item{mimeType}{The MIME type of the text snippet}
}
\value{
GoogleCloudAiplatformV1SchemaPredictInstanceTextSentimentPredictionInstance object
}
\description{
GoogleCloudAiplatformV1SchemaPredictInstanceTextSentimentPredictionInstance Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Prediction input format for Text Sentiment.
}
\concept{GoogleCloudAiplatformV1SchemaPredictInstanceTextSentimentPredictionInstance functions}
