% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaPredictInstanceVideoActionRecognitionPredictionInstance}
\alias{GoogleCloudAiplatformV1SchemaPredictInstanceVideoActionRecognitionPredictionInstance}
\title{GoogleCloudAiplatformV1SchemaPredictInstanceVideoActionRecognitionPredictionInstance Object}
\usage{
GoogleCloudAiplatformV1SchemaPredictInstanceVideoActionRecognitionPredictionInstance(
  mimeType = NULL,
  timeSegmentStart = NULL,
  content = NULL,
  timeSegmentEnd = NULL
)
}
\arguments{
\item{mimeType}{The MIME type of the content of the video}

\item{timeSegmentStart}{The beginning, inclusive, of the video's time segment on which to perform the prediction}

\item{content}{The Google Cloud Storage location of the video on which to perform the prediction}

\item{timeSegmentEnd}{The end, exclusive, of the video's time segment on which to perform the prediction}
}
\value{
GoogleCloudAiplatformV1SchemaPredictInstanceVideoActionRecognitionPredictionInstance object
}
\description{
GoogleCloudAiplatformV1SchemaPredictInstanceVideoActionRecognitionPredictionInstance Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Prediction input format for Video Action Recognition.
}
\concept{GoogleCloudAiplatformV1SchemaPredictInstanceVideoActionRecognitionPredictionInstance functions}
