% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaPredictInstanceVideoObjectTrackingPredictionInstance}
\alias{GoogleCloudAiplatformV1SchemaPredictInstanceVideoObjectTrackingPredictionInstance}
\title{GoogleCloudAiplatformV1SchemaPredictInstanceVideoObjectTrackingPredictionInstance Object}
\usage{
GoogleCloudAiplatformV1SchemaPredictInstanceVideoObjectTrackingPredictionInstance(
  timeSegmentEnd = NULL,
  timeSegmentStart = NULL,
  content = NULL,
  mimeType = NULL
)
}
\arguments{
\item{timeSegmentEnd}{The end, exclusive, of the video's time segment on which to perform the prediction}

\item{timeSegmentStart}{The beginning, inclusive, of the video's time segment on which to perform the prediction}

\item{content}{The Google Cloud Storage location of the video on which to perform the prediction}

\item{mimeType}{The MIME type of the content of the video}
}
\value{
GoogleCloudAiplatformV1SchemaPredictInstanceVideoObjectTrackingPredictionInstance object
}
\description{
GoogleCloudAiplatformV1SchemaPredictInstanceVideoObjectTrackingPredictionInstance Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Prediction input format for Video Object Tracking.
}
\concept{GoogleCloudAiplatformV1SchemaPredictInstanceVideoObjectTrackingPredictionInstance functions}
