% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaPredictionResult.instance}
\alias{GoogleCloudAiplatformV1SchemaPredictionResult.instance}
\title{GoogleCloudAiplatformV1SchemaPredictionResult.instance Object}
\usage{
GoogleCloudAiplatformV1SchemaPredictionResult.instance()
}
\value{
GoogleCloudAiplatformV1SchemaPredictionResult.instance object
}
\description{
GoogleCloudAiplatformV1SchemaPredictionResult.instance Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
User's input instance. Struct is used here instead of Any so that JsonFormat does not append an extra '@type' field when we convert the proto to JSON.
}
\seealso{
Other GoogleCloudAiplatformV1SchemaPredictionResult functions: 
\code{\link{GoogleCloudAiplatformV1SchemaPredictionResult}()}
}
\concept{GoogleCloudAiplatformV1SchemaPredictionResult functions}
