% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaPredictionResult}
\alias{GoogleCloudAiplatformV1SchemaPredictionResult}
\title{GoogleCloudAiplatformV1SchemaPredictionResult Object}
\usage{
GoogleCloudAiplatformV1SchemaPredictionResult(
  GoogleCloudAiplatformV1SchemaPredictionResult.instance = NULL,
  instance = NULL,
  prediction = NULL,
  key = NULL,
  error = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1SchemaPredictionResult.instance}{The \link{GoogleCloudAiplatformV1SchemaPredictionResult.instance} object or list of objects}

\item{instance}{User's input instance}

\item{prediction}{The prediction result}

\item{key}{Optional user-provided key from the input instance}

\item{error}{The error result}
}
\value{
GoogleCloudAiplatformV1SchemaPredictionResult object
}
\description{
GoogleCloudAiplatformV1SchemaPredictionResult Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a line of JSONL in the batch prediction output file.
}
\seealso{
Other GoogleCloudAiplatformV1SchemaPredictionResult functions: 
\code{\link{GoogleCloudAiplatformV1SchemaPredictionResult.instance}()}
}
\concept{GoogleCloudAiplatformV1SchemaPredictionResult functions}
