% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaPredictionResultError}
\alias{GoogleCloudAiplatformV1SchemaPredictionResultError}
\title{GoogleCloudAiplatformV1SchemaPredictionResultError Object}
\usage{
GoogleCloudAiplatformV1SchemaPredictionResultError(
  message = NULL,
  status = NULL
)
}
\arguments{
\item{message}{Error message with additional details}

\item{status}{Error status}
}
\value{
GoogleCloudAiplatformV1SchemaPredictionResultError object
}
\description{
GoogleCloudAiplatformV1SchemaPredictionResultError Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\concept{GoogleCloudAiplatformV1SchemaPredictionResultError functions}
