% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaPredictParamsImageClassificationPredictionParams}
\alias{GoogleCloudAiplatformV1SchemaPredictParamsImageClassificationPredictionParams}
\title{GoogleCloudAiplatformV1SchemaPredictParamsImageClassificationPredictionParams Object}
\usage{
GoogleCloudAiplatformV1SchemaPredictParamsImageClassificationPredictionParams(
  confidenceThreshold = NULL,
  maxPredictions = NULL
)
}
\arguments{
\item{confidenceThreshold}{The Model only returns predictions with at least this confidence score}

\item{maxPredictions}{The Model only returns up to that many top, by confidence score, predictions per instance}
}
\value{
GoogleCloudAiplatformV1SchemaPredictParamsImageClassificationPredictionParams object
}
\description{
GoogleCloudAiplatformV1SchemaPredictParamsImageClassificationPredictionParams Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Prediction model parameters for Image Classification.
}
\concept{GoogleCloudAiplatformV1SchemaPredictParamsImageClassificationPredictionParams functions}
