% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaPredictParamsImageObjectDetectionPredictionParams}
\alias{GoogleCloudAiplatformV1SchemaPredictParamsImageObjectDetectionPredictionParams}
\title{GoogleCloudAiplatformV1SchemaPredictParamsImageObjectDetectionPredictionParams Object}
\usage{
GoogleCloudAiplatformV1SchemaPredictParamsImageObjectDetectionPredictionParams(
  confidenceThreshold = NULL,
  maxPredictions = NULL
)
}
\arguments{
\item{confidenceThreshold}{The Model only returns predictions with at least this confidence score}

\item{maxPredictions}{The Model only returns up to that many top, by confidence score, predictions per instance}
}
\value{
GoogleCloudAiplatformV1SchemaPredictParamsImageObjectDetectionPredictionParams object
}
\description{
GoogleCloudAiplatformV1SchemaPredictParamsImageObjectDetectionPredictionParams Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Prediction model parameters for Image Object Detection.
}
\concept{GoogleCloudAiplatformV1SchemaPredictParamsImageObjectDetectionPredictionParams functions}
