% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaPredictParamsImageSegmentationPredictionParams}
\alias{GoogleCloudAiplatformV1SchemaPredictParamsImageSegmentationPredictionParams}
\title{GoogleCloudAiplatformV1SchemaPredictParamsImageSegmentationPredictionParams Object}
\usage{
GoogleCloudAiplatformV1SchemaPredictParamsImageSegmentationPredictionParams(
  confidenceThreshold = NULL
)
}
\arguments{
\item{confidenceThreshold}{When the model predicts category of pixels of the image, it will only provide predictions for pixels that it is at least this much confident about}
}
\value{
GoogleCloudAiplatformV1SchemaPredictParamsImageSegmentationPredictionParams object
}
\description{
GoogleCloudAiplatformV1SchemaPredictParamsImageSegmentationPredictionParams Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Prediction model parameters for Image Segmentation.
}
\concept{GoogleCloudAiplatformV1SchemaPredictParamsImageSegmentationPredictionParams functions}
