% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaPredictParamsVideoActionRecognitionPredictionParams}
\alias{GoogleCloudAiplatformV1SchemaPredictParamsVideoActionRecognitionPredictionParams}
\title{GoogleCloudAiplatformV1SchemaPredictParamsVideoActionRecognitionPredictionParams Object}
\usage{
GoogleCloudAiplatformV1SchemaPredictParamsVideoActionRecognitionPredictionParams(
  maxPredictions = NULL,
  confidenceThreshold = NULL
)
}
\arguments{
\item{maxPredictions}{The model only returns up to that many top, by confidence score, predictions per frame of the video}

\item{confidenceThreshold}{The Model only returns predictions with at least this confidence score}
}
\value{
GoogleCloudAiplatformV1SchemaPredictParamsVideoActionRecognitionPredictionParams object
}
\description{
GoogleCloudAiplatformV1SchemaPredictParamsVideoActionRecognitionPredictionParams Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Prediction model parameters for Video Action Recognition.
}
\concept{GoogleCloudAiplatformV1SchemaPredictParamsVideoActionRecognitionPredictionParams functions}
