% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaPredictParamsVideoClassificationPredictionParams}
\alias{GoogleCloudAiplatformV1SchemaPredictParamsVideoClassificationPredictionParams}
\title{GoogleCloudAiplatformV1SchemaPredictParamsVideoClassificationPredictionParams Object}
\usage{
GoogleCloudAiplatformV1SchemaPredictParamsVideoClassificationPredictionParams(
  confidenceThreshold = NULL,
  maxPredictions = NULL,
  shotClassification = NULL,
  segmentClassification = NULL,
  oneSecIntervalClassification = NULL
)
}
\arguments{
\item{confidenceThreshold}{The Model only returns predictions with at least this confidence score}

\item{maxPredictions}{The Model only returns up to that many top, by confidence score, predictions per instance}

\item{shotClassification}{Set to true to request shot-level classification}

\item{segmentClassification}{Set to true to request segment-level classification}

\item{oneSecIntervalClassification}{Set to true to request classification for a video at one-second intervals}
}
\value{
GoogleCloudAiplatformV1SchemaPredictParamsVideoClassificationPredictionParams object
}
\description{
GoogleCloudAiplatformV1SchemaPredictParamsVideoClassificationPredictionParams Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Prediction model parameters for Video Classification.
}
\concept{GoogleCloudAiplatformV1SchemaPredictParamsVideoClassificationPredictionParams functions}
