% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaPredictParamsVideoObjectTrackingPredictionParams}
\alias{GoogleCloudAiplatformV1SchemaPredictParamsVideoObjectTrackingPredictionParams}
\title{GoogleCloudAiplatformV1SchemaPredictParamsVideoObjectTrackingPredictionParams Object}
\usage{
GoogleCloudAiplatformV1SchemaPredictParamsVideoObjectTrackingPredictionParams(
  confidenceThreshold = NULL,
  maxPredictions = NULL,
  minBoundingBoxSize = NULL
)
}
\arguments{
\item{confidenceThreshold}{The Model only returns predictions with at least this confidence score}

\item{maxPredictions}{The model only returns up to that many top, by confidence score, predictions per frame of the video}

\item{minBoundingBoxSize}{Only bounding boxes with shortest edge at least that long as a relative value of video frame size are returned}
}
\value{
GoogleCloudAiplatformV1SchemaPredictParamsVideoObjectTrackingPredictionParams object
}
\description{
GoogleCloudAiplatformV1SchemaPredictParamsVideoObjectTrackingPredictionParams Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Prediction model parameters for Video Object Tracking.
}
\concept{GoogleCloudAiplatformV1SchemaPredictParamsVideoObjectTrackingPredictionParams functions}
