% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaPredictPredictionClassificationPredictionResult}
\alias{GoogleCloudAiplatformV1SchemaPredictPredictionClassificationPredictionResult}
\title{GoogleCloudAiplatformV1SchemaPredictPredictionClassificationPredictionResult Object}
\usage{
GoogleCloudAiplatformV1SchemaPredictPredictionClassificationPredictionResult(
  ids = NULL,
  confidences = NULL,
  displayNames = NULL
)
}
\arguments{
\item{ids}{The resource IDs of the AnnotationSpecs that had been identified}

\item{confidences}{The Model's confidences in correctness of the predicted IDs, higher value means higher confidence}

\item{displayNames}{The display names of the AnnotationSpecs that had been identified, order matches the IDs}
}
\value{
GoogleCloudAiplatformV1SchemaPredictPredictionClassificationPredictionResult object
}
\description{
GoogleCloudAiplatformV1SchemaPredictPredictionClassificationPredictionResult Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Prediction output format for Image and Text Classification.
}
\concept{GoogleCloudAiplatformV1SchemaPredictPredictionClassificationPredictionResult functions}
