% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaPredictPredictionImageObjectDetectionPredictionResult}
\alias{GoogleCloudAiplatformV1SchemaPredictPredictionImageObjectDetectionPredictionResult}
\title{GoogleCloudAiplatformV1SchemaPredictPredictionImageObjectDetectionPredictionResult Object}
\usage{
GoogleCloudAiplatformV1SchemaPredictPredictionImageObjectDetectionPredictionResult(
  bboxes = NULL,
  confidences = NULL,
  displayNames = NULL,
  ids = NULL
)
}
\arguments{
\item{bboxes}{Bounding boxes, i}

\item{confidences}{The Model's confidences in correctness of the predicted IDs, higher value means higher confidence}

\item{displayNames}{The display names of the AnnotationSpecs that had been identified, order matches the IDs}

\item{ids}{The resource IDs of the AnnotationSpecs that had been identified, ordered by the confidence score descendingly}
}
\value{
GoogleCloudAiplatformV1SchemaPredictPredictionImageObjectDetectionPredictionResult object
}
\description{
GoogleCloudAiplatformV1SchemaPredictPredictionImageObjectDetectionPredictionResult Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Prediction output format for Image Object Detection.
}
\concept{GoogleCloudAiplatformV1SchemaPredictPredictionImageObjectDetectionPredictionResult functions}
