% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaPredictPredictionImageSegmentationPredictionResult}
\alias{GoogleCloudAiplatformV1SchemaPredictPredictionImageSegmentationPredictionResult}
\title{GoogleCloudAiplatformV1SchemaPredictPredictionImageSegmentationPredictionResult Object}
\usage{
GoogleCloudAiplatformV1SchemaPredictPredictionImageSegmentationPredictionResult(
  confidenceMask = NULL,
  categoryMask = NULL
)
}
\arguments{
\item{confidenceMask}{A one channel image which is encoded as an 8bit lossless PNG}

\item{categoryMask}{A PNG image where each pixel in the mask represents the category in which the pixel in the original image was predicted to belong to}
}
\value{
GoogleCloudAiplatformV1SchemaPredictPredictionImageSegmentationPredictionResult object
}
\description{
GoogleCloudAiplatformV1SchemaPredictPredictionImageSegmentationPredictionResult Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Prediction output format for Image Segmentation.
}
\concept{GoogleCloudAiplatformV1SchemaPredictPredictionImageSegmentationPredictionResult functions}
