% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaPredictPredictionTabularClassificationPredictionResult}
\alias{GoogleCloudAiplatformV1SchemaPredictPredictionTabularClassificationPredictionResult}
\title{GoogleCloudAiplatformV1SchemaPredictPredictionTabularClassificationPredictionResult Object}
\usage{
GoogleCloudAiplatformV1SchemaPredictPredictionTabularClassificationPredictionResult(
  scores = NULL,
  classes = NULL
)
}
\arguments{
\item{scores}{The model's confidence in each class being correct, higher value means higher confidence}

\item{classes}{The name of the classes being classified, contains all possible values of the target column}
}
\value{
GoogleCloudAiplatformV1SchemaPredictPredictionTabularClassificationPredictionResult object
}
\description{
GoogleCloudAiplatformV1SchemaPredictPredictionTabularClassificationPredictionResult Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Prediction output format for Tabular Classification.
}
\concept{GoogleCloudAiplatformV1SchemaPredictPredictionTabularClassificationPredictionResult functions}
