% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaPredictPredictionTabularRegressionPredictionResult}
\alias{GoogleCloudAiplatformV1SchemaPredictPredictionTabularRegressionPredictionResult}
\title{GoogleCloudAiplatformV1SchemaPredictPredictionTabularRegressionPredictionResult Object}
\usage{
GoogleCloudAiplatformV1SchemaPredictPredictionTabularRegressionPredictionResult(
  lowerBound = NULL,
  quantileValues = NULL,
  upperBound = NULL,
  value = NULL,
  quantilePredictions = NULL
)
}
\arguments{
\item{lowerBound}{The lower bound of the prediction interval}

\item{quantileValues}{Quantile values}

\item{upperBound}{The upper bound of the prediction interval}

\item{value}{The regression value}

\item{quantilePredictions}{Quantile predictions, in 1-1 correspondence with quantile_values}
}
\value{
GoogleCloudAiplatformV1SchemaPredictPredictionTabularRegressionPredictionResult object
}
\description{
GoogleCloudAiplatformV1SchemaPredictPredictionTabularRegressionPredictionResult Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Prediction output format for Tabular Regression.
}
\concept{GoogleCloudAiplatformV1SchemaPredictPredictionTabularRegressionPredictionResult functions}
