% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaPredictPredictionTextSentimentPredictionResult}
\alias{GoogleCloudAiplatformV1SchemaPredictPredictionTextSentimentPredictionResult}
\title{GoogleCloudAiplatformV1SchemaPredictPredictionTextSentimentPredictionResult Object}
\usage{
GoogleCloudAiplatformV1SchemaPredictPredictionTextSentimentPredictionResult(
  sentiment = NULL
)
}
\arguments{
\item{sentiment}{The integer sentiment labels between 0 (inclusive) and sentimentMax label (inclusive), while 0 maps to the least positive sentiment and sentimentMax maps to the most positive one}
}
\value{
GoogleCloudAiplatformV1SchemaPredictPredictionTextSentimentPredictionResult object
}
\description{
GoogleCloudAiplatformV1SchemaPredictPredictionTextSentimentPredictionResult Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Prediction output format for Text Sentiment
}
\concept{GoogleCloudAiplatformV1SchemaPredictPredictionTextSentimentPredictionResult functions}
