% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaPredictPredictionTftFeatureImportance}
\alias{GoogleCloudAiplatformV1SchemaPredictPredictionTftFeatureImportance}
\title{GoogleCloudAiplatformV1SchemaPredictPredictionTftFeatureImportance Object}
\usage{
GoogleCloudAiplatformV1SchemaPredictPredictionTftFeatureImportance(
  horizonWeights = NULL,
  contextColumns = NULL,
  horizonColumns = NULL,
  attributeWeights = NULL,
  attributeColumns = NULL,
  contextWeights = NULL
)
}
\arguments{
\item{horizonWeights}{No description}

\item{contextColumns}{No description}

\item{horizonColumns}{No description}

\item{attributeWeights}{No description}

\item{attributeColumns}{No description}

\item{contextWeights}{TFT feature importance values}
}
\value{
GoogleCloudAiplatformV1SchemaPredictPredictionTftFeatureImportance object
}
\description{
GoogleCloudAiplatformV1SchemaPredictPredictionTftFeatureImportance Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\concept{GoogleCloudAiplatformV1SchemaPredictPredictionTftFeatureImportance functions}
