% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaPredictPredictionTimeSeriesForecastingPredictionResult}
\alias{GoogleCloudAiplatformV1SchemaPredictPredictionTimeSeriesForecastingPredictionResult}
\title{GoogleCloudAiplatformV1SchemaPredictPredictionTimeSeriesForecastingPredictionResult Object}
\usage{
GoogleCloudAiplatformV1SchemaPredictPredictionTimeSeriesForecastingPredictionResult(
  quantileValues = NULL,
  tftFeatureImportance = NULL,
  value = NULL,
  quantilePredictions = NULL
)
}
\arguments{
\item{quantileValues}{Quantile values}

\item{tftFeatureImportance}{Only use these if TFt is enabled}

\item{value}{The regression value}

\item{quantilePredictions}{Quantile predictions, in 1-1 correspondence with quantile_values}
}
\value{
GoogleCloudAiplatformV1SchemaPredictPredictionTimeSeriesForecastingPredictionResult object
}
\description{
GoogleCloudAiplatformV1SchemaPredictPredictionTimeSeriesForecastingPredictionResult Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Prediction output format for Time Series Forecasting.
}
\concept{GoogleCloudAiplatformV1SchemaPredictPredictionTimeSeriesForecastingPredictionResult functions}
