% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaPredictPredictionVideoActionRecognitionPredictionResult}
\alias{GoogleCloudAiplatformV1SchemaPredictPredictionVideoActionRecognitionPredictionResult}
\title{GoogleCloudAiplatformV1SchemaPredictPredictionVideoActionRecognitionPredictionResult Object}
\usage{
GoogleCloudAiplatformV1SchemaPredictPredictionVideoActionRecognitionPredictionResult(
  displayName = NULL,
  timeSegmentStart = NULL,
  id = NULL,
  timeSegmentEnd = NULL,
  confidence = NULL
)
}
\arguments{
\item{displayName}{The display name of the AnnotationSpec that had been identified}

\item{timeSegmentStart}{The beginning, inclusive, of the video's time segment in which the AnnotationSpec has been identified}

\item{id}{The resource ID of the AnnotationSpec that had been identified}

\item{timeSegmentEnd}{The end, exclusive, of the video's time segment in which the AnnotationSpec has been identified}

\item{confidence}{The Model's confidence in correction of this prediction, higher value means higher confidence}
}
\value{
GoogleCloudAiplatformV1SchemaPredictPredictionVideoActionRecognitionPredictionResult object
}
\description{
GoogleCloudAiplatformV1SchemaPredictPredictionVideoActionRecognitionPredictionResult Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Prediction output format for Video Action Recognition.
}
\concept{GoogleCloudAiplatformV1SchemaPredictPredictionVideoActionRecognitionPredictionResult functions}
