% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaPredictPredictionVideoClassificationPredictionResult}
\alias{GoogleCloudAiplatformV1SchemaPredictPredictionVideoClassificationPredictionResult}
\title{GoogleCloudAiplatformV1SchemaPredictPredictionVideoClassificationPredictionResult Object}
\usage{
GoogleCloudAiplatformV1SchemaPredictPredictionVideoClassificationPredictionResult(
  id = NULL,
  timeSegmentStart = NULL,
  timeSegmentEnd = NULL,
  confidence = NULL,
  type = NULL,
  displayName = NULL
)
}
\arguments{
\item{id}{The resource ID of the AnnotationSpec that had been identified}

\item{timeSegmentStart}{The beginning, inclusive, of the video's time segment in which the AnnotationSpec has been identified}

\item{timeSegmentEnd}{The end, exclusive, of the video's time segment in which the AnnotationSpec has been identified}

\item{confidence}{The Model's confidence in correction of this prediction, higher value means higher confidence}

\item{type}{The type of the prediction}

\item{displayName}{The display name of the AnnotationSpec that had been identified}
}
\value{
GoogleCloudAiplatformV1SchemaPredictPredictionVideoClassificationPredictionResult object
}
\description{
GoogleCloudAiplatformV1SchemaPredictPredictionVideoClassificationPredictionResult Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Prediction output format for Video Classification.
}
\concept{GoogleCloudAiplatformV1SchemaPredictPredictionVideoClassificationPredictionResult functions}
