% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaPredictPredictionVideoObjectTrackingPredictionResult}
\alias{GoogleCloudAiplatformV1SchemaPredictPredictionVideoObjectTrackingPredictionResult}
\title{GoogleCloudAiplatformV1SchemaPredictPredictionVideoObjectTrackingPredictionResult Object}
\usage{
GoogleCloudAiplatformV1SchemaPredictPredictionVideoObjectTrackingPredictionResult(
  id = NULL,
  confidence = NULL,
  timeSegmentEnd = NULL,
  displayName = NULL,
  frames = NULL,
  timeSegmentStart = NULL
)
}
\arguments{
\item{id}{The resource ID of the AnnotationSpec that had been identified}

\item{confidence}{The Model's confidence in correction of this prediction, higher value means higher confidence}

\item{timeSegmentEnd}{The end, inclusive, of the video's time segment in which the object instance has been detected}

\item{displayName}{The display name of the AnnotationSpec that had been identified}

\item{frames}{All of the frames of the video in which a single object instance has been detected}

\item{timeSegmentStart}{The beginning, inclusive, of the video's time segment in which the object instance has been detected}
}
\value{
GoogleCloudAiplatformV1SchemaPredictPredictionVideoObjectTrackingPredictionResult object
}
\description{
GoogleCloudAiplatformV1SchemaPredictPredictionVideoObjectTrackingPredictionResult Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Prediction output format for Video Object Tracking.
}
\concept{GoogleCloudAiplatformV1SchemaPredictPredictionVideoObjectTrackingPredictionResult functions}
