% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaPredictPredictionVideoObjectTrackingPredictionResultFrame}
\alias{GoogleCloudAiplatformV1SchemaPredictPredictionVideoObjectTrackingPredictionResultFrame}
\title{GoogleCloudAiplatformV1SchemaPredictPredictionVideoObjectTrackingPredictionResultFrame Object}
\usage{
GoogleCloudAiplatformV1SchemaPredictPredictionVideoObjectTrackingPredictionResultFrame(
  yMin = NULL,
  yMax = NULL,
  xMax = NULL,
  xMin = NULL,
  timeOffset = NULL
)
}
\arguments{
\item{yMin}{The topmost coordinate of the bounding box}

\item{yMax}{The bottommost coordinate of the bounding box}

\item{xMax}{The rightmost coordinate of the bounding box}

\item{xMin}{The leftmost coordinate of the bounding box}

\item{timeOffset}{A time (frame) of a video in which the object has been detected}
}
\value{
GoogleCloudAiplatformV1SchemaPredictPredictionVideoObjectTrackingPredictionResultFrame object
}
\description{
GoogleCloudAiplatformV1SchemaPredictPredictionVideoObjectTrackingPredictionResultFrame Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The fields \code{xMin}, \code{xMax}, \code{yMin}, and \code{yMax} refer to a bounding box, i.e. the rectangle over the video frame pinpointing the found AnnotationSpec. The coordinates are relative to the frame size, and the point 0,0 is in the top left of the frame.
}
\concept{GoogleCloudAiplatformV1SchemaPredictPredictionVideoObjectTrackingPredictionResultFrame functions}
