% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTablesDatasetMetadata}
\alias{GoogleCloudAiplatformV1SchemaTablesDatasetMetadata}
\title{GoogleCloudAiplatformV1SchemaTablesDatasetMetadata Object}
\usage{
GoogleCloudAiplatformV1SchemaTablesDatasetMetadata(inputConfig = NULL)
}
\arguments{
\item{inputConfig}{No description}
}
\value{
GoogleCloudAiplatformV1SchemaTablesDatasetMetadata object
}
\description{
GoogleCloudAiplatformV1SchemaTablesDatasetMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The metadata of Datasets that contain tables data.
}
\concept{GoogleCloudAiplatformV1SchemaTablesDatasetMetadata functions}
