% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTablesDatasetMetadataBigQuerySource}
\alias{GoogleCloudAiplatformV1SchemaTablesDatasetMetadataBigQuerySource}
\title{GoogleCloudAiplatformV1SchemaTablesDatasetMetadataBigQuerySource Object}
\usage{
GoogleCloudAiplatformV1SchemaTablesDatasetMetadataBigQuerySource(uri = NULL)
}
\arguments{
\item{uri}{The URI of a BigQuery table}
}
\value{
GoogleCloudAiplatformV1SchemaTablesDatasetMetadataBigQuerySource object
}
\description{
GoogleCloudAiplatformV1SchemaTablesDatasetMetadataBigQuerySource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\concept{GoogleCloudAiplatformV1SchemaTablesDatasetMetadataBigQuerySource functions}
