% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTablesDatasetMetadataGcsSource}
\alias{GoogleCloudAiplatformV1SchemaTablesDatasetMetadataGcsSource}
\title{GoogleCloudAiplatformV1SchemaTablesDatasetMetadataGcsSource Object}
\usage{
GoogleCloudAiplatformV1SchemaTablesDatasetMetadataGcsSource(uri = NULL)
}
\arguments{
\item{uri}{Cloud Storage URI of one or more files}
}
\value{
GoogleCloudAiplatformV1SchemaTablesDatasetMetadataGcsSource object
}
\description{
GoogleCloudAiplatformV1SchemaTablesDatasetMetadataGcsSource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\concept{GoogleCloudAiplatformV1SchemaTablesDatasetMetadataGcsSource functions}
