% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTablesDatasetMetadataInputConfig}
\alias{GoogleCloudAiplatformV1SchemaTablesDatasetMetadataInputConfig}
\title{GoogleCloudAiplatformV1SchemaTablesDatasetMetadataInputConfig Object}
\usage{
GoogleCloudAiplatformV1SchemaTablesDatasetMetadataInputConfig(
  gcsSource = NULL,
  bigquerySource = NULL
)
}
\arguments{
\item{gcsSource}{No description}

\item{bigquerySource}{No description}
}
\value{
GoogleCloudAiplatformV1SchemaTablesDatasetMetadataInputConfig object
}
\description{
GoogleCloudAiplatformV1SchemaTablesDatasetMetadataInputConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The tables Dataset's data source. The Dataset doesn't store the data directly, but only pointer(s) to its data.
}
\concept{GoogleCloudAiplatformV1SchemaTablesDatasetMetadataInputConfig functions}
