% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTextClassificationAnnotation}
\alias{GoogleCloudAiplatformV1SchemaTextClassificationAnnotation}
\title{GoogleCloudAiplatformV1SchemaTextClassificationAnnotation Object}
\usage{
GoogleCloudAiplatformV1SchemaTextClassificationAnnotation(
  displayName = NULL,
  annotationSpecId = NULL
)
}
\arguments{
\item{displayName}{The display name of the AnnotationSpec that this Annotation pertains to}

\item{annotationSpecId}{The resource Id of the AnnotationSpec that this Annotation pertains to}
}
\value{
GoogleCloudAiplatformV1SchemaTextClassificationAnnotation object
}
\description{
GoogleCloudAiplatformV1SchemaTextClassificationAnnotation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Annotation details specific to text classification.
}
\concept{GoogleCloudAiplatformV1SchemaTextClassificationAnnotation functions}
