% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTextDataItem}
\alias{GoogleCloudAiplatformV1SchemaTextDataItem}
\title{GoogleCloudAiplatformV1SchemaTextDataItem Object}
\usage{
GoogleCloudAiplatformV1SchemaTextDataItem()
}
\value{
GoogleCloudAiplatformV1SchemaTextDataItem object
}
\description{
GoogleCloudAiplatformV1SchemaTextDataItem Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Payload of Text DataItem.
}
\concept{GoogleCloudAiplatformV1SchemaTextDataItem functions}
