% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTextDatasetMetadata}
\alias{GoogleCloudAiplatformV1SchemaTextDatasetMetadata}
\title{GoogleCloudAiplatformV1SchemaTextDatasetMetadata Object}
\usage{
GoogleCloudAiplatformV1SchemaTextDatasetMetadata(
  dataItemSchemaUri = NULL,
  gcsBucket = NULL
)
}
\arguments{
\item{dataItemSchemaUri}{Points to a YAML file stored on Google Cloud Storage describing payload of the Text DataItems that belong to this Dataset}

\item{gcsBucket}{Google Cloud Storage Bucket name that contains the blob data of this Dataset}
}
\value{
GoogleCloudAiplatformV1SchemaTextDatasetMetadata object
}
\description{
GoogleCloudAiplatformV1SchemaTextDatasetMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The metadata of Datasets that contain Text DataItems.
}
\concept{GoogleCloudAiplatformV1SchemaTextDatasetMetadata functions}
