% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTextExtractionAnnotation}
\alias{GoogleCloudAiplatformV1SchemaTextExtractionAnnotation}
\title{GoogleCloudAiplatformV1SchemaTextExtractionAnnotation Object}
\usage{
GoogleCloudAiplatformV1SchemaTextExtractionAnnotation(
  annotationSpecId = NULL,
  textSegment = NULL,
  displayName = NULL
)
}
\arguments{
\item{annotationSpecId}{The resource Id of the AnnotationSpec that this Annotation pertains to}

\item{textSegment}{The segment of the text content}

\item{displayName}{The display name of the AnnotationSpec that this Annotation pertains to}
}
\value{
GoogleCloudAiplatformV1SchemaTextExtractionAnnotation object
}
\description{
GoogleCloudAiplatformV1SchemaTextExtractionAnnotation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Annotation details specific to text extraction.
}
\concept{GoogleCloudAiplatformV1SchemaTextExtractionAnnotation functions}
