% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTextPromptDatasetMetadata}
\alias{GoogleCloudAiplatformV1SchemaTextPromptDatasetMetadata}
\title{GoogleCloudAiplatformV1SchemaTextPromptDatasetMetadata Object}
\usage{
GoogleCloudAiplatformV1SchemaTextPromptDatasetMetadata(
  topP = NULL,
  maxOutputTokens = NULL,
  gcsUri = NULL,
  promptType = NULL,
  topK = NULL,
  text = NULL,
  temperature = NULL
)
}
\arguments{
\item{topP}{Top P value set when the dataset was saved}

\item{maxOutputTokens}{Value of the maximum number of tokens generated set when the dataset was saved}

\item{gcsUri}{The Google Cloud Storage URI that stores the prompt data}

\item{promptType}{Type of the prompt dataset}

\item{topK}{Top K value set when the dataset was saved}

\item{text}{The content of the prompt dataset}

\item{temperature}{Temperature value used for sampling set when the dataset was saved}
}
\value{
GoogleCloudAiplatformV1SchemaTextPromptDatasetMetadata object
}
\description{
GoogleCloudAiplatformV1SchemaTextPromptDatasetMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The metadata of Datasets that contain Text Prompt data.
}
\concept{GoogleCloudAiplatformV1SchemaTextPromptDatasetMetadata functions}
