% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTextSegment}
\alias{GoogleCloudAiplatformV1SchemaTextSegment}
\title{GoogleCloudAiplatformV1SchemaTextSegment Object}
\usage{
GoogleCloudAiplatformV1SchemaTextSegment(
  endOffset = NULL,
  startOffset = NULL,
  content = NULL
)
}
\arguments{
\item{endOffset}{Zero-based character index of the first character past the end of the text segment (counting character from the beginning of the text)}

\item{startOffset}{Zero-based character index of the first character of the text segment (counting characters from the beginning of the text)}

\item{content}{The text content in the segment for output only}
}
\value{
GoogleCloudAiplatformV1SchemaTextSegment object
}
\description{
GoogleCloudAiplatformV1SchemaTextSegment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The text segment inside of DataItem.
}
\concept{GoogleCloudAiplatformV1SchemaTextSegment functions}
