% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTextSentimentAnnotation}
\alias{GoogleCloudAiplatformV1SchemaTextSentimentAnnotation}
\title{GoogleCloudAiplatformV1SchemaTextSentimentAnnotation Object}
\usage{
GoogleCloudAiplatformV1SchemaTextSentimentAnnotation(
  sentimentMax = NULL,
  displayName = NULL,
  annotationSpecId = NULL,
  sentiment = NULL
)
}
\arguments{
\item{sentimentMax}{The sentiment max score for text}

\item{displayName}{The display name of the AnnotationSpec that this Annotation pertains to}

\item{annotationSpecId}{The resource Id of the AnnotationSpec that this Annotation pertains to}

\item{sentiment}{The sentiment score for text}
}
\value{
GoogleCloudAiplatformV1SchemaTextSentimentAnnotation object
}
\description{
GoogleCloudAiplatformV1SchemaTextSentimentAnnotation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Annotation details specific to text sentiment.
}
\concept{GoogleCloudAiplatformV1SchemaTextSentimentAnnotation functions}
