% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTextSentimentSavedQueryMetadata}
\alias{GoogleCloudAiplatformV1SchemaTextSentimentSavedQueryMetadata}
\title{GoogleCloudAiplatformV1SchemaTextSentimentSavedQueryMetadata Object}
\usage{
GoogleCloudAiplatformV1SchemaTextSentimentSavedQueryMetadata(
  sentimentMax = NULL
)
}
\arguments{
\item{sentimentMax}{The maximum sentiment of sentiment Anntoation in this SavedQuery}
}
\value{
GoogleCloudAiplatformV1SchemaTextSentimentSavedQueryMetadata object
}
\description{
GoogleCloudAiplatformV1SchemaTextSentimentSavedQueryMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The metadata of SavedQuery contains TextSentiment Annotations.
}
\concept{GoogleCloudAiplatformV1SchemaTextSentimentSavedQueryMetadata functions}
