% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTimeSegment}
\alias{GoogleCloudAiplatformV1SchemaTimeSegment}
\title{GoogleCloudAiplatformV1SchemaTimeSegment Object}
\usage{
GoogleCloudAiplatformV1SchemaTimeSegment(
  startTimeOffset = NULL,
  endTimeOffset = NULL
)
}
\arguments{
\item{startTimeOffset}{Start of the time segment (inclusive), represented as the duration since the start of the DataItem}

\item{endTimeOffset}{End of the time segment (exclusive), represented as the duration since the start of the DataItem}
}
\value{
GoogleCloudAiplatformV1SchemaTimeSegment object
}
\description{
GoogleCloudAiplatformV1SchemaTimeSegment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A time period inside of a DataItem that has a time dimension (e.g. video).
}
\concept{GoogleCloudAiplatformV1SchemaTimeSegment functions}
