% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTimeSeriesDatasetMetadata}
\alias{GoogleCloudAiplatformV1SchemaTimeSeriesDatasetMetadata}
\title{GoogleCloudAiplatformV1SchemaTimeSeriesDatasetMetadata Object}
\usage{
GoogleCloudAiplatformV1SchemaTimeSeriesDatasetMetadata(
  inputConfig = NULL,
  timeColumn = NULL,
  timeSeriesIdentifierColumn = NULL
)
}
\arguments{
\item{inputConfig}{No description}

\item{timeColumn}{The column name of the time column that identifies time order in the time series}

\item{timeSeriesIdentifierColumn}{The column name of the time series identifier column that identifies the time series}
}
\value{
GoogleCloudAiplatformV1SchemaTimeSeriesDatasetMetadata object
}
\description{
GoogleCloudAiplatformV1SchemaTimeSeriesDatasetMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The metadata of Datasets that contain time series data.
}
\concept{GoogleCloudAiplatformV1SchemaTimeSeriesDatasetMetadata functions}
