% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTimeSeriesDatasetMetadataBigQuerySource}
\alias{GoogleCloudAiplatformV1SchemaTimeSeriesDatasetMetadataBigQuerySource}
\title{GoogleCloudAiplatformV1SchemaTimeSeriesDatasetMetadataBigQuerySource Object}
\usage{
GoogleCloudAiplatformV1SchemaTimeSeriesDatasetMetadataBigQuerySource(
  uri = NULL
)
}
\arguments{
\item{uri}{The URI of a BigQuery table}
}
\value{
GoogleCloudAiplatformV1SchemaTimeSeriesDatasetMetadataBigQuerySource object
}
\description{
GoogleCloudAiplatformV1SchemaTimeSeriesDatasetMetadataBigQuerySource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\concept{GoogleCloudAiplatformV1SchemaTimeSeriesDatasetMetadataBigQuerySource functions}
