% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTimeSeriesDatasetMetadataGcsSource}
\alias{GoogleCloudAiplatformV1SchemaTimeSeriesDatasetMetadataGcsSource}
\title{GoogleCloudAiplatformV1SchemaTimeSeriesDatasetMetadataGcsSource Object}
\usage{
GoogleCloudAiplatformV1SchemaTimeSeriesDatasetMetadataGcsSource(uri = NULL)
}
\arguments{
\item{uri}{Cloud Storage URI of one or more files}
}
\value{
GoogleCloudAiplatformV1SchemaTimeSeriesDatasetMetadataGcsSource object
}
\description{
GoogleCloudAiplatformV1SchemaTimeSeriesDatasetMetadataGcsSource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\concept{GoogleCloudAiplatformV1SchemaTimeSeriesDatasetMetadataGcsSource functions}
