% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTimeSeriesDatasetMetadataInputConfig}
\alias{GoogleCloudAiplatformV1SchemaTimeSeriesDatasetMetadataInputConfig}
\title{GoogleCloudAiplatformV1SchemaTimeSeriesDatasetMetadataInputConfig Object}
\usage{
GoogleCloudAiplatformV1SchemaTimeSeriesDatasetMetadataInputConfig(
  gcsSource = NULL,
  bigquerySource = NULL
)
}
\arguments{
\item{gcsSource}{No description}

\item{bigquerySource}{No description}
}
\value{
GoogleCloudAiplatformV1SchemaTimeSeriesDatasetMetadataInputConfig object
}
\description{
GoogleCloudAiplatformV1SchemaTimeSeriesDatasetMetadataInputConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The time series Dataset's data source. The Dataset doesn't store the data directly, but only pointer(s) to its data.
}
\concept{GoogleCloudAiplatformV1SchemaTimeSeriesDatasetMetadataInputConfig functions}
