% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlForecasting}
\alias{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlForecasting}
\title{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlForecasting Object}
\usage{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlForecasting(
  inputs = NULL,
  metadata = NULL
)
}
\arguments{
\item{inputs}{The input parameters of this TrainingJob}

\item{metadata}{The metadata information}
}
\value{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlForecasting object
}
\description{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlForecasting Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A TrainingJob that trains and uploads an AutoML Forecasting Model.
}
\concept{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlForecasting functions}
