% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlForecastingInputs}
\alias{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlForecastingInputs}
\title{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlForecastingInputs Object}
\usage{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlForecastingInputs(
  exportEvaluatedDataItemsConfig = NULL,
  windowConfig = NULL,
  validationOptions = NULL,
  dataGranularity = NULL,
  quantiles = NULL,
  weightColumn = NULL,
  hierarchyConfig = NULL,
  availableAtForecastColumns = NULL,
  contextWindow = NULL,
  timeColumn = NULL,
  optimizationObjective = NULL,
  additionalExperiments = NULL,
  timeSeriesIdentifierColumn = NULL,
  targetColumn = NULL,
  timeSeriesAttributeColumns = NULL,
  enableProbabilisticInference = NULL,
  unavailableAtForecastColumns = NULL,
  trainBudgetMilliNodeHours = NULL,
  forecastHorizon = NULL,
  transformations = NULL,
  holidayRegions = NULL
)
}
\arguments{
\item{exportEvaluatedDataItemsConfig}{Configuration for exporting test set predictions to a BigQuery table}

\item{windowConfig}{Config containing strategy for generating sliding windows}

\item{validationOptions}{Validation options for the data validation component}

\item{dataGranularity}{Expected difference in time granularity between rows in the data}

\item{quantiles}{Quantiles to use for minimize-quantile-loss \code{optimization_objective}, or for probabilistic inference}

\item{weightColumn}{Column name that should be used as the weight column}

\item{hierarchyConfig}{Configuration that defines the hierarchical relationship of time series and parameters for hierarchical forecasting strategies}

\item{availableAtForecastColumns}{Names of columns that are available and provided when a forecast is requested}

\item{contextWindow}{The amount of time into the past training and prediction data is used for model training and prediction respectively}

\item{timeColumn}{The name of the column that identifies time order in the time series}

\item{optimizationObjective}{Objective function the model is optimizing towards}

\item{additionalExperiments}{Additional experiment flags for the time series forcasting training}

\item{timeSeriesIdentifierColumn}{The name of the column that identifies the time series}

\item{targetColumn}{The name of the column that the Model is to predict values for}

\item{timeSeriesAttributeColumns}{Column names that should be used as attribute columns}

\item{enableProbabilisticInference}{If probabilistic inference is enabled, the model will fit a distribution that captures the uncertainty of a prediction}

\item{unavailableAtForecastColumns}{Names of columns that are unavailable when a forecast is requested}

\item{trainBudgetMilliNodeHours}{Required}

\item{forecastHorizon}{The amount of time into the future for which forecasted values for the target are returned}

\item{transformations}{Each transformation will apply transform function to given input column}

\item{holidayRegions}{The geographical region based on which the holiday effect is applied in modeling by adding holiday categorical array feature that include all holidays matching the date}
}
\value{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlForecastingInputs object
}
\description{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlForecastingInputs Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\concept{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlForecastingInputs functions}
