% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlForecastingInputsGranularity}
\alias{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlForecastingInputsGranularity}
\title{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlForecastingInputsGranularity Object}
\usage{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlForecastingInputsGranularity(
  unit = NULL,
  quantity = NULL
)
}
\arguments{
\item{unit}{The time granularity unit of this time period}

\item{quantity}{The number of granularity_units between data points in the training data}
}
\value{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlForecastingInputsGranularity object
}
\description{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlForecastingInputsGranularity Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A duration of time expressed in time granularity units.
}
\concept{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlForecastingInputsGranularity functions}
