% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlForecastingMetadata}
\alias{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlForecastingMetadata}
\title{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlForecastingMetadata Object}
\usage{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlForecastingMetadata(
  trainCostMilliNodeHours = NULL,
  evaluatedDataItemsBigqueryUri = NULL
)
}
\arguments{
\item{trainCostMilliNodeHours}{Output only}

\item{evaluatedDataItemsBigqueryUri}{BigQuery destination uri for exported evaluated examples}
}
\value{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlForecastingMetadata object
}
\description{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlForecastingMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Model metadata specific to AutoML Forecasting.
}
\concept{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlForecastingMetadata functions}
