% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageClassification}
\alias{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageClassification}
\title{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageClassification Object}
\usage{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageClassification(
  metadata = NULL,
  inputs = NULL
)
}
\arguments{
\item{metadata}{The metadata information}

\item{inputs}{The input parameters of this TrainingJob}
}
\value{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageClassification object
}
\description{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageClassification Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A TrainingJob that trains and uploads an AutoML Image Classification Model.
}
\concept{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageClassification functions}
