% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageClassificationInputs}
\alias{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageClassificationInputs}
\title{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageClassificationInputs Object}
\usage{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageClassificationInputs(
  budgetMilliNodeHours = NULL,
  tunableParameter = NULL,
  multiLabel = NULL,
  uptrainBaseModelId = NULL,
  disableEarlyStopping = NULL,
  baseModelId = NULL,
  modelType = NULL
)
}
\arguments{
\item{budgetMilliNodeHours}{The training budget of creating this model, expressed in milli node hours i}

\item{tunableParameter}{Trainer type for Vision TrainRequest}

\item{multiLabel}{If false, a single-label (multi-class) Model will be trained (i}

\item{uptrainBaseModelId}{The ID of \code{base} model for upTraining}

\item{disableEarlyStopping}{Use the entire training budget}

\item{baseModelId}{The ID of the \code{base} model}

\item{modelType}{No description}
}
\value{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageClassificationInputs object
}
\description{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageClassificationInputs Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\concept{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageClassificationInputs functions}
