% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageClassificationMetadata}
\alias{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageClassificationMetadata}
\title{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageClassificationMetadata Object}
\usage{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageClassificationMetadata(
  costMilliNodeHours = NULL,
  successfulStopReason = NULL
)
}
\arguments{
\item{costMilliNodeHours}{The actual training cost of creating this model, expressed in milli node hours, i}

\item{successfulStopReason}{For successful job completions, this is the reason why the job has finished}
}
\value{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageClassificationMetadata object
}
\description{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageClassificationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\concept{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageClassificationMetadata functions}
