% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageObjectDetection}
\alias{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageObjectDetection}
\title{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageObjectDetection Object}
\usage{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageObjectDetection(
  inputs = NULL,
  metadata = NULL
)
}
\arguments{
\item{inputs}{The input parameters of this TrainingJob}

\item{metadata}{The metadata information}
}
\value{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageObjectDetection object
}
\description{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageObjectDetection Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A TrainingJob that trains and uploads an AutoML Image Object Detection Model.
}
\concept{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageObjectDetection functions}
