% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageObjectDetectionInputs}
\alias{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageObjectDetectionInputs}
\title{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageObjectDetectionInputs Object}
\usage{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageObjectDetectionInputs(
  disableEarlyStopping = NULL,
  uptrainBaseModelId = NULL,
  tunableParameter = NULL,
  modelType = NULL,
  budgetMilliNodeHours = NULL
)
}
\arguments{
\item{disableEarlyStopping}{Use the entire training budget}

\item{uptrainBaseModelId}{The ID of \code{base} model for upTraining}

\item{tunableParameter}{Trainer type for Vision TrainRequest}

\item{modelType}{No description}

\item{budgetMilliNodeHours}{The training budget of creating this model, expressed in milli node hours i}
}
\value{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageObjectDetectionInputs object
}
\description{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageObjectDetectionInputs Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\concept{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageObjectDetectionInputs functions}
