% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageSegmentation}
\alias{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageSegmentation}
\title{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageSegmentation Object}
\usage{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageSegmentation(
  metadata = NULL,
  inputs = NULL
)
}
\arguments{
\item{metadata}{The metadata information}

\item{inputs}{The input parameters of this TrainingJob}
}
\value{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageSegmentation object
}
\description{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageSegmentation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A TrainingJob that trains and uploads an AutoML Image Segmentation Model.
}
\concept{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageSegmentation functions}
