% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageSegmentationInputs}
\alias{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageSegmentationInputs}
\title{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageSegmentationInputs Object}
\usage{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageSegmentationInputs(
  baseModelId = NULL,
  modelType = NULL,
  budgetMilliNodeHours = NULL
)
}
\arguments{
\item{baseModelId}{The ID of the \code{base} model}

\item{modelType}{No description}

\item{budgetMilliNodeHours}{The training budget of creating this model, expressed in milli node hours i}
}
\value{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageSegmentationInputs object
}
\description{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageSegmentationInputs Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\concept{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlImageSegmentationInputs functions}
