% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutomlImageTrainingTunableParameter}
\alias{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutomlImageTrainingTunableParameter}
\title{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutomlImageTrainingTunableParameter Object}
\usage{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutomlImageTrainingTunableParameter(
 
    GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutomlImageTrainingTunableParameter.datasetConfig
    = NULL,
 
    GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutomlImageTrainingTunableParameter.trainerConfig
    = NULL,
  studySpec = NULL,
  trainerType = NULL,
  datasetConfig = NULL,
  checkpointName = NULL,
  trainerConfig = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutomlImageTrainingTunableParameter.datasetConfig}{The \link{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutomlImageTrainingTunableParameter.datasetConfig} object or list of objects}

\item{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutomlImageTrainingTunableParameter.trainerConfig}{The \link{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutomlImageTrainingTunableParameter.trainerConfig} object or list of objects}

\item{studySpec}{Optioinal}

\item{trainerType}{No description}

\item{datasetConfig}{Customizable dataset settings, used in the \code{model_garden_trainer}}

\item{checkpointName}{Optional}

\item{trainerConfig}{Customizable trainer settings, used in the \code{model_garden_trainer}}
}
\value{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutomlImageTrainingTunableParameter object
}
\description{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutomlImageTrainingTunableParameter Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A wrapper class which contains the tunable parameters in an AutoML Image training job.
}
\seealso{
Other GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutomlImageTrainingTunableParameter functions: 
\code{\link{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutomlImageTrainingTunableParameter.datasetConfig}()},
\code{\link{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutomlImageTrainingTunableParameter.trainerConfig}()}
}
\concept{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutomlImageTrainingTunableParameter functions}
