% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTables}
\alias{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTables}
\title{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTables Object}
\usage{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTables(
  metadata = NULL,
  inputs = NULL
)
}
\arguments{
\item{metadata}{The metadata information}

\item{inputs}{The input parameters of this TrainingJob}
}
\value{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTables object
}
\description{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTables Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A TrainingJob that trains and uploads an AutoML Tables Model.
}
\concept{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTables functions}
