% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTablesInputs}
\alias{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTablesInputs}
\title{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTablesInputs Object}
\usage{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTablesInputs(
  weightColumnName = NULL,
  optimizationObjectivePrecisionValue = NULL,
  additionalExperiments = NULL,
  optimizationObjectiveRecallValue = NULL,
  disableEarlyStopping = NULL,
  optimizationObjective = NULL,
  transformations = NULL,
  exportEvaluatedDataItemsConfig = NULL,
  targetColumn = NULL,
  trainBudgetMilliNodeHours = NULL,
  predictionType = NULL
)
}
\arguments{
\item{weightColumnName}{Column name that should be used as the weight column}

\item{optimizationObjectivePrecisionValue}{Required when optimization_objective is 'maximize-recall-at-precision'}

\item{additionalExperiments}{Additional experiment flags for the Tables training pipeline}

\item{optimizationObjectiveRecallValue}{Required when optimization_objective is 'maximize-precision-at-recall'}

\item{disableEarlyStopping}{Use the entire training budget}

\item{optimizationObjective}{Objective function the model is optimizing towards}

\item{transformations}{Each transformation will apply transform function to given input column}

\item{exportEvaluatedDataItemsConfig}{Configuration for exporting test set predictions to a BigQuery table}

\item{targetColumn}{The column name of the target column that the model is to predict}

\item{trainBudgetMilliNodeHours}{Required}

\item{predictionType}{The type of prediction the Model is to produce}
}
\value{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTablesInputs object
}
\description{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTablesInputs Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\concept{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTablesInputs functions}
