% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTablesInputsTransformation}
\alias{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTablesInputsTransformation}
\title{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTablesInputsTransformation Object}
\usage{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTablesInputsTransformation(
  text = NULL,
  repeatedCategorical = NULL,
  timestamp = NULL,
  repeatedText = NULL,
  repeatedNumeric = NULL,
  numeric = NULL,
  auto = NULL,
  categorical = NULL
)
}
\arguments{
\item{text}{No description}

\item{repeatedCategorical}{No description}

\item{timestamp}{No description}

\item{repeatedText}{No description}

\item{repeatedNumeric}{No description}

\item{numeric}{No description}

\item{auto}{No description}

\item{categorical}{No description}
}
\value{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTablesInputsTransformation object
}
\description{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTablesInputsTransformation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\concept{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTablesInputsTransformation functions}
