% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTablesMetadata}
\alias{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTablesMetadata}
\title{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTablesMetadata Object}
\usage{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTablesMetadata(
  evaluatedDataItemsBigqueryUri = NULL,
  trainCostMilliNodeHours = NULL
)
}
\arguments{
\item{evaluatedDataItemsBigqueryUri}{BigQuery destination uri for exported evaluated examples}

\item{trainCostMilliNodeHours}{Output only}
}
\value{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTablesMetadata object
}
\description{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTablesMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Model metadata specific to AutoML Tables.
}
\concept{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTablesMetadata functions}
