% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTextClassification}
\alias{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTextClassification}
\title{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTextClassification Object}
\usage{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTextClassification(
  inputs = NULL
)
}
\arguments{
\item{inputs}{The input parameters of this TrainingJob}
}
\value{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTextClassification object
}
\description{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTextClassification Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A TrainingJob that trains and uploads an AutoML Text Classification Model.
}
\concept{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTextClassification functions}
