% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTextClassificationInputs}
\alias{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTextClassificationInputs}
\title{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTextClassificationInputs Object}
\usage{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTextClassificationInputs(
  multiLabel = NULL
)
}
\arguments{
\item{multiLabel}{No description}
}
\value{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTextClassificationInputs object
}
\description{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTextClassificationInputs Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\concept{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTextClassificationInputs functions}
