% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTextExtraction}
\alias{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTextExtraction}
\title{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTextExtraction Object}
\usage{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTextExtraction(
  inputs = NULL
)
}
\arguments{
\item{inputs}{The input parameters of this TrainingJob}
}
\value{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTextExtraction object
}
\description{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTextExtraction Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A TrainingJob that trains and uploads an AutoML Text Extraction Model.
}
\concept{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTextExtraction functions}
