% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTextSentiment}
\alias{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTextSentiment}
\title{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTextSentiment Object}
\usage{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTextSentiment(
  inputs = NULL
)
}
\arguments{
\item{inputs}{The input parameters of this TrainingJob}
}
\value{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTextSentiment object
}
\description{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTextSentiment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A TrainingJob that trains and uploads an AutoML Text Sentiment Model.
}
\concept{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTextSentiment functions}
