% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTextSentimentInputs}
\alias{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTextSentimentInputs}
\title{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTextSentimentInputs Object}
\usage{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTextSentimentInputs(
  sentimentMax = NULL
)
}
\arguments{
\item{sentimentMax}{A sentiment is expressed as an integer ordinal, where higher value means a more positive sentiment}
}
\value{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTextSentimentInputs object
}
\description{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTextSentimentInputs Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\concept{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlTextSentimentInputs functions}
