% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlVideoActionRecognition}
\alias{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlVideoActionRecognition}
\title{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlVideoActionRecognition Object}
\usage{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlVideoActionRecognition(
  inputs = NULL
)
}
\arguments{
\item{inputs}{The input parameters of this TrainingJob}
}
\value{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlVideoActionRecognition object
}
\description{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlVideoActionRecognition Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A TrainingJob that trains and uploads an AutoML Video Action Recognition Model.
}
\concept{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlVideoActionRecognition functions}
