% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlVideoClassification}
\alias{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlVideoClassification}
\title{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlVideoClassification Object}
\usage{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlVideoClassification(
  inputs = NULL
)
}
\arguments{
\item{inputs}{The input parameters of this TrainingJob}
}
\value{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlVideoClassification object
}
\description{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlVideoClassification Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A TrainingJob that trains and uploads an AutoML Video Classification Model.
}
\concept{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlVideoClassification functions}
