% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlVideoClassificationInputs}
\alias{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlVideoClassificationInputs}
\title{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlVideoClassificationInputs Object}
\usage{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlVideoClassificationInputs(
  modelType = NULL
)
}
\arguments{
\item{modelType}{No description}
}
\value{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlVideoClassificationInputs object
}
\description{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlVideoClassificationInputs Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\concept{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlVideoClassificationInputs functions}
