% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlVideoObjectTracking}
\alias{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlVideoObjectTracking}
\title{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlVideoObjectTracking Object}
\usage{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlVideoObjectTracking(
  inputs = NULL
)
}
\arguments{
\item{inputs}{The input parameters of this TrainingJob}
}
\value{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlVideoObjectTracking object
}
\description{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlVideoObjectTracking Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A TrainingJob that trains and uploads an AutoML Video ObjectTracking Model.
}
\concept{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionAutoMlVideoObjectTracking functions}
