% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionCustomJobMetadata}
\alias{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionCustomJobMetadata}
\title{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionCustomJobMetadata Object}
\usage{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionCustomJobMetadata(
  backingCustomJob = NULL
)
}
\arguments{
\item{backingCustomJob}{The resource name of the CustomJob that has been created to carry out this custom task}
}
\value{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionCustomJobMetadata object
}
\description{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionCustomJobMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\concept{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionCustomJobMetadata functions}
