% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionCustomTask}
\alias{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionCustomTask}
\title{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionCustomTask Object}
\usage{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionCustomTask(
  metadata = NULL,
  inputs = NULL
)
}
\arguments{
\item{metadata}{The metadata information}

\item{inputs}{The input parameters of this CustomTask}
}
\value{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionCustomTask object
}
\description{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionCustomTask Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A TrainingJob that trains a custom code Model.
}
\concept{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionCustomTask functions}
