% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionHierarchyConfig}
\alias{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionHierarchyConfig}
\title{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionHierarchyConfig Object}
\usage{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionHierarchyConfig(
  groupTemporalTotalWeight = NULL,
  groupColumns = NULL,
  groupTotalWeight = NULL,
  temporalTotalWeight = NULL
)
}
\arguments{
\item{groupTemporalTotalWeight}{The weight of the loss for predictions aggregated over both the horizon and time series in the same hierarchy group}

\item{groupColumns}{A list of time series attribute column names that define the time series hierarchy}

\item{groupTotalWeight}{The weight of the loss for predictions aggregated over time series in the same group}

\item{temporalTotalWeight}{The weight of the loss for predictions aggregated over the horizon for a single time series}
}
\value{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionHierarchyConfig object
}
\description{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionHierarchyConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration that defines the hierarchical relationship of time series and parameters for hierarchical forecasting strategies.
}
\concept{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionHierarchyConfig functions}
