% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionHyperparameterTuningJobMetadata}
\alias{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionHyperparameterTuningJobMetadata}
\title{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionHyperparameterTuningJobMetadata Object}
\usage{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionHyperparameterTuningJobMetadata(
  bestTrialBackingCustomJob = NULL,
  backingHyperparameterTuningJob = NULL
)
}
\arguments{
\item{bestTrialBackingCustomJob}{The resource name of the CustomJob that has been created to run the best Trial of this HyperparameterTuning task}

\item{backingHyperparameterTuningJob}{The resource name of the HyperparameterTuningJob that has been created to carry out this HyperparameterTuning task}
}
\value{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionHyperparameterTuningJobMetadata object
}
\description{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionHyperparameterTuningJobMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\concept{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionHyperparameterTuningJobMetadata functions}
