% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionHyperparameterTuningJobSpec}
\alias{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionHyperparameterTuningJobSpec}
\title{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionHyperparameterTuningJobSpec Object}
\usage{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionHyperparameterTuningJobSpec(
  trialJobSpec = NULL,
  maxTrialCount = NULL,
  maxFailedTrialCount = NULL,
  studySpec = NULL,
  parallelTrialCount = NULL
)
}
\arguments{
\item{trialJobSpec}{The spec of a trial job}

\item{maxTrialCount}{The desired total number of Trials}

\item{maxFailedTrialCount}{The number of failed Trials that need to be seen before failing the HyperparameterTuningJob}

\item{studySpec}{Study configuration of the HyperparameterTuningJob}

\item{parallelTrialCount}{The desired number of Trials to run in parallel}
}
\value{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionHyperparameterTuningJobSpec object
}
\description{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionHyperparameterTuningJobSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\concept{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionHyperparameterTuningJobSpec functions}
