% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionHyperparameterTuningTask}
\alias{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionHyperparameterTuningTask}
\title{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionHyperparameterTuningTask Object}
\usage{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionHyperparameterTuningTask(
  inputs = NULL,
  metadata = NULL
)
}
\arguments{
\item{inputs}{The input parameters of this HyperparameterTuningTask}

\item{metadata}{The metadata information}
}
\value{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionHyperparameterTuningTask object
}
\description{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionHyperparameterTuningTask Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A TrainingJob that tunes Hypererparameters of a custom code Model.
}
\concept{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionHyperparameterTuningTask functions}
