% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionSeq2SeqPlusForecastingInputsGranularity}
\alias{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionSeq2SeqPlusForecastingInputsGranularity}
\title{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionSeq2SeqPlusForecastingInputsGranularity Object}
\usage{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionSeq2SeqPlusForecastingInputsGranularity(
  quantity = NULL,
  unit = NULL
)
}
\arguments{
\item{quantity}{The number of granularity_units between data points in the training data}

\item{unit}{The time granularity unit of this time period}
}
\value{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionSeq2SeqPlusForecastingInputsGranularity object
}
\description{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionSeq2SeqPlusForecastingInputsGranularity Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A duration of time expressed in time granularity units.
}
\concept{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionSeq2SeqPlusForecastingInputsGranularity functions}
