% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionSeq2SeqPlusForecastingMetadata}
\alias{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionSeq2SeqPlusForecastingMetadata}
\title{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionSeq2SeqPlusForecastingMetadata Object}
\usage{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionSeq2SeqPlusForecastingMetadata(
  evaluatedDataItemsBigqueryUri = NULL,
  trainCostMilliNodeHours = NULL
)
}
\arguments{
\item{evaluatedDataItemsBigqueryUri}{BigQuery destination uri for exported evaluated examples}

\item{trainCostMilliNodeHours}{Output only}
}
\value{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionSeq2SeqPlusForecastingMetadata object
}
\description{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionSeq2SeqPlusForecastingMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Model metadata specific to Seq2Seq Plus Forecasting.
}
\concept{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionSeq2SeqPlusForecastingMetadata functions}
