% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionTftForecasting}
\alias{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionTftForecasting}
\title{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionTftForecasting Object}
\usage{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionTftForecasting(
  inputs = NULL,
  metadata = NULL
)
}
\arguments{
\item{inputs}{The input parameters of this TrainingJob}

\item{metadata}{The metadata information}
}
\value{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionTftForecasting object
}
\description{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionTftForecasting Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A TrainingJob that trains and uploads an AutoML Forecasting Model.
}
\concept{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionTftForecasting functions}
