% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionTftForecastingInputsGranularity}
\alias{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionTftForecastingInputsGranularity}
\title{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionTftForecastingInputsGranularity Object}
\usage{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionTftForecastingInputsGranularity(
  unit = NULL,
  quantity = NULL
)
}
\arguments{
\item{unit}{The time granularity unit of this time period}

\item{quantity}{The number of granularity_units between data points in the training data}
}
\value{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionTftForecastingInputsGranularity object
}
\description{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionTftForecastingInputsGranularity Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A duration of time expressed in time granularity units.
}
\concept{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionTftForecastingInputsGranularity functions}
