% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionTftForecastingInputsTransformation}
\alias{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionTftForecastingInputsTransformation}
\title{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionTftForecastingInputsTransformation Object}
\usage{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionTftForecastingInputsTransformation(
  timestamp = NULL,
  text = NULL,
  numeric = NULL,
  auto = NULL,
  categorical = NULL
)
}
\arguments{
\item{timestamp}{No description}

\item{text}{No description}

\item{numeric}{No description}

\item{auto}{No description}

\item{categorical}{No description}
}
\value{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionTftForecastingInputsTransformation object
}
\description{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionTftForecastingInputsTransformation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\concept{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionTftForecastingInputsTransformation functions}
