% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionTftForecastingMetadata}
\alias{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionTftForecastingMetadata}
\title{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionTftForecastingMetadata Object}
\usage{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionTftForecastingMetadata(
  evaluatedDataItemsBigqueryUri = NULL,
  trainCostMilliNodeHours = NULL
)
}
\arguments{
\item{evaluatedDataItemsBigqueryUri}{BigQuery destination uri for exported evaluated examples}

\item{trainCostMilliNodeHours}{Output only}
}
\value{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionTftForecastingMetadata object
}
\description{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionTftForecastingMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Model metadata specific to TFT Forecasting.
}
\concept{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionTftForecastingMetadata functions}
