% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionWindowConfig}
\alias{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionWindowConfig}
\title{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionWindowConfig Object}
\usage{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionWindowConfig(
  column = NULL,
  strideLength = NULL,
  maxCount = NULL
)
}
\arguments{
\item{column}{Name of the column that should be used to generate sliding windows}

\item{strideLength}{Stride length used to generate input examples}

\item{maxCount}{Maximum number of windows that should be generated across all time series}
}
\value{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionWindowConfig object
}
\description{
GoogleCloudAiplatformV1SchemaTrainingjobDefinitionWindowConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Config that contains the strategy used to generate sliding windows in time series training. A window is a series of rows that comprise the context up to the time of prediction, and the horizon following. The corresponding row for each window marks the start of the forecast horizon. Each window is used as an input example for training/evaluation.
}
\concept{GoogleCloudAiplatformV1SchemaTrainingjobDefinitionWindowConfig functions}
