% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaVertex}
\alias{GoogleCloudAiplatformV1SchemaVertex}
\title{GoogleCloudAiplatformV1SchemaVertex Object}
\usage{
GoogleCloudAiplatformV1SchemaVertex(y = NULL, x = NULL)
}
\arguments{
\item{y}{Y coordinate}

\item{x}{X coordinate}
}
\value{
GoogleCloudAiplatformV1SchemaVertex object
}
\description{
GoogleCloudAiplatformV1SchemaVertex Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A vertex represents a 2D point in the image. NOTE: the normalized vertex coordinates are relative to the original image and range from 0 to 1.
}
\concept{GoogleCloudAiplatformV1SchemaVertex functions}
