% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaVideoActionRecognitionAnnotation}
\alias{GoogleCloudAiplatformV1SchemaVideoActionRecognitionAnnotation}
\title{GoogleCloudAiplatformV1SchemaVideoActionRecognitionAnnotation Object}
\usage{
GoogleCloudAiplatformV1SchemaVideoActionRecognitionAnnotation(
  annotationSpecId = NULL,
  timeSegment = NULL,
  displayName = NULL
)
}
\arguments{
\item{annotationSpecId}{The resource Id of the AnnotationSpec that this Annotation pertains to}

\item{timeSegment}{This Annotation applies to the time period represented by the TimeSegment}

\item{displayName}{The display name of the AnnotationSpec that this Annotation pertains to}
}
\value{
GoogleCloudAiplatformV1SchemaVideoActionRecognitionAnnotation object
}
\description{
GoogleCloudAiplatformV1SchemaVideoActionRecognitionAnnotation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Annotation details specific to video action recognition.
}
\concept{GoogleCloudAiplatformV1SchemaVideoActionRecognitionAnnotation functions}
