% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaVideoClassificationAnnotation}
\alias{GoogleCloudAiplatformV1SchemaVideoClassificationAnnotation}
\title{GoogleCloudAiplatformV1SchemaVideoClassificationAnnotation Object}
\usage{
GoogleCloudAiplatformV1SchemaVideoClassificationAnnotation(
  annotationSpecId = NULL,
  displayName = NULL,
  timeSegment = NULL
)
}
\arguments{
\item{annotationSpecId}{The resource Id of the AnnotationSpec that this Annotation pertains to}

\item{displayName}{The display name of the AnnotationSpec that this Annotation pertains to}

\item{timeSegment}{This Annotation applies to the time period represented by the TimeSegment}
}
\value{
GoogleCloudAiplatformV1SchemaVideoClassificationAnnotation object
}
\description{
GoogleCloudAiplatformV1SchemaVideoClassificationAnnotation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Annotation details specific to video classification.
}
\concept{GoogleCloudAiplatformV1SchemaVideoClassificationAnnotation functions}
