% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaVideoDataItem}
\alias{GoogleCloudAiplatformV1SchemaVideoDataItem}
\title{GoogleCloudAiplatformV1SchemaVideoDataItem Object}
\usage{
GoogleCloudAiplatformV1SchemaVideoDataItem(gcsUri = NULL)
}
\arguments{
\item{gcsUri}{Required}
}
\value{
GoogleCloudAiplatformV1SchemaVideoDataItem object
}
\description{
GoogleCloudAiplatformV1SchemaVideoDataItem Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Payload of Video DataItem.
}
\concept{GoogleCloudAiplatformV1SchemaVideoDataItem functions}
