% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaVideoDatasetMetadata}
\alias{GoogleCloudAiplatformV1SchemaVideoDatasetMetadata}
\title{GoogleCloudAiplatformV1SchemaVideoDatasetMetadata Object}
\usage{
GoogleCloudAiplatformV1SchemaVideoDatasetMetadata(
  gcsBucket = NULL,
  dataItemSchemaUri = NULL
)
}
\arguments{
\item{gcsBucket}{Google Cloud Storage Bucket name that contains the blob data of this Dataset}

\item{dataItemSchemaUri}{Points to a YAML file stored on Google Cloud Storage describing payload of the Video DataItems that belong to this Dataset}
}
\value{
GoogleCloudAiplatformV1SchemaVideoDatasetMetadata object
}
\description{
GoogleCloudAiplatformV1SchemaVideoDatasetMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The metadata of Datasets that contain Video DataItems.
}
\concept{GoogleCloudAiplatformV1SchemaVideoDatasetMetadata functions}
