% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaVideoObjectTrackingAnnotation}
\alias{GoogleCloudAiplatformV1SchemaVideoObjectTrackingAnnotation}
\title{GoogleCloudAiplatformV1SchemaVideoObjectTrackingAnnotation Object}
\usage{
GoogleCloudAiplatformV1SchemaVideoObjectTrackingAnnotation(
  instanceId = NULL,
  xMin = NULL,
  displayName = NULL,
  yMax = NULL,
  yMin = NULL,
  timeOffset = NULL,
  annotationSpecId = NULL,
  xMax = NULL
)
}
\arguments{
\item{instanceId}{The instance of the object, expressed as a positive integer}

\item{xMin}{The leftmost coordinate of the bounding box}

\item{displayName}{The display name of the AnnotationSpec that this Annotation pertains to}

\item{yMax}{The bottommost coordinate of the bounding box}

\item{yMin}{The topmost coordinate of the bounding box}

\item{timeOffset}{A time (frame) of a video to which this annotation pertains}

\item{annotationSpecId}{The resource Id of the AnnotationSpec that this Annotation pertains to}

\item{xMax}{The rightmost coordinate of the bounding box}
}
\value{
GoogleCloudAiplatformV1SchemaVideoObjectTrackingAnnotation object
}
\description{
GoogleCloudAiplatformV1SchemaVideoObjectTrackingAnnotation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Annotation details specific to video object tracking.
}
\concept{GoogleCloudAiplatformV1SchemaVideoObjectTrackingAnnotation functions}
