% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaVisualInspectionClassificationLabelSavedQueryMetadata}
\alias{GoogleCloudAiplatformV1SchemaVisualInspectionClassificationLabelSavedQueryMetadata}
\title{GoogleCloudAiplatformV1SchemaVisualInspectionClassificationLabelSavedQueryMetadata Object}
\usage{
GoogleCloudAiplatformV1SchemaVisualInspectionClassificationLabelSavedQueryMetadata(
  multiLabel = NULL
)
}
\arguments{
\item{multiLabel}{Whether or not the classification label is multi_label}
}
\value{
GoogleCloudAiplatformV1SchemaVisualInspectionClassificationLabelSavedQueryMetadata object
}
\description{
GoogleCloudAiplatformV1SchemaVisualInspectionClassificationLabelSavedQueryMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\concept{GoogleCloudAiplatformV1SchemaVisualInspectionClassificationLabelSavedQueryMetadata functions}
