% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SchemaVisualInspectionMaskSavedQueryMetadata}
\alias{GoogleCloudAiplatformV1SchemaVisualInspectionMaskSavedQueryMetadata}
\title{GoogleCloudAiplatformV1SchemaVisualInspectionMaskSavedQueryMetadata Object}
\usage{
GoogleCloudAiplatformV1SchemaVisualInspectionMaskSavedQueryMetadata()
}
\value{
GoogleCloudAiplatformV1SchemaVisualInspectionMaskSavedQueryMetadata object
}
\description{
GoogleCloudAiplatformV1SchemaVisualInspectionMaskSavedQueryMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\concept{GoogleCloudAiplatformV1SchemaVisualInspectionMaskSavedQueryMetadata functions}
