% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SearchDataItemsResponse}
\alias{GoogleCloudAiplatformV1SearchDataItemsResponse}
\title{GoogleCloudAiplatformV1SearchDataItemsResponse Object}
\usage{
GoogleCloudAiplatformV1SearchDataItemsResponse(
  dataItemViews = NULL,
  nextPageToken = NULL
)
}
\arguments{
\item{dataItemViews}{The DataItemViews read}

\item{nextPageToken}{A token to retrieve next page of results}
}
\value{
GoogleCloudAiplatformV1SearchDataItemsResponse object
}
\description{
GoogleCloudAiplatformV1SearchDataItemsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for DatasetService.SearchDataItems.
}
\concept{GoogleCloudAiplatformV1SearchDataItemsResponse functions}
