% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SearchFeaturesResponse}
\alias{GoogleCloudAiplatformV1SearchFeaturesResponse}
\title{GoogleCloudAiplatformV1SearchFeaturesResponse Object}
\usage{
GoogleCloudAiplatformV1SearchFeaturesResponse(
  features = NULL,
  nextPageToken = NULL
)
}
\arguments{
\item{features}{The Features matching the request}

\item{nextPageToken}{A token, which can be sent as SearchFeaturesRequest}
}
\value{
GoogleCloudAiplatformV1SearchFeaturesResponse object
}
\description{
GoogleCloudAiplatformV1SearchFeaturesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for FeaturestoreService.SearchFeatures.
}
\concept{GoogleCloudAiplatformV1SearchFeaturesResponse functions}
