% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SearchMigratableResourcesRequest}
\alias{GoogleCloudAiplatformV1SearchMigratableResourcesRequest}
\title{GoogleCloudAiplatformV1SearchMigratableResourcesRequest Object}
\usage{
GoogleCloudAiplatformV1SearchMigratableResourcesRequest(
  pageSize = NULL,
  filter = NULL,
  pageToken = NULL
)
}
\arguments{
\item{pageSize}{The standard page size}

\item{filter}{A filter for your search}

\item{pageToken}{The standard page token}
}
\value{
GoogleCloudAiplatformV1SearchMigratableResourcesRequest object
}
\description{
GoogleCloudAiplatformV1SearchMigratableResourcesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for MigrationService.SearchMigratableResources.
}
\seealso{
Other GoogleCloudAiplatformV1SearchMigratableResourcesRequest functions: 
\code{\link{projects.locations.migratableResources.search}()}
}
\concept{GoogleCloudAiplatformV1SearchMigratableResourcesRequest functions}
