% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SearchMigratableResourcesResponse}
\alias{GoogleCloudAiplatformV1SearchMigratableResourcesResponse}
\title{GoogleCloudAiplatformV1SearchMigratableResourcesResponse Object}
\usage{
GoogleCloudAiplatformV1SearchMigratableResourcesResponse(
  migratableResources = NULL,
  nextPageToken = NULL
)
}
\arguments{
\item{migratableResources}{All migratable resources that can be migrated to the location specified in the request}

\item{nextPageToken}{The standard next-page token}
}
\value{
GoogleCloudAiplatformV1SearchMigratableResourcesResponse object
}
\description{
GoogleCloudAiplatformV1SearchMigratableResourcesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for MigrationService.SearchMigratableResources.
}
\concept{GoogleCloudAiplatformV1SearchMigratableResourcesResponse functions}
