% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SearchModelDeploymentMonitoringStatsAnomaliesRequest}
\alias{GoogleCloudAiplatformV1SearchModelDeploymentMonitoringStatsAnomaliesRequest}
\title{GoogleCloudAiplatformV1SearchModelDeploymentMonitoringStatsAnomaliesRequest Object}
\usage{
GoogleCloudAiplatformV1SearchModelDeploymentMonitoringStatsAnomaliesRequest(
  endTime = NULL,
  pageSize = NULL,
  objectives = NULL,
  deployedModelId = NULL,
  featureDisplayName = NULL,
  pageToken = NULL,
  startTime = NULL
)
}
\arguments{
\item{endTime}{The latest timestamp of stats being generated}

\item{pageSize}{The standard list page size}

\item{objectives}{Required}

\item{deployedModelId}{Required}

\item{featureDisplayName}{The feature display name}

\item{pageToken}{A page token received from a previous JobService}

\item{startTime}{The earliest timestamp of stats being generated}
}
\value{
GoogleCloudAiplatformV1SearchModelDeploymentMonitoringStatsAnomaliesRequest object
}
\description{
GoogleCloudAiplatformV1SearchModelDeploymentMonitoringStatsAnomaliesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for JobService.SearchModelDeploymentMonitoringStatsAnomalies.
}
\seealso{
Other GoogleCloudAiplatformV1SearchModelDeploymentMonitoringStatsAnomaliesRequest functions: 
\code{\link{projects.locations.modelDeploymentMonitoringJobs.searchModelDeploymentMonitoringStatsAnomalies}()}
}
\concept{GoogleCloudAiplatformV1SearchModelDeploymentMonitoringStatsAnomaliesRequest functions}
