% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SearchModelDeploymentMonitoringStatsAnomaliesRequestStatsAnomaliesObjective}
\alias{GoogleCloudAiplatformV1SearchModelDeploymentMonitoringStatsAnomaliesRequestStatsAnomaliesObjective}
\title{GoogleCloudAiplatformV1SearchModelDeploymentMonitoringStatsAnomaliesRequestStatsAnomaliesObjective Object}
\usage{
GoogleCloudAiplatformV1SearchModelDeploymentMonitoringStatsAnomaliesRequestStatsAnomaliesObjective(
  type = NULL,
  topFeatureCount = NULL
)
}
\arguments{
\item{type}{No description}

\item{topFeatureCount}{If set, all attribution scores between SearchModelDeploymentMonitoringStatsAnomaliesRequest}
}
\value{
GoogleCloudAiplatformV1SearchModelDeploymentMonitoringStatsAnomaliesRequestStatsAnomaliesObjective object
}
\description{
GoogleCloudAiplatformV1SearchModelDeploymentMonitoringStatsAnomaliesRequestStatsAnomaliesObjective Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Stats requested for specific objective.
}
\concept{GoogleCloudAiplatformV1SearchModelDeploymentMonitoringStatsAnomaliesRequestStatsAnomaliesObjective functions}
