% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SearchModelDeploymentMonitoringStatsAnomaliesResponse}
\alias{GoogleCloudAiplatformV1SearchModelDeploymentMonitoringStatsAnomaliesResponse}
\title{GoogleCloudAiplatformV1SearchModelDeploymentMonitoringStatsAnomaliesResponse Object}
\usage{
GoogleCloudAiplatformV1SearchModelDeploymentMonitoringStatsAnomaliesResponse(
  nextPageToken = NULL,
  monitoringStats = NULL
)
}
\arguments{
\item{nextPageToken}{The page token that can be used by the next JobService}

\item{monitoringStats}{Stats retrieved for requested objectives}
}
\value{
GoogleCloudAiplatformV1SearchModelDeploymentMonitoringStatsAnomaliesResponse object
}
\description{
GoogleCloudAiplatformV1SearchModelDeploymentMonitoringStatsAnomaliesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for JobService.SearchModelDeploymentMonitoringStatsAnomalies.
}
\concept{GoogleCloudAiplatformV1SearchModelDeploymentMonitoringStatsAnomaliesResponse functions}
