% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SmoothGradConfig}
\alias{GoogleCloudAiplatformV1SmoothGradConfig}
\title{GoogleCloudAiplatformV1SmoothGradConfig Object}
\usage{
GoogleCloudAiplatformV1SmoothGradConfig(
  featureNoiseSigma = NULL,
  noisySampleCount = NULL,
  noiseSigma = NULL
)
}
\arguments{
\item{featureNoiseSigma}{This is similar to noise_sigma, but provides additional flexibility}

\item{noisySampleCount}{The number of gradient samples to use for approximation}

\item{noiseSigma}{This is a single float value and will be used to add noise to all the features}
}
\value{
GoogleCloudAiplatformV1SmoothGradConfig object
}
\description{
GoogleCloudAiplatformV1SmoothGradConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Config for SmoothGrad approximation of gradients. When enabled, the gradients are approximated by averaging the gradients from noisy samples in the vicinity of the inputs. Adding noise can help improve the computed gradients. Refer to this paper for more details: https://arxiv.org/pdf/1706.03825.pdf
}
\concept{GoogleCloudAiplatformV1SmoothGradConfig functions}
