% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SpecialistPool}
\alias{GoogleCloudAiplatformV1SpecialistPool}
\title{GoogleCloudAiplatformV1SpecialistPool Object}
\usage{
GoogleCloudAiplatformV1SpecialistPool(
  specialistWorkerEmails = NULL,
  specialistManagerEmails = NULL,
  name = NULL,
  displayName = NULL
)
}
\arguments{
\item{specialistWorkerEmails}{The email addresses of workers in the SpecialistPool}

\item{specialistManagerEmails}{The email addresses of the managers in the SpecialistPool}

\item{name}{Required}

\item{displayName}{Required}
}
\value{
GoogleCloudAiplatformV1SpecialistPool object
}
\description{
GoogleCloudAiplatformV1SpecialistPool Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
SpecialistPool represents customers' own workforce to work on their data labeling jobs. It includes a group of specialist managers and workers. Managers are responsible for managing the workers in this pool as well as customers' data labeling jobs associated with this pool. Customers create specialist pool as well as start data labeling jobs on Cloud, managers and workers handle the jobs using CrowdCompute console.
}
\seealso{
Other GoogleCloudAiplatformV1SpecialistPool functions: 
\code{\link{projects.locations.specialistPools.create}()},
\code{\link{projects.locations.specialistPools.patch}()}
}
\concept{GoogleCloudAiplatformV1SpecialistPool functions}
