% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1StopTrialRequest}
\alias{GoogleCloudAiplatformV1StopTrialRequest}
\title{GoogleCloudAiplatformV1StopTrialRequest Object}
\usage{
GoogleCloudAiplatformV1StopTrialRequest()
}
\value{
GoogleCloudAiplatformV1StopTrialRequest object
}
\description{
GoogleCloudAiplatformV1StopTrialRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for VizierService.StopTrial.
}
\seealso{
Other GoogleCloudAiplatformV1StopTrialRequest functions: 
\code{\link{projects.locations.studies.trials.stop}()}
}
\concept{GoogleCloudAiplatformV1StopTrialRequest functions}
