% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1StreamingPredictRequest}
\alias{GoogleCloudAiplatformV1StreamingPredictRequest}
\title{GoogleCloudAiplatformV1StreamingPredictRequest Object}
\usage{
GoogleCloudAiplatformV1StreamingPredictRequest(
  inputs = NULL,
  parameters = NULL
)
}
\arguments{
\item{inputs}{The prediction input}

\item{parameters}{The parameters that govern the prediction}
}
\value{
GoogleCloudAiplatformV1StreamingPredictRequest object
}
\description{
GoogleCloudAiplatformV1StreamingPredictRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for PredictionService.StreamingPredict. The first message must contain endpoint field and optionally input. The subsequent messages must contain input.
}
\seealso{
Other GoogleCloudAiplatformV1StreamingPredictRequest functions: 
\code{\link{projects.locations.endpoints.serverStreamingPredict}()},
\code{\link{projects.locations.publishers.models.serverStreamingPredict}()}
}
\concept{GoogleCloudAiplatformV1StreamingPredictRequest functions}
