% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1StreamingPredictResponse}
\alias{GoogleCloudAiplatformV1StreamingPredictResponse}
\title{GoogleCloudAiplatformV1StreamingPredictResponse Object}
\usage{
GoogleCloudAiplatformV1StreamingPredictResponse(
  outputs = NULL,
  parameters = NULL
)
}
\arguments{
\item{outputs}{The prediction output}

\item{parameters}{The parameters that govern the prediction}
}
\value{
GoogleCloudAiplatformV1StreamingPredictResponse object
}
\description{
GoogleCloudAiplatformV1StreamingPredictResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for PredictionService.StreamingPredict.
}
\concept{GoogleCloudAiplatformV1StreamingPredictResponse functions}
