% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1StreamingReadFeatureValuesRequest}
\alias{GoogleCloudAiplatformV1StreamingReadFeatureValuesRequest}
\title{GoogleCloudAiplatformV1StreamingReadFeatureValuesRequest Object}
\usage{
GoogleCloudAiplatformV1StreamingReadFeatureValuesRequest(
  featureSelector = NULL,
  entityIds = NULL
)
}
\arguments{
\item{featureSelector}{Required}

\item{entityIds}{Required}
}
\value{
GoogleCloudAiplatformV1StreamingReadFeatureValuesRequest object
}
\description{
GoogleCloudAiplatformV1StreamingReadFeatureValuesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for FeaturestoreOnlineServingService.StreamingFeatureValuesRead.
}
\seealso{
Other GoogleCloudAiplatformV1StreamingReadFeatureValuesRequest functions: 
\code{\link{projects.locations.featurestores.entityTypes.streamingReadFeatureValues}()}
}
\concept{GoogleCloudAiplatformV1StreamingReadFeatureValuesRequest functions}
