% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1StringArray}
\alias{GoogleCloudAiplatformV1StringArray}
\title{GoogleCloudAiplatformV1StringArray Object}
\usage{
GoogleCloudAiplatformV1StringArray(values = NULL)
}
\arguments{
\item{values}{A list of string values}
}
\value{
GoogleCloudAiplatformV1StringArray object
}
\description{
GoogleCloudAiplatformV1StringArray Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of string values.
}
\concept{GoogleCloudAiplatformV1StringArray functions}
