% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1Study}
\alias{GoogleCloudAiplatformV1Study}
\title{GoogleCloudAiplatformV1Study Object}
\usage{
GoogleCloudAiplatformV1Study(displayName = NULL, studySpec = NULL)
}
\arguments{
\item{displayName}{Required}

\item{studySpec}{Required}
}
\value{
GoogleCloudAiplatformV1Study object
}
\description{
GoogleCloudAiplatformV1Study Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A message representing a Study.
}
\seealso{
Other GoogleCloudAiplatformV1Study functions: 
\code{\link{projects.locations.studies.create}()}
}
\concept{GoogleCloudAiplatformV1Study functions}
