% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1StudySpec}
\alias{GoogleCloudAiplatformV1StudySpec}
\title{GoogleCloudAiplatformV1StudySpec Object}
\usage{
GoogleCloudAiplatformV1StudySpec(
  observationNoise = NULL,
  decayCurveStoppingSpec = NULL,
  measurementSelectionType = NULL,
  medianAutomatedStoppingSpec = NULL,
  metrics = NULL,
  convexAutomatedStoppingSpec = NULL,
  algorithm = NULL,
  parameters = NULL
)
}
\arguments{
\item{observationNoise}{The observation noise level of the study}

\item{decayCurveStoppingSpec}{The automated early stopping spec using decay curve rule}

\item{measurementSelectionType}{Describe which measurement selection type will be used}

\item{medianAutomatedStoppingSpec}{The automated early stopping spec using median rule}

\item{metrics}{Required}

\item{convexAutomatedStoppingSpec}{The automated early stopping spec using convex stopping rule}

\item{algorithm}{The search algorithm specified for the Study}

\item{parameters}{Required}
}
\value{
GoogleCloudAiplatformV1StudySpec object
}
\description{
GoogleCloudAiplatformV1StudySpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents specification of a Study.
}
\concept{GoogleCloudAiplatformV1StudySpec functions}
