% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1StudySpecDecayCurveAutomatedStoppingSpec}
\alias{GoogleCloudAiplatformV1StudySpecDecayCurveAutomatedStoppingSpec}
\title{GoogleCloudAiplatformV1StudySpecDecayCurveAutomatedStoppingSpec Object}
\usage{
GoogleCloudAiplatformV1StudySpecDecayCurveAutomatedStoppingSpec(
  useElapsedDuration = NULL
)
}
\arguments{
\item{useElapsedDuration}{True if Measurement}
}
\value{
GoogleCloudAiplatformV1StudySpecDecayCurveAutomatedStoppingSpec object
}
\description{
GoogleCloudAiplatformV1StudySpecDecayCurveAutomatedStoppingSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The decay curve automated stopping rule builds a Gaussian Process Regressor to predict the final objective value of a Trial based on the already completed Trials and the intermediate measurements of the current Trial. Early stopping is requested for the current Trial if there is very low probability to exceed the optimal value found so far.
}
\concept{GoogleCloudAiplatformV1StudySpecDecayCurveAutomatedStoppingSpec functions}
