% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1StudySpecMedianAutomatedStoppingSpec}
\alias{GoogleCloudAiplatformV1StudySpecMedianAutomatedStoppingSpec}
\title{GoogleCloudAiplatformV1StudySpecMedianAutomatedStoppingSpec Object}
\usage{
GoogleCloudAiplatformV1StudySpecMedianAutomatedStoppingSpec(
  useElapsedDuration = NULL
)
}
\arguments{
\item{useElapsedDuration}{True if median automated stopping rule applies on Measurement}
}
\value{
GoogleCloudAiplatformV1StudySpecMedianAutomatedStoppingSpec object
}
\description{
GoogleCloudAiplatformV1StudySpecMedianAutomatedStoppingSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The median automated stopping rule stops a pending Trial if the Trial's best objective_value is strictly below the median 'performance' of all completed Trials reported up to the Trial's last measurement. Currently, 'performance' refers to the running average of the objective values reported by the Trial in each measurement.
}
\concept{GoogleCloudAiplatformV1StudySpecMedianAutomatedStoppingSpec functions}
