% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1StudySpecMetricSpec}
\alias{GoogleCloudAiplatformV1StudySpecMetricSpec}
\title{GoogleCloudAiplatformV1StudySpecMetricSpec Object}
\usage{
GoogleCloudAiplatformV1StudySpecMetricSpec(
  safetyConfig = NULL,
  goal = NULL,
  metricId = NULL
)
}
\arguments{
\item{safetyConfig}{Used for safe search}

\item{goal}{Required}

\item{metricId}{Required}
}
\value{
GoogleCloudAiplatformV1StudySpecMetricSpec object
}
\description{
GoogleCloudAiplatformV1StudySpecMetricSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a metric to optimize.
}
\concept{GoogleCloudAiplatformV1StudySpecMetricSpec functions}
