% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1StudySpecParameterSpec}
\alias{GoogleCloudAiplatformV1StudySpecParameterSpec}
\title{GoogleCloudAiplatformV1StudySpecParameterSpec Object}
\usage{
GoogleCloudAiplatformV1StudySpecParameterSpec(
  doubleValueSpec = NULL,
  categoricalValueSpec = NULL,
  conditionalParameterSpecs = NULL,
  scaleType = NULL,
  parameterId = NULL,
  discreteValueSpec = NULL,
  integerValueSpec = NULL
)
}
\arguments{
\item{doubleValueSpec}{The value spec for a 'DOUBLE' parameter}

\item{categoricalValueSpec}{The value spec for a 'CATEGORICAL' parameter}

\item{conditionalParameterSpecs}{A conditional parameter node is active if the parameter's value matches the conditional node's parent_value_condition}

\item{scaleType}{How the parameter should be scaled}

\item{parameterId}{Required}

\item{discreteValueSpec}{The value spec for a 'DISCRETE' parameter}

\item{integerValueSpec}{The value spec for an 'INTEGER' parameter}
}
\value{
GoogleCloudAiplatformV1StudySpecParameterSpec object
}
\description{
GoogleCloudAiplatformV1StudySpecParameterSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a single parameter to optimize.
}
\concept{GoogleCloudAiplatformV1StudySpecParameterSpec functions}
