% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1StudySpecParameterSpecCategoricalValueSpec}
\alias{GoogleCloudAiplatformV1StudySpecParameterSpecCategoricalValueSpec}
\title{GoogleCloudAiplatformV1StudySpecParameterSpecCategoricalValueSpec Object}
\usage{
GoogleCloudAiplatformV1StudySpecParameterSpecCategoricalValueSpec(
  values = NULL,
  defaultValue = NULL
)
}
\arguments{
\item{values}{Required}

\item{defaultValue}{A default value for a \code{CATEGORICAL} parameter that is assumed to be a relatively good starting point}
}
\value{
GoogleCloudAiplatformV1StudySpecParameterSpecCategoricalValueSpec object
}
\description{
GoogleCloudAiplatformV1StudySpecParameterSpecCategoricalValueSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Value specification for a parameter in \code{CATEGORICAL} type.
}
\concept{GoogleCloudAiplatformV1StudySpecParameterSpecCategoricalValueSpec functions}
