% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1StudySpecParameterSpecConditionalParameterSpec}
\alias{GoogleCloudAiplatformV1StudySpecParameterSpecConditionalParameterSpec}
\title{GoogleCloudAiplatformV1StudySpecParameterSpecConditionalParameterSpec Object}
\usage{
GoogleCloudAiplatformV1StudySpecParameterSpecConditionalParameterSpec(
  parentIntValues = NULL,
  parameterSpec = NULL,
  parentCategoricalValues = NULL,
  parentDiscreteValues = NULL
)
}
\arguments{
\item{parentIntValues}{The spec for matching values from a parent parameter of \code{INTEGER} type}

\item{parameterSpec}{Required}

\item{parentCategoricalValues}{The spec for matching values from a parent parameter of \code{CATEGORICAL} type}

\item{parentDiscreteValues}{The spec for matching values from a parent parameter of \code{DISCRETE} type}
}
\value{
GoogleCloudAiplatformV1StudySpecParameterSpecConditionalParameterSpec object
}
\description{
GoogleCloudAiplatformV1StudySpecParameterSpecConditionalParameterSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a parameter spec with condition from its parent parameter.
}
\concept{GoogleCloudAiplatformV1StudySpecParameterSpecConditionalParameterSpec functions}
