% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1StudySpecParameterSpecDiscreteValueSpec}
\alias{GoogleCloudAiplatformV1StudySpecParameterSpecDiscreteValueSpec}
\title{GoogleCloudAiplatformV1StudySpecParameterSpecDiscreteValueSpec Object}
\usage{
GoogleCloudAiplatformV1StudySpecParameterSpecDiscreteValueSpec(
  defaultValue = NULL,
  values = NULL
)
}
\arguments{
\item{defaultValue}{A default value for a \code{DISCRETE} parameter that is assumed to be a relatively good starting point}

\item{values}{Required}
}
\value{
GoogleCloudAiplatformV1StudySpecParameterSpecDiscreteValueSpec object
}
\description{
GoogleCloudAiplatformV1StudySpecParameterSpecDiscreteValueSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Value specification for a parameter in \code{DISCRETE} type.
}
\concept{GoogleCloudAiplatformV1StudySpecParameterSpecDiscreteValueSpec functions}
