% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1StudySpecParameterSpecDoubleValueSpec}
\alias{GoogleCloudAiplatformV1StudySpecParameterSpecDoubleValueSpec}
\title{GoogleCloudAiplatformV1StudySpecParameterSpecDoubleValueSpec Object}
\usage{
GoogleCloudAiplatformV1StudySpecParameterSpecDoubleValueSpec(
  maxValue = NULL,
  defaultValue = NULL,
  minValue = NULL
)
}
\arguments{
\item{maxValue}{Required}

\item{defaultValue}{A default value for a \code{DOUBLE} parameter that is assumed to be a relatively good starting point}

\item{minValue}{Required}
}
\value{
GoogleCloudAiplatformV1StudySpecParameterSpecDoubleValueSpec object
}
\description{
GoogleCloudAiplatformV1StudySpecParameterSpecDoubleValueSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Value specification for a parameter in \code{DOUBLE} type.
}
\concept{GoogleCloudAiplatformV1StudySpecParameterSpecDoubleValueSpec functions}
