% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1StudySpecParameterSpecIntegerValueSpec}
\alias{GoogleCloudAiplatformV1StudySpecParameterSpecIntegerValueSpec}
\title{GoogleCloudAiplatformV1StudySpecParameterSpecIntegerValueSpec Object}
\usage{
GoogleCloudAiplatformV1StudySpecParameterSpecIntegerValueSpec(
  maxValue = NULL,
  minValue = NULL,
  defaultValue = NULL
)
}
\arguments{
\item{maxValue}{Required}

\item{minValue}{Required}

\item{defaultValue}{A default value for an \code{INTEGER} parameter that is assumed to be a relatively good starting point}
}
\value{
GoogleCloudAiplatformV1StudySpecParameterSpecIntegerValueSpec object
}
\description{
GoogleCloudAiplatformV1StudySpecParameterSpecIntegerValueSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Value specification for a parameter in \code{INTEGER} type.
}
\concept{GoogleCloudAiplatformV1StudySpecParameterSpecIntegerValueSpec functions}
