% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SuggestTrialsMetadata}
\alias{GoogleCloudAiplatformV1SuggestTrialsMetadata}
\title{GoogleCloudAiplatformV1SuggestTrialsMetadata Object}
\usage{
GoogleCloudAiplatformV1SuggestTrialsMetadata(
  genericMetadata = NULL,
  clientId = NULL
)
}
\arguments{
\item{genericMetadata}{Operation metadata for suggesting Trials}

\item{clientId}{The identifier of the client that is requesting the suggestion}
}
\value{
GoogleCloudAiplatformV1SuggestTrialsMetadata object
}
\description{
GoogleCloudAiplatformV1SuggestTrialsMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details of operations that perform Trials suggestion.
}
\concept{GoogleCloudAiplatformV1SuggestTrialsMetadata functions}
