% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SuggestTrialsRequest}
\alias{GoogleCloudAiplatformV1SuggestTrialsRequest}
\title{GoogleCloudAiplatformV1SuggestTrialsRequest Object}
\usage{
GoogleCloudAiplatformV1SuggestTrialsRequest(
  clientId = NULL,
  suggestionCount = NULL
)
}
\arguments{
\item{clientId}{Required}

\item{suggestionCount}{Required}
}
\value{
GoogleCloudAiplatformV1SuggestTrialsRequest object
}
\description{
GoogleCloudAiplatformV1SuggestTrialsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for VizierService.SuggestTrials.
}
\seealso{
Other GoogleCloudAiplatformV1SuggestTrialsRequest functions: 
\code{\link{projects.locations.studies.trials.suggest}()}
}
\concept{GoogleCloudAiplatformV1SuggestTrialsRequest functions}
