% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1SuggestTrialsResponse}
\alias{GoogleCloudAiplatformV1SuggestTrialsResponse}
\title{GoogleCloudAiplatformV1SuggestTrialsResponse Object}
\usage{
GoogleCloudAiplatformV1SuggestTrialsResponse(
  endTime = NULL,
  trials = NULL,
  startTime = NULL,
  studyState = NULL
)
}
\arguments{
\item{endTime}{The time at which operation processing completed}

\item{trials}{A list of Trials}

\item{startTime}{The time at which the operation was started}

\item{studyState}{The state of the Study}
}
\value{
GoogleCloudAiplatformV1SuggestTrialsResponse object
}
\description{
GoogleCloudAiplatformV1SuggestTrialsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for VizierService.SuggestTrials.
}
\concept{GoogleCloudAiplatformV1SuggestTrialsResponse functions}
