% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1Tensor}
\alias{GoogleCloudAiplatformV1Tensor}
\title{GoogleCloudAiplatformV1Tensor Object}
\usage{
GoogleCloudAiplatformV1Tensor(
  GoogleCloudAiplatformV1Tensor.structVal = NULL,
  structVal = NULL,
  dtype = NULL,
  floatVal = NULL,
  shape = NULL,
  intVal = NULL,
  doubleVal = NULL,
  uint64Val = NULL,
  boolVal = NULL,
  tensorVal = NULL,
  stringVal = NULL,
  bytesVal = NULL,
  listVal = NULL,
  int64Val = NULL,
  uintVal = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1Tensor.structVal}{The \link{GoogleCloudAiplatformV1Tensor.structVal} object or list of objects}

\item{structVal}{A map of string to tensor}

\item{dtype}{The data type of tensor}

\item{floatVal}{FLOAT}

\item{shape}{Shape of the tensor}

\item{intVal}{INT_8 INT_16 INT_32}

\item{doubleVal}{DOUBLE}

\item{uint64Val}{UINT64}

\item{boolVal}{Type specific representations that make it easy to create tensor protos in all languages}

\item{tensorVal}{Serialized raw tensor content}

\item{stringVal}{STRING}

\item{bytesVal}{STRING}

\item{listVal}{A list of tensor values}

\item{int64Val}{INT64}

\item{uintVal}{UINT8 UINT16 UINT32}
}
\value{
GoogleCloudAiplatformV1Tensor object
}
\description{
GoogleCloudAiplatformV1Tensor Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A tensor value type.
}
\seealso{
Other GoogleCloudAiplatformV1Tensor functions: 
\code{\link{GoogleCloudAiplatformV1Tensor.structVal}()}
}
\concept{GoogleCloudAiplatformV1Tensor functions}
