% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1Tensor.structVal}
\alias{GoogleCloudAiplatformV1Tensor.structVal}
\title{GoogleCloudAiplatformV1Tensor.structVal Object}
\usage{
GoogleCloudAiplatformV1Tensor.structVal()
}
\value{
GoogleCloudAiplatformV1Tensor.structVal object
}
\description{
GoogleCloudAiplatformV1Tensor.structVal Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A map of string to tensor.
}
\seealso{
Other GoogleCloudAiplatformV1Tensor functions: 
\code{\link{GoogleCloudAiplatformV1Tensor}()}
}
\concept{GoogleCloudAiplatformV1Tensor functions}
