% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1Tensorboard.labels}
\alias{GoogleCloudAiplatformV1Tensorboard.labels}
\title{GoogleCloudAiplatformV1Tensorboard.labels Object}
\usage{
GoogleCloudAiplatformV1Tensorboard.labels()
}
\value{
GoogleCloudAiplatformV1Tensorboard.labels object
}
\description{
GoogleCloudAiplatformV1Tensorboard.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The labels with user-defined metadata to organize your Tensorboards. Label keys and values can be no longer than 64 characters (Unicode codepoints), can only contain lowercase letters, numeric characters, underscores and dashes. International characters are allowed. No more than 64 user labels can be associated with one Tensorboard (System labels are excluded). See https://goo.gl/xmQnxf for more information and examples of labels. System reserved label keys are prefixed with 'aiplatform.googleapis.com/' and are immutable.
}
\seealso{
Other GoogleCloudAiplatformV1Tensorboard functions: 
\code{\link{GoogleCloudAiplatformV1Tensorboard}()},
\code{\link{projects.locations.tensorboards.create}()},
\code{\link{projects.locations.tensorboards.patch}()}
}
\concept{GoogleCloudAiplatformV1Tensorboard functions}
