% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1Tensorboard}
\alias{GoogleCloudAiplatformV1Tensorboard}
\title{GoogleCloudAiplatformV1Tensorboard Object}
\usage{
GoogleCloudAiplatformV1Tensorboard(
  GoogleCloudAiplatformV1Tensorboard.labels = NULL,
  etag = NULL,
  encryptionSpec = NULL,
  displayName = NULL,
  isDefault = NULL,
  description = NULL,
  labels = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1Tensorboard.labels}{The \link{GoogleCloudAiplatformV1Tensorboard.labels} object or list of objects}

\item{etag}{Used to perform a consistent read-modify-write updates}

\item{encryptionSpec}{Customer-managed encryption key spec for a Tensorboard}

\item{displayName}{Required}

\item{isDefault}{Used to indicate if the TensorBoard instance is the default one}

\item{description}{Description of this Tensorboard}

\item{labels}{The labels with user-defined metadata to organize your Tensorboards}
}
\value{
GoogleCloudAiplatformV1Tensorboard object
}
\description{
GoogleCloudAiplatformV1Tensorboard Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Tensorboard is a physical database that stores users' training metrics. A default Tensorboard is provided in each region of a Google Cloud project. If needed users can also create extra Tensorboards in their projects.
}
\seealso{
Other GoogleCloudAiplatformV1Tensorboard functions: 
\code{\link{GoogleCloudAiplatformV1Tensorboard.labels}()},
\code{\link{projects.locations.tensorboards.create}()},
\code{\link{projects.locations.tensorboards.patch}()}
}
\concept{GoogleCloudAiplatformV1Tensorboard functions}
