% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1TensorboardBlob}
\alias{GoogleCloudAiplatformV1TensorboardBlob}
\title{GoogleCloudAiplatformV1TensorboardBlob Object}
\usage{
GoogleCloudAiplatformV1TensorboardBlob(data = NULL)
}
\arguments{
\item{data}{Optional}
}
\value{
GoogleCloudAiplatformV1TensorboardBlob object
}
\description{
GoogleCloudAiplatformV1TensorboardBlob Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
One blob (e.g, image, graph) viewable on a blob metric plot.
}
\concept{GoogleCloudAiplatformV1TensorboardBlob functions}
