% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1TensorboardBlobSequence}
\alias{GoogleCloudAiplatformV1TensorboardBlobSequence}
\title{GoogleCloudAiplatformV1TensorboardBlobSequence Object}
\usage{
GoogleCloudAiplatformV1TensorboardBlobSequence(values = NULL)
}
\arguments{
\item{values}{List of blobs contained within the sequence}
}
\value{
GoogleCloudAiplatformV1TensorboardBlobSequence object
}
\description{
GoogleCloudAiplatformV1TensorboardBlobSequence Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
One point viewable on a blob metric plot, but mostly just a wrapper message to work around repeated fields can't be used directly within \code{oneof} fields.
}
\concept{GoogleCloudAiplatformV1TensorboardBlobSequence functions}
