% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1TensorboardExperiment.labels}
\alias{GoogleCloudAiplatformV1TensorboardExperiment.labels}
\title{GoogleCloudAiplatformV1TensorboardExperiment.labels Object}
\usage{
GoogleCloudAiplatformV1TensorboardExperiment.labels()
}
\value{
GoogleCloudAiplatformV1TensorboardExperiment.labels object
}
\description{
GoogleCloudAiplatformV1TensorboardExperiment.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The labels with user-defined metadata to organize your TensorboardExperiment. Label keys and values cannot be longer than 64 characters (Unicode codepoints), can only contain lowercase letters, numeric characters, underscores and dashes. International characters are allowed. No more than 64 user labels can be associated with one Dataset (System labels are excluded). See https://goo.gl/xmQnxf for more information and examples of labels. System reserved label keys are prefixed with \verb{aiplatform.googleapis.com/} and are immutable. The following system labels exist for each Dataset: * \code{aiplatform.googleapis.com/dataset_metadata_schema}: output only. Its value is the metadata_schema's title.
}
\seealso{
Other GoogleCloudAiplatformV1TensorboardExperiment functions: 
\code{\link{GoogleCloudAiplatformV1TensorboardExperiment}()},
\code{\link{projects.locations.tensorboards.experiments.create}()},
\code{\link{projects.locations.tensorboards.experiments.patch}()}
}
\concept{GoogleCloudAiplatformV1TensorboardExperiment functions}
