% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1TensorboardExperiment}
\alias{GoogleCloudAiplatformV1TensorboardExperiment}
\title{GoogleCloudAiplatformV1TensorboardExperiment Object}
\usage{
GoogleCloudAiplatformV1TensorboardExperiment(
  GoogleCloudAiplatformV1TensorboardExperiment.labels = NULL,
  source = NULL,
  labels = NULL,
  displayName = NULL,
  etag = NULL,
  description = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1TensorboardExperiment.labels}{The \link{GoogleCloudAiplatformV1TensorboardExperiment.labels} object or list of objects}

\item{source}{Immutable}

\item{labels}{The labels with user-defined metadata to organize your TensorboardExperiment}

\item{displayName}{User provided name of this TensorboardExperiment}

\item{etag}{Used to perform consistent read-modify-write updates}

\item{description}{Description of this TensorboardExperiment}
}
\value{
GoogleCloudAiplatformV1TensorboardExperiment object
}
\description{
GoogleCloudAiplatformV1TensorboardExperiment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A TensorboardExperiment is a group of TensorboardRuns, that are typically the results of a training job run, in a Tensorboard.
}
\seealso{
Other GoogleCloudAiplatformV1TensorboardExperiment functions: 
\code{\link{GoogleCloudAiplatformV1TensorboardExperiment.labels}()},
\code{\link{projects.locations.tensorboards.experiments.create}()},
\code{\link{projects.locations.tensorboards.experiments.patch}()}
}
\concept{GoogleCloudAiplatformV1TensorboardExperiment functions}
