% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1TensorboardRun.labels}
\alias{GoogleCloudAiplatformV1TensorboardRun.labels}
\title{GoogleCloudAiplatformV1TensorboardRun.labels Object}
\usage{
GoogleCloudAiplatformV1TensorboardRun.labels()
}
\value{
GoogleCloudAiplatformV1TensorboardRun.labels object
}
\description{
GoogleCloudAiplatformV1TensorboardRun.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The labels with user-defined metadata to organize your TensorboardRuns. This field will be used to filter and visualize Runs in the Tensorboard UI. For example, a Vertex AI training job can set a label aiplatform.googleapis.com/training_job_id=xxxxx to all the runs created within that job. An end user can set a label experiment_id=xxxxx for all the runs produced in a Jupyter notebook. These runs can be grouped by a label value and visualized together in the Tensorboard UI. Label keys and values can be no longer than 64 characters (Unicode codepoints), can only contain lowercase letters, numeric characters, underscores and dashes. International characters are allowed. No more than 64 user labels can be associated with one TensorboardRun (System labels are excluded). See https://goo.gl/xmQnxf for more information and examples of labels. System reserved label keys are prefixed with 'aiplatform.googleapis.com/' and are immutable.
}
\seealso{
Other GoogleCloudAiplatformV1TensorboardRun functions: 
\code{\link{GoogleCloudAiplatformV1TensorboardRun}()},
\code{\link{projects.locations.tensorboards.experiments.runs.create}()},
\code{\link{projects.locations.tensorboards.experiments.runs.patch}()}
}
\concept{GoogleCloudAiplatformV1TensorboardRun functions}
