% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1TensorboardRun}
\alias{GoogleCloudAiplatformV1TensorboardRun}
\title{GoogleCloudAiplatformV1TensorboardRun Object}
\usage{
GoogleCloudAiplatformV1TensorboardRun(
  GoogleCloudAiplatformV1TensorboardRun.labels = NULL,
  description = NULL,
  etag = NULL,
  labels = NULL,
  displayName = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1TensorboardRun.labels}{The \link{GoogleCloudAiplatformV1TensorboardRun.labels} object or list of objects}

\item{description}{Description of this TensorboardRun}

\item{etag}{Used to perform a consistent read-modify-write updates}

\item{labels}{The labels with user-defined metadata to organize your TensorboardRuns}

\item{displayName}{Required}
}
\value{
GoogleCloudAiplatformV1TensorboardRun object
}
\description{
GoogleCloudAiplatformV1TensorboardRun Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
TensorboardRun maps to a specific execution of a training job with a given set of hyperparameter values, model definition, dataset, etc
}
\seealso{
Other GoogleCloudAiplatformV1TensorboardRun functions: 
\code{\link{GoogleCloudAiplatformV1TensorboardRun.labels}()},
\code{\link{projects.locations.tensorboards.experiments.runs.create}()},
\code{\link{projects.locations.tensorboards.experiments.runs.patch}()}
}
\concept{GoogleCloudAiplatformV1TensorboardRun functions}
