% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1TensorboardTensor}
\alias{GoogleCloudAiplatformV1TensorboardTensor}
\title{GoogleCloudAiplatformV1TensorboardTensor Object}
\usage{
GoogleCloudAiplatformV1TensorboardTensor(value = NULL, versionNumber = NULL)
}
\arguments{
\item{value}{Required}

\item{versionNumber}{Optional}
}
\value{
GoogleCloudAiplatformV1TensorboardTensor object
}
\description{
GoogleCloudAiplatformV1TensorboardTensor Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
One point viewable on a tensor metric plot.
}
\concept{GoogleCloudAiplatformV1TensorboardTensor functions}
