% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1TensorboardTimeSeries}
\alias{GoogleCloudAiplatformV1TensorboardTimeSeries}
\title{GoogleCloudAiplatformV1TensorboardTimeSeries Object}
\usage{
GoogleCloudAiplatformV1TensorboardTimeSeries(
  etag = NULL,
  pluginData = NULL,
  pluginName = NULL,
  valueType = NULL,
  description = NULL,
  displayName = NULL
)
}
\arguments{
\item{etag}{Used to perform a consistent read-modify-write updates}

\item{pluginData}{Data of the current plugin, with the size limited to 65KB}

\item{pluginName}{Immutable}

\item{valueType}{Required}

\item{description}{Description of this TensorboardTimeSeries}

\item{displayName}{Required}
}
\value{
GoogleCloudAiplatformV1TensorboardTimeSeries object
}
\description{
GoogleCloudAiplatformV1TensorboardTimeSeries Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
TensorboardTimeSeries maps to times series produced in training runs
}
\seealso{
Other GoogleCloudAiplatformV1TensorboardTimeSeries functions: 
\code{\link{projects.locations.tensorboards.experiments.runs.timeSeries.create}()},
\code{\link{projects.locations.tensorboards.experiments.runs.timeSeries.patch}()}
}
\concept{GoogleCloudAiplatformV1TensorboardTimeSeries functions}
