% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1TensorboardTimeSeriesMetadata}
\alias{GoogleCloudAiplatformV1TensorboardTimeSeriesMetadata}
\title{GoogleCloudAiplatformV1TensorboardTimeSeriesMetadata Object}
\usage{
GoogleCloudAiplatformV1TensorboardTimeSeriesMetadata()
}
\value{
GoogleCloudAiplatformV1TensorboardTimeSeriesMetadata object
}
\description{
GoogleCloudAiplatformV1TensorboardTimeSeriesMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describes metadata for a TensorboardTimeSeries.
}
\concept{GoogleCloudAiplatformV1TensorboardTimeSeriesMetadata functions}
