% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1TFRecordDestination}
\alias{GoogleCloudAiplatformV1TFRecordDestination}
\title{GoogleCloudAiplatformV1TFRecordDestination Object}
\usage{
GoogleCloudAiplatformV1TFRecordDestination(gcsDestination = NULL)
}
\arguments{
\item{gcsDestination}{Required}
}
\value{
GoogleCloudAiplatformV1TFRecordDestination object
}
\description{
GoogleCloudAiplatformV1TFRecordDestination Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The storage details for TFRecord output content.
}
\concept{GoogleCloudAiplatformV1TFRecordDestination functions}
