% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1ThresholdConfig}
\alias{GoogleCloudAiplatformV1ThresholdConfig}
\title{GoogleCloudAiplatformV1ThresholdConfig Object}
\usage{
GoogleCloudAiplatformV1ThresholdConfig(value = NULL)
}
\arguments{
\item{value}{Specify a threshold value that can trigger the alert}
}
\value{
GoogleCloudAiplatformV1ThresholdConfig object
}
\description{
GoogleCloudAiplatformV1ThresholdConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The config for feature monitoring threshold.
}
\concept{GoogleCloudAiplatformV1ThresholdConfig functions}
