% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1TimeSeriesData}
\alias{GoogleCloudAiplatformV1TimeSeriesData}
\title{GoogleCloudAiplatformV1TimeSeriesData Object}
\usage{
GoogleCloudAiplatformV1TimeSeriesData(
  valueType = NULL,
  values = NULL,
  tensorboardTimeSeriesId = NULL
)
}
\arguments{
\item{valueType}{Required}

\item{values}{Required}

\item{tensorboardTimeSeriesId}{Required}
}
\value{
GoogleCloudAiplatformV1TimeSeriesData object
}
\description{
GoogleCloudAiplatformV1TimeSeriesData Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
All the data stored in a TensorboardTimeSeries.
}
\concept{GoogleCloudAiplatformV1TimeSeriesData functions}
