% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1TimeSeriesDataPoint}
\alias{GoogleCloudAiplatformV1TimeSeriesDataPoint}
\title{GoogleCloudAiplatformV1TimeSeriesDataPoint Object}
\usage{
GoogleCloudAiplatformV1TimeSeriesDataPoint(
  scalar = NULL,
  step = NULL,
  wallTime = NULL,
  blobs = NULL,
  tensor = NULL
)
}
\arguments{
\item{scalar}{A scalar value}

\item{step}{Step index of this data point within the run}

\item{wallTime}{Wall clock timestamp when this data point is generated by the end user}

\item{blobs}{A blob sequence value}

\item{tensor}{A tensor value}
}
\value{
GoogleCloudAiplatformV1TimeSeriesDataPoint object
}
\description{
GoogleCloudAiplatformV1TimeSeriesDataPoint Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A TensorboardTimeSeries data point.
}
\concept{GoogleCloudAiplatformV1TimeSeriesDataPoint functions}
