% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1TimestampSplit}
\alias{GoogleCloudAiplatformV1TimestampSplit}
\title{GoogleCloudAiplatformV1TimestampSplit Object}
\usage{
GoogleCloudAiplatformV1TimestampSplit(
  key = NULL,
  trainingFraction = NULL,
  validationFraction = NULL,
  testFraction = NULL
)
}
\arguments{
\item{key}{Required}

\item{trainingFraction}{The fraction of the input data that is to be used to train the Model}

\item{validationFraction}{The fraction of the input data that is to be used to validate the Model}

\item{testFraction}{The fraction of the input data that is to be used to evaluate the Model}
}
\value{
GoogleCloudAiplatformV1TimestampSplit object
}
\description{
GoogleCloudAiplatformV1TimestampSplit Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Assigns input data to training, validation, and test sets based on a provided timestamps. The youngest data pieces are assigned to training set, next to validation set, and the oldest to the test set. Supported only for tabular Datasets.
}
\concept{GoogleCloudAiplatformV1TimestampSplit functions}
