% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1TrainingConfig}
\alias{GoogleCloudAiplatformV1TrainingConfig}
\title{GoogleCloudAiplatformV1TrainingConfig Object}
\usage{
GoogleCloudAiplatformV1TrainingConfig(timeoutTrainingMilliHours = NULL)
}
\arguments{
\item{timeoutTrainingMilliHours}{The timeout hours for the CMLE training job, expressed in milli hours i}
}
\value{
GoogleCloudAiplatformV1TrainingConfig object
}
\description{
GoogleCloudAiplatformV1TrainingConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
CMLE training config. For every active learning labeling iteration, system will train a machine learning model on CMLE. The trained model will be used by data sampling algorithm to select DataItems.
}
\concept{GoogleCloudAiplatformV1TrainingConfig functions}
