% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1TrainingPipeline}
\alias{GoogleCloudAiplatformV1TrainingPipeline}
\title{GoogleCloudAiplatformV1TrainingPipeline Object}
\usage{
GoogleCloudAiplatformV1TrainingPipeline(
  GoogleCloudAiplatformV1TrainingPipeline.labels = NULL,
  trainingTaskDefinition = NULL,
  trainingTaskInputs = NULL,
  labels = NULL,
  inputDataConfig = NULL,
  parentModel = NULL,
  modelToUpload = NULL,
  modelId = NULL,
  encryptionSpec = NULL,
  displayName = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1TrainingPipeline.labels}{The \link{GoogleCloudAiplatformV1TrainingPipeline.labels} object or list of objects}

\item{trainingTaskDefinition}{Required}

\item{trainingTaskInputs}{Required}

\item{labels}{The labels with user-defined metadata to organize TrainingPipelines}

\item{inputDataConfig}{Specifies Vertex AI owned input data that may be used for training the Model}

\item{parentModel}{Optional}

\item{modelToUpload}{Describes the Model that may be uploaded (via ModelService}

\item{modelId}{Optional}

\item{encryptionSpec}{Customer-managed encryption key spec for a TrainingPipeline}

\item{displayName}{Required}
}
\value{
GoogleCloudAiplatformV1TrainingPipeline object
}
\description{
GoogleCloudAiplatformV1TrainingPipeline Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The TrainingPipeline orchestrates tasks associated with training a Model. It always executes the training task, and optionally may also export data from Vertex AI's Dataset which becomes the training input, upload the Model to Vertex AI, and evaluate the Model.
}
\seealso{
Other GoogleCloudAiplatformV1TrainingPipeline functions: 
\code{\link{GoogleCloudAiplatformV1TrainingPipeline.labels}()},
\code{\link{projects.locations.trainingPipelines.create}()}
}
\concept{GoogleCloudAiplatformV1TrainingPipeline functions}
