% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1Trial}
\alias{GoogleCloudAiplatformV1Trial}
\title{GoogleCloudAiplatformV1Trial Object}
\usage{
GoogleCloudAiplatformV1Trial(GoogleCloudAiplatformV1Trial.webAccessUris = NULL)
}
\arguments{
\item{GoogleCloudAiplatformV1Trial.webAccessUris}{The \link{GoogleCloudAiplatformV1Trial.webAccessUris} object or list of objects}
}
\value{
GoogleCloudAiplatformV1Trial object
}
\description{
GoogleCloudAiplatformV1Trial Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A message representing a Trial. A Trial contains a unique set of Parameters that has been or will be evaluated, along with the objective metrics got by running the Trial.
}
\seealso{
Other GoogleCloudAiplatformV1Trial functions: 
\code{\link{GoogleCloudAiplatformV1Trial.webAccessUris}()},
\code{\link{projects.locations.studies.trials.create}()}
}
\concept{GoogleCloudAiplatformV1Trial functions}
