% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1Trial.webAccessUris}
\alias{GoogleCloudAiplatformV1Trial.webAccessUris}
\title{GoogleCloudAiplatformV1Trial.webAccessUris Object}
\usage{
GoogleCloudAiplatformV1Trial.webAccessUris()
}
\value{
GoogleCloudAiplatformV1Trial.webAccessUris object
}
\description{
GoogleCloudAiplatformV1Trial.webAccessUris Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Output only. URIs for accessing \href{https://cloud.google.com/vertex-ai/docs/training/monitor-debug-interactive-shell}{interactive shells} (one URI for each training node). Only available if this trial is part of a HyperparameterTuningJob and the job's trial_job_spec.enable_web_access field is \code{true}. The keys are names of each node used for the trial; for example, \code{workerpool0-0} for the primary node, \code{workerpool1-0} for the first node in the second worker pool, and \code{workerpool1-1} for the second node in the second worker pool. The values are the URIs for each node's interactive shell.
}
\seealso{
Other GoogleCloudAiplatformV1Trial functions: 
\code{\link{GoogleCloudAiplatformV1Trial}()},
\code{\link{projects.locations.studies.trials.create}()}
}
\concept{GoogleCloudAiplatformV1Trial functions}
