% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1UndeployIndexOperationMetadata}
\alias{GoogleCloudAiplatformV1UndeployIndexOperationMetadata}
\title{GoogleCloudAiplatformV1UndeployIndexOperationMetadata Object}
\usage{
GoogleCloudAiplatformV1UndeployIndexOperationMetadata(genericMetadata = NULL)
}
\arguments{
\item{genericMetadata}{The operation generic information}
}
\value{
GoogleCloudAiplatformV1UndeployIndexOperationMetadata object
}
\description{
GoogleCloudAiplatformV1UndeployIndexOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Runtime operation information for IndexEndpointService.UndeployIndex.
}
\concept{GoogleCloudAiplatformV1UndeployIndexOperationMetadata functions}
