% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1UndeployIndexRequest}
\alias{GoogleCloudAiplatformV1UndeployIndexRequest}
\title{GoogleCloudAiplatformV1UndeployIndexRequest Object}
\usage{
GoogleCloudAiplatformV1UndeployIndexRequest(deployedIndexId = NULL)
}
\arguments{
\item{deployedIndexId}{Required}
}
\value{
GoogleCloudAiplatformV1UndeployIndexRequest object
}
\description{
GoogleCloudAiplatformV1UndeployIndexRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for IndexEndpointService.UndeployIndex.
}
\seealso{
Other GoogleCloudAiplatformV1UndeployIndexRequest functions: 
\code{\link{projects.locations.indexEndpoints.undeployIndex}()}
}
\concept{GoogleCloudAiplatformV1UndeployIndexRequest functions}
