% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1UndeployModelOperationMetadata}
\alias{GoogleCloudAiplatformV1UndeployModelOperationMetadata}
\title{GoogleCloudAiplatformV1UndeployModelOperationMetadata Object}
\usage{
GoogleCloudAiplatformV1UndeployModelOperationMetadata(genericMetadata = NULL)
}
\arguments{
\item{genericMetadata}{The operation generic information}
}
\value{
GoogleCloudAiplatformV1UndeployModelOperationMetadata object
}
\description{
GoogleCloudAiplatformV1UndeployModelOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Runtime operation information for EndpointService.UndeployModel.
}
\concept{GoogleCloudAiplatformV1UndeployModelOperationMetadata functions}
