% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1UndeployModelRequest}
\alias{GoogleCloudAiplatformV1UndeployModelRequest}
\title{GoogleCloudAiplatformV1UndeployModelRequest Object}
\usage{
GoogleCloudAiplatformV1UndeployModelRequest(
  GoogleCloudAiplatformV1UndeployModelRequest.trafficSplit = NULL,
  deployedModelId = NULL,
  trafficSplit = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1UndeployModelRequest.trafficSplit}{The \link{GoogleCloudAiplatformV1UndeployModelRequest.trafficSplit} object or list of objects}

\item{deployedModelId}{Required}

\item{trafficSplit}{If this field is provided, then the Endpoint's traffic_split will be overwritten with it}
}
\value{
GoogleCloudAiplatformV1UndeployModelRequest object
}
\description{
GoogleCloudAiplatformV1UndeployModelRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for EndpointService.UndeployModel.
}
\seealso{
Other GoogleCloudAiplatformV1UndeployModelRequest functions: 
\code{\link{GoogleCloudAiplatformV1UndeployModelRequest.trafficSplit}()},
\code{\link{projects.locations.endpoints.undeployModel}()}
}
\concept{GoogleCloudAiplatformV1UndeployModelRequest functions}
