% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1UndeployModelRequest.trafficSplit}
\alias{GoogleCloudAiplatformV1UndeployModelRequest.trafficSplit}
\title{GoogleCloudAiplatformV1UndeployModelRequest.trafficSplit Object}
\usage{
GoogleCloudAiplatformV1UndeployModelRequest.trafficSplit()
}
\value{
GoogleCloudAiplatformV1UndeployModelRequest.trafficSplit object
}
\description{
GoogleCloudAiplatformV1UndeployModelRequest.trafficSplit Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
If this field is provided, then the Endpoint's traffic_split will be overwritten with it. If last DeployedModel is being undeployed from the Endpoint, the \link{Endpoint.traffic_split} will always end up empty when this call returns. A DeployedModel will be successfully undeployed only if it doesn't have any traffic assigned to it when this method executes, or if this field unassigns any traffic to it.
}
\seealso{
Other GoogleCloudAiplatformV1UndeployModelRequest functions: 
\code{\link{GoogleCloudAiplatformV1UndeployModelRequest}()},
\code{\link{projects.locations.endpoints.undeployModel}()}
}
\concept{GoogleCloudAiplatformV1UndeployModelRequest functions}
