% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1UnmanagedContainerModel}
\alias{GoogleCloudAiplatformV1UnmanagedContainerModel}
\title{GoogleCloudAiplatformV1UnmanagedContainerModel Object}
\usage{
GoogleCloudAiplatformV1UnmanagedContainerModel(
  predictSchemata = NULL,
  artifactUri = NULL,
  containerSpec = NULL
)
}
\arguments{
\item{predictSchemata}{Contains the schemata used in Model's predictions and explanations}

\item{artifactUri}{The path to the directory containing the Model artifact and any of its supporting files}

\item{containerSpec}{Input only}
}
\value{
GoogleCloudAiplatformV1UnmanagedContainerModel object
}
\description{
GoogleCloudAiplatformV1UnmanagedContainerModel Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains model information necessary to perform batch prediction without requiring a full model import.
}
\concept{GoogleCloudAiplatformV1UnmanagedContainerModel functions}
