% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1UpdateExplanationDatasetOperationMetadata}
\alias{GoogleCloudAiplatformV1UpdateExplanationDatasetOperationMetadata}
\title{GoogleCloudAiplatformV1UpdateExplanationDatasetOperationMetadata Object}
\usage{
GoogleCloudAiplatformV1UpdateExplanationDatasetOperationMetadata(
  genericMetadata = NULL
)
}
\arguments{
\item{genericMetadata}{The common part of the operation metadata}
}
\value{
GoogleCloudAiplatformV1UpdateExplanationDatasetOperationMetadata object
}
\description{
GoogleCloudAiplatformV1UpdateExplanationDatasetOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Runtime operation information for ModelService.UpdateExplanationDataset.
}
\concept{GoogleCloudAiplatformV1UpdateExplanationDatasetOperationMetadata functions}
