% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1UpdateExplanationDatasetRequest}
\alias{GoogleCloudAiplatformV1UpdateExplanationDatasetRequest}
\title{GoogleCloudAiplatformV1UpdateExplanationDatasetRequest Object}
\usage{
GoogleCloudAiplatformV1UpdateExplanationDatasetRequest(examples = NULL)
}
\arguments{
\item{examples}{The example config containing the location of the dataset}
}
\value{
GoogleCloudAiplatformV1UpdateExplanationDatasetRequest object
}
\description{
GoogleCloudAiplatformV1UpdateExplanationDatasetRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for ModelService.UpdateExplanationDataset.
}
\seealso{
Other GoogleCloudAiplatformV1UpdateExplanationDatasetRequest functions: 
\code{\link{projects.locations.models.updateExplanationDataset}()}
}
\concept{GoogleCloudAiplatformV1UpdateExplanationDatasetRequest functions}
