% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1UpdateFeaturestoreOperationMetadata}
\alias{GoogleCloudAiplatformV1UpdateFeaturestoreOperationMetadata}
\title{GoogleCloudAiplatformV1UpdateFeaturestoreOperationMetadata Object}
\usage{
GoogleCloudAiplatformV1UpdateFeaturestoreOperationMetadata(
  genericMetadata = NULL
)
}
\arguments{
\item{genericMetadata}{Operation metadata for Featurestore}
}
\value{
GoogleCloudAiplatformV1UpdateFeaturestoreOperationMetadata object
}
\description{
GoogleCloudAiplatformV1UpdateFeaturestoreOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details of operations that perform update Featurestore.
}
\concept{GoogleCloudAiplatformV1UpdateFeaturestoreOperationMetadata functions}
