% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1UpdateIndexOperationMetadata}
\alias{GoogleCloudAiplatformV1UpdateIndexOperationMetadata}
\title{GoogleCloudAiplatformV1UpdateIndexOperationMetadata Object}
\usage{
GoogleCloudAiplatformV1UpdateIndexOperationMetadata(
  nearestNeighborSearchOperationMetadata = NULL,
  genericMetadata = NULL
)
}
\arguments{
\item{nearestNeighborSearchOperationMetadata}{The operation metadata with regard to Matching Engine Index operation}

\item{genericMetadata}{The operation generic information}
}
\value{
GoogleCloudAiplatformV1UpdateIndexOperationMetadata object
}
\description{
GoogleCloudAiplatformV1UpdateIndexOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Runtime operation information for IndexService.UpdateIndex.
}
\concept{GoogleCloudAiplatformV1UpdateIndexOperationMetadata functions}
