% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1UpdateModelDeploymentMonitoringJobOperationMetadata}
\alias{GoogleCloudAiplatformV1UpdateModelDeploymentMonitoringJobOperationMetadata}
\title{GoogleCloudAiplatformV1UpdateModelDeploymentMonitoringJobOperationMetadata Object}
\usage{
GoogleCloudAiplatformV1UpdateModelDeploymentMonitoringJobOperationMetadata(
  genericMetadata = NULL
)
}
\arguments{
\item{genericMetadata}{The operation generic information}
}
\value{
GoogleCloudAiplatformV1UpdateModelDeploymentMonitoringJobOperationMetadata object
}
\description{
GoogleCloudAiplatformV1UpdateModelDeploymentMonitoringJobOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Runtime operation information for JobService.UpdateModelDeploymentMonitoringJob.
}
\concept{GoogleCloudAiplatformV1UpdateModelDeploymentMonitoringJobOperationMetadata functions}
