% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1UpdateSpecialistPoolOperationMetadata}
\alias{GoogleCloudAiplatformV1UpdateSpecialistPoolOperationMetadata}
\title{GoogleCloudAiplatformV1UpdateSpecialistPoolOperationMetadata Object}
\usage{
GoogleCloudAiplatformV1UpdateSpecialistPoolOperationMetadata(
  genericMetadata = NULL
)
}
\arguments{
\item{genericMetadata}{The operation generic information}
}
\value{
GoogleCloudAiplatformV1UpdateSpecialistPoolOperationMetadata object
}
\description{
GoogleCloudAiplatformV1UpdateSpecialistPoolOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Runtime operation metadata for SpecialistPoolService.UpdateSpecialistPool.
}
\concept{GoogleCloudAiplatformV1UpdateSpecialistPoolOperationMetadata functions}
