% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1UpdateTensorboardOperationMetadata}
\alias{GoogleCloudAiplatformV1UpdateTensorboardOperationMetadata}
\title{GoogleCloudAiplatformV1UpdateTensorboardOperationMetadata Object}
\usage{
GoogleCloudAiplatformV1UpdateTensorboardOperationMetadata(
  genericMetadata = NULL
)
}
\arguments{
\item{genericMetadata}{Operation metadata for Tensorboard}
}
\value{
GoogleCloudAiplatformV1UpdateTensorboardOperationMetadata object
}
\description{
GoogleCloudAiplatformV1UpdateTensorboardOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details of operations that perform update Tensorboard.
}
\concept{GoogleCloudAiplatformV1UpdateTensorboardOperationMetadata functions}
