% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1UploadModelOperationMetadata}
\alias{GoogleCloudAiplatformV1UploadModelOperationMetadata}
\title{GoogleCloudAiplatformV1UploadModelOperationMetadata Object}
\usage{
GoogleCloudAiplatformV1UploadModelOperationMetadata(genericMetadata = NULL)
}
\arguments{
\item{genericMetadata}{The common part of the operation metadata}
}
\value{
GoogleCloudAiplatformV1UploadModelOperationMetadata object
}
\description{
GoogleCloudAiplatformV1UploadModelOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details of ModelService.UploadModel operation.
}
\concept{GoogleCloudAiplatformV1UploadModelOperationMetadata functions}
