% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1UploadModelRequest}
\alias{GoogleCloudAiplatformV1UploadModelRequest}
\title{GoogleCloudAiplatformV1UploadModelRequest Object}
\usage{
GoogleCloudAiplatformV1UploadModelRequest(
  model = NULL,
  modelId = NULL,
  parentModel = NULL,
  serviceAccount = NULL
)
}
\arguments{
\item{model}{Required}

\item{modelId}{Optional}

\item{parentModel}{Optional}

\item{serviceAccount}{Optional}
}
\value{
GoogleCloudAiplatformV1UploadModelRequest object
}
\description{
GoogleCloudAiplatformV1UploadModelRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for ModelService.UploadModel.
}
\seealso{
Other GoogleCloudAiplatformV1UploadModelRequest functions: 
\code{\link{projects.locations.models.upload}()}
}
\concept{GoogleCloudAiplatformV1UploadModelRequest functions}
