% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1UploadModelResponse}
\alias{GoogleCloudAiplatformV1UploadModelResponse}
\title{GoogleCloudAiplatformV1UploadModelResponse Object}
\usage{
GoogleCloudAiplatformV1UploadModelResponse(model = NULL)
}
\arguments{
\item{model}{The name of the uploaded Model resource}
}
\value{
GoogleCloudAiplatformV1UploadModelResponse object
}
\description{
GoogleCloudAiplatformV1UploadModelResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message of ModelService.UploadModel operation.
}
\concept{GoogleCloudAiplatformV1UploadModelResponse functions}
