% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1UserActionReference}
\alias{GoogleCloudAiplatformV1UserActionReference}
\title{GoogleCloudAiplatformV1UserActionReference Object}
\usage{
GoogleCloudAiplatformV1UserActionReference(
  method = NULL,
  operation = NULL,
  dataLabelingJob = NULL
)
}
\arguments{
\item{method}{The method name of the API RPC call}

\item{operation}{For API calls that return a long running operation}

\item{dataLabelingJob}{For API calls that start a LabelingJob}
}
\value{
GoogleCloudAiplatformV1UserActionReference object
}
\description{
GoogleCloudAiplatformV1UserActionReference Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
References an API call. It contains more information about long running operation and Jobs that are triggered by the API call.
}
\concept{GoogleCloudAiplatformV1UserActionReference functions}
