% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1Value}
\alias{GoogleCloudAiplatformV1Value}
\title{GoogleCloudAiplatformV1Value Object}
\usage{
GoogleCloudAiplatformV1Value(
  intValue = NULL,
  stringValue = NULL,
  doubleValue = NULL
)
}
\arguments{
\item{intValue}{An integer value}

\item{stringValue}{A string value}

\item{doubleValue}{A double value}
}
\value{
GoogleCloudAiplatformV1Value object
}
\description{
GoogleCloudAiplatformV1Value Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Value is the value of the field.
}
\concept{GoogleCloudAiplatformV1Value functions}
