% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1WorkerPoolSpec}
\alias{GoogleCloudAiplatformV1WorkerPoolSpec}
\title{GoogleCloudAiplatformV1WorkerPoolSpec Object}
\usage{
GoogleCloudAiplatformV1WorkerPoolSpec(
  machineSpec = NULL,
  containerSpec = NULL,
  diskSpec = NULL,
  replicaCount = NULL,
  pythonPackageSpec = NULL,
  nfsMounts = NULL
)
}
\arguments{
\item{machineSpec}{Optional}

\item{containerSpec}{The custom container task}

\item{diskSpec}{Disk spec}

\item{replicaCount}{Optional}

\item{pythonPackageSpec}{The Python packaged task}

\item{nfsMounts}{Optional}
}
\value{
GoogleCloudAiplatformV1WorkerPoolSpec object
}
\description{
GoogleCloudAiplatformV1WorkerPoolSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents the spec of a worker pool in a job.
}
\concept{GoogleCloudAiplatformV1WorkerPoolSpec functions}
