% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1WriteTensorboardRunDataRequest}
\alias{GoogleCloudAiplatformV1WriteTensorboardRunDataRequest}
\title{GoogleCloudAiplatformV1WriteTensorboardRunDataRequest Object}
\usage{
GoogleCloudAiplatformV1WriteTensorboardRunDataRequest(
  timeSeriesData = NULL,
  tensorboardRun = NULL
)
}
\arguments{
\item{timeSeriesData}{Required}

\item{tensorboardRun}{Required}
}
\value{
GoogleCloudAiplatformV1WriteTensorboardRunDataRequest object
}
\description{
GoogleCloudAiplatformV1WriteTensorboardRunDataRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for TensorboardService.WriteTensorboardRunData.
}
\seealso{
Other GoogleCloudAiplatformV1WriteTensorboardRunDataRequest functions: 
\code{\link{projects.locations.tensorboards.experiments.runs.write}()}
}
\concept{GoogleCloudAiplatformV1WriteTensorboardRunDataRequest functions}
