% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudAiplatformV1XraiAttribution}
\alias{GoogleCloudAiplatformV1XraiAttribution}
\title{GoogleCloudAiplatformV1XraiAttribution Object}
\usage{
GoogleCloudAiplatformV1XraiAttribution(
  smoothGradConfig = NULL,
  blurBaselineConfig = NULL,
  stepCount = NULL
)
}
\arguments{
\item{smoothGradConfig}{Config for SmoothGrad approximation of gradients}

\item{blurBaselineConfig}{Config for XRAI with blur baseline}

\item{stepCount}{Required}
}
\value{
GoogleCloudAiplatformV1XraiAttribution object
}
\description{
GoogleCloudAiplatformV1XraiAttribution Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An explanation method that redistributes Integrated Gradients attributions to segmented regions, taking advantage of the model's fully differentiable structure. Refer to this paper for more details: https://arxiv.org/abs/1906.02825 Supported only by image Models.
}
\concept{GoogleCloudAiplatformV1XraiAttribution functions}
