% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudLocationListLocationsResponse}
\alias{GoogleCloudLocationListLocationsResponse}
\title{GoogleCloudLocationListLocationsResponse Object}
\usage{
GoogleCloudLocationListLocationsResponse(
  nextPageToken = NULL,
  locations = NULL
)
}
\arguments{
\item{nextPageToken}{The standard List next-page token}

\item{locations}{A list of locations that matches the specified filter in the request}
}
\value{
GoogleCloudLocationListLocationsResponse object
}
\description{
GoogleCloudLocationListLocationsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response message for Locations.ListLocations.
}
\concept{GoogleCloudLocationListLocationsResponse functions}
