% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudLocationLocation.labels}
\alias{GoogleCloudLocationLocation.labels}
\title{GoogleCloudLocationLocation.labels Object}
\usage{
GoogleCloudLocationLocation.labels()
}
\value{
GoogleCloudLocationLocation.labels object
}
\description{
GoogleCloudLocationLocation.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Cross-service attributes for the location. For example {'cloud.googleapis.com/region': 'us-east1'}
}
\seealso{
Other GoogleCloudLocationLocation functions: 
\code{\link{GoogleCloudLocationLocation.metadata}()},
\code{\link{GoogleCloudLocationLocation}()}
}
\concept{GoogleCloudLocationLocation functions}
