% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudLocationLocation.metadata}
\alias{GoogleCloudLocationLocation.metadata}
\title{GoogleCloudLocationLocation.metadata Object}
\usage{
GoogleCloudLocationLocation.metadata()
}
\value{
GoogleCloudLocationLocation.metadata object
}
\description{
GoogleCloudLocationLocation.metadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Service-specific metadata. For example the available capacity at the given location.
}
\seealso{
Other GoogleCloudLocationLocation functions: 
\code{\link{GoogleCloudLocationLocation.labels}()},
\code{\link{GoogleCloudLocationLocation}()}
}
\concept{GoogleCloudLocationLocation functions}
