% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleCloudLocationLocation}
\alias{GoogleCloudLocationLocation}
\title{GoogleCloudLocationLocation Object}
\usage{
GoogleCloudLocationLocation(
  GoogleCloudLocationLocation.labels = NULL,
  GoogleCloudLocationLocation.metadata = NULL,
  name = NULL,
  labels = NULL,
  metadata = NULL,
  displayName = NULL,
  locationId = NULL
)
}
\arguments{
\item{GoogleCloudLocationLocation.labels}{The \link{GoogleCloudLocationLocation.labels} object or list of objects}

\item{GoogleCloudLocationLocation.metadata}{The \link{GoogleCloudLocationLocation.metadata} object or list of objects}

\item{name}{Resource name for the location, which may vary between implementations}

\item{labels}{Cross-service attributes for the location}

\item{metadata}{Service-specific metadata}

\item{displayName}{The friendly name for this location, typically a nearby city name}

\item{locationId}{The canonical id for this location}
}
\value{
GoogleCloudLocationLocation object
}
\description{
GoogleCloudLocationLocation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A resource that represents a Google Cloud location.
}
\seealso{
Other GoogleCloudLocationLocation functions: 
\code{\link{GoogleCloudLocationLocation.labels}()},
\code{\link{GoogleCloudLocationLocation.metadata}()}
}
\concept{GoogleCloudLocationLocation functions}
