% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleIamV1Binding}
\alias{GoogleIamV1Binding}
\title{GoogleIamV1Binding Object}
\usage{
GoogleIamV1Binding(members = NULL, condition = NULL, role = NULL)
}
\arguments{
\item{members}{Specifies the principals requesting access for a Google Cloud resource}

\item{condition}{The condition that is associated with this binding}

\item{role}{Role that is assigned to the list of \code{members}, or principals}
}
\value{
GoogleIamV1Binding object
}
\description{
GoogleIamV1Binding Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Associates \code{members}, or principals, with a \code{role}.
}
\concept{GoogleIamV1Binding functions}
