% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleLongrunningListOperationsResponse}
\alias{GoogleLongrunningListOperationsResponse}
\title{GoogleLongrunningListOperationsResponse Object}
\usage{
GoogleLongrunningListOperationsResponse(
  nextPageToken = NULL,
  operations = NULL
)
}
\arguments{
\item{nextPageToken}{The standard List next-page token}

\item{operations}{A list of operations that matches the specified filter in the request}
}
\value{
GoogleLongrunningListOperationsResponse object
}
\description{
GoogleLongrunningListOperationsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response message for Operations.ListOperations.
}
\concept{GoogleLongrunningListOperationsResponse functions}
