% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleLongrunningOperation.metadata}
\alias{GoogleLongrunningOperation.metadata}
\title{GoogleLongrunningOperation.metadata Object}
\usage{
GoogleLongrunningOperation.metadata()
}
\value{
GoogleLongrunningOperation.metadata object
}
\description{
GoogleLongrunningOperation.metadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Service-specific metadata associated with the operation. It typically contains progress information and common metadata such as create time. Some services might not provide such metadata. Any method that returns a long-running operation should document the metadata type, if any.
}
\seealso{
Other GoogleLongrunningOperation functions: 
\code{\link{GoogleLongrunningOperation.response}()},
\code{\link{GoogleLongrunningOperation}()}
}
\concept{GoogleLongrunningOperation functions}
