% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleLongrunningOperation.response}
\alias{GoogleLongrunningOperation.response}
\title{GoogleLongrunningOperation.response Object}
\usage{
GoogleLongrunningOperation.response()
}
\value{
GoogleLongrunningOperation.response object
}
\description{
GoogleLongrunningOperation.response Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The normal response of the operation in case of success. If the original method returns no data on success, such as \code{Delete}, the response is \code{google.protobuf.Empty}. If the original method is standard \code{Get}/\code{Create}/\code{Update}, the response should be the resource. For other methods, the response should have the type \code{XxxResponse}, where \code{Xxx} is the original method name. For example, if the original method name is \code{TakeSnapshot()}, the inferred response type is \code{TakeSnapshotResponse}.
}
\seealso{
Other GoogleLongrunningOperation functions: 
\code{\link{GoogleLongrunningOperation.metadata}()},
\code{\link{GoogleLongrunningOperation}()}
}
\concept{GoogleLongrunningOperation functions}
