% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleRpcStatus}
\alias{GoogleRpcStatus}
\title{GoogleRpcStatus Object}
\usage{
GoogleRpcStatus(message = NULL, details = NULL, code = NULL)
}
\arguments{
\item{message}{A developer-facing error message, which should be in English}

\item{details}{A list of messages that carry the error details}

\item{code}{The status code, which should be an enum value of google}
}
\value{
GoogleRpcStatus object
}
\description{
GoogleRpcStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The \code{Status} type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by \href{https://github.com/grpc}{gRPC}. Each \code{Status} message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the \href{https://cloud.google.com/apis/design/errors}{API Design Guide}.
}
\concept{GoogleRpcStatus functions}
