% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleTypeColor}
\alias{GoogleTypeColor}
\title{GoogleTypeColor Object}
\usage{
GoogleTypeColor(green = NULL, alpha = NULL, blue = NULL, red = NULL)
}
\arguments{
\item{green}{The amount of green in the color as a value in the interval \link{0, 1}}

\item{alpha}{The fraction of this color that should be applied to the pixel}

\item{blue}{The amount of blue in the color as a value in the interval \link{0, 1}}

\item{red}{The amount of red in the color as a value in the interval \link{0, 1}}
}
\value{
GoogleTypeColor object
}
\description{
GoogleTypeColor Object
}
\details{

}
\concept{GoogleTypeColor functions}
