% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleTypeInterval}
\alias{GoogleTypeInterval}
\title{GoogleTypeInterval Object}
\usage{
GoogleTypeInterval(endTime = NULL, startTime = NULL)
}
\arguments{
\item{endTime}{Optional}

\item{startTime}{Optional}
}
\value{
GoogleTypeInterval object
}
\description{
GoogleTypeInterval Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a time interval, encoded as a Timestamp start (inclusive) and a Timestamp end (exclusive). The start must be less than or equal to the end. When the start equals the end, the interval is empty (matches no time). When both start and end are unspecified, the interval matches any time.
}
\concept{GoogleTypeInterval functions}
