% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_objects.R
\name{GoogleTypeMoney}
\alias{GoogleTypeMoney}
\title{GoogleTypeMoney Object}
\usage{
GoogleTypeMoney(nanos = NULL, units = NULL, currencyCode = NULL)
}
\arguments{
\item{nanos}{Number of nano (10^-9) units of the amount}

\item{units}{The whole units of the amount}

\item{currencyCode}{The three-letter currency code defined in ISO 4217}
}
\value{
GoogleTypeMoney object
}
\description{
GoogleTypeMoney Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents an amount of money with its currency type.
}
\concept{GoogleTypeMoney functions}
