% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.batchPredictionJobs.cancel}
\alias{projects.locations.batchPredictionJobs.cancel}
\title{Cancels a BatchPredictionJob. Starts asynchronous cancellation on the BatchPredictionJob. The server makes the best effort to cancel the job, but success is not guaranteed. Clients can use JobService.GetBatchPredictionJob or other methods to check whether the cancellation succeeded or whether the job completed despite cancellation. On a successful cancellation, the BatchPredictionJob is not deleted;instead its BatchPredictionJob.state is set to \code{CANCELLED}. Any files already outputted by the job are not deleted.}
\usage{
projects.locations.batchPredictionJobs.cancel(
  GoogleCloudAiplatformV1CancelBatchPredictionJobRequest,
  name
)
}
\arguments{
\item{GoogleCloudAiplatformV1CancelBatchPredictionJobRequest}{The \link{GoogleCloudAiplatformV1CancelBatchPredictionJobRequest} object to pass to this method}

\item{name}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1CancelBatchPredictionJobRequest functions: 
\code{\link{GoogleCloudAiplatformV1CancelBatchPredictionJobRequest}()}
}
\concept{GoogleCloudAiplatformV1CancelBatchPredictionJobRequest functions}
