% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.batchPredictionJobs.create}
\alias{projects.locations.batchPredictionJobs.create}
\title{Creates a BatchPredictionJob. A BatchPredictionJob once created will right away be attempted to start.}
\usage{
projects.locations.batchPredictionJobs.create(
  GoogleCloudAiplatformV1BatchPredictionJob,
  parent
)
}
\arguments{
\item{GoogleCloudAiplatformV1BatchPredictionJob}{The \link{GoogleCloudAiplatformV1BatchPredictionJob} object to pass to this method}

\item{parent}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1BatchPredictionJob functions: 
\code{\link{GoogleCloudAiplatformV1BatchPredictionJob.labels}()},
\code{\link{GoogleCloudAiplatformV1BatchPredictionJob}()}
}
\concept{GoogleCloudAiplatformV1BatchPredictionJob functions}
