% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.customJobs.cancel}
\alias{projects.locations.customJobs.cancel}
\title{Cancels a CustomJob. Starts asynchronous cancellation on the CustomJob. The server makes a best effort to cancel the job, but success is not guaranteed. Clients can use JobService.GetCustomJob or other methods to check whether the cancellation succeeded or whether the job completed despite cancellation. On successful cancellation, the CustomJob is not deleted; instead it becomes a job with a CustomJob.error value with a google.rpc.Status.code of 1, corresponding to \code{Code.CANCELLED}, and CustomJob.state is set to \code{CANCELLED}.}
\usage{
projects.locations.customJobs.cancel(
  GoogleCloudAiplatformV1CancelCustomJobRequest,
  name
)
}
\arguments{
\item{GoogleCloudAiplatformV1CancelCustomJobRequest}{The \link{GoogleCloudAiplatformV1CancelCustomJobRequest} object to pass to this method}

\item{name}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1CancelCustomJobRequest functions: 
\code{\link{GoogleCloudAiplatformV1CancelCustomJobRequest}()}
}
\concept{GoogleCloudAiplatformV1CancelCustomJobRequest functions}
