% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.customJobs.create}
\alias{projects.locations.customJobs.create}
\title{Creates a CustomJob. A created CustomJob right away will be attempted to be run.}
\usage{
projects.locations.customJobs.create(GoogleCloudAiplatformV1CustomJob, parent)
}
\arguments{
\item{GoogleCloudAiplatformV1CustomJob}{The \link{GoogleCloudAiplatformV1CustomJob} object to pass to this method}

\item{parent}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1CustomJob functions: 
\code{\link{GoogleCloudAiplatformV1CustomJob.labels}()},
\code{\link{GoogleCloudAiplatformV1CustomJob.webAccessUris}()},
\code{\link{GoogleCloudAiplatformV1CustomJob}()}
}
\concept{GoogleCloudAiplatformV1CustomJob functions}
