% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.dataLabelingJobs.cancel}
\alias{projects.locations.dataLabelingJobs.cancel}
\title{Cancels a DataLabelingJob. Success of cancellation is not guaranteed.}
\usage{
projects.locations.dataLabelingJobs.cancel(
  GoogleCloudAiplatformV1CancelDataLabelingJobRequest,
  name
)
}
\arguments{
\item{GoogleCloudAiplatformV1CancelDataLabelingJobRequest}{The \link{GoogleCloudAiplatformV1CancelDataLabelingJobRequest} object to pass to this method}

\item{name}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1CancelDataLabelingJobRequest functions: 
\code{\link{GoogleCloudAiplatformV1CancelDataLabelingJobRequest}()}
}
\concept{GoogleCloudAiplatformV1CancelDataLabelingJobRequest functions}
