% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.dataLabelingJobs.create}
\alias{projects.locations.dataLabelingJobs.create}
\title{Creates a DataLabelingJob.}
\usage{
projects.locations.dataLabelingJobs.create(
  GoogleCloudAiplatformV1DataLabelingJob,
  parent
)
}
\arguments{
\item{GoogleCloudAiplatformV1DataLabelingJob}{The \link{GoogleCloudAiplatformV1DataLabelingJob} object to pass to this method}

\item{parent}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1DataLabelingJob functions: 
\code{\link{GoogleCloudAiplatformV1DataLabelingJob.annotationLabels}()},
\code{\link{GoogleCloudAiplatformV1DataLabelingJob.labels}()},
\code{\link{GoogleCloudAiplatformV1DataLabelingJob}()}
}
\concept{GoogleCloudAiplatformV1DataLabelingJob functions}
