% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.datasets.create}
\alias{projects.locations.datasets.create}
\title{Creates a Dataset.}
\usage{
projects.locations.datasets.create(GoogleCloudAiplatformV1Dataset, parent)
}
\arguments{
\item{GoogleCloudAiplatformV1Dataset}{The \link{GoogleCloudAiplatformV1Dataset} object to pass to this method}

\item{parent}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1Dataset functions: 
\code{\link{GoogleCloudAiplatformV1Dataset.labels}()},
\code{\link{GoogleCloudAiplatformV1Dataset}()},
\code{\link{projects.locations.datasets.patch}()}
}
\concept{GoogleCloudAiplatformV1Dataset functions}
