% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.datasets.searchDataItems}
\alias{projects.locations.datasets.searchDataItems}
\title{Searches DataItems in a Dataset.}
\usage{
projects.locations.datasets.searchDataItems(
  dataset,
  pageSize = NULL,
  annotationFilters = NULL,
  dataItemFilter = NULL,
  orderByAnnotation.orderBy = NULL,
  orderByDataItem = NULL,
  pageToken = NULL,
  orderByAnnotation.savedQuery = NULL,
  annotationsLimit = NULL,
  annotationsFilter = NULL,
  dataLabelingJob = NULL,
  fieldMask = NULL,
  savedQuery = NULL,
  orderBy = NULL
)
}
\arguments{
\item{dataset}{Required}

\item{pageSize}{Requested page size}

\item{annotationFilters}{An expression that specifies what Annotations will be returned per DataItem}

\item{dataItemFilter}{An expression for filtering the DataItem that will be returned}

\item{orderByAnnotation.orderBy}{A comma-separated list of annotation fields to order by, sorted in ascending order}

\item{orderByDataItem}{A comma-separated list of data item fields to order by, sorted in ascending order}

\item{pageToken}{A token identifying a page of results for the server to return Typically obtained via SearchDataItemsResponse}

\item{orderByAnnotation.savedQuery}{Required}

\item{annotationsLimit}{If set, only up to this many of Annotations will be returned per DataItemView}

\item{annotationsFilter}{An expression for filtering the Annotations that will be returned per DataItem}

\item{dataLabelingJob}{The resource name of a DataLabelingJob}

\item{fieldMask}{Mask specifying which fields of DataItemView to read}

\item{savedQuery}{The resource name of a SavedQuery(annotation set in UI)}

\item{orderBy}{A comma-separated list of fields to order by, sorted in ascending order}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}
}
