% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.endpoints.deployModel}
\alias{projects.locations.endpoints.deployModel}
\title{Deploys a Model into this Endpoint, creating a DeployedModel within it.}
\usage{
projects.locations.endpoints.deployModel(
  GoogleCloudAiplatformV1DeployModelRequest,
  endpoint
)
}
\arguments{
\item{GoogleCloudAiplatformV1DeployModelRequest}{The \link{GoogleCloudAiplatformV1DeployModelRequest} object to pass to this method}

\item{endpoint}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1DeployModelRequest functions: 
\code{\link{GoogleCloudAiplatformV1DeployModelRequest.trafficSplit}()},
\code{\link{GoogleCloudAiplatformV1DeployModelRequest}()}
}
\concept{GoogleCloudAiplatformV1DeployModelRequest functions}
