% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.endpoints.explain}
\alias{projects.locations.endpoints.explain}
\title{Perform an online explanation. If deployed_model_id is specified, the corresponding DeployModel must have explanation_spec populated. If deployed_model_id is not specified, all DeployedModels must have explanation_spec populated.}
\usage{
projects.locations.endpoints.explain(
  GoogleCloudAiplatformV1ExplainRequest,
  endpoint
)
}
\arguments{
\item{GoogleCloudAiplatformV1ExplainRequest}{The \link{GoogleCloudAiplatformV1ExplainRequest} object to pass to this method}

\item{endpoint}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1ExplainRequest functions: 
\code{\link{GoogleCloudAiplatformV1ExplainRequest}()}
}
\concept{GoogleCloudAiplatformV1ExplainRequest functions}
