% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.endpoints.mutateDeployedModel}
\alias{projects.locations.endpoints.mutateDeployedModel}
\title{Updates an existing deployed model. Updatable fields include \code{min_replica_count}, \code{max_replica_count}, \code{autoscaling_metric_specs}, \code{disable_container_logging} (v1 only), and \code{enable_container_logging} (v1beta1 only).}
\usage{
projects.locations.endpoints.mutateDeployedModel(
  GoogleCloudAiplatformV1MutateDeployedModelRequest,
  endpoint
)
}
\arguments{
\item{GoogleCloudAiplatformV1MutateDeployedModelRequest}{The \link{GoogleCloudAiplatformV1MutateDeployedModelRequest} object to pass to this method}

\item{endpoint}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1MutateDeployedModelRequest functions: 
\code{\link{GoogleCloudAiplatformV1MutateDeployedModelRequest}()}
}
\concept{GoogleCloudAiplatformV1MutateDeployedModelRequest functions}
