% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.endpoints.patch}
\alias{projects.locations.endpoints.patch}
\title{Updates an Endpoint.}
\usage{
projects.locations.endpoints.patch(
  GoogleCloudAiplatformV1Endpoint,
  name,
  updateMask = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1Endpoint}{The \link{GoogleCloudAiplatformV1Endpoint} object to pass to this method}

\item{name}{Output only}

\item{updateMask}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1Endpoint functions: 
\code{\link{GoogleCloudAiplatformV1Endpoint.labels}()},
\code{\link{GoogleCloudAiplatformV1Endpoint.trafficSplit}()},
\code{\link{GoogleCloudAiplatformV1Endpoint}()},
\code{\link{projects.locations.endpoints.create}()}
}
\concept{GoogleCloudAiplatformV1Endpoint functions}
