% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.endpoints.undeployModel}
\alias{projects.locations.endpoints.undeployModel}
\title{Undeploys a Model from an Endpoint, removing a DeployedModel from it, and freeing all resources it's using.}
\usage{
projects.locations.endpoints.undeployModel(
  GoogleCloudAiplatformV1UndeployModelRequest,
  endpoint
)
}
\arguments{
\item{GoogleCloudAiplatformV1UndeployModelRequest}{The \link{GoogleCloudAiplatformV1UndeployModelRequest} object to pass to this method}

\item{endpoint}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1UndeployModelRequest functions: 
\code{\link{GoogleCloudAiplatformV1UndeployModelRequest.trafficSplit}()},
\code{\link{GoogleCloudAiplatformV1UndeployModelRequest}()}
}
\concept{GoogleCloudAiplatformV1UndeployModelRequest functions}
