% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.featurestores.entityTypes.create}
\alias{projects.locations.featurestores.entityTypes.create}
\title{Creates a new EntityType in a given Featurestore.}
\usage{
projects.locations.featurestores.entityTypes.create(
  GoogleCloudAiplatformV1EntityType,
  parent,
  entityTypeId = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1EntityType}{The \link{GoogleCloudAiplatformV1EntityType} object to pass to this method}

\item{parent}{Required}

\item{entityTypeId}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1EntityType functions: 
\code{\link{GoogleCloudAiplatformV1EntityType.labels}()},
\code{\link{GoogleCloudAiplatformV1EntityType}()},
\code{\link{projects.locations.featurestores.entityTypes.patch}()}
}
\concept{GoogleCloudAiplatformV1EntityType functions}
