% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.featurestores.entityTypes.delete}
\alias{projects.locations.featurestores.entityTypes.delete}
\title{Deletes a single EntityType. The EntityType must not have any Features or \code{force} must be set to true for the request to succeed.}
\usage{
projects.locations.featurestores.entityTypes.delete(name, force = NULL)
}
\arguments{
\item{name}{Required}

\item{force}{If set to true, any Features for this EntityType will also be deleted}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}
}
