% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.featurestores.entityTypes.deleteFeatureValues}
\alias{projects.locations.featurestores.entityTypes.deleteFeatureValues}
\title{Delete Feature values from Featurestore. The progress of the deletion is tracked by the returned operation. The deleted feature values are guaranteed to be invisible to subsequent read operations after the operation is marked as successfully done. If a delete feature values operation fails, the feature values returned from reads and exports may be inconsistent. If consistency is required, the caller must retry the same delete request again and wait till the new operation returned is marked as successfully done.}
\usage{
projects.locations.featurestores.entityTypes.deleteFeatureValues(
  GoogleCloudAiplatformV1DeleteFeatureValuesRequest,
  entityType
)
}
\arguments{
\item{GoogleCloudAiplatformV1DeleteFeatureValuesRequest}{The \link{GoogleCloudAiplatformV1DeleteFeatureValuesRequest} object to pass to this method}

\item{entityType}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1DeleteFeatureValuesRequest functions: 
\code{\link{GoogleCloudAiplatformV1DeleteFeatureValuesRequest}()}
}
\concept{GoogleCloudAiplatformV1DeleteFeatureValuesRequest functions}
