% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.featurestores.entityTypes.exportFeatureValues}
\alias{projects.locations.featurestores.entityTypes.exportFeatureValues}
\title{Exports Feature values from all the entities of a target EntityType.}
\usage{
projects.locations.featurestores.entityTypes.exportFeatureValues(
  GoogleCloudAiplatformV1ExportFeatureValuesRequest,
  entityType
)
}
\arguments{
\item{GoogleCloudAiplatformV1ExportFeatureValuesRequest}{The \link{GoogleCloudAiplatformV1ExportFeatureValuesRequest} object to pass to this method}

\item{entityType}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1ExportFeatureValuesRequest functions: 
\code{\link{GoogleCloudAiplatformV1ExportFeatureValuesRequest}()}
}
\concept{GoogleCloudAiplatformV1ExportFeatureValuesRequest functions}
