% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.featurestores.entityTypes.features.list}
\alias{projects.locations.featurestores.entityTypes.features.list}
\title{Lists Features in a given EntityType.}
\usage{
projects.locations.featurestores.entityTypes.features.list(
  parent,
  pageToken = NULL,
  orderBy = NULL,
  filter = NULL,
  latestStatsCount = NULL,
  pageSize = NULL,
  readMask = NULL
)
}
\arguments{
\item{parent}{Required}

\item{pageToken}{A page token, received from a previous FeaturestoreService}

\item{orderBy}{A comma-separated list of fields to order by, sorted in ascending order}

\item{filter}{Lists the Features that match the filter expression}

\item{latestStatsCount}{If set, return the most recent ListFeaturesRequest}

\item{pageSize}{The maximum number of Features to return}

\item{readMask}{Mask specifying which fields to read}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}
}
