% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.featurestores.entityTypes.patch}
\alias{projects.locations.featurestores.entityTypes.patch}
\title{Updates the parameters of a single EntityType.}
\usage{
projects.locations.featurestores.entityTypes.patch(
  GoogleCloudAiplatformV1EntityType,
  name,
  updateMask = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1EntityType}{The \link{GoogleCloudAiplatformV1EntityType} object to pass to this method}

\item{name}{Immutable}

\item{updateMask}{Field mask is used to specify the fields to be overwritten in the EntityType resource by the update}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1EntityType functions: 
\code{\link{GoogleCloudAiplatformV1EntityType.labels}()},
\code{\link{GoogleCloudAiplatformV1EntityType}()},
\code{\link{projects.locations.featurestores.entityTypes.create}()}
}
\concept{GoogleCloudAiplatformV1EntityType functions}
