% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.featurestores.entityTypes.readFeatureValues}
\alias{projects.locations.featurestores.entityTypes.readFeatureValues}
\title{Reads Feature values of a specific entity of an EntityType. For reading feature values of multiple entities of an EntityType, please use StreamingReadFeatureValues.}
\usage{
projects.locations.featurestores.entityTypes.readFeatureValues(
  GoogleCloudAiplatformV1ReadFeatureValuesRequest,
  entityType
)
}
\arguments{
\item{GoogleCloudAiplatformV1ReadFeatureValuesRequest}{The \link{GoogleCloudAiplatformV1ReadFeatureValuesRequest} object to pass to this method}

\item{entityType}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1ReadFeatureValuesRequest functions: 
\code{\link{GoogleCloudAiplatformV1ReadFeatureValuesRequest}()}
}
\concept{GoogleCloudAiplatformV1ReadFeatureValuesRequest functions}
