% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.featurestores.entityTypes.setIamPolicy}
\alias{projects.locations.featurestores.entityTypes.setIamPolicy}
\title{Sets the access control policy on the specified resource. Replaces any existing policy. Can return \code{NOT_FOUND}, \code{INVALID_ARGUMENT}, and \code{PERMISSION_DENIED} errors.}
\usage{
projects.locations.featurestores.entityTypes.setIamPolicy(
  GoogleIamV1SetIamPolicyRequest,
  resource
)
}
\arguments{
\item{GoogleIamV1SetIamPolicyRequest}{The \link{GoogleIamV1SetIamPolicyRequest} object to pass to this method}

\item{resource}{REQUIRED: The resource for which the policy is being specified}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleIamV1SetIamPolicyRequest functions: 
\code{\link{GoogleIamV1SetIamPolicyRequest}()},
\code{\link{projects.locations.featurestores.setIamPolicy}()}
}
\concept{GoogleIamV1SetIamPolicyRequest functions}
