% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.featurestores.entityTypes.streamingReadFeatureValues}
\alias{projects.locations.featurestores.entityTypes.streamingReadFeatureValues}
\title{Reads Feature values for multiple entities. Depending on their size, data for different entities may be broken up across multiple responses.}
\usage{
projects.locations.featurestores.entityTypes.streamingReadFeatureValues(
  GoogleCloudAiplatformV1StreamingReadFeatureValuesRequest,
  entityType
)
}
\arguments{
\item{GoogleCloudAiplatformV1StreamingReadFeatureValuesRequest}{The \link{GoogleCloudAiplatformV1StreamingReadFeatureValuesRequest} object to pass to this method}

\item{entityType}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1StreamingReadFeatureValuesRequest functions: 
\code{\link{GoogleCloudAiplatformV1StreamingReadFeatureValuesRequest}()}
}
\concept{GoogleCloudAiplatformV1StreamingReadFeatureValuesRequest functions}
