% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.featurestores.entityTypes.testIamPermissions}
\alias{projects.locations.featurestores.entityTypes.testIamPermissions}
\title{Returns permissions that a caller has on the specified resource. If the resource does not exist, this will return an empty set of permissions, not a \code{NOT_FOUND} error. Note: This operation is designed to be used for building permission-aware UIs and command-line tools, not for authorization checking. This operation may 'fail open' without warning.}
\usage{
projects.locations.featurestores.entityTypes.testIamPermissions(
  resource,
  permissions = NULL
)
}
\arguments{
\item{resource}{REQUIRED: The resource for which the policy detail is being requested}

\item{permissions}{The set of permissions to check for the \code{resource}}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}
}
