% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.featurestores.entityTypes.writeFeatureValues}
\alias{projects.locations.featurestores.entityTypes.writeFeatureValues}
\title{Writes Feature values of one or more entities of an EntityType. The Feature values are merged into existing entities if any. The Feature values to be written must have timestamp within the online storage retention.}
\usage{
projects.locations.featurestores.entityTypes.writeFeatureValues(
  GoogleCloudAiplatformV1WriteFeatureValuesRequest,
  entityType
)
}
\arguments{
\item{GoogleCloudAiplatformV1WriteFeatureValuesRequest}{The \link{GoogleCloudAiplatformV1WriteFeatureValuesRequest} object to pass to this method}

\item{entityType}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1WriteFeatureValuesRequest functions: 
\code{\link{GoogleCloudAiplatformV1WriteFeatureValuesRequest}()}
}
\concept{GoogleCloudAiplatformV1WriteFeatureValuesRequest functions}
