% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.featurestores.searchFeatures}
\alias{projects.locations.featurestores.searchFeatures}
\title{Searches Features matching a query in a given project.}
\usage{
projects.locations.featurestores.searchFeatures(
  location,
  pageSize = NULL,
  pageToken = NULL,
  query = NULL
)
}
\arguments{
\item{location}{Required}

\item{pageSize}{The maximum number of Features to return}

\item{pageToken}{A page token, received from a previous FeaturestoreService}

\item{query}{Query string that is a conjunction of field-restricted queries and/or field-restricted filters}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}
}
