% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.indexEndpoints.findNeighbors}
\alias{projects.locations.indexEndpoints.findNeighbors}
\title{Finds the nearest neighbors of each vector within the request.}
\usage{
projects.locations.indexEndpoints.findNeighbors(
  GoogleCloudAiplatformV1FindNeighborsRequest,
  indexEndpoint
)
}
\arguments{
\item{GoogleCloudAiplatformV1FindNeighborsRequest}{The \link{GoogleCloudAiplatformV1FindNeighborsRequest} object to pass to this method}

\item{indexEndpoint}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1FindNeighborsRequest functions: 
\code{\link{GoogleCloudAiplatformV1FindNeighborsRequest}()}
}
\concept{GoogleCloudAiplatformV1FindNeighborsRequest functions}
