% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.indexEndpoints.mutateDeployedIndex}
\alias{projects.locations.indexEndpoints.mutateDeployedIndex}
\title{Update an existing DeployedIndex under an IndexEndpoint.}
\usage{
projects.locations.indexEndpoints.mutateDeployedIndex(
  GoogleCloudAiplatformV1DeployedIndex,
  indexEndpoint
)
}
\arguments{
\item{GoogleCloudAiplatformV1DeployedIndex}{The \link{GoogleCloudAiplatformV1DeployedIndex} object to pass to this method}

\item{indexEndpoint}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1DeployedIndex functions: 
\code{\link{GoogleCloudAiplatformV1DeployedIndex}()}
}
\concept{GoogleCloudAiplatformV1DeployedIndex functions}
