% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.indexEndpoints.readIndexDatapoints}
\alias{projects.locations.indexEndpoints.readIndexDatapoints}
\title{Reads the datapoints/vectors of the given IDs. A maximum of 1000 datapoints can be retrieved in a batch.}
\usage{
projects.locations.indexEndpoints.readIndexDatapoints(
  GoogleCloudAiplatformV1ReadIndexDatapointsRequest,
  indexEndpoint
)
}
\arguments{
\item{GoogleCloudAiplatformV1ReadIndexDatapointsRequest}{The \link{GoogleCloudAiplatformV1ReadIndexDatapointsRequest} object to pass to this method}

\item{indexEndpoint}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1ReadIndexDatapointsRequest functions: 
\code{\link{GoogleCloudAiplatformV1ReadIndexDatapointsRequest}()}
}
\concept{GoogleCloudAiplatformV1ReadIndexDatapointsRequest functions}
