% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.indexEndpoints.undeployIndex}
\alias{projects.locations.indexEndpoints.undeployIndex}
\title{Undeploys an Index from an IndexEndpoint, removing a DeployedIndex from it, and freeing all resources it's using.}
\usage{
projects.locations.indexEndpoints.undeployIndex(
  GoogleCloudAiplatformV1UndeployIndexRequest,
  indexEndpoint
)
}
\arguments{
\item{GoogleCloudAiplatformV1UndeployIndexRequest}{The \link{GoogleCloudAiplatformV1UndeployIndexRequest} object to pass to this method}

\item{indexEndpoint}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1UndeployIndexRequest functions: 
\code{\link{GoogleCloudAiplatformV1UndeployIndexRequest}()}
}
\concept{GoogleCloudAiplatformV1UndeployIndexRequest functions}
