% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.indexes.patch}
\alias{projects.locations.indexes.patch}
\title{Updates an Index.}
\usage{
projects.locations.indexes.patch(
  GoogleCloudAiplatformV1Index,
  name,
  updateMask = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1Index}{The \link{GoogleCloudAiplatformV1Index} object to pass to this method}

\item{name}{Output only}

\item{updateMask}{The update mask applies to the resource}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1Index functions: 
\code{\link{GoogleCloudAiplatformV1Index.labels}()},
\code{\link{GoogleCloudAiplatformV1Index}()},
\code{\link{projects.locations.indexes.create}()}
}
\concept{GoogleCloudAiplatformV1Index functions}
