% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.indexes.removeDatapoints}
\alias{projects.locations.indexes.removeDatapoints}
\title{Remove Datapoints from an Index.}
\usage{
projects.locations.indexes.removeDatapoints(
  GoogleCloudAiplatformV1RemoveDatapointsRequest,
  index
)
}
\arguments{
\item{GoogleCloudAiplatformV1RemoveDatapointsRequest}{The \link{GoogleCloudAiplatformV1RemoveDatapointsRequest} object to pass to this method}

\item{index}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1RemoveDatapointsRequest functions: 
\code{\link{GoogleCloudAiplatformV1RemoveDatapointsRequest}()}
}
\concept{GoogleCloudAiplatformV1RemoveDatapointsRequest functions}
