% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.metadataStores.artifacts.patch}
\alias{projects.locations.metadataStores.artifacts.patch}
\title{Updates a stored Artifact.}
\usage{
projects.locations.metadataStores.artifacts.patch(
  GoogleCloudAiplatformV1Artifact,
  name,
  allowMissing = NULL,
  updateMask = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1Artifact}{The \link{GoogleCloudAiplatformV1Artifact} object to pass to this method}

\item{name}{Output only}

\item{allowMissing}{If set to true, and the Artifact is not found, a new Artifact is created}

\item{updateMask}{Optional}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1Artifact functions: 
\code{\link{GoogleCloudAiplatformV1Artifact.labels}()},
\code{\link{GoogleCloudAiplatformV1Artifact.metadata}()},
\code{\link{GoogleCloudAiplatformV1Artifact}()},
\code{\link{projects.locations.metadataStores.artifacts.create}()}
}
\concept{GoogleCloudAiplatformV1Artifact functions}
