% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.metadataStores.artifacts.queryArtifactLineageSubgraph}
\alias{projects.locations.metadataStores.artifacts.queryArtifactLineageSubgraph}
\title{Retrieves lineage of an Artifact represented through Artifacts and Executions connected by Event edges and returned as a LineageSubgraph.}
\usage{
projects.locations.metadataStores.artifacts.queryArtifactLineageSubgraph(
  artifact,
  maxHops = NULL,
  filter = NULL
)
}
\arguments{
\item{artifact}{Required}

\item{maxHops}{Specifies the size of the lineage graph in terms of number of hops from the specified artifact}

\item{filter}{Filter specifying the boolean condition for the Artifacts to satisfy in order to be part of the Lineage Subgraph}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}
}
