% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.metadataStores.contexts.addContextChildren}
\alias{projects.locations.metadataStores.contexts.addContextChildren}
\title{Adds a set of Contexts as children to a parent Context. If any of the child Contexts have already been added to the parent Context, they are simply skipped. If this call would create a cycle or cause any Context to have more than 10 parents, the request will fail with an INVALID_ARGUMENT error.}
\usage{
projects.locations.metadataStores.contexts.addContextChildren(
  GoogleCloudAiplatformV1AddContextChildrenRequest,
  context
)
}
\arguments{
\item{GoogleCloudAiplatformV1AddContextChildrenRequest}{The \link{GoogleCloudAiplatformV1AddContextChildrenRequest} object to pass to this method}

\item{context}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1AddContextChildrenRequest functions: 
\code{\link{GoogleCloudAiplatformV1AddContextChildrenRequest}()}
}
\concept{GoogleCloudAiplatformV1AddContextChildrenRequest functions}
