% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.metadataStores.contexts.create}
\alias{projects.locations.metadataStores.contexts.create}
\title{Creates a Context associated with a MetadataStore.}
\usage{
projects.locations.metadataStores.contexts.create(
  GoogleCloudAiplatformV1Context,
  parent,
  contextId = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1Context}{The \link{GoogleCloudAiplatformV1Context} object to pass to this method}

\item{parent}{Required}

\item{contextId}{The {context} portion of the resource name with the format: \code{projects/{project}/locations/{location}/metadataStores/{metadatastore}/contexts/{context}}}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1Context functions: 
\code{\link{GoogleCloudAiplatformV1Context.labels}()},
\code{\link{GoogleCloudAiplatformV1Context.metadata}()},
\code{\link{GoogleCloudAiplatformV1Context}()},
\code{\link{projects.locations.metadataStores.contexts.patch}()}
}
\concept{GoogleCloudAiplatformV1Context functions}
