% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.metadataStores.contexts.patch}
\alias{projects.locations.metadataStores.contexts.patch}
\title{Updates a stored Context.}
\usage{
projects.locations.metadataStores.contexts.patch(
  GoogleCloudAiplatformV1Context,
  name,
  updateMask = NULL,
  allowMissing = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1Context}{The \link{GoogleCloudAiplatformV1Context} object to pass to this method}

\item{name}{Immutable}

\item{updateMask}{Optional}

\item{allowMissing}{If set to true, and the Context is not found, a new Context is created}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1Context functions: 
\code{\link{GoogleCloudAiplatformV1Context.labels}()},
\code{\link{GoogleCloudAiplatformV1Context.metadata}()},
\code{\link{GoogleCloudAiplatformV1Context}()},
\code{\link{projects.locations.metadataStores.contexts.create}()}
}
\concept{GoogleCloudAiplatformV1Context functions}
