% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.metadataStores.contexts.removeContextChildren}
\alias{projects.locations.metadataStores.contexts.removeContextChildren}
\title{Remove a set of children contexts from a parent Context. If any of the child Contexts were NOT added to the parent Context, they are simply skipped.}
\usage{
projects.locations.metadataStores.contexts.removeContextChildren(
  GoogleCloudAiplatformV1RemoveContextChildrenRequest,
  context
)
}
\arguments{
\item{GoogleCloudAiplatformV1RemoveContextChildrenRequest}{The \link{GoogleCloudAiplatformV1RemoveContextChildrenRequest} object to pass to this method}

\item{context}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1RemoveContextChildrenRequest functions: 
\code{\link{GoogleCloudAiplatformV1RemoveContextChildrenRequest}()}
}
\concept{GoogleCloudAiplatformV1RemoveContextChildrenRequest functions}
