% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.metadataStores.create}
\alias{projects.locations.metadataStores.create}
\title{Initializes a MetadataStore, including allocation of resources.}
\usage{
projects.locations.metadataStores.create(
  GoogleCloudAiplatformV1MetadataStore,
  parent,
  metadataStoreId = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1MetadataStore}{The \link{GoogleCloudAiplatformV1MetadataStore} object to pass to this method}

\item{parent}{Required}

\item{metadataStoreId}{The {metadatastore} portion of the resource name with the format: \code{projects/{project}/locations/{location}/metadataStores/{metadatastore}} If not provided, the MetadataStore's ID will be a UUID generated by the service}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1MetadataStore functions: 
\code{\link{GoogleCloudAiplatformV1MetadataStore}()}
}
\concept{GoogleCloudAiplatformV1MetadataStore functions}
