% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.metadataStores.executions.addExecutionEvents}
\alias{projects.locations.metadataStores.executions.addExecutionEvents}
\title{Adds Events to the specified Execution. An Event indicates whether an Artifact was used as an input or output for an Execution. If an Event already exists between the Execution and the Artifact, the Event is skipped.}
\usage{
projects.locations.metadataStores.executions.addExecutionEvents(
  GoogleCloudAiplatformV1AddExecutionEventsRequest,
  execution
)
}
\arguments{
\item{GoogleCloudAiplatformV1AddExecutionEventsRequest}{The \link{GoogleCloudAiplatformV1AddExecutionEventsRequest} object to pass to this method}

\item{execution}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1AddExecutionEventsRequest functions: 
\code{\link{GoogleCloudAiplatformV1AddExecutionEventsRequest}()}
}
\concept{GoogleCloudAiplatformV1AddExecutionEventsRequest functions}
