% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.metadataStores.executions.list}
\alias{projects.locations.metadataStores.executions.list}
\title{Lists Executions in the MetadataStore.}
\usage{
projects.locations.metadataStores.executions.list(
  parent,
  orderBy = NULL,
  filter = NULL,
  pageToken = NULL,
  pageSize = NULL
)
}
\arguments{
\item{parent}{Required}

\item{orderBy}{How the list of messages is ordered}

\item{filter}{Filter specifying the boolean condition for the Executions to satisfy in order to be part of the result set}

\item{pageToken}{A page token, received from a previous MetadataService}

\item{pageSize}{The maximum number of Executions to return}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}
}
