% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.metadataStores.executions.patch}
\alias{projects.locations.metadataStores.executions.patch}
\title{Updates a stored Execution.}
\usage{
projects.locations.metadataStores.executions.patch(
  GoogleCloudAiplatformV1Execution,
  name,
  allowMissing = NULL,
  updateMask = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1Execution}{The \link{GoogleCloudAiplatformV1Execution} object to pass to this method}

\item{name}{Output only}

\item{allowMissing}{If set to true, and the Execution is not found, a new Execution is created}

\item{updateMask}{Optional}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1Execution functions: 
\code{\link{GoogleCloudAiplatformV1Execution.labels}()},
\code{\link{GoogleCloudAiplatformV1Execution.metadata}()},
\code{\link{GoogleCloudAiplatformV1Execution}()},
\code{\link{projects.locations.metadataStores.executions.create}()}
}
\concept{GoogleCloudAiplatformV1Execution functions}
