% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.metadataStores.executions.purge}
\alias{projects.locations.metadataStores.executions.purge}
\title{Purges Executions.}
\usage{
projects.locations.metadataStores.executions.purge(
  GoogleCloudAiplatformV1PurgeExecutionsRequest,
  parent
)
}
\arguments{
\item{GoogleCloudAiplatformV1PurgeExecutionsRequest}{The \link{GoogleCloudAiplatformV1PurgeExecutionsRequest} object to pass to this method}

\item{parent}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1PurgeExecutionsRequest functions: 
\code{\link{GoogleCloudAiplatformV1PurgeExecutionsRequest}()}
}
\concept{GoogleCloudAiplatformV1PurgeExecutionsRequest functions}
