% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.metadataStores.metadataSchemas.create}
\alias{projects.locations.metadataStores.metadataSchemas.create}
\title{Creates a MetadataSchema.}
\usage{
projects.locations.metadataStores.metadataSchemas.create(
  GoogleCloudAiplatformV1MetadataSchema,
  parent,
  metadataSchemaId = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1MetadataSchema}{The \link{GoogleCloudAiplatformV1MetadataSchema} object to pass to this method}

\item{parent}{Required}

\item{metadataSchemaId}{The {metadata_schema} portion of the resource name with the format: \code{projects/{project}/locations/{location}/metadataStores/{metadatastore}/metadataSchemas/{metadataschema}} If not provided, the MetadataStore's ID will be a UUID generated by the service}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1MetadataSchema functions: 
\code{\link{GoogleCloudAiplatformV1MetadataSchema}()}
}
\concept{GoogleCloudAiplatformV1MetadataSchema functions}
