% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.metadataStores.metadataSchemas.list}
\alias{projects.locations.metadataStores.metadataSchemas.list}
\title{Lists MetadataSchemas.}
\usage{
projects.locations.metadataStores.metadataSchemas.list(
  parent,
  filter = NULL,
  pageSize = NULL,
  pageToken = NULL
)
}
\arguments{
\item{parent}{Required}

\item{filter}{A query to filter available MetadataSchemas for matching results}

\item{pageSize}{The maximum number of MetadataSchemas to return}

\item{pageToken}{A page token, received from a previous MetadataService}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}
}
