% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.migratableResources.search}
\alias{projects.locations.migratableResources.search}
\title{Searches all of the resources in automl.googleapis.com, datalabeling.googleapis.com and ml.googleapis.com that can be migrated to Vertex AI's given location.}
\usage{
projects.locations.migratableResources.search(
  GoogleCloudAiplatformV1SearchMigratableResourcesRequest,
  parent
)
}
\arguments{
\item{GoogleCloudAiplatformV1SearchMigratableResourcesRequest}{The \link{GoogleCloudAiplatformV1SearchMigratableResourcesRequest} object to pass to this method}

\item{parent}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1SearchMigratableResourcesRequest functions: 
\code{\link{GoogleCloudAiplatformV1SearchMigratableResourcesRequest}()}
}
\concept{GoogleCloudAiplatformV1SearchMigratableResourcesRequest functions}
