% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.modelDeploymentMonitoringJobs.create}
\alias{projects.locations.modelDeploymentMonitoringJobs.create}
\title{Creates a ModelDeploymentMonitoringJob. It will run periodically on a configured interval.}
\usage{
projects.locations.modelDeploymentMonitoringJobs.create(
  GoogleCloudAiplatformV1ModelDeploymentMonitoringJob,
  parent
)
}
\arguments{
\item{GoogleCloudAiplatformV1ModelDeploymentMonitoringJob}{The \link{GoogleCloudAiplatformV1ModelDeploymentMonitoringJob} object to pass to this method}

\item{parent}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1ModelDeploymentMonitoringJob functions: 
\code{\link{GoogleCloudAiplatformV1ModelDeploymentMonitoringJob.labels}()},
\code{\link{GoogleCloudAiplatformV1ModelDeploymentMonitoringJob}()},
\code{\link{projects.locations.modelDeploymentMonitoringJobs.patch}()}
}
\concept{GoogleCloudAiplatformV1ModelDeploymentMonitoringJob functions}
