% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.modelDeploymentMonitoringJobs.pause}
\alias{projects.locations.modelDeploymentMonitoringJobs.pause}
\title{Pauses a ModelDeploymentMonitoringJob. If the job is running, the server makes a best effort to cancel the job. Will mark ModelDeploymentMonitoringJob.state to 'PAUSED'.}
\usage{
projects.locations.modelDeploymentMonitoringJobs.pause(
  GoogleCloudAiplatformV1PauseModelDeploymentMonitoringJobRequest,
  name
)
}
\arguments{
\item{GoogleCloudAiplatformV1PauseModelDeploymentMonitoringJobRequest}{The \link{GoogleCloudAiplatformV1PauseModelDeploymentMonitoringJobRequest} object to pass to this method}

\item{name}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1PauseModelDeploymentMonitoringJobRequest functions: 
\code{\link{GoogleCloudAiplatformV1PauseModelDeploymentMonitoringJobRequest}()}
}
\concept{GoogleCloudAiplatformV1PauseModelDeploymentMonitoringJobRequest functions}
