% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.modelDeploymentMonitoringJobs.resume}
\alias{projects.locations.modelDeploymentMonitoringJobs.resume}
\title{Resumes a paused ModelDeploymentMonitoringJob. It will start to run from next scheduled time. A deleted ModelDeploymentMonitoringJob can't be resumed.}
\usage{
projects.locations.modelDeploymentMonitoringJobs.resume(
  GoogleCloudAiplatformV1ResumeModelDeploymentMonitoringJobRequest,
  name
)
}
\arguments{
\item{GoogleCloudAiplatformV1ResumeModelDeploymentMonitoringJobRequest}{The \link{GoogleCloudAiplatformV1ResumeModelDeploymentMonitoringJobRequest} object to pass to this method}

\item{name}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1ResumeModelDeploymentMonitoringJobRequest functions: 
\code{\link{GoogleCloudAiplatformV1ResumeModelDeploymentMonitoringJobRequest}()}
}
\concept{GoogleCloudAiplatformV1ResumeModelDeploymentMonitoringJobRequest functions}
