% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.models.copy}
\alias{projects.locations.models.copy}
\title{Copies an already existing Vertex AI Model into the specified Location. The source Model must exist in the same Project. When copying custom Models, the users themselves are responsible for Model.metadata content to be region-agnostic, as well as making sure that any resources (e.g. files) it depends on remain accessible.}
\usage{
projects.locations.models.copy(GoogleCloudAiplatformV1CopyModelRequest, parent)
}
\arguments{
\item{GoogleCloudAiplatformV1CopyModelRequest}{The \link{GoogleCloudAiplatformV1CopyModelRequest} object to pass to this method}

\item{parent}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1CopyModelRequest functions: 
\code{\link{GoogleCloudAiplatformV1CopyModelRequest}()}
}
\concept{GoogleCloudAiplatformV1CopyModelRequest functions}
