% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.models.evaluations.import}
\alias{projects.locations.models.evaluations.import}
\title{Imports an externally generated ModelEvaluation.}
\usage{
projects.locations.models.evaluations.import(
  GoogleCloudAiplatformV1ImportModelEvaluationRequest,
  parent
)
}
\arguments{
\item{GoogleCloudAiplatformV1ImportModelEvaluationRequest}{The \link{GoogleCloudAiplatformV1ImportModelEvaluationRequest} object to pass to this method}

\item{parent}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1ImportModelEvaluationRequest functions: 
\code{\link{GoogleCloudAiplatformV1ImportModelEvaluationRequest}()}
}
\concept{GoogleCloudAiplatformV1ImportModelEvaluationRequest functions}
