% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.models.evaluations.slices.batchImport}
\alias{projects.locations.models.evaluations.slices.batchImport}
\title{Imports a list of externally generated EvaluatedAnnotations.}
\usage{
projects.locations.models.evaluations.slices.batchImport(
  GoogleCloudAiplatformV1BatchImportEvaluatedAnnotationsRequest,
  parent
)
}
\arguments{
\item{GoogleCloudAiplatformV1BatchImportEvaluatedAnnotationsRequest}{The \link{GoogleCloudAiplatformV1BatchImportEvaluatedAnnotationsRequest} object to pass to this method}

\item{parent}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1BatchImportEvaluatedAnnotationsRequest functions: 
\code{\link{GoogleCloudAiplatformV1BatchImportEvaluatedAnnotationsRequest}()}
}
\concept{GoogleCloudAiplatformV1BatchImportEvaluatedAnnotationsRequest functions}
