% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.models.export}
\alias{projects.locations.models.export}
\title{Exports a trained, exportable Model to a location specified by the user. A Model is considered to be exportable if it has at least one supported export format.}
\usage{
projects.locations.models.export(
  GoogleCloudAiplatformV1ExportModelRequest,
  name
)
}
\arguments{
\item{GoogleCloudAiplatformV1ExportModelRequest}{The \link{GoogleCloudAiplatformV1ExportModelRequest} object to pass to this method}

\item{name}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1ExportModelRequest functions: 
\code{\link{GoogleCloudAiplatformV1ExportModelRequest}()}
}
\concept{GoogleCloudAiplatformV1ExportModelRequest functions}
