% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.models.listVersions}
\alias{projects.locations.models.listVersions}
\title{Lists versions of the specified model.}
\usage{
projects.locations.models.listVersions(
  name,
  readMask = NULL,
  pageToken = NULL,
  filter = NULL,
  pageSize = NULL,
  orderBy = NULL
)
}
\arguments{
\item{name}{Required}

\item{readMask}{Mask specifying which fields to read}

\item{pageToken}{The standard list page token}

\item{filter}{An expression for filtering the results of the request}

\item{pageSize}{The standard list page size}

\item{orderBy}{A comma-separated list of fields to order by, sorted in ascending order}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}
}
