% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.models.mergeVersionAliases}
\alias{projects.locations.models.mergeVersionAliases}
\title{Merges a set of aliases for a Model version.}
\usage{
projects.locations.models.mergeVersionAliases(
  GoogleCloudAiplatformV1MergeVersionAliasesRequest,
  name
)
}
\arguments{
\item{GoogleCloudAiplatformV1MergeVersionAliasesRequest}{The \link{GoogleCloudAiplatformV1MergeVersionAliasesRequest} object to pass to this method}

\item{name}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1MergeVersionAliasesRequest functions: 
\code{\link{GoogleCloudAiplatformV1MergeVersionAliasesRequest}()}
}
\concept{GoogleCloudAiplatformV1MergeVersionAliasesRequest functions}
