% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.models.patch}
\alias{projects.locations.models.patch}
\title{Updates a Model.}
\usage{
projects.locations.models.patch(
  GoogleCloudAiplatformV1Model,
  name,
  updateMask = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1Model}{The \link{GoogleCloudAiplatformV1Model} object to pass to this method}

\item{name}{The resource name of the Model}

\item{updateMask}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1Model functions: 
\code{\link{GoogleCloudAiplatformV1Model.labels}()},
\code{\link{GoogleCloudAiplatformV1Model}()}
}
\concept{GoogleCloudAiplatformV1Model functions}
