% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.models.updateExplanationDataset}
\alias{projects.locations.models.updateExplanationDataset}
\title{Incrementally update the dataset used for an examples model.}
\usage{
projects.locations.models.updateExplanationDataset(
  GoogleCloudAiplatformV1UpdateExplanationDatasetRequest,
  model
)
}
\arguments{
\item{GoogleCloudAiplatformV1UpdateExplanationDatasetRequest}{The \link{GoogleCloudAiplatformV1UpdateExplanationDatasetRequest} object to pass to this method}

\item{model}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1UpdateExplanationDatasetRequest functions: 
\code{\link{GoogleCloudAiplatformV1UpdateExplanationDatasetRequest}()}
}
\concept{GoogleCloudAiplatformV1UpdateExplanationDatasetRequest functions}
