% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.nasJobs.cancel}
\alias{projects.locations.nasJobs.cancel}
\title{Cancels a NasJob. Starts asynchronous cancellation on the NasJob. The server makes a best effort to cancel the job, but success is not guaranteed. Clients can use JobService.GetNasJob or other methods to check whether the cancellation succeeded or whether the job completed despite cancellation. On successful cancellation, the NasJob is not deleted; instead it becomes a job with a NasJob.error value with a google.rpc.Status.code of 1, corresponding to \code{Code.CANCELLED}, and NasJob.state is set to \code{CANCELLED}.}
\usage{
projects.locations.nasJobs.cancel(
  GoogleCloudAiplatformV1CancelNasJobRequest,
  name
)
}
\arguments{
\item{GoogleCloudAiplatformV1CancelNasJobRequest}{The \link{GoogleCloudAiplatformV1CancelNasJobRequest} object to pass to this method}

\item{name}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1CancelNasJobRequest functions: 
\code{\link{GoogleCloudAiplatformV1CancelNasJobRequest}()}
}
\concept{GoogleCloudAiplatformV1CancelNasJobRequest functions}
