% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.nasJobs.create}
\alias{projects.locations.nasJobs.create}
\title{Creates a NasJob}
\usage{
projects.locations.nasJobs.create(GoogleCloudAiplatformV1NasJob, parent)
}
\arguments{
\item{GoogleCloudAiplatformV1NasJob}{The \link{GoogleCloudAiplatformV1NasJob} object to pass to this method}

\item{parent}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1NasJob functions: 
\code{\link{GoogleCloudAiplatformV1NasJob.labels}()},
\code{\link{GoogleCloudAiplatformV1NasJob}()}
}
\concept{GoogleCloudAiplatformV1NasJob functions}
