% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.pipelineJobs.create}
\alias{projects.locations.pipelineJobs.create}
\title{Creates a PipelineJob. A PipelineJob will run immediately when created.}
\usage{
projects.locations.pipelineJobs.create(
  GoogleCloudAiplatformV1PipelineJob,
  parent,
  pipelineJobId = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1PipelineJob}{The \link{GoogleCloudAiplatformV1PipelineJob} object to pass to this method}

\item{parent}{Required}

\item{pipelineJobId}{The ID to use for the PipelineJob, which will become the final component of the PipelineJob name}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1PipelineJob functions: 
\code{\link{GoogleCloudAiplatformV1PipelineJob.labels}()},
\code{\link{GoogleCloudAiplatformV1PipelineJob.pipelineSpec}()},
\code{\link{GoogleCloudAiplatformV1PipelineJob}()}
}
\concept{GoogleCloudAiplatformV1PipelineJob functions}
