% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.publishers.models.predict}
\alias{projects.locations.publishers.models.predict}
\title{Perform an online prediction.}
\usage{
projects.locations.publishers.models.predict(
  GoogleCloudAiplatformV1PredictRequest,
  endpoint
)
}
\arguments{
\item{GoogleCloudAiplatformV1PredictRequest}{The \link{GoogleCloudAiplatformV1PredictRequest} object to pass to this method}

\item{endpoint}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1PredictRequest functions: 
\code{\link{GoogleCloudAiplatformV1PredictRequest}()},
\code{\link{projects.locations.endpoints.predict}()}
}
\concept{GoogleCloudAiplatformV1PredictRequest functions}
