% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.publishers.models.rawPredict}
\alias{projects.locations.publishers.models.rawPredict}
\title{Perform an online prediction with an arbitrary HTTP payload. The response includes the following HTTP headers: * \code{X-Vertex-AI-Endpoint-Id}: ID of the Endpoint that served this prediction. * \code{X-Vertex-AI-Deployed-Model-Id}: ID of the Endpoint's DeployedModel that served this prediction.}
\usage{
projects.locations.publishers.models.rawPredict(
  GoogleCloudAiplatformV1RawPredictRequest,
  endpoint
)
}
\arguments{
\item{GoogleCloudAiplatformV1RawPredictRequest}{The \link{GoogleCloudAiplatformV1RawPredictRequest} object to pass to this method}

\item{endpoint}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1RawPredictRequest functions: 
\code{\link{GoogleCloudAiplatformV1RawPredictRequest}()},
\code{\link{projects.locations.endpoints.rawPredict}()}
}
\concept{GoogleCloudAiplatformV1RawPredictRequest functions}
