% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.publishers.models.serverStreamingPredict}
\alias{projects.locations.publishers.models.serverStreamingPredict}
\title{Perform a server-side streaming online prediction request for Vertex LLM streaming.}
\usage{
projects.locations.publishers.models.serverStreamingPredict(
  GoogleCloudAiplatformV1StreamingPredictRequest,
  endpoint
)
}
\arguments{
\item{GoogleCloudAiplatformV1StreamingPredictRequest}{The \link{GoogleCloudAiplatformV1StreamingPredictRequest} object to pass to this method}

\item{endpoint}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1StreamingPredictRequest functions: 
\code{\link{GoogleCloudAiplatformV1StreamingPredictRequest}()},
\code{\link{projects.locations.endpoints.serverStreamingPredict}()}
}
\concept{GoogleCloudAiplatformV1StreamingPredictRequest functions}
