% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.schedules.patch}
\alias{projects.locations.schedules.patch}
\title{Updates an active or paused Schedule. When the Schedule is updated, new runs will be scheduled starting from the updated next execution time after the update time based on the time_specification in the updated Schedule. All unstarted runs before the update time will be skipped while already created runs will NOT be paused or canceled.}
\usage{
projects.locations.schedules.patch(
  GoogleCloudAiplatformV1Schedule,
  name,
  updateMask = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1Schedule}{The \link{GoogleCloudAiplatformV1Schedule} object to pass to this method}

\item{name}{Immutable}

\item{updateMask}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1Schedule functions: 
\code{\link{GoogleCloudAiplatformV1Schedule}()},
\code{\link{projects.locations.schedules.create}()}
}
\concept{GoogleCloudAiplatformV1Schedule functions}
