% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.specialistPools.list}
\alias{projects.locations.specialistPools.list}
\title{Lists SpecialistPools in a Location.}
\usage{
projects.locations.specialistPools.list(
  parent,
  pageToken = NULL,
  pageSize = NULL,
  readMask = NULL
)
}
\arguments{
\item{parent}{Required}

\item{pageToken}{The standard list page token}

\item{pageSize}{The standard list page size}

\item{readMask}{Mask specifying which fields to read}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}
}
