% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.studies.lookup}
\alias{projects.locations.studies.lookup}
\title{Looks a study up using the user-defined display_name field instead of the fully qualified resource name.}
\usage{
projects.locations.studies.lookup(
  GoogleCloudAiplatformV1LookupStudyRequest,
  parent
)
}
\arguments{
\item{GoogleCloudAiplatformV1LookupStudyRequest}{The \link{GoogleCloudAiplatformV1LookupStudyRequest} object to pass to this method}

\item{parent}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1LookupStudyRequest functions: 
\code{\link{GoogleCloudAiplatformV1LookupStudyRequest}()}
}
\concept{GoogleCloudAiplatformV1LookupStudyRequest functions}
