% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.studies.trials.addTrialMeasurement}
\alias{projects.locations.studies.trials.addTrialMeasurement}
\title{Adds a measurement of the objective metrics to a Trial. This measurement is assumed to have been taken before the Trial is complete.}
\usage{
projects.locations.studies.trials.addTrialMeasurement(
  GoogleCloudAiplatformV1AddTrialMeasurementRequest,
  trialName
)
}
\arguments{
\item{GoogleCloudAiplatformV1AddTrialMeasurementRequest}{The \link{GoogleCloudAiplatformV1AddTrialMeasurementRequest} object to pass to this method}

\item{trialName}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1AddTrialMeasurementRequest functions: 
\code{\link{GoogleCloudAiplatformV1AddTrialMeasurementRequest}()}
}
\concept{GoogleCloudAiplatformV1AddTrialMeasurementRequest functions}
