% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.studies.trials.complete}
\alias{projects.locations.studies.trials.complete}
\title{Marks a Trial as complete.}
\usage{
projects.locations.studies.trials.complete(
  GoogleCloudAiplatformV1CompleteTrialRequest,
  name
)
}
\arguments{
\item{GoogleCloudAiplatformV1CompleteTrialRequest}{The \link{GoogleCloudAiplatformV1CompleteTrialRequest} object to pass to this method}

\item{name}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1CompleteTrialRequest functions: 
\code{\link{GoogleCloudAiplatformV1CompleteTrialRequest}()}
}
\concept{GoogleCloudAiplatformV1CompleteTrialRequest functions}
