% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.studies.trials.create}
\alias{projects.locations.studies.trials.create}
\title{Adds a user provided Trial to a Study.}
\usage{
projects.locations.studies.trials.create(GoogleCloudAiplatformV1Trial, parent)
}
\arguments{
\item{GoogleCloudAiplatformV1Trial}{The \link{GoogleCloudAiplatformV1Trial} object to pass to this method}

\item{parent}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1Trial functions: 
\code{\link{GoogleCloudAiplatformV1Trial.webAccessUris}()},
\code{\link{GoogleCloudAiplatformV1Trial}()}
}
\concept{GoogleCloudAiplatformV1Trial functions}
