% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.studies.trials.list}
\alias{projects.locations.studies.trials.list}
\title{Lists the Trials associated with a Study.}
\usage{
projects.locations.studies.trials.list(
  parent,
  pageToken = NULL,
  pageSize = NULL
)
}
\arguments{
\item{parent}{Required}

\item{pageToken}{Optional}

\item{pageSize}{Optional}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}
}
