% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.studies.trials.listOptimalTrials}
\alias{projects.locations.studies.trials.listOptimalTrials}
\title{Lists the pareto-optimal Trials for multi-objective Study or the optimal Trials for single-objective Study. The definition of pareto-optimal can be checked in wiki page. https://en.wikipedia.org/wiki/Pareto_efficiency}
\usage{
projects.locations.studies.trials.listOptimalTrials(
  GoogleCloudAiplatformV1ListOptimalTrialsRequest,
  parent
)
}
\arguments{
\item{GoogleCloudAiplatformV1ListOptimalTrialsRequest}{The \link{GoogleCloudAiplatformV1ListOptimalTrialsRequest} object to pass to this method}

\item{parent}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1ListOptimalTrialsRequest functions: 
\code{\link{GoogleCloudAiplatformV1ListOptimalTrialsRequest}()}
}
\concept{GoogleCloudAiplatformV1ListOptimalTrialsRequest functions}
