% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.studies.trials.stop}
\alias{projects.locations.studies.trials.stop}
\title{Stops a Trial.}
\usage{
projects.locations.studies.trials.stop(
  GoogleCloudAiplatformV1StopTrialRequest,
  name
)
}
\arguments{
\item{GoogleCloudAiplatformV1StopTrialRequest}{The \link{GoogleCloudAiplatformV1StopTrialRequest} object to pass to this method}

\item{name}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1StopTrialRequest functions: 
\code{\link{GoogleCloudAiplatformV1StopTrialRequest}()}
}
\concept{GoogleCloudAiplatformV1StopTrialRequest functions}
