% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.studies.trials.suggest}
\alias{projects.locations.studies.trials.suggest}
\title{Adds one or more Trials to a Study, with parameter values suggested by Vertex AI Vizier. Returns a long-running operation associated with the generation of Trial suggestions. When this long-running operation succeeds, it will contain a SuggestTrialsResponse.}
\usage{
projects.locations.studies.trials.suggest(
  GoogleCloudAiplatformV1SuggestTrialsRequest,
  parent
)
}
\arguments{
\item{GoogleCloudAiplatformV1SuggestTrialsRequest}{The \link{GoogleCloudAiplatformV1SuggestTrialsRequest} object to pass to this method}

\item{parent}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1SuggestTrialsRequest functions: 
\code{\link{GoogleCloudAiplatformV1SuggestTrialsRequest}()}
}
\concept{GoogleCloudAiplatformV1SuggestTrialsRequest functions}
