% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.tensorboards.create}
\alias{projects.locations.tensorboards.create}
\title{Creates a Tensorboard.}
\usage{
projects.locations.tensorboards.create(
  GoogleCloudAiplatformV1Tensorboard,
  parent
)
}
\arguments{
\item{GoogleCloudAiplatformV1Tensorboard}{The \link{GoogleCloudAiplatformV1Tensorboard} object to pass to this method}

\item{parent}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1Tensorboard functions: 
\code{\link{GoogleCloudAiplatformV1Tensorboard.labels}()},
\code{\link{GoogleCloudAiplatformV1Tensorboard}()},
\code{\link{projects.locations.tensorboards.patch}()}
}
\concept{GoogleCloudAiplatformV1Tensorboard functions}
