% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.tensorboards.experiments.create}
\alias{projects.locations.tensorboards.experiments.create}
\title{Creates a TensorboardExperiment.}
\usage{
projects.locations.tensorboards.experiments.create(
  GoogleCloudAiplatformV1TensorboardExperiment,
  parent,
  tensorboardExperimentId = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1TensorboardExperiment}{The \link{GoogleCloudAiplatformV1TensorboardExperiment} object to pass to this method}

\item{parent}{Required}

\item{tensorboardExperimentId}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1TensorboardExperiment functions: 
\code{\link{GoogleCloudAiplatformV1TensorboardExperiment.labels}()},
\code{\link{GoogleCloudAiplatformV1TensorboardExperiment}()},
\code{\link{projects.locations.tensorboards.experiments.patch}()}
}
\concept{GoogleCloudAiplatformV1TensorboardExperiment functions}
