% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.tensorboards.experiments.runs.batchCreate}
\alias{projects.locations.tensorboards.experiments.runs.batchCreate}
\title{Batch create TensorboardRuns.}
\usage{
projects.locations.tensorboards.experiments.runs.batchCreate(
  GoogleCloudAiplatformV1BatchCreateTensorboardRunsRequest,
  parent
)
}
\arguments{
\item{GoogleCloudAiplatformV1BatchCreateTensorboardRunsRequest}{The \link{GoogleCloudAiplatformV1BatchCreateTensorboardRunsRequest} object to pass to this method}

\item{parent}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1BatchCreateTensorboardRunsRequest functions: 
\code{\link{GoogleCloudAiplatformV1BatchCreateTensorboardRunsRequest}()}
}
\concept{GoogleCloudAiplatformV1BatchCreateTensorboardRunsRequest functions}
