% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.tensorboards.experiments.runs.create}
\alias{projects.locations.tensorboards.experiments.runs.create}
\title{Creates a TensorboardRun.}
\usage{
projects.locations.tensorboards.experiments.runs.create(
  GoogleCloudAiplatformV1TensorboardRun,
  parent,
  tensorboardRunId = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1TensorboardRun}{The \link{GoogleCloudAiplatformV1TensorboardRun} object to pass to this method}

\item{parent}{Required}

\item{tensorboardRunId}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1TensorboardRun functions: 
\code{\link{GoogleCloudAiplatformV1TensorboardRun.labels}()},
\code{\link{GoogleCloudAiplatformV1TensorboardRun}()},
\code{\link{projects.locations.tensorboards.experiments.runs.patch}()}
}
\concept{GoogleCloudAiplatformV1TensorboardRun functions}
