% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.tensorboards.experiments.runs.timeSeries.batchCreate}
\alias{projects.locations.tensorboards.experiments.runs.timeSeries.batchCreate}
\title{Batch create TensorboardTimeSeries that belong to a TensorboardExperiment.}
\usage{
projects.locations.tensorboards.experiments.runs.timeSeries.batchCreate(
  GoogleCloudAiplatformV1BatchCreateTensorboardTimeSeriesRequest,
  parent,
  runsId
)
}
\arguments{
\item{GoogleCloudAiplatformV1BatchCreateTensorboardTimeSeriesRequest}{The \link{GoogleCloudAiplatformV1BatchCreateTensorboardTimeSeriesRequest} object to pass to this method}

\item{parent}{Required}

\item{runsId}{}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1BatchCreateTensorboardTimeSeriesRequest functions: 
\code{\link{GoogleCloudAiplatformV1BatchCreateTensorboardTimeSeriesRequest}()}
}
\concept{GoogleCloudAiplatformV1BatchCreateTensorboardTimeSeriesRequest functions}
