% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.tensorboards.experiments.runs.timeSeries.batchRead}
\alias{projects.locations.tensorboards.experiments.runs.timeSeries.batchRead}
\title{Reads multiple TensorboardTimeSeries' data. The data point number limit is 1000 for scalars, 100 for tensors and blob references. If the number of data points stored is less than the limit, all data is returned. Otherwise, the number limit of data points is randomly selected from this time series and returned.}
\usage{
projects.locations.tensorboards.experiments.runs.timeSeries.batchRead(
  tensorboard,
  experimentsId,
  runsId,
  timeSeries = NULL
)
}
\arguments{
\item{tensorboard}{Required}

\item{experimentsId}{}

\item{runsId}{}

\item{timeSeries}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}
}
