% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.tensorboards.experiments.runs.timeSeries.create}
\alias{projects.locations.tensorboards.experiments.runs.timeSeries.create}
\title{Creates a TensorboardTimeSeries.}
\usage{
projects.locations.tensorboards.experiments.runs.timeSeries.create(
  GoogleCloudAiplatformV1TensorboardTimeSeries,
  parent,
  tensorboardTimeSeriesId = NULL
)
}
\arguments{
\item{GoogleCloudAiplatformV1TensorboardTimeSeries}{The \link{GoogleCloudAiplatformV1TensorboardTimeSeries} object to pass to this method}

\item{parent}{Required}

\item{tensorboardTimeSeriesId}{Optional}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1TensorboardTimeSeries functions: 
\code{\link{GoogleCloudAiplatformV1TensorboardTimeSeries}()},
\code{\link{projects.locations.tensorboards.experiments.runs.timeSeries.patch}()}
}
\concept{GoogleCloudAiplatformV1TensorboardTimeSeries functions}
