% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.tensorboards.experiments.runs.timeSeries.exportTensorboardTimeSeries}
\alias{projects.locations.tensorboards.experiments.runs.timeSeries.exportTensorboardTimeSeries}
\title{Exports a TensorboardTimeSeries' data. Data is returned in paginated responses.}
\usage{
projects.locations.tensorboards.experiments.runs.timeSeries.exportTensorboardTimeSeries(
  GoogleCloudAiplatformV1ExportTensorboardTimeSeriesDataRequest,
  tensorboardTimeSeries
)
}
\arguments{
\item{GoogleCloudAiplatformV1ExportTensorboardTimeSeriesDataRequest}{The \link{GoogleCloudAiplatformV1ExportTensorboardTimeSeriesDataRequest} object to pass to this method}

\item{tensorboardTimeSeries}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1ExportTensorboardTimeSeriesDataRequest functions: 
\code{\link{GoogleCloudAiplatformV1ExportTensorboardTimeSeriesDataRequest}()}
}
\concept{GoogleCloudAiplatformV1ExportTensorboardTimeSeriesDataRequest functions}
