% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.tensorboards.experiments.runs.timeSeries.read}
\alias{projects.locations.tensorboards.experiments.runs.timeSeries.read}
\title{Reads a TensorboardTimeSeries' data. By default, if the number of data points stored is less than 1000, all data is returned. Otherwise, 1000 data points is randomly selected from this time series and returned. This value can be changed by changing max_data_points, which can't be greater than 10k.}
\usage{
projects.locations.tensorboards.experiments.runs.timeSeries.read(
  tensorboardTimeSeries,
  filter = NULL,
  maxDataPoints = NULL
)
}
\arguments{
\item{tensorboardTimeSeries}{Required}

\item{filter}{Reads the TensorboardTimeSeries' data that match the filter expression}

\item{maxDataPoints}{The maximum number of TensorboardTimeSeries' data to return}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloud-platform.read-only
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/cloud-platform.read-only)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}
}
