% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.tensorboards.experiments.runs.write}
\alias{projects.locations.tensorboards.experiments.runs.write}
\title{Write time series data points into multiple TensorboardTimeSeries under a TensorboardRun. If any data fail to be ingested, an error is returned.}
\usage{
projects.locations.tensorboards.experiments.runs.write(
  GoogleCloudAiplatformV1WriteTensorboardRunDataRequest,
  tensorboardRun
)
}
\arguments{
\item{GoogleCloudAiplatformV1WriteTensorboardRunDataRequest}{The \link{GoogleCloudAiplatformV1WriteTensorboardRunDataRequest} object to pass to this method}

\item{tensorboardRun}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1WriteTensorboardRunDataRequest functions: 
\code{\link{GoogleCloudAiplatformV1WriteTensorboardRunDataRequest}()}
}
\concept{GoogleCloudAiplatformV1WriteTensorboardRunDataRequest functions}
