% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.tensorboards.experiments.write}
\alias{projects.locations.tensorboards.experiments.write}
\title{Write time series data points of multiple TensorboardTimeSeries in multiple TensorboardRun's. If any data fail to be ingested, an error is returned.}
\usage{
projects.locations.tensorboards.experiments.write(
  GoogleCloudAiplatformV1WriteTensorboardExperimentDataRequest,
  tensorboardExperiment
)
}
\arguments{
\item{GoogleCloudAiplatformV1WriteTensorboardExperimentDataRequest}{The \link{GoogleCloudAiplatformV1WriteTensorboardExperimentDataRequest} object to pass to this method}

\item{tensorboardExperiment}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1WriteTensorboardExperimentDataRequest functions: 
\code{\link{GoogleCloudAiplatformV1WriteTensorboardExperimentDataRequest}()}
}
\concept{GoogleCloudAiplatformV1WriteTensorboardExperimentDataRequest functions}
