% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.tensorboards.list}
\alias{projects.locations.tensorboards.list}
\title{Lists Tensorboards in a Location.}
\usage{
projects.locations.tensorboards.list(
  parent,
  filter = NULL,
  orderBy = NULL,
  readMask = NULL,
  pageSize = NULL,
  pageToken = NULL
)
}
\arguments{
\item{parent}{Required}

\item{filter}{Lists the Tensorboards that match the filter expression}

\item{orderBy}{Field to use to sort the list}

\item{readMask}{Mask specifying which fields to read}

\item{pageSize}{The maximum number of Tensorboards to return}

\item{pageToken}{A page token, received from a previous TensorboardService}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloud-platform.read-only
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/cloud-platform.read-only)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}
}
