% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.trainingPipelines.cancel}
\alias{projects.locations.trainingPipelines.cancel}
\title{Cancels a TrainingPipeline. Starts asynchronous cancellation on the TrainingPipeline. The server makes a best effort to cancel the pipeline, but success is not guaranteed. Clients can use PipelineService.GetTrainingPipeline or other methods to check whether the cancellation succeeded or whether the pipeline completed despite cancellation. On successful cancellation, the TrainingPipeline is not deleted; instead it becomes a pipeline with a TrainingPipeline.error value with a google.rpc.Status.code of 1, corresponding to \code{Code.CANCELLED}, and TrainingPipeline.state is set to \code{CANCELLED}.}
\usage{
projects.locations.trainingPipelines.cancel(
  GoogleCloudAiplatformV1CancelTrainingPipelineRequest,
  name
)
}
\arguments{
\item{GoogleCloudAiplatformV1CancelTrainingPipelineRequest}{The \link{GoogleCloudAiplatformV1CancelTrainingPipelineRequest} object to pass to this method}

\item{name}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1CancelTrainingPipelineRequest functions: 
\code{\link{GoogleCloudAiplatformV1CancelTrainingPipelineRequest}()}
}
\concept{GoogleCloudAiplatformV1CancelTrainingPipelineRequest functions}
