% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiplatform_functions.R
\name{projects.locations.trainingPipelines.create}
\alias{projects.locations.trainingPipelines.create}
\title{Creates a TrainingPipeline. A created TrainingPipeline right away will be attempted to be run.}
\usage{
projects.locations.trainingPipelines.create(
  GoogleCloudAiplatformV1TrainingPipeline,
  parent
)
}
\arguments{
\item{GoogleCloudAiplatformV1TrainingPipeline}{The \link{GoogleCloudAiplatformV1TrainingPipeline} object to pass to this method}

\item{parent}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vertex-ai/}{Google Documentation}

Other GoogleCloudAiplatformV1TrainingPipeline functions: 
\code{\link{GoogleCloudAiplatformV1TrainingPipeline.labels}()},
\code{\link{GoogleCloudAiplatformV1TrainingPipeline}()}
}
\concept{GoogleCloudAiplatformV1TrainingPipeline functions}
